Store all files types in hidden uploads field
