+++ django_file_form/static/file_form/file_form.js
@@ -1,4 +1,4 @@
-!function(){var t="undefined"!=typeof globalThis?globalThis:"undefined"!=typeof window?window:"undefined"!=typeof global?global:"undefined"!=typeof self?self:{};function e(t){var e={exports:{}};return t(e,e.exports),e.exports}var n=function(t){return t&&t.Math==Math&&t},r=n("object"==typeof globalThis&&globalThis)||n("object"==typeof window&&window)||n("object"==typeof self&&self)||n("object"==typeof t&&t)||function(){return this}()||Function("return this")(),o=function(t){try{return!!t()}catch(t){return!0}},i=!o((function(){return 7!=Object.defineProperty({},1,{get:function(){return 7}})[1]})),a={}.propertyIsEnumerable,u=Object.getOwnPropertyDescriptor,s={f:u&&!a.call({1:2},1)?function(t){var e=u(this,t);return!!e&&e.enumerable}:a},c=function(t,e){return{enumerable:!(1&t),configurable:!(2&t),writable:!(4&t),value:e}},l={}.toString,f=function(t){return l.call(t).slice(8,-1)},p="".split,d=o((function(){return!Object("z").propertyIsEnumerable(0)}))?function(t){return"String"==f(t)?p.call(t,""):Object(t)}:Object,h=function(t){if(null==t)throw TypeError("Can't call method on "+t);return t},v=function(t){return d(h(t))},y=function(t){return"object"==typeof t?null!==t:"function"==typeof t},g=function(t,e){if(!y(t))return t;var n,r;if(e&&"function"==typeof(n=t.toString)&&!y(r=n.call(t)))return r;if("function"==typeof(n=t.valueOf)&&!y(r=n.call(t)))return r;if(!e&&"function"==typeof(n=t.toString)&&!y(r=n.call(t)))return r;throw TypeError("Can't convert object to primitive value")},m={}.hasOwnProperty,x=function(t,e){return m.call(t,e)},b=r.document,w=y(b)&&y(b.createElement),_=function(t){return w?b.createElement(t):{}},S=!i&&!o((function(){return 7!=Object.defineProperty(_("div"),"a",{get:function(){return 7}}).a})),E=Object.getOwnPropertyDescriptor,A={f:i?E:function(t,e){if(t=v(t),e=g(e,!0),S)try{return E(t,e)}catch(t){}if(x(t,e))return c(!s.f.call(t,e),t[e])}},k=function(t){if(!y(t))throw TypeError(String(t)+" is not an object");return t},R=Object.defineProperty,O={f:i?R:function(t,e,n){if(k(t),e=g(e,!0),k(n),S)try{return R(t,e,n)}catch(t){}if("get"in n||"set"in n)throw TypeError("Accessors not supported");return"value"in n&&(t[e]=n.value),t}},T=i?function(t,e,n){return O.f(t,e,c(1,n))}:function(t,e,n){return t[e]=n,t},j=function(t,e){try{T(r,t,e)}catch(n){r[t]=e}return e},P="__core-js_shared__",C=r[P]||j(P,{}),I=Function.toString;"function"!=typeof C.inspectSource&&(C.inspectSource=function(t){return I.call(t)});var U,L,D,H=C.inspectSource,F=r.WeakMap,M="function"==typeof F&&/native code/.test(H(F)),N=e((function(t){(t.exports=function(t,e){return C[t]||(C[t]=void 0!==e?e:{})})("versions",[]).push({version:"3.9.1",mode:"global",copyright:"© 2021 Denis Pushkarev (zloirock.ru)"})})),B=0,z=Math.random(),q=function(t){return"Symbol("+String(void 0===t?"":t)+")_"+(++B+z).toString(36)},G=N("keys"),$=function(t){return G[t]||(G[t]=q(t))},K={},W=r.WeakMap;if(M){var X=C.state||(C.state=new W),V=X.get,Y=X.has,Q=X.set;U=function(t,e){return e.facade=t,Q.call(X,t,e),e},L=function(t){return V.call(X,t)||{}},D=function(t){return Y.call(X,t)}}else{var J=$("state");K[J]=!0,U=function(t,e){return e.facade=t,T(t,J,e),e},L=function(t){return x(t,J)?t[J]:{}},D=function(t){return x(t,J)}}var Z,tt,et={set:U,get:L,has:D,enforce:function(t){return D(t)?L(t):U(t,{})},getterFor:function(t){return function(e){var n;if(!y(e)||(n=L(e)).type!==t)throw TypeError("Incompatible receiver, "+t+" required");return n}}},nt=e((function(t){var e=et.get,n=et.enforce,o=String(String).split("String");(t.exports=function(t,e,i,a){var u,s=!!a&&!!a.unsafe,c=!!a&&!!a.enumerable,l=!!a&&!!a.noTargetGet;"function"==typeof i&&("string"!=typeof e||x(i,"name")||T(i,"name",e),(u=n(i)).source||(u.source=o.join("string"==typeof e?e:""))),t!==r?(s?!l&&t[e]&&(c=!0):delete t[e],c?t[e]=i:T(t,e,i)):c?t[e]=i:j(e,i)})(Function.prototype,"toString",(function(){return"function"==typeof this&&e(this).source||H(this)}))})),rt=r,ot=function(t){return"function"==typeof t?t:void 0},it=function(t,e){return arguments.length<2?ot(rt[t])||ot(r[t]):rt[t]&&rt[t][e]||r[t]&&r[t][e]},at=Math.ceil,ut=Math.floor,st=function(t){return isNaN(t=+t)?0:(t>0?ut:at)(t)},ct=Math.min,lt=function(t){return t>0?ct(st(t),9007199254740991):0},ft=Math.max,pt=Math.min,dt=function(t,e){var n=st(t);return n<0?ft(n+e,0):pt(n,e)},ht=function(t){return function(e,n,r){var o,i=v(e),a=lt(i.length),u=dt(r,a);if(t&&n!=n){for(;a>u;)if((o=i[u++])!=o)return!0}else for(;a>u;u++)if((t||u in i)&&i[u]===n)return t||u||0;return!t&&-1}},vt={includes:ht(!0),indexOf:ht(!1)},yt=vt.indexOf,gt=function(t,e){var n,r=v(t),o=0,i=[];for(n in r)!x(K,n)&&x(r,n)&&i.push(n);for(;e.length>o;)x(r,n=e[o++])&&(~yt(i,n)||i.push(n));return i},mt=["constructor","hasOwnProperty","isPrototypeOf","propertyIsEnumerable","toLocaleString","toString","valueOf"],xt=mt.concat("length","prototype"),bt={f:Object.getOwnPropertyNames||function(t){return gt(t,xt)}},wt={f:Object.getOwnPropertySymbols},_t=it("Reflect","ownKeys")||function(t){var e=bt.f(k(t)),n=wt.f;return n?e.concat(n(t)):e},St=function(t,e){for(var n=_t(e),r=O.f,o=A.f,i=0;i<n.length;i++){var a=n[i];x(t,a)||r(t,a,o(e,a))}},Et=/#|\.prototype\./,At=function(t,e){var n=Rt[kt(t)];return n==Tt||n!=Ot&&("function"==typeof e?o(e):!!e)},kt=At.normalize=function(t){return String(t).replace(Et,".").toLowerCase()},Rt=At.data={},Ot=At.NATIVE="N",Tt=At.POLYFILL="P",jt=At,Pt=A.f,Ct=function(t,e){var n,o,i,a,u,s=t.target,c=t.global,l=t.stat;if(n=c?r:l?r[s]||j(s,{}):(r[s]||{}).prototype)for(o in e){if(a=e[o],i=t.noTargetGet?(u=Pt(n,o))&&u.value:n[o],!jt(c?o:s+(l?".":"#")+o,t.forced)&&void 0!==i){if(typeof a==typeof i)continue;St(a,i)}(t.sham||i&&i.sham)&&T(a,"sham",!0),nt(n,o,a,t)}},It=function(t){if("function"!=typeof t)throw TypeError(String(t)+" is not a function");return t},Ut=function(t,e,n){if(It(t),void 0===e)return t;switch(n){case 0:return function(){return t.call(e)};case 1:return function(n){return t.call(e,n)};case 2:return function(n,r){return t.call(e,n,r)};case 3:return function(n,r,o){return t.call(e,n,r,o)}}return function(){return t.apply(e,arguments)}},Lt=function(t){return Object(h(t))},Dt=function(t){var e=t.return;if(void 0!==e)return k(e.call(t)).value},Ht=function(t,e,n,r){try{return r?e(k(n)[0],n[1]):e(n)}catch(e){throw Dt(t),e}},Ft="process"==f(r.process),Mt=it("navigator","userAgent")||"",Nt=r.process,Bt=Nt&&Nt.versions,zt=Bt&&Bt.v8;zt?tt=(Z=zt.split("."))[0]+Z[1]:Mt&&(!(Z=Mt.match(/Edge\/(\d+)/))||Z[1]>=74)&&(Z=Mt.match(/Chrome\/(\d+)/))&&(tt=Z[1]);var qt=tt&&+tt,Gt=!!Object.getOwnPropertySymbols&&!o((function(){return!Symbol.sham&&(Ft?38===qt:qt>37&&qt<41)})),$t=Gt&&!Symbol.sham&&"symbol"==typeof Symbol.iterator,Kt=N("wks"),Wt=r.Symbol,Xt=$t?Wt:Wt&&Wt.withoutSetter||q,Vt=function(t){return x(Kt,t)&&(Gt||"string"==typeof Kt[t])||(Gt&&x(Wt,t)?Kt[t]=Wt[t]:Kt[t]=Xt("Symbol."+t)),Kt[t]},Yt={},Qt=Vt("iterator"),Jt=Array.prototype,Zt=function(t){return void 0!==t&&(Yt.Array===t||Jt[Qt]===t)},te=function(t,e,n){var r=g(e);r in t?O.f(t,r,c(0,n)):t[r]=n},ee={};ee[Vt("toStringTag")]="z";var ne="[object z]"===String(ee),re=Vt("toStringTag"),oe="Arguments"==f(function(){return arguments}()),ie=ne?f:function(t){var e,n,r;return void 0===t?"Undefined":null===t?"Null":"string"==typeof(n=function(t,e){try{return t[e]}catch(t){}}(e=Object(t),re))?n:oe?f(e):"Object"==(r=f(e))&&"function"==typeof e.callee?"Arguments":r},ae=Vt("iterator"),ue=function(t){if(null!=t)return t[ae]||t["@@iterator"]||Yt[ie(t)]},se=Vt("iterator"),ce=!1;try{var le=0,fe={next:function(){return{done:!!le++}},return:function(){ce=!0}};fe[se]=function(){return this},Array.from(fe,(function(){throw 2}))}catch(t){}var pe=function(t,e){if(!e&&!ce)return!1;var n=!1;try{var r={};r[se]=function(){return{next:function(){return{done:n=!0}}}},t(r)}catch(t){}return n},de=!pe((function(t){Array.from(t)}));Ct({target:"Array",stat:!0,forced:de},{from:function(t){var e,n,r,o,i,a,u=Lt(t),s="function"==typeof this?this:Array,c=arguments.length,l=c>1?arguments[1]:void 0,f=void 0!==l,p=ue(u),d=0;if(f&&(l=Ut(l,c>2?arguments[2]:void 0,2)),null==p||s==Array&&Zt(p))for(n=new s(e=lt(u.length));e>d;d++)a=f?l(u[d],d):u[d],te(n,d,a);else for(i=(o=p.call(u)).next,n=new s;!(r=i.call(o)).done;d++)a=f?Ht(o,l,[r.value,d],!0):r.value,te(n,d,a);return n.length=d,n}});var he,ve,ye,ge=function(t){return function(e,n){var r,o,i=String(h(e)),a=st(n),u=i.length;return a<0||a>=u?t?"":void 0:(r=i.charCodeAt(a))<55296||r>56319||a+1===u||(o=i.charCodeAt(a+1))<56320||o>57343?t?i.charAt(a):r:t?i.slice(a,a+2):o-56320+(r-55296<<10)+65536}},me={codeAt:ge(!1),charAt:ge(!0)},xe=!o((function(){function t(){}return t.prototype.constructor=null,Object.getPrototypeOf(new t)!==t.prototype})),be=$("IE_PROTO"),we=Object.prototype,_e=xe?Object.getPrototypeOf:function(t){return t=Lt(t),x(t,be)?t[be]:"function"==typeof t.constructor&&t instanceof t.constructor?t.constructor.prototype:t instanceof Object?we:null},Se=Vt("iterator"),Ee=!1;[].keys&&("next"in(ye=[].keys())?(ve=_e(_e(ye)))!==Object.prototype&&(he=ve):Ee=!0),(null==he||o((function(){var t={};return he[Se].call(t)!==t})))&&(he={}),x(he,Se)||T(he,Se,(function(){return this}));var Ae,ke={IteratorPrototype:he,BUGGY_SAFARI_ITERATORS:Ee},Re=Object.keys||function(t){return gt(t,mt)},Oe=i?Object.defineProperties:function(t,e){k(t);for(var n,r=Re(e),o=r.length,i=0;o>i;)O.f(t,n=r[i++],e[n]);return t},Te=it("document","documentElement"),je=$("IE_PROTO"),Pe=function(){},Ce=function(t){return"<script>"+t+"</"+"script>"},Ie=function(){try{Ae=document.domain&&new ActiveXObject("htmlfile")}catch(t){}var t,e;Ie=Ae?function(t){t.write(Ce("")),t.close();var e=t.parentWindow.Object;return t=null,e}(Ae):((e=_("iframe")).style.display="none",Te.appendChild(e),e.src=String("javascript:"),(t=e.contentWindow.document).open(),t.write(Ce("document.F=Object")),t.close(),t.F);for(var n=mt.length;n--;)delete Ie.prototype[mt[n]];return Ie()};K[je]=!0;var Ue=Object.create||function(t,e){var n;return null!==t?(Pe.prototype=k(t),n=new Pe,Pe.prototype=null,n[je]=t):n=Ie(),void 0===e?n:Oe(n,e)},Le=O.f,De=Vt("toStringTag"),He=function(t,e,n){t&&!x(t=n?t:t.prototype,De)&&Le(t,De,{configurable:!0,value:e})},Fe=ke.IteratorPrototype,Me=function(){return this},Ne=Object.setPrototypeOf||("__proto__"in{}?function(){var t,e=!1,n={};try{(t=Object.getOwnPropertyDescriptor(Object.prototype,"__proto__").set).call(n,[]),e=n instanceof Array}catch(t){}return function(n,r){return k(n),function(t){if(!y(t)&&null!==t)throw TypeError("Can't set "+String(t)+" as a prototype")}(r),e?t.call(n,r):n.__proto__=r,n}}():void 0),Be=ke.IteratorPrototype,ze=ke.BUGGY_SAFARI_ITERATORS,qe=Vt("iterator"),Ge="keys",$e="values",Ke="entries",We=function(){return this},Xe=function(t,e,n,r,o,i,a){!function(t,e,n){var r=e+" Iterator";t.prototype=Ue(Fe,{next:c(1,n)}),He(t,r,!1),Yt[r]=Me}(n,e,r);var u,s,l,f=function(t){if(t===o&&y)return y;if(!ze&&t in h)return h[t];switch(t){case Ge:case $e:case Ke:return function(){return new n(this,t)}}return function(){return new n(this)}},p=e+" Iterator",d=!1,h=t.prototype,v=h[qe]||h["@@iterator"]||o&&h[o],y=!ze&&v||f(o),g="Array"==e&&h.entries||v;if(g&&(u=_e(g.call(new t)),Be!==Object.prototype&&u.next&&(_e(u)!==Be&&(Ne?Ne(u,Be):"function"!=typeof u[qe]&&T(u,qe,We)),He(u,p,!0))),o==$e&&v&&v.name!==$e&&(d=!0,y=function(){return v.call(this)}),h[qe]!==y&&T(h,qe,y),Yt[e]=y,o)if(s={values:f($e),keys:i?y:f(Ge),entries:f(Ke)},a)for(l in s)(ze||d||!(l in h))&&nt(h,l,s[l]);else Ct({target:e,proto:!0,forced:ze||d},s);return s},Ve=me.charAt,Ye="String Iterator",Qe=et.set,Je=et.getterFor(Ye);Xe(String,"String",(function(t){Qe(this,{type:Ye,string:String(t),index:0})}),(function(){var t,e=Je(this),n=e.string,r=e.index;return r>=n.length?{value:void 0,done:!0}:(t=Ve(n,r),e.index+=t.length,{value:t,done:!1})}));var Ze=ne?{}.toString:function(){return"[object "+ie(this)+"]"};ne||nt(Object.prototype,"toString",Ze,{unsafe:!0});var tn=Vt("unscopables"),en=Array.prototype;null==en[tn]&&O.f(en,tn,{configurable:!0,value:Ue(null)});var nn=function(t){en[tn][t]=!0},rn="Array Iterator",on=et.set,an=et.getterFor(rn),un=Xe(Array,"Array",(function(t,e){on(this,{type:rn,target:v(t),index:0,kind:e})}),(function(){var t=an(this),e=t.target,n=t.kind,r=t.index++;return!e||r>=e.length?(t.target=void 0,{value:void 0,done:!0}):"keys"==n?{value:r,done:!1}:"values"==n?{value:e[r],done:!1}:{value:[r,e[r]],done:!1}}),"values");Yt.Arguments=Yt.Array,nn("keys"),nn("values"),nn("entries");var sn={CSSRuleList:0,CSSStyleDeclaration:0,CSSValueList:0,ClientRectList:0,DOMRectList:0,DOMStringList:0,DOMTokenList:1,DataTransferItemList:0,FileList:0,HTMLAllCollection:0,HTMLCollection:0,HTMLFormElement:0,HTMLSelectElement:0,MediaList:0,MimeTypeArray:0,NamedNodeMap:0,NodeList:1,PaintRequestList:0,Plugin:0,PluginArray:0,SVGLengthList:0,SVGNumberList:0,SVGPathSegList:0,SVGPointList:0,SVGStringList:0,SVGTransformList:0,SourceBufferList:0,StyleSheetList:0,TextTrackCueList:0,TextTrackList:0,TouchList:0},cn=Vt("iterator"),ln=Vt("toStringTag"),fn=un.values;for(var pn in sn){var dn=r[pn],hn=dn&&dn.prototype;if(hn){if(hn[cn]!==fn)try{T(hn,cn,fn)}catch(t){hn[cn]=fn}if(hn[ln]||T(hn,ln,pn),sn[pn])for(var vn in un)if(hn[vn]!==un[vn])try{T(hn,vn,un[vn])}catch(t){hn[vn]=un[vn]}}}var yn=!o((function(){return Object.isExtensible(Object.preventExtensions({}))})),gn=e((function(t){var e=O.f,n=q("meta"),r=0,o=Object.isExtensible||function(){return!0},i=function(t){e(t,n,{value:{objectID:"O"+ ++r,weakData:{}}})},a=t.exports={REQUIRED:!1,fastKey:function(t,e){if(!y(t))return"symbol"==typeof t?t:("string"==typeof t?"S":"P")+t;if(!x(t,n)){if(!o(t))return"F";if(!e)return"E";i(t)}return t[n].objectID},getWeakData:function(t,e){if(!x(t,n)){if(!o(t))return!0;if(!e)return!1;i(t)}return t[n].weakData},onFreeze:function(t){return yn&&a.REQUIRED&&o(t)&&!x(t,n)&&i(t),t}};K[n]=!0})),mn=function(t,e){this.stopped=t,this.result=e},xn=function(t,e,n){var r,o,i,a,u,s,c,l=n&&n.that,f=!(!n||!n.AS_ENTRIES),p=!(!n||!n.IS_ITERATOR),d=!(!n||!n.INTERRUPTED),h=Ut(e,l,1+f+d),v=function(t){return r&&Dt(r),new mn(!0,t)},y=function(t){return f?(k(t),d?h(t[0],t[1],v):h(t[0],t[1])):d?h(t,v):h(t)};if(p)r=t;else{if("function"!=typeof(o=ue(t)))throw TypeError("Target is not iterable");if(Zt(o)){for(i=0,a=lt(t.length);a>i;i++)if((u=y(t[i]))&&u instanceof mn)return u;return new mn(!1)}r=o.call(t)}for(s=r.next;!(c=s.call(r)).done;){try{u=y(c.value)}catch(t){throw Dt(r),t}if("object"==typeof u&&u&&u instanceof mn)return u}return new mn(!1)},bn=function(t,e,n){if(!(t instanceof e))throw TypeError("Incorrect "+(n?n+" ":"")+"invocation");return t},wn=function(t,e,n){var r,o;return Ne&&"function"==typeof(r=e.constructor)&&r!==n&&y(o=r.prototype)&&o!==n.prototype&&Ne(t,o),t},_n=function(t,e,n){var i=-1!==t.indexOf("Map"),a=-1!==t.indexOf("Weak"),u=i?"set":"add",s=r[t],c=s&&s.prototype,l=s,f={},p=function(t){var e=c[t];nt(c,t,"add"==t?function(t){return e.call(this,0===t?0:t),this}:"delete"==t?function(t){return!(a&&!y(t))&&e.call(this,0===t?0:t)}:"get"==t?function(t){return a&&!y(t)?void 0:e.call(this,0===t?0:t)}:"has"==t?function(t){return!(a&&!y(t))&&e.call(this,0===t?0:t)}:function(t,n){return e.call(this,0===t?0:t,n),this})};if(jt(t,"function"!=typeof s||!(a||c.forEach&&!o((function(){(new s).entries().next()})))))l=n.getConstructor(e,t,i,u),gn.REQUIRED=!0;else if(jt(t,!0)){var d=new l,h=d[u](a?{}:-0,1)!=d,v=o((function(){d.has(1)})),g=pe((function(t){new s(t)})),m=!a&&o((function(){for(var t=new s,e=5;e--;)t[u](e,e);return!t.has(-0)}));g||((l=e((function(e,n){bn(e,l,t);var r=wn(new s,e,l);return null!=n&&xn(n,r[u],{that:r,AS_ENTRIES:i}),r}))).prototype=c,c.constructor=l),(v||m)&&(p("delete"),p("has"),i&&p("get")),(m||h)&&p(u),a&&c.clear&&delete c.clear}return f[t]=l,Ct({global:!0,forced:l!=s},f),He(l,t),a||n.setStrong(l,t,i),l},Sn=function(t,e,n){for(var r in e)nt(t,r,e[r],n);return t},En=Vt("species"),An=function(t){var e=it(t),n=O.f;i&&e&&!e[En]&&n(e,En,{configurable:!0,get:function(){return this}})},kn=O.f,Rn=gn.fastKey,On=et.set,Tn=et.getterFor,jn={getConstructor:function(t,e,n,r){var o=t((function(t,a){bn(t,o,e),On(t,{type:e,index:Ue(null),first:void 0,last:void 0,size:0}),i||(t.size=0),null!=a&&xn(a,t[r],{that:t,AS_ENTRIES:n})})),a=Tn(e),u=function(t,e,n){var r,o,u=a(t),c=s(t,e);return c?c.value=n:(u.last=c={index:o=Rn(e,!0),key:e,value:n,previous:r=u.last,next:void 0,removed:!1},u.first||(u.first=c),r&&(r.next=c),i?u.size++:t.size++,"F"!==o&&(u.index[o]=c)),t},s=function(t,e){var n,r=a(t),o=Rn(e);if("F"!==o)return r.index[o];for(n=r.first;n;n=n.next)if(n.key==e)return n};return Sn(o.prototype,{clear:function(){for(var t=a(this),e=t.index,n=t.first;n;)n.removed=!0,n.previous&&(n.previous=n.previous.next=void 0),delete e[n.index],n=n.next;t.first=t.last=void 0,i?t.size=0:this.size=0},delete:function(t){var e=this,n=a(e),r=s(e,t);if(r){var o=r.next,u=r.previous;delete n.index[r.index],r.removed=!0,u&&(u.next=o),o&&(o.previous=u),n.first==r&&(n.first=o),n.last==r&&(n.last=u),i?n.size--:e.size--}return!!r},forEach:function(t){for(var e,n=a(this),r=Ut(t,arguments.length>1?arguments[1]:void 0,3);e=e?e.next:n.first;)for(r(e.value,e.key,this);e&&e.removed;)e=e.previous},has:function(t){return!!s(this,t)}}),Sn(o.prototype,n?{get:function(t){var e=s(this,t);return e&&e.value},set:function(t,e){return u(this,0===t?0:t,e)}}:{add:function(t){return u(this,t=0===t?0:t,t)}}),i&&kn(o.prototype,"size",{get:function(){return a(this).size}}),o},setStrong:function(t,e,n){var r=e+" Iterator",o=Tn(e),i=Tn(r);Xe(t,e,(function(t,e){On(this,{type:r,target:t,state:o(t),kind:e,last:void 0})}),(function(){for(var t=i(this),e=t.kind,n=t.last;n&&n.removed;)n=n.previous;return t.target&&(t.last=n=n?n.next:t.state.first)?"keys"==e?{value:n.key,done:!1}:"values"==e?{value:n.value,done:!1}:{value:[n.key,n.value],done:!1}:(t.target=void 0,{value:void 0,done:!0})}),n?"entries":"values",!n,!0),An(e)}};_n("Set",(function(t){return function(){return t(this,arguments.length?arguments[0]:void 0)}}),jn);var Pn=Array.isArray||function(t){return"Array"==f(t)},Cn=Vt("species"),In=function(t,e){var n;return Pn(t)&&("function"!=typeof(n=t.constructor)||n!==Array&&!Pn(n.prototype)?y(n)&&null===(n=n[Cn])&&(n=void 0):n=void 0),new(void 0===n?Array:n)(0===e?0:e)},Un=[].push,Ln=function(t){var e=1==t,n=2==t,r=3==t,o=4==t,i=6==t,a=7==t,u=5==t||i;return function(s,c,l,f){for(var p,h,v=Lt(s),y=d(v),g=Ut(c,l,3),m=lt(y.length),x=0,b=f||In,w=e?b(s,m):n||a?b(s,0):void 0;m>x;x++)if((u||x in y)&&(h=g(p=y[x],x,v),t))if(e)w[x]=h;else if(h)switch(t){case 3:return!0;case 5:return p;case 6:return x;case 2:Un.call(w,p)}else switch(t){case 4:return!1;case 7:Un.call(w,p)}return i?-1:r||o?o:w}},Dn={forEach:Ln(0),map:Ln(1),filter:Ln(2),some:Ln(3),every:Ln(4),find:Ln(5),findIndex:Ln(6),filterOut:Ln(7)},Hn=Vt("species"),Fn=function(t){return qt>=51||!o((function(){var e=[];return(e.constructor={})[Hn]=function(){return{foo:1}},1!==e[t](Boolean).foo}))},Mn=Dn.map,Nn=Fn("map");Ct({target:"Array",proto:!0,forced:!Nn},{map:function(t){return Mn(this,t,arguments.length>1?arguments[1]:void 0)}});var Bn=function(t,e){var n=[][t];return!!n&&o((function(){n.call(null,e||function(){throw 1},1)}))},zn=Dn.forEach,qn=Bn("forEach")?[].forEach:function(t){return zn(this,t,arguments.length>1?arguments[1]:void 0)};for(var Gn in sn){var $n=r[Gn],Kn=$n&&$n.prototype;if(Kn&&Kn.forEach!==qn)try{T(Kn,"forEach",qn)}catch(t){Kn.forEach=qn}}var Wn=function t(e){var n=e.parentElement;return n?"FORM"===n.tagName?n:t(n):null},Xn=o((function(){Re(1)}));Ct({target:"Object",stat:!0,forced:Xn},{keys:function(t){return Re(Lt(t))}});var Vn=bt.f,Yn={}.toString,Qn="object"==typeof window&&window&&Object.getOwnPropertyNames?Object.getOwnPropertyNames(window):[],Jn={f:function(t){return Qn&&"[object Window]"==Yn.call(t)?function(t){try{return Vn(t)}catch(t){return Qn.slice()}}(t):Vn(v(t))}},Zn={f:Vt},tr=O.f,er=function(t){var e=rt.Symbol||(rt.Symbol={});x(e,t)||tr(e,t,{value:Zn.f(t)})},nr=Dn.forEach,rr=$("hidden"),or="Symbol",ir=Vt("toPrimitive"),ar=et.set,ur=et.getterFor(or),sr=Object.prototype,cr=r.Symbol,lr=it("JSON","stringify"),fr=A.f,pr=O.f,dr=Jn.f,hr=s.f,vr=N("symbols"),yr=N("op-symbols"),gr=N("string-to-symbol-registry"),mr=N("symbol-to-string-registry"),xr=N("wks"),br=r.QObject,wr=!br||!br.prototype||!br.prototype.findChild,_r=i&&o((function(){return 7!=Ue(pr({},"a",{get:function(){return pr(this,"a",{value:7}).a}})).a}))?function(t,e,n){var r=fr(sr,e);r&&delete sr[e],pr(t,e,n),r&&t!==sr&&pr(sr,e,r)}:pr,Sr=function(t,e){var n=vr[t]=Ue(cr.prototype);return ar(n,{type:or,tag:t,description:e}),i||(n.description=e),n},Er=$t?function(t){return"symbol"==typeof t}:function(t){return Object(t)instanceof cr},Ar=function(t,e,n){t===sr&&Ar(yr,e,n),k(t);var r=g(e,!0);return k(n),x(vr,r)?(n.enumerable?(x(t,rr)&&t[rr][r]&&(t[rr][r]=!1),n=Ue(n,{enumerable:c(0,!1)})):(x(t,rr)||pr(t,rr,c(1,{})),t[rr][r]=!0),_r(t,r,n)):pr(t,r,n)},kr=function(t,e){k(t);var n=v(e),r=Re(n).concat(jr(n));return nr(r,(function(e){i&&!Rr.call(n,e)||Ar(t,e,n[e])})),t},Rr=function(t){var e=g(t,!0),n=hr.call(this,e);return!(this===sr&&x(vr,e)&&!x(yr,e))&&(!(n||!x(this,e)||!x(vr,e)||x(this,rr)&&this[rr][e])||n)},Or=function(t,e){var n=v(t),r=g(e,!0);if(n!==sr||!x(vr,r)||x(yr,r)){var o=fr(n,r);return!o||!x(vr,r)||x(n,rr)&&n[rr][r]||(o.enumerable=!0),o}},Tr=function(t){var e=dr(v(t)),n=[];return nr(e,(function(t){x(vr,t)||x(K,t)||n.push(t)})),n},jr=function(t){var e=t===sr,n=dr(e?yr:v(t)),r=[];return nr(n,(function(t){!x(vr,t)||e&&!x(sr,t)||r.push(vr[t])})),r};if(Gt||(nt((cr=function(){if(this instanceof cr)throw TypeError("Symbol is not a constructor");var t=arguments.length&&void 0!==arguments[0]?String(arguments[0]):void 0,e=q(t),n=function(t){this===sr&&n.call(yr,t),x(this,rr)&&x(this[rr],e)&&(this[rr][e]=!1),_r(this,e,c(1,t))};return i&&wr&&_r(sr,e,{configurable:!0,set:n}),Sr(e,t)}).prototype,"toString",(function(){return ur(this).tag})),nt(cr,"withoutSetter",(function(t){return Sr(q(t),t)})),s.f=Rr,O.f=Ar,A.f=Or,bt.f=Jn.f=Tr,wt.f=jr,Zn.f=function(t){return Sr(Vt(t),t)},i&&(pr(cr.prototype,"description",{configurable:!0,get:function(){return ur(this).description}}),nt(sr,"propertyIsEnumerable",Rr,{unsafe:!0}))),Ct({global:!0,wrap:!0,forced:!Gt,sham:!Gt},{Symbol:cr}),nr(Re(xr),(function(t){er(t)})),Ct({target:or,stat:!0,forced:!Gt},{for:function(t){var e=String(t);if(x(gr,e))return gr[e];var n=cr(e);return gr[e]=n,mr[n]=e,n},keyFor:function(t){if(!Er(t))throw TypeError(t+" is not a symbol");if(x(mr,t))return mr[t]},useSetter:function(){wr=!0},useSimple:function(){wr=!1}}),Ct({target:"Object",stat:!0,forced:!Gt,sham:!i},{create:function(t,e){return void 0===e?Ue(t):kr(Ue(t),e)},defineProperty:Ar,defineProperties:kr,getOwnPropertyDescriptor:Or}),Ct({target:"Object",stat:!0,forced:!Gt},{getOwnPropertyNames:Tr,getOwnPropertySymbols:jr}),Ct({target:"Object",stat:!0,forced:o((function(){wt.f(1)}))},{getOwnPropertySymbols:function(t){return wt.f(Lt(t))}}),lr){var Pr=!Gt||o((function(){var t=cr();return"[null]"!=lr([t])||"{}"!=lr({a:t})||"{}"!=lr(Object(t))}));Ct({target:"JSON",stat:!0,forced:Pr},{stringify:function(t,e,n){for(var r,o=[t],i=1;arguments.length>i;)o.push(arguments[i++]);if(r=e,(y(e)||void 0!==t)&&!Er(t))return Pn(e)||(e=function(t,e){if("function"==typeof r&&(e=r.call(this,t,e)),!Er(e))return e}),o[1]=e,lr.apply(null,o)}})}cr.prototype[ir]||T(cr.prototype,ir,cr.prototype.valueOf),He(cr,or),K[rr]=!0;var Cr=Dn.filter,Ir=Fn("filter");Ct({target:"Array",proto:!0,forced:!Ir},{filter:function(t){return Cr(this,t,arguments.length>1?arguments[1]:void 0)}});var Ur=A.f,Lr=o((function(){Ur(1)}));function Dr(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}Ct({target:"Object",stat:!0,forced:!i||Lr,sham:!i},{getOwnPropertyDescriptor:function(t,e){return Ur(v(t),e)}}),Ct({target:"Object",stat:!0,sham:!i},{getOwnPropertyDescriptors:function(t){for(var e,n,r=v(t),o=A.f,i=_t(r),a={},u=0;i.length>u;)void 0!==(n=o(r,e=i[u++]))&&te(a,e,n);return a}});var Hr=Vt("isConcatSpreadable"),Fr=9007199254740991,Mr="Maximum allowed index exceeded",Nr=qt>=51||!o((function(){var t=[];return t[Hr]=!1,t.concat()[0]!==t})),Br=Fn("concat"),zr=function(t){if(!y(t))return!1;var e=t[Hr];return void 0!==e?!!e:Pn(t)};Ct({target:"Array",proto:!0,forced:!Nr||!Br},{concat:function(t){var e,n,r,o,i,a=Lt(this),u=In(a,0),s=0;for(e=-1,r=arguments.length;e<r;e++)if(zr(i=-1===e?a:arguments[e])){if(s+(o=lt(i.length))>Fr)throw TypeError(Mr);for(n=0;n<o;n++,s++)n in i&&te(u,s,i[n])}else{if(s>=Fr)throw TypeError(Mr);te(u,s++,i)}return u.length=s,u}});var qr,Gr=Vt("match"),$r=function(t){var e;return y(t)&&(void 0!==(e=t[Gr])?!!e:"RegExp"==f(t))},Kr=function(t){if($r(t))throw TypeError("The method doesn't accept regular expressions");return t},Wr=Vt("match"),Xr=function(t){var e=/./;try{"/./"[t](e)}catch(n){try{return e[Wr]=!1,"/./"[t](e)}catch(t){}}return!1},Vr=A.f,Yr="".startsWith,Qr=Math.min,Jr=Xr("startsWith"),Zr=!(Jr||(qr=Vr(String.prototype,"startsWith"),!qr||qr.writable));Ct({target:"String",proto:!0,forced:!Zr&&!Jr},{startsWith:function(t){var e=String(h(this));Kr(t);var n=lt(Qr(arguments.length>1?arguments[1]:void 0,e.length)),r=String(t);return Yr?Yr.call(e,r,n):e.slice(n,n+r.length)===r}});var to=O.f,eo=Function.prototype,no=eo.toString,ro=/^\s*function ([^ (]*)/,oo="name";function io(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,r=new Array(e);n<e;n++)r[n]=t[n];return r}i&&!(oo in eo)&&to(eo,oo,{configurable:!0,get:function(){try{return no.call(this).match(ro)[1]}catch(t){return""}}});var ao=O.f,uo=r.Symbol;if(i&&"function"==typeof uo&&(!("description"in uo.prototype)||void 0!==uo().description)){var so={},co=function(){var t=arguments.length<1||void 0===arguments[0]?void 0:String(arguments[0]),e=this instanceof co?new uo(t):void 0===t?uo():uo(t);return""===t&&(so[e]=!0),e};St(co,uo);var lo=co.prototype=uo.prototype;lo.constructor=co;var fo=lo.toString,po="Symbol(test)"==String(uo("test")),ho=/^Symbol\((.*)\)[^)]+$/;ao(lo,"description",{configurable:!0,get:function(){var t=y(this)?this.valueOf():this,e=fo.call(t);if(x(so,t))return"";var n=po?e.slice(7,-1):e.replace(ho,"$1");return""===n?void 0:n}}),Ct({global:!0,forced:!0},{Symbol:co})}er("iterator");var vo=Fn("slice"),yo=Vt("species"),go=[].slice,mo=Math.max;function xo(t,e){if(t){if("string"==typeof t)return io(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);return"Object"===n&&t.constructor&&(n=t.constructor.name),"Map"===n||"Set"===n?Array.from(t):"Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)?io(t,e):void 0}}function bo(t){return function(t){if(Array.isArray(t))return io(t)}(t)||function(t){if("undefined"!=typeof Symbol&&Symbol.iterator in Object(t))return Array.from(t)}(t)||xo(t)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}Ct({target:"Array",proto:!0,forced:!vo},{slice:function(t,e){var n,r,o,i=v(this),a=lt(i.length),u=dt(t,a),s=dt(void 0===e?a:e,a);if(Pn(i)&&("function"!=typeof(n=i.constructor)||n!==Array&&!Pn(n.prototype)?y(n)&&null===(n=n[yo])&&(n=void 0):n=void 0,n===Array||void 0===n))return go.call(i,u,s);for(r=new(void 0===n?Array:n)(mo(s-u,0)),o=0;u<s;u++,o++)u in i&&te(r,o,i[u]);return r.length=o,r}});var wo,_o,So,Eo=r.Promise,Ao=Vt("species"),ko=function(t,e){var n,r=k(t).constructor;return void 0===r||null==(n=k(r)[Ao])?e:It(n)},Ro=/(iphone|ipod|ipad).*applewebkit/i.test(Mt),Oo=r.location,To=r.setImmediate,jo=r.clearImmediate,Po=r.process,Co=r.MessageChannel,Io=r.Dispatch,Uo=0,Lo={},Do="onreadystatechange",Ho=function(t){if(Lo.hasOwnProperty(t)){var e=Lo[t];delete Lo[t],e()}},Fo=function(t){return function(){Ho(t)}},Mo=function(t){Ho(t.data)},No=function(t){r.postMessage(t+"",Oo.protocol+"//"+Oo.host)};To&&jo||(To=function(t){for(var e=[],n=1;arguments.length>n;)e.push(arguments[n++]);return Lo[++Uo]=function(){("function"==typeof t?t:Function(t)).apply(void 0,e)},wo(Uo),Uo},jo=function(t){delete Lo[t]},Ft?wo=function(t){Po.nextTick(Fo(t))}:Io&&Io.now?wo=function(t){Io.now(Fo(t))}:Co&&!Ro?(So=(_o=new Co).port2,_o.port1.onmessage=Mo,wo=Ut(So.postMessage,So,1)):r.addEventListener&&"function"==typeof postMessage&&!r.importScripts&&Oo&&"file:"!==Oo.protocol&&!o(No)?(wo=No,r.addEventListener("message",Mo,!1)):wo=Do in _("script")?function(t){Te.appendChild(_("script")).onreadystatechange=function(){Te.removeChild(this),Ho(t)}}:function(t){setTimeout(Fo(t),0)});var Bo,zo,qo,Go,$o,Ko,Wo,Xo,Vo={set:To,clear:jo},Yo=/web0s(?!.*chrome)/i.test(Mt),Qo=A.f,Jo=Vo.set,Zo=r.MutationObserver||r.WebKitMutationObserver,ti=r.document,ei=r.process,ni=r.Promise,ri=Qo(r,"queueMicrotask"),oi=ri&&ri.value;oi||(Bo=function(){var t,e;for(Ft&&(t=ei.domain)&&t.exit();zo;){e=zo.fn,zo=zo.next;try{e()}catch(t){throw zo?Go():qo=void 0,t}}qo=void 0,t&&t.enter()},Ro||Ft||Yo||!Zo||!ti?ni&&ni.resolve?(Wo=ni.resolve(void 0),Xo=Wo.then,Go=function(){Xo.call(Wo,Bo)}):Go=Ft?function(){ei.nextTick(Bo)}:function(){Jo.call(r,Bo)}:($o=!0,Ko=ti.createTextNode(""),new Zo(Bo).observe(Ko,{characterData:!0}),Go=function(){Ko.data=$o=!$o}));var ii,ai,ui,si,ci=oi||function(t){var e={fn:t,next:void 0};qo&&(qo.next=e),zo||(zo=e,Go()),qo=e},li=function(t){var e,n;this.promise=new t((function(t,r){if(void 0!==e||void 0!==n)throw TypeError("Bad Promise constructor");e=t,n=r})),this.resolve=It(e),this.reject=It(n)},fi={f:function(t){return new li(t)}},pi=function(t,e){if(k(t),y(e)&&e.constructor===t)return e;var n=fi.f(t);return(0,n.resolve)(e),n.promise},di=function(t){try{return{error:!1,value:t()}}catch(t){return{error:!0,value:t}}},hi=Vo.set,vi=Vt("species"),yi="Promise",gi=et.get,mi=et.set,xi=et.getterFor(yi),bi=Eo,wi=r.TypeError,_i=r.document,Si=r.process,Ei=it("fetch"),Ai=fi.f,ki=Ai,Ri=!!(_i&&_i.createEvent&&r.dispatchEvent),Oi="function"==typeof PromiseRejectionEvent,Ti="unhandledrejection",ji=jt(yi,(function(){if(!(H(bi)!==String(bi))){if(66===qt)return!0;if(!Ft&&!Oi)return!0}if(qt>=51&&/native code/.test(bi))return!1;var t=bi.resolve(1),e=function(t){t((function(){}),(function(){}))};return(t.constructor={})[vi]=e,!(t.then((function(){}))instanceof e)})),Pi=ji||!pe((function(t){bi.all(t).catch((function(){}))})),Ci=function(t){var e;return!(!y(t)||"function"!=typeof(e=t.then))&&e},Ii=function(t,e){if(!t.notified){t.notified=!0;var n=t.reactions;ci((function(){for(var r=t.value,o=1==t.state,i=0;n.length>i;){var a,u,s,c=n[i++],l=o?c.ok:c.fail,f=c.resolve,p=c.reject,d=c.domain;try{l?(o||(2===t.rejection&&Hi(t),t.rejection=1),!0===l?a=r:(d&&d.enter(),a=l(r),d&&(d.exit(),s=!0)),a===c.promise?p(wi("Promise-chain cycle")):(u=Ci(a))?u.call(a,f,p):f(a)):p(r)}catch(t){d&&!s&&d.exit(),p(t)}}t.reactions=[],t.notified=!1,e&&!t.rejection&&Li(t)}))}},Ui=function(t,e,n){var o,i;Ri?((o=_i.createEvent("Event")).promise=e,o.reason=n,o.initEvent(t,!1,!0),r.dispatchEvent(o)):o={promise:e,reason:n},!Oi&&(i=r["on"+t])?i(o):t===Ti&&function(t,e){var n=r.console;n&&n.error&&(1===arguments.length?n.error(t):n.error(t,e))}("Unhandled promise rejection",n)},Li=function(t){hi.call(r,(function(){var e,n=t.facade,r=t.value;if(Di(t)&&(e=di((function(){Ft?Si.emit("unhandledRejection",r,n):Ui(Ti,n,r)})),t.rejection=Ft||Di(t)?2:1,e.error))throw e.value}))},Di=function(t){return 1!==t.rejection&&!t.parent},Hi=function(t){hi.call(r,(function(){var e=t.facade;Ft?Si.emit("rejectionHandled",e):Ui("rejectionhandled",e,t.value)}))},Fi=function(t,e,n){return function(r){t(e,r,n)}},Mi=function(t,e,n){t.done||(t.done=!0,n&&(t=n),t.value=e,t.state=2,Ii(t,!0))},Ni=function(t,e,n){if(!t.done){t.done=!0,n&&(t=n);try{if(t.facade===e)throw wi("Promise can't be resolved itself");var r=Ci(e);r?ci((function(){var n={done:!1};try{r.call(e,Fi(Ni,n,t),Fi(Mi,n,t))}catch(e){Mi(n,e,t)}})):(t.value=e,t.state=1,Ii(t,!1))}catch(e){Mi({done:!1},e,t)}}};function Bi(t,e,n,r,o,i,a){try{var u=t[i](a),s=u.value}catch(t){return void n(t)}u.done?e(s):Promise.resolve(s).then(r,o)}function zi(t){return function(){var e=this,n=arguments;return new Promise((function(r,o){var i=t.apply(e,n);function a(t){Bi(i,r,o,a,u,"next",t)}function u(t){Bi(i,r,o,a,u,"throw",t)}a(void 0)}))}}function qi(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}function Gi(t,e){for(var n=0;n<e.length;n++){var r=e[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(t,r.key,r)}}function $i(t,e,n){return e&&Gi(t.prototype,e),n&&Gi(t,n),t}function Ki(t){var e;if("undefined"!=typeof Symbol){if(Symbol.asyncIterator&&null!=(e=t[Symbol.asyncIterator]))return e.call(t);if(Symbol.iterator&&null!=(e=t[Symbol.iterator]))return e.call(t)}throw new TypeError("Object is not async iterable")}ji&&(bi=function(t){bn(this,bi,yi),It(t),ii.call(this);var e=gi(this);try{t(Fi(Ni,e),Fi(Mi,e))}catch(t){Mi(e,t)}},(ii=function(t){mi(this,{type:yi,done:!1,notified:!1,parent:!1,reactions:[],rejection:!1,state:0,value:void 0})}).prototype=Sn(bi.prototype,{then:function(t,e){var n=xi(this),r=Ai(ko(this,bi));return r.ok="function"!=typeof t||t,r.fail="function"==typeof e&&e,r.domain=Ft?Si.domain:void 0,n.parent=!0,n.reactions.push(r),0!=n.state&&Ii(n,!1),r.promise},catch:function(t){return this.then(void 0,t)}}),ai=function(){var t=new ii,e=gi(t);this.promise=t,this.resolve=Fi(Ni,e),this.reject=Fi(Mi,e)},fi.f=Ai=function(t){return t===bi||t===ui?new ai(t):ki(t)},"function"==typeof Eo&&(si=Eo.prototype.then,nt(Eo.prototype,"then",(function(t,e){var n=this;return new bi((function(t,e){si.call(n,t,e)})).then(t,e)}),{unsafe:!0}),"function"==typeof Ei&&Ct({global:!0,enumerable:!0,forced:!0},{fetch:function(t){return pi(bi,Ei.apply(r,arguments))}}))),Ct({global:!0,wrap:!0,forced:ji},{Promise:bi}),He(bi,yi,!1),An(yi),ui=it(yi),Ct({target:yi,stat:!0,forced:ji},{reject:function(t){var e=Ai(this);return e.reject.call(void 0,t),e.promise}}),Ct({target:yi,stat:!0,forced:ji},{resolve:function(t){return pi(this,t)}}),Ct({target:yi,stat:!0,forced:Pi},{all:function(t){var e=this,n=Ai(e),r=n.resolve,o=n.reject,i=di((function(){var n=It(e.resolve),i=[],a=0,u=1;xn(t,(function(t){var s=a++,c=!1;i.push(void 0),u++,n.call(e,t).then((function(t){c||(c=!0,i[s]=t,--u||r(i))}),o)})),--u||r(i)}));return i.error&&o(i.value),n.promise},race:function(t){var e=this,n=Ai(e),r=n.reject,o=di((function(){var o=It(e.resolve);xn(t,(function(t){o.call(e,t).then(n.resolve,r)}))}));return o.error&&r(o.value),n.promise}}),er("asyncIterator");var Wi="".repeat||function(t){var e=String(h(this)),n="",r=st(t);if(r<0||r==1/0)throw RangeError("Wrong number of repetitions");for(;r>0;(r>>>=1)&&(e+=e))1&r&&(n+=e);return n},Xi=1..toFixed,Vi=Math.floor,Yi=function(t,e,n){return 0===e?n:e%2==1?Yi(t,e-1,n*t):Yi(t*t,e/2,n)},Qi=function(t,e,n){for(var r=-1,o=n;++r<6;)o+=e*t[r],t[r]=o%1e7,o=Vi(o/1e7)},Ji=function(t,e){for(var n=6,r=0;--n>=0;)r+=t[n],t[n]=Vi(r/e),r=r%e*1e7},Zi=function(t){for(var e=6,n="";--e>=0;)if(""!==n||0===e||0!==t[e]){var r=String(t[e]);n=""===n?r:n+Wi.call("0",7-r.length)+r}return n},ta=Xi&&("0.000"!==8e-5.toFixed(3)||"1"!==.9.toFixed(0)||"1.25"!==1.255.toFixed(2)||"1000000000000000128"!==(0xde0b6b3a7640080).toFixed(0))||!o((function(){Xi.call({})}));Ct({target:"Number",proto:!0,forced:ta},{toFixed:function(t){var e,n,r,o,i=function(t){if("number"!=typeof t&&"Number"!=f(t))throw TypeError("Incorrect invocation");return+t}(this),a=st(t),u=[0,0,0,0,0,0],s="",c="0";if(a<0||a>20)throw RangeError("Incorrect fraction digits");if(i!=i)return"NaN";if(i<=-1e21||i>=1e21)return String(i);if(i<0&&(s="-",i=-i),i>1e-21)if(n=(e=function(t){for(var e=0,n=t;n>=4096;)e+=12,n/=4096;for(;n>=2;)e+=1,n/=2;return e}(i*Yi(2,69,1))-69)<0?i*Yi(2,-e,1):i/Yi(2,e,1),n*=4503599627370496,(e=52-e)>0){for(Qi(u,0,n),r=a;r>=7;)Qi(u,1e7,0),r-=7;for(Qi(u,Yi(10,r,1),0),r=e-1;r>=23;)Ji(u,1<<23),r-=23;Ji(u,1<<r),Qi(u,1,1),Ji(u,2),c=Zi(u)}else Qi(u,0,n),Qi(u,1<<-e,0),c=Zi(u)+Wi.call("0",a);return c=a>0?s+((o=c.length)<=a?"0."+Wi.call("0",a-o)+c:c.slice(0,o-a)+"."+c.slice(o-a)):s+c}});var ea=Dn.find,na="find",ra=!0;na in[]&&Array(1).find((function(){ra=!1})),Ct({target:"Array",proto:!0,forced:ra},{find:function(t){return ea(this,t,arguments.length>1?arguments[1]:void 0)}}),nn(na);var oa=Fn("splice"),ia=Math.max,aa=Math.min,ua=9007199254740991,sa="Maximum allowed length exceeded";function ca(t){return(ca="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}Ct({target:"Array",proto:!0,forced:!oa},{splice:function(t,e){var n,r,o,i,a,u,s=Lt(this),c=lt(s.length),l=dt(t,c),f=arguments.length;if(0===f?n=r=0:1===f?(n=0,r=c-l):(n=f-2,r=aa(ia(st(e),0),c-l)),c+n-r>ua)throw TypeError(sa);for(o=In(s,r),i=0;i<r;i++)(a=l+i)in s&&te(o,i,s[a]);if(o.length=r,n<r){for(i=l;i<c-r;i++)u=i+n,(a=i+r)in s?s[u]=s[a]:delete s[u];for(i=c;i>c-r+n;i--)delete s[i-1]}else if(n>r)for(i=c-r;i>l;i--)u=i+n-1,(a=i+r-1)in s?s[u]=s[a]:delete s[u];for(i=0;i<n;i++)s[i+l]=arguments[i+2];return s.length=c-r+n,o}}),er("toStringTag"),He(Math,"Math",!0),He(r.JSON,"JSON",!0);var la=o((function(){_e(1)}));Ct({target:"Object",stat:!0,forced:la,sham:!xe},{getPrototypeOf:function(t){return _e(Lt(t))}});var fa=function(){var t=k(this),e="";return t.global&&(e+="g"),t.ignoreCase&&(e+="i"),t.multiline&&(e+="m"),t.dotAll&&(e+="s"),t.unicode&&(e+="u"),t.sticky&&(e+="y"),e},pa="toString",da=RegExp.prototype,ha=da.toString,va=o((function(){return"/a/b"!=ha.call({source:"a",flags:"b"})})),ya=ha.name!=pa;(va||ya)&&nt(RegExp.prototype,pa,(function(){var t=k(this),e=String(t.source),n=t.flags;return"/"+e+"/"+String(void 0===n&&t instanceof RegExp&&!("flags"in da)?fa.call(t):n)}),{unsafe:!0});var ga=e((function(t){var e=function(t){var e,n=Object.prototype,r=n.hasOwnProperty,o="function"==typeof Symbol?Symbol:{},i=o.iterator||"@@iterator",a=o.asyncIterator||"@@asyncIterator",u=o.toStringTag||"@@toStringTag";function s(t,e,n){return Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}),t[e]}try{s({},"")}catch(t){s=function(t,e,n){return t[e]=n}}function c(t,e,n,r){var o=e&&e.prototype instanceof y?e:y,i=Object.create(o.prototype),a=new O(r||[]);return i._invoke=function(t,e,n){var r=f;return function(o,i){if(r===d)throw new Error("Generator is already running");if(r===h){if("throw"===o)throw i;return j()}for(n.method=o,n.arg=i;;){var a=n.delegate;if(a){var u=A(a,n);if(u){if(u===v)continue;return u}}if("next"===n.method)n.sent=n._sent=n.arg;else if("throw"===n.method){if(r===f)throw r=h,n.arg;n.dispatchException(n.arg)}else"return"===n.method&&n.abrupt("return",n.arg);r=d;var s=l(t,e,n);if("normal"===s.type){if(r=n.done?h:p,s.arg===v)continue;return{value:s.arg,done:n.done}}"throw"===s.type&&(r=h,n.method="throw",n.arg=s.arg)}}}(t,n,a),i}function l(t,e,n){try{return{type:"normal",arg:t.call(e,n)}}catch(t){return{type:"throw",arg:t}}}t.wrap=c;var f="suspendedStart",p="suspendedYield",d="executing",h="completed",v={};function y(){}function g(){}function m(){}var x={};x[i]=function(){return this};var b=Object.getPrototypeOf,w=b&&b(b(T([])));w&&w!==n&&r.call(w,i)&&(x=w);var _=m.prototype=y.prototype=Object.create(x);function S(t){["next","throw","return"].forEach((function(e){s(t,e,(function(t){return this._invoke(e,t)}))}))}function E(t,e){function n(o,i,a,u){var s=l(t[o],t,i);if("throw"!==s.type){var c=s.arg,f=c.value;return f&&"object"===ca(f)&&r.call(f,"__await")?e.resolve(f.__await).then((function(t){n("next",t,a,u)}),(function(t){n("throw",t,a,u)})):e.resolve(f).then((function(t){c.value=t,a(c)}),(function(t){return n("throw",t,a,u)}))}u(s.arg)}var o;this._invoke=function(t,r){function i(){return new e((function(e,o){n(t,r,e,o)}))}return o=o?o.then(i,i):i()}}function A(t,n){var r=t.iterator[n.method];if(r===e){if(n.delegate=null,"throw"===n.method){if(t.iterator.return&&(n.method="return",n.arg=e,A(t,n),"throw"===n.method))return v;n.method="throw",n.arg=new TypeError("The iterator does not provide a 'throw' method")}return v}var o=l(r,t.iterator,n.arg);if("throw"===o.type)return n.method="throw",n.arg=o.arg,n.delegate=null,v;var i=o.arg;return i?i.done?(n[t.resultName]=i.value,n.next=t.nextLoc,"return"!==n.method&&(n.method="next",n.arg=e),n.delegate=null,v):i:(n.method="throw",n.arg=new TypeError("iterator result is not an object"),n.delegate=null,v)}function k(t){var e={tryLoc:t[0]};1 in t&&(e.catchLoc=t[1]),2 in t&&(e.finallyLoc=t[2],e.afterLoc=t[3]),this.tryEntries.push(e)}function R(t){var e=t.completion||{};e.type="normal",delete e.arg,t.completion=e}function O(t){this.tryEntries=[{tryLoc:"root"}],t.forEach(k,this),this.reset(!0)}function T(t){if(t){var n=t[i];if(n)return n.call(t);if("function"==typeof t.next)return t;if(!isNaN(t.length)){var o=-1,a=function n(){for(;++o<t.length;)if(r.call(t,o))return n.value=t[o],n.done=!1,n;return n.value=e,n.done=!0,n};return a.next=a}}return{next:j}}function j(){return{value:e,done:!0}}return g.prototype=_.constructor=m,m.constructor=g,g.displayName=s(m,u,"GeneratorFunction"),t.isGeneratorFunction=function(t){var e="function"==typeof t&&t.constructor;return!!e&&(e===g||"GeneratorFunction"===(e.displayName||e.name))},t.mark=function(t){return Object.setPrototypeOf?Object.setPrototypeOf(t,m):(t.__proto__=m,s(t,u,"GeneratorFunction")),t.prototype=Object.create(_),t},t.awrap=function(t){return{__await:t}},S(E.prototype),E.prototype[a]=function(){return this},t.AsyncIterator=E,t.async=function(e,n,r,o,i){void 0===i&&(i=Promise);var a=new E(c(e,n,r,o),i);return t.isGeneratorFunction(n)?a:a.next().then((function(t){return t.done?t.value:a.next()}))},S(_),s(_,u,"Generator"),_[i]=function(){return this},_.toString=function(){return"[object Generator]"},t.keys=function(t){var e=[];for(var n in t)e.push(n);return e.reverse(),function n(){for(;e.length;){var r=e.pop();if(r in t)return n.value=r,n.done=!1,n}return n.done=!0,n}},t.values=T,O.prototype={constructor:O,reset:function(t){if(this.prev=0,this.next=0,this.sent=this._sent=e,this.done=!1,this.delegate=null,this.method="next",this.arg=e,this.tryEntries.forEach(R),!t)for(var n in this)"t"===n.charAt(0)&&r.call(this,n)&&!isNaN(+n.slice(1))&&(this[n]=e)},stop:function(){this.done=!0;var t=this.tryEntries[0].completion;if("throw"===t.type)throw t.arg;return this.rval},dispatchException:function(t){if(this.done)throw t;var n=this;function o(r,o){return u.type="throw",u.arg=t,n.next=r,o&&(n.method="next",n.arg=e),!!o}for(var i=this.tryEntries.length-1;i>=0;--i){var a=this.tryEntries[i],u=a.completion;if("root"===a.tryLoc)return o("end");if(a.tryLoc<=this.prev){var s=r.call(a,"catchLoc"),c=r.call(a,"finallyLoc");if(s&&c){if(this.prev<a.catchLoc)return o(a.catchLoc,!0);if(this.prev<a.finallyLoc)return o(a.finallyLoc)}else if(s){if(this.prev<a.catchLoc)return o(a.catchLoc,!0)}else{if(!c)throw new Error("try statement without catch or finally");if(this.prev<a.finallyLoc)return o(a.finallyLoc)}}}},abrupt:function(t,e){for(var n=this.tryEntries.length-1;n>=0;--n){var o=this.tryEntries[n];if(o.tryLoc<=this.prev&&r.call(o,"finallyLoc")&&this.prev<o.finallyLoc){var i=o;break}}i&&("break"===t||"continue"===t)&&i.tryLoc<=e&&e<=i.finallyLoc&&(i=null);var a=i?i.completion:{};return a.type=t,a.arg=e,i?(this.method="next",this.next=i.finallyLoc,v):this.complete(a)},complete:function(t,e){if("throw"===t.type)throw t.arg;return"break"===t.type||"continue"===t.type?this.next=t.arg:"return"===t.type?(this.rval=this.arg=t.arg,this.method="return",this.next="end"):"normal"===t.type&&e&&(this.next=e),v},finish:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var n=this.tryEntries[e];if(n.finallyLoc===t)return this.complete(n.completion,n.afterLoc),R(n),v}},catch:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var n=this.tryEntries[e];if(n.tryLoc===t){var r=n.completion;if("throw"===r.type){var o=r.arg;R(n)}return o}}throw new Error("illegal catch attempt")},delegateYield:function(t,n,r){return this.delegate={iterator:T(t),resultName:n,nextLoc:r},"next"===this.method&&(this.arg=e),v}},t}(t.exports);try{regeneratorRuntime=e}catch(t){Function("r","regeneratorRuntime = r")(e)}})),ma=function(t,e){return e?"".concat(e,"-").concat(t):t},xa=function(t,e){return e?t.slice(e.length+1):t},ba=function(t,e,n){var r=ma(e,n),o=t.querySelector('[name="'.concat(r,'"]'));return o||null},wa=function(t,e){return"".concat(xa(t,e),"-uploads")},_a=function(t,e,n){var r;return null===(r=ba(t,e,n))||void 0===r?void 0:r.value};function Sa(t,e){return RegExp(t,e)}var Ea={UNSUPPORTED_Y:o((function(){var t=Sa("a","y");return t.lastIndex=2,null!=t.exec("abcd")})),BROKEN_CARET:o((function(){var t=Sa("^r","gy");return t.lastIndex=2,null!=t.exec("str")}))},Aa=RegExp.prototype.exec,ka=String.prototype.replace,Ra=Aa,Oa=function(){var t=/a/,e=/b*/g;return Aa.call(t,"a"),Aa.call(e,"a"),0!==t.lastIndex||0!==e.lastIndex}(),Ta=Ea.UNSUPPORTED_Y||Ea.BROKEN_CARET,ja=void 0!==/()??/.exec("")[1];(Oa||ja||Ta)&&(Ra=function(t){var e,n,r,o,i=this,a=Ta&&i.sticky,u=fa.call(i),s=i.source,c=0,l=t;return a&&(-1===(u=u.replace("y","")).indexOf("g")&&(u+="g"),l=String(t).slice(i.lastIndex),i.lastIndex>0&&(!i.multiline||i.multiline&&"\n"!==t[i.lastIndex-1])&&(s="(?: "+s+")",l=" "+l,c++),n=new RegExp("^(?:"+s+")",u)),ja&&(n=new RegExp("^"+s+"$(?!\\s)",u)),Oa&&(e=i.lastIndex),r=Aa.call(a?n:i,l),a?r?(r.input=r.input.slice(c),r[0]=r[0].slice(c),r.index=i.lastIndex,i.lastIndex+=r[0].length):i.lastIndex=0:Oa&&r&&(i.lastIndex=i.global?r.index+r[0].length:e),ja&&r&&r.length>1&&ka.call(r[0],n,(function(){for(o=1;o<arguments.length-2;o++)void 0===arguments[o]&&(r[o]=void 0)})),r});var Pa=Ra;Ct({target:"RegExp",proto:!0,forced:/./.exec!==Pa},{exec:Pa});
+!function(){var t="undefined"!=typeof globalThis?globalThis:"undefined"!=typeof window?window:"undefined"!=typeof global?global:"undefined"!=typeof self?self:{};function e(t){var e={exports:{}};return t(e,e.exports),e.exports}var n=function(t){return t&&t.Math==Math&&t},r=n("object"==typeof globalThis&&globalThis)||n("object"==typeof window&&window)||n("object"==typeof self&&self)||n("object"==typeof t&&t)||function(){return this}()||Function("return this")(),o=function(t){try{return!!t()}catch(t){return!0}},i=!o((function(){return 7!=Object.defineProperty({},1,{get:function(){return 7}})[1]})),a={}.propertyIsEnumerable,u=Object.getOwnPropertyDescriptor,s={f:u&&!a.call({1:2},1)?function(t){var e=u(this,t);return!!e&&e.enumerable}:a},c=function(t,e){return{enumerable:!(1&t),configurable:!(2&t),writable:!(4&t),value:e}},l={}.toString,f=function(t){return l.call(t).slice(8,-1)},p="".split,d=o((function(){return!Object("z").propertyIsEnumerable(0)}))?function(t){return"String"==f(t)?p.call(t,""):Object(t)}:Object,h=function(t){if(null==t)throw TypeError("Can't call method on "+t);return t},v=function(t){return d(h(t))},y=function(t){return"object"==typeof t?null!==t:"function"==typeof t},g=function(t,e){if(!y(t))return t;var n,r;if(e&&"function"==typeof(n=t.toString)&&!y(r=n.call(t)))return r;if("function"==typeof(n=t.valueOf)&&!y(r=n.call(t)))return r;if(!e&&"function"==typeof(n=t.toString)&&!y(r=n.call(t)))return r;throw TypeError("Can't convert object to primitive value")},m={}.hasOwnProperty,x=function(t,e){return m.call(t,e)},b=r.document,w=y(b)&&y(b.createElement),_=function(t){return w?b.createElement(t):{}},E=!i&&!o((function(){return 7!=Object.defineProperty(_("div"),"a",{get:function(){return 7}}).a})),S=Object.getOwnPropertyDescriptor,A={f:i?S:function(t,e){if(t=v(t),e=g(e,!0),E)try{return S(t,e)}catch(t){}if(x(t,e))return c(!s.f.call(t,e),t[e])}},k=function(t){if(!y(t))throw TypeError(String(t)+" is not an object");return t},R=Object.defineProperty,O={f:i?R:function(t,e,n){if(k(t),e=g(e,!0),k(n),E)try{return R(t,e,n)}catch(t){}if("get"in n||"set"in n)throw TypeError("Accessors not supported");return"value"in n&&(t[e]=n.value),t}},T=i?function(t,e,n){return O.f(t,e,c(1,n))}:function(t,e,n){return t[e]=n,t},j=function(t,e){try{T(r,t,e)}catch(n){r[t]=e}return e},P="__core-js_shared__",C=r[P]||j(P,{}),I=Function.toString;"function"!=typeof C.inspectSource&&(C.inspectSource=function(t){return I.call(t)});var U,L,D,H=C.inspectSource,F=r.WeakMap,M="function"==typeof F&&/native code/.test(H(F)),N=e((function(t){(t.exports=function(t,e){return C[t]||(C[t]=void 0!==e?e:{})})("versions",[]).push({version:"3.9.1",mode:"global",copyright:"© 2021 Denis Pushkarev (zloirock.ru)"})})),B=0,z=Math.random(),q=function(t){return"Symbol("+String(void 0===t?"":t)+")_"+(++B+z).toString(36)},G=N("keys"),$=function(t){return G[t]||(G[t]=q(t))},K={},W=r.WeakMap;if(M){var X=C.state||(C.state=new W),V=X.get,Y=X.has,Q=X.set;U=function(t,e){return e.facade=t,Q.call(X,t,e),e},L=function(t){return V.call(X,t)||{}},D=function(t){return Y.call(X,t)}}else{var J=$("state");K[J]=!0,U=function(t,e){return e.facade=t,T(t,J,e),e},L=function(t){return x(t,J)?t[J]:{}},D=function(t){return x(t,J)}}var Z,tt,et={set:U,get:L,has:D,enforce:function(t){return D(t)?L(t):U(t,{})},getterFor:function(t){return function(e){var n;if(!y(e)||(n=L(e)).type!==t)throw TypeError("Incompatible receiver, "+t+" required");return n}}},nt=e((function(t){var e=et.get,n=et.enforce,o=String(String).split("String");(t.exports=function(t,e,i,a){var u,s=!!a&&!!a.unsafe,c=!!a&&!!a.enumerable,l=!!a&&!!a.noTargetGet;"function"==typeof i&&("string"!=typeof e||x(i,"name")||T(i,"name",e),(u=n(i)).source||(u.source=o.join("string"==typeof e?e:""))),t!==r?(s?!l&&t[e]&&(c=!0):delete t[e],c?t[e]=i:T(t,e,i)):c?t[e]=i:j(e,i)})(Function.prototype,"toString",(function(){return"function"==typeof this&&e(this).source||H(this)}))})),rt=r,ot=function(t){return"function"==typeof t?t:void 0},it=function(t,e){return arguments.length<2?ot(rt[t])||ot(r[t]):rt[t]&&rt[t][e]||r[t]&&r[t][e]},at=Math.ceil,ut=Math.floor,st=function(t){return isNaN(t=+t)?0:(t>0?ut:at)(t)},ct=Math.min,lt=function(t){return t>0?ct(st(t),9007199254740991):0},ft=Math.max,pt=Math.min,dt=function(t,e){var n=st(t);return n<0?ft(n+e,0):pt(n,e)},ht=function(t){return function(e,n,r){var o,i=v(e),a=lt(i.length),u=dt(r,a);if(t&&n!=n){for(;a>u;)if((o=i[u++])!=o)return!0}else for(;a>u;u++)if((t||u in i)&&i[u]===n)return t||u||0;return!t&&-1}},vt={includes:ht(!0),indexOf:ht(!1)},yt=vt.indexOf,gt=function(t,e){var n,r=v(t),o=0,i=[];for(n in r)!x(K,n)&&x(r,n)&&i.push(n);for(;e.length>o;)x(r,n=e[o++])&&(~yt(i,n)||i.push(n));return i},mt=["constructor","hasOwnProperty","isPrototypeOf","propertyIsEnumerable","toLocaleString","toString","valueOf"],xt=mt.concat("length","prototype"),bt={f:Object.getOwnPropertyNames||function(t){return gt(t,xt)}},wt={f:Object.getOwnPropertySymbols},_t=it("Reflect","ownKeys")||function(t){var e=bt.f(k(t)),n=wt.f;return n?e.concat(n(t)):e},Et=function(t,e){for(var n=_t(e),r=O.f,o=A.f,i=0;i<n.length;i++){var a=n[i];x(t,a)||r(t,a,o(e,a))}},St=/#|\.prototype\./,At=function(t,e){var n=Rt[kt(t)];return n==Tt||n!=Ot&&("function"==typeof e?o(e):!!e)},kt=At.normalize=function(t){return String(t).replace(St,".").toLowerCase()},Rt=At.data={},Ot=At.NATIVE="N",Tt=At.POLYFILL="P",jt=At,Pt=A.f,Ct=function(t,e){var n,o,i,a,u,s=t.target,c=t.global,l=t.stat;if(n=c?r:l?r[s]||j(s,{}):(r[s]||{}).prototype)for(o in e){if(a=e[o],i=t.noTargetGet?(u=Pt(n,o))&&u.value:n[o],!jt(c?o:s+(l?".":"#")+o,t.forced)&&void 0!==i){if(typeof a==typeof i)continue;Et(a,i)}(t.sham||i&&i.sham)&&T(a,"sham",!0),nt(n,o,a,t)}},It=function(t){if("function"!=typeof t)throw TypeError(String(t)+" is not a function");return t},Ut=function(t,e,n){if(It(t),void 0===e)return t;switch(n){case 0:return function(){return t.call(e)};case 1:return function(n){return t.call(e,n)};case 2:return function(n,r){return t.call(e,n,r)};case 3:return function(n,r,o){return t.call(e,n,r,o)}}return function(){return t.apply(e,arguments)}},Lt=function(t){return Object(h(t))},Dt=function(t){var e=t.return;if(void 0!==e)return k(e.call(t)).value},Ht=function(t,e,n,r){try{return r?e(k(n)[0],n[1]):e(n)}catch(e){throw Dt(t),e}},Ft="process"==f(r.process),Mt=it("navigator","userAgent")||"",Nt=r.process,Bt=Nt&&Nt.versions,zt=Bt&&Bt.v8;zt?tt=(Z=zt.split("."))[0]+Z[1]:Mt&&(!(Z=Mt.match(/Edge\/(\d+)/))||Z[1]>=74)&&(Z=Mt.match(/Chrome\/(\d+)/))&&(tt=Z[1]);var qt=tt&&+tt,Gt=!!Object.getOwnPropertySymbols&&!o((function(){return!Symbol.sham&&(Ft?38===qt:qt>37&&qt<41)})),$t=Gt&&!Symbol.sham&&"symbol"==typeof Symbol.iterator,Kt=N("wks"),Wt=r.Symbol,Xt=$t?Wt:Wt&&Wt.withoutSetter||q,Vt=function(t){return x(Kt,t)&&(Gt||"string"==typeof Kt[t])||(Gt&&x(Wt,t)?Kt[t]=Wt[t]:Kt[t]=Xt("Symbol."+t)),Kt[t]},Yt={},Qt=Vt("iterator"),Jt=Array.prototype,Zt=function(t){return void 0!==t&&(Yt.Array===t||Jt[Qt]===t)},te=function(t,e,n){var r=g(e);r in t?O.f(t,r,c(0,n)):t[r]=n},ee={};ee[Vt("toStringTag")]="z";var ne="[object z]"===String(ee),re=Vt("toStringTag"),oe="Arguments"==f(function(){return arguments}()),ie=ne?f:function(t){var e,n,r;return void 0===t?"Undefined":null===t?"Null":"string"==typeof(n=function(t,e){try{return t[e]}catch(t){}}(e=Object(t),re))?n:oe?f(e):"Object"==(r=f(e))&&"function"==typeof e.callee?"Arguments":r},ae=Vt("iterator"),ue=function(t){if(null!=t)return t[ae]||t["@@iterator"]||Yt[ie(t)]},se=Vt("iterator"),ce=!1;try{var le=0,fe={next:function(){return{done:!!le++}},return:function(){ce=!0}};fe[se]=function(){return this},Array.from(fe,(function(){throw 2}))}catch(t){}var pe=function(t,e){if(!e&&!ce)return!1;var n=!1;try{var r={};r[se]=function(){return{next:function(){return{done:n=!0}}}},t(r)}catch(t){}return n},de=!pe((function(t){Array.from(t)}));Ct({target:"Array",stat:!0,forced:de},{from:function(t){var e,n,r,o,i,a,u=Lt(t),s="function"==typeof this?this:Array,c=arguments.length,l=c>1?arguments[1]:void 0,f=void 0!==l,p=ue(u),d=0;if(f&&(l=Ut(l,c>2?arguments[2]:void 0,2)),null==p||s==Array&&Zt(p))for(n=new s(e=lt(u.length));e>d;d++)a=f?l(u[d],d):u[d],te(n,d,a);else for(i=(o=p.call(u)).next,n=new s;!(r=i.call(o)).done;d++)a=f?Ht(o,l,[r.value,d],!0):r.value,te(n,d,a);return n.length=d,n}});var he,ve,ye,ge=function(t){return function(e,n){var r,o,i=String(h(e)),a=st(n),u=i.length;return a<0||a>=u?t?"":void 0:(r=i.charCodeAt(a))<55296||r>56319||a+1===u||(o=i.charCodeAt(a+1))<56320||o>57343?t?i.charAt(a):r:t?i.slice(a,a+2):o-56320+(r-55296<<10)+65536}},me={codeAt:ge(!1),charAt:ge(!0)},xe=!o((function(){function t(){}return t.prototype.constructor=null,Object.getPrototypeOf(new t)!==t.prototype})),be=$("IE_PROTO"),we=Object.prototype,_e=xe?Object.getPrototypeOf:function(t){return t=Lt(t),x(t,be)?t[be]:"function"==typeof t.constructor&&t instanceof t.constructor?t.constructor.prototype:t instanceof Object?we:null},Ee=Vt("iterator"),Se=!1;[].keys&&("next"in(ye=[].keys())?(ve=_e(_e(ye)))!==Object.prototype&&(he=ve):Se=!0),(null==he||o((function(){var t={};return he[Ee].call(t)!==t})))&&(he={}),x(he,Ee)||T(he,Ee,(function(){return this}));var Ae,ke={IteratorPrototype:he,BUGGY_SAFARI_ITERATORS:Se},Re=Object.keys||function(t){return gt(t,mt)},Oe=i?Object.defineProperties:function(t,e){k(t);for(var n,r=Re(e),o=r.length,i=0;o>i;)O.f(t,n=r[i++],e[n]);return t},Te=it("document","documentElement"),je=$("IE_PROTO"),Pe=function(){},Ce=function(t){return"<script>"+t+"</"+"script>"},Ie=function(){try{Ae=document.domain&&new ActiveXObject("htmlfile")}catch(t){}var t,e;Ie=Ae?function(t){t.write(Ce("")),t.close();var e=t.parentWindow.Object;return t=null,e}(Ae):((e=_("iframe")).style.display="none",Te.appendChild(e),e.src=String("javascript:"),(t=e.contentWindow.document).open(),t.write(Ce("document.F=Object")),t.close(),t.F);for(var n=mt.length;n--;)delete Ie.prototype[mt[n]];return Ie()};K[je]=!0;var Ue=Object.create||function(t,e){var n;return null!==t?(Pe.prototype=k(t),n=new Pe,Pe.prototype=null,n[je]=t):n=Ie(),void 0===e?n:Oe(n,e)},Le=O.f,De=Vt("toStringTag"),He=function(t,e,n){t&&!x(t=n?t:t.prototype,De)&&Le(t,De,{configurable:!0,value:e})},Fe=ke.IteratorPrototype,Me=function(){return this},Ne=Object.setPrototypeOf||("__proto__"in{}?function(){var t,e=!1,n={};try{(t=Object.getOwnPropertyDescriptor(Object.prototype,"__proto__").set).call(n,[]),e=n instanceof Array}catch(t){}return function(n,r){return k(n),function(t){if(!y(t)&&null!==t)throw TypeError("Can't set "+String(t)+" as a prototype")}(r),e?t.call(n,r):n.__proto__=r,n}}():void 0),Be=ke.IteratorPrototype,ze=ke.BUGGY_SAFARI_ITERATORS,qe=Vt("iterator"),Ge="keys",$e="values",Ke="entries",We=function(){return this},Xe=function(t,e,n,r,o,i,a){!function(t,e,n){var r=e+" Iterator";t.prototype=Ue(Fe,{next:c(1,n)}),He(t,r,!1),Yt[r]=Me}(n,e,r);var u,s,l,f=function(t){if(t===o&&y)return y;if(!ze&&t in h)return h[t];switch(t){case Ge:case $e:case Ke:return function(){return new n(this,t)}}return function(){return new n(this)}},p=e+" Iterator",d=!1,h=t.prototype,v=h[qe]||h["@@iterator"]||o&&h[o],y=!ze&&v||f(o),g="Array"==e&&h.entries||v;if(g&&(u=_e(g.call(new t)),Be!==Object.prototype&&u.next&&(_e(u)!==Be&&(Ne?Ne(u,Be):"function"!=typeof u[qe]&&T(u,qe,We)),He(u,p,!0))),o==$e&&v&&v.name!==$e&&(d=!0,y=function(){return v.call(this)}),h[qe]!==y&&T(h,qe,y),Yt[e]=y,o)if(s={values:f($e),keys:i?y:f(Ge),entries:f(Ke)},a)for(l in s)(ze||d||!(l in h))&&nt(h,l,s[l]);else Ct({target:e,proto:!0,forced:ze||d},s);return s},Ve=me.charAt,Ye="String Iterator",Qe=et.set,Je=et.getterFor(Ye);Xe(String,"String",(function(t){Qe(this,{type:Ye,string:String(t),index:0})}),(function(){var t,e=Je(this),n=e.string,r=e.index;return r>=n.length?{value:void 0,done:!0}:(t=Ve(n,r),e.index+=t.length,{value:t,done:!1})}));var Ze=ne?{}.toString:function(){return"[object "+ie(this)+"]"};ne||nt(Object.prototype,"toString",Ze,{unsafe:!0});var tn=Vt("unscopables"),en=Array.prototype;null==en[tn]&&O.f(en,tn,{configurable:!0,value:Ue(null)});var nn=function(t){en[tn][t]=!0},rn="Array Iterator",on=et.set,an=et.getterFor(rn),un=Xe(Array,"Array",(function(t,e){on(this,{type:rn,target:v(t),index:0,kind:e})}),(function(){var t=an(this),e=t.target,n=t.kind,r=t.index++;return!e||r>=e.length?(t.target=void 0,{value:void 0,done:!0}):"keys"==n?{value:r,done:!1}:"values"==n?{value:e[r],done:!1}:{value:[r,e[r]],done:!1}}),"values");Yt.Arguments=Yt.Array,nn("keys"),nn("values"),nn("entries");var sn={CSSRuleList:0,CSSStyleDeclaration:0,CSSValueList:0,ClientRectList:0,DOMRectList:0,DOMStringList:0,DOMTokenList:1,DataTransferItemList:0,FileList:0,HTMLAllCollection:0,HTMLCollection:0,HTMLFormElement:0,HTMLSelectElement:0,MediaList:0,MimeTypeArray:0,NamedNodeMap:0,NodeList:1,PaintRequestList:0,Plugin:0,PluginArray:0,SVGLengthList:0,SVGNumberList:0,SVGPathSegList:0,SVGPointList:0,SVGStringList:0,SVGTransformList:0,SourceBufferList:0,StyleSheetList:0,TextTrackCueList:0,TextTrackList:0,TouchList:0},cn=Vt("iterator"),ln=Vt("toStringTag"),fn=un.values;for(var pn in sn){var dn=r[pn],hn=dn&&dn.prototype;if(hn){if(hn[cn]!==fn)try{T(hn,cn,fn)}catch(t){hn[cn]=fn}if(hn[ln]||T(hn,ln,pn),sn[pn])for(var vn in un)if(hn[vn]!==un[vn])try{T(hn,vn,un[vn])}catch(t){hn[vn]=un[vn]}}}var yn=!o((function(){return Object.isExtensible(Object.preventExtensions({}))})),gn=e((function(t){var e=O.f,n=q("meta"),r=0,o=Object.isExtensible||function(){return!0},i=function(t){e(t,n,{value:{objectID:"O"+ ++r,weakData:{}}})},a=t.exports={REQUIRED:!1,fastKey:function(t,e){if(!y(t))return"symbol"==typeof t?t:("string"==typeof t?"S":"P")+t;if(!x(t,n)){if(!o(t))return"F";if(!e)return"E";i(t)}return t[n].objectID},getWeakData:function(t,e){if(!x(t,n)){if(!o(t))return!0;if(!e)return!1;i(t)}return t[n].weakData},onFreeze:function(t){return yn&&a.REQUIRED&&o(t)&&!x(t,n)&&i(t),t}};K[n]=!0})),mn=function(t,e){this.stopped=t,this.result=e},xn=function(t,e,n){var r,o,i,a,u,s,c,l=n&&n.that,f=!(!n||!n.AS_ENTRIES),p=!(!n||!n.IS_ITERATOR),d=!(!n||!n.INTERRUPTED),h=Ut(e,l,1+f+d),v=function(t){return r&&Dt(r),new mn(!0,t)},y=function(t){return f?(k(t),d?h(t[0],t[1],v):h(t[0],t[1])):d?h(t,v):h(t)};if(p)r=t;else{if("function"!=typeof(o=ue(t)))throw TypeError("Target is not iterable");if(Zt(o)){for(i=0,a=lt(t.length);a>i;i++)if((u=y(t[i]))&&u instanceof mn)return u;return new mn(!1)}r=o.call(t)}for(s=r.next;!(c=s.call(r)).done;){try{u=y(c.value)}catch(t){throw Dt(r),t}if("object"==typeof u&&u&&u instanceof mn)return u}return new mn(!1)},bn=function(t,e,n){if(!(t instanceof e))throw TypeError("Incorrect "+(n?n+" ":"")+"invocation");return t},wn=function(t,e,n){var r,o;return Ne&&"function"==typeof(r=e.constructor)&&r!==n&&y(o=r.prototype)&&o!==n.prototype&&Ne(t,o),t},_n=function(t,e,n){var i=-1!==t.indexOf("Map"),a=-1!==t.indexOf("Weak"),u=i?"set":"add",s=r[t],c=s&&s.prototype,l=s,f={},p=function(t){var e=c[t];nt(c,t,"add"==t?function(t){return e.call(this,0===t?0:t),this}:"delete"==t?function(t){return!(a&&!y(t))&&e.call(this,0===t?0:t)}:"get"==t?function(t){return a&&!y(t)?void 0:e.call(this,0===t?0:t)}:"has"==t?function(t){return!(a&&!y(t))&&e.call(this,0===t?0:t)}:function(t,n){return e.call(this,0===t?0:t,n),this})};if(jt(t,"function"!=typeof s||!(a||c.forEach&&!o((function(){(new s).entries().next()})))))l=n.getConstructor(e,t,i,u),gn.REQUIRED=!0;else if(jt(t,!0)){var d=new l,h=d[u](a?{}:-0,1)!=d,v=o((function(){d.has(1)})),g=pe((function(t){new s(t)})),m=!a&&o((function(){for(var t=new s,e=5;e--;)t[u](e,e);return!t.has(-0)}));g||((l=e((function(e,n){bn(e,l,t);var r=wn(new s,e,l);return null!=n&&xn(n,r[u],{that:r,AS_ENTRIES:i}),r}))).prototype=c,c.constructor=l),(v||m)&&(p("delete"),p("has"),i&&p("get")),(m||h)&&p(u),a&&c.clear&&delete c.clear}return f[t]=l,Ct({global:!0,forced:l!=s},f),He(l,t),a||n.setStrong(l,t,i),l},En=function(t,e,n){for(var r in e)nt(t,r,e[r],n);return t},Sn=Vt("species"),An=function(t){var e=it(t),n=O.f;i&&e&&!e[Sn]&&n(e,Sn,{configurable:!0,get:function(){return this}})},kn=O.f,Rn=gn.fastKey,On=et.set,Tn=et.getterFor,jn={getConstructor:function(t,e,n,r){var o=t((function(t,a){bn(t,o,e),On(t,{type:e,index:Ue(null),first:void 0,last:void 0,size:0}),i||(t.size=0),null!=a&&xn(a,t[r],{that:t,AS_ENTRIES:n})})),a=Tn(e),u=function(t,e,n){var r,o,u=a(t),c=s(t,e);return c?c.value=n:(u.last=c={index:o=Rn(e,!0),key:e,value:n,previous:r=u.last,next:void 0,removed:!1},u.first||(u.first=c),r&&(r.next=c),i?u.size++:t.size++,"F"!==o&&(u.index[o]=c)),t},s=function(t,e){var n,r=a(t),o=Rn(e);if("F"!==o)return r.index[o];for(n=r.first;n;n=n.next)if(n.key==e)return n};return En(o.prototype,{clear:function(){for(var t=a(this),e=t.index,n=t.first;n;)n.removed=!0,n.previous&&(n.previous=n.previous.next=void 0),delete e[n.index],n=n.next;t.first=t.last=void 0,i?t.size=0:this.size=0},delete:function(t){var e=this,n=a(e),r=s(e,t);if(r){var o=r.next,u=r.previous;delete n.index[r.index],r.removed=!0,u&&(u.next=o),o&&(o.previous=u),n.first==r&&(n.first=o),n.last==r&&(n.last=u),i?n.size--:e.size--}return!!r},forEach:function(t){for(var e,n=a(this),r=Ut(t,arguments.length>1?arguments[1]:void 0,3);e=e?e.next:n.first;)for(r(e.value,e.key,this);e&&e.removed;)e=e.previous},has:function(t){return!!s(this,t)}}),En(o.prototype,n?{get:function(t){var e=s(this,t);return e&&e.value},set:function(t,e){return u(this,0===t?0:t,e)}}:{add:function(t){return u(this,t=0===t?0:t,t)}}),i&&kn(o.prototype,"size",{get:function(){return a(this).size}}),o},setStrong:function(t,e,n){var r=e+" Iterator",o=Tn(e),i=Tn(r);Xe(t,e,(function(t,e){On(this,{type:r,target:t,state:o(t),kind:e,last:void 0})}),(function(){for(var t=i(this),e=t.kind,n=t.last;n&&n.removed;)n=n.previous;return t.target&&(t.last=n=n?n.next:t.state.first)?"keys"==e?{value:n.key,done:!1}:"values"==e?{value:n.value,done:!1}:{value:[n.key,n.value],done:!1}:(t.target=void 0,{value:void 0,done:!0})}),n?"entries":"values",!n,!0),An(e)}};_n("Set",(function(t){return function(){return t(this,arguments.length?arguments[0]:void 0)}}),jn);var Pn=Array.isArray||function(t){return"Array"==f(t)},Cn=Vt("species"),In=function(t,e){var n;return Pn(t)&&("function"!=typeof(n=t.constructor)||n!==Array&&!Pn(n.prototype)?y(n)&&null===(n=n[Cn])&&(n=void 0):n=void 0),new(void 0===n?Array:n)(0===e?0:e)},Un=[].push,Ln=function(t){var e=1==t,n=2==t,r=3==t,o=4==t,i=6==t,a=7==t,u=5==t||i;return function(s,c,l,f){for(var p,h,v=Lt(s),y=d(v),g=Ut(c,l,3),m=lt(y.length),x=0,b=f||In,w=e?b(s,m):n||a?b(s,0):void 0;m>x;x++)if((u||x in y)&&(h=g(p=y[x],x,v),t))if(e)w[x]=h;else if(h)switch(t){case 3:return!0;case 5:return p;case 6:return x;case 2:Un.call(w,p)}else switch(t){case 4:return!1;case 7:Un.call(w,p)}return i?-1:r||o?o:w}},Dn={forEach:Ln(0),map:Ln(1),filter:Ln(2),some:Ln(3),every:Ln(4),find:Ln(5),findIndex:Ln(6),filterOut:Ln(7)},Hn=Vt("species"),Fn=function(t){return qt>=51||!o((function(){var e=[];return(e.constructor={})[Hn]=function(){return{foo:1}},1!==e[t](Boolean).foo}))},Mn=Dn.map,Nn=Fn("map");Ct({target:"Array",proto:!0,forced:!Nn},{map:function(t){return Mn(this,t,arguments.length>1?arguments[1]:void 0)}});var Bn=function(t,e){var n=[][t];return!!n&&o((function(){n.call(null,e||function(){throw 1},1)}))},zn=Dn.forEach,qn=Bn("forEach")?[].forEach:function(t){return zn(this,t,arguments.length>1?arguments[1]:void 0)};for(var Gn in sn){var $n=r[Gn],Kn=$n&&$n.prototype;if(Kn&&Kn.forEach!==qn)try{T(Kn,"forEach",qn)}catch(t){Kn.forEach=qn}}var Wn=function t(e){var n=e.parentElement;return n?"FORM"===n.tagName?n:t(n):null},Xn=o((function(){Re(1)}));Ct({target:"Object",stat:!0,forced:Xn},{keys:function(t){return Re(Lt(t))}});var Vn=bt.f,Yn={}.toString,Qn="object"==typeof window&&window&&Object.getOwnPropertyNames?Object.getOwnPropertyNames(window):[],Jn={f:function(t){return Qn&&"[object Window]"==Yn.call(t)?function(t){try{return Vn(t)}catch(t){return Qn.slice()}}(t):Vn(v(t))}},Zn={f:Vt},tr=O.f,er=function(t){var e=rt.Symbol||(rt.Symbol={});x(e,t)||tr(e,t,{value:Zn.f(t)})},nr=Dn.forEach,rr=$("hidden"),or="Symbol",ir=Vt("toPrimitive"),ar=et.set,ur=et.getterFor(or),sr=Object.prototype,cr=r.Symbol,lr=it("JSON","stringify"),fr=A.f,pr=O.f,dr=Jn.f,hr=s.f,vr=N("symbols"),yr=N("op-symbols"),gr=N("string-to-symbol-registry"),mr=N("symbol-to-string-registry"),xr=N("wks"),br=r.QObject,wr=!br||!br.prototype||!br.prototype.findChild,_r=i&&o((function(){return 7!=Ue(pr({},"a",{get:function(){return pr(this,"a",{value:7}).a}})).a}))?function(t,e,n){var r=fr(sr,e);r&&delete sr[e],pr(t,e,n),r&&t!==sr&&pr(sr,e,r)}:pr,Er=function(t,e){var n=vr[t]=Ue(cr.prototype);return ar(n,{type:or,tag:t,description:e}),i||(n.description=e),n},Sr=$t?function(t){return"symbol"==typeof t}:function(t){return Object(t)instanceof cr},Ar=function(t,e,n){t===sr&&Ar(yr,e,n),k(t);var r=g(e,!0);return k(n),x(vr,r)?(n.enumerable?(x(t,rr)&&t[rr][r]&&(t[rr][r]=!1),n=Ue(n,{enumerable:c(0,!1)})):(x(t,rr)||pr(t,rr,c(1,{})),t[rr][r]=!0),_r(t,r,n)):pr(t,r,n)},kr=function(t,e){k(t);var n=v(e),r=Re(n).concat(jr(n));return nr(r,(function(e){i&&!Rr.call(n,e)||Ar(t,e,n[e])})),t},Rr=function(t){var e=g(t,!0),n=hr.call(this,e);return!(this===sr&&x(vr,e)&&!x(yr,e))&&(!(n||!x(this,e)||!x(vr,e)||x(this,rr)&&this[rr][e])||n)},Or=function(t,e){var n=v(t),r=g(e,!0);if(n!==sr||!x(vr,r)||x(yr,r)){var o=fr(n,r);return!o||!x(vr,r)||x(n,rr)&&n[rr][r]||(o.enumerable=!0),o}},Tr=function(t){var e=dr(v(t)),n=[];return nr(e,(function(t){x(vr,t)||x(K,t)||n.push(t)})),n},jr=function(t){var e=t===sr,n=dr(e?yr:v(t)),r=[];return nr(n,(function(t){!x(vr,t)||e&&!x(sr,t)||r.push(vr[t])})),r};if(Gt||(nt((cr=function(){if(this instanceof cr)throw TypeError("Symbol is not a constructor");var t=arguments.length&&void 0!==arguments[0]?String(arguments[0]):void 0,e=q(t),n=function(t){this===sr&&n.call(yr,t),x(this,rr)&&x(this[rr],e)&&(this[rr][e]=!1),_r(this,e,c(1,t))};return i&&wr&&_r(sr,e,{configurable:!0,set:n}),Er(e,t)}).prototype,"toString",(function(){return ur(this).tag})),nt(cr,"withoutSetter",(function(t){return Er(q(t),t)})),s.f=Rr,O.f=Ar,A.f=Or,bt.f=Jn.f=Tr,wt.f=jr,Zn.f=function(t){return Er(Vt(t),t)},i&&(pr(cr.prototype,"description",{configurable:!0,get:function(){return ur(this).description}}),nt(sr,"propertyIsEnumerable",Rr,{unsafe:!0}))),Ct({global:!0,wrap:!0,forced:!Gt,sham:!Gt},{Symbol:cr}),nr(Re(xr),(function(t){er(t)})),Ct({target:or,stat:!0,forced:!Gt},{for:function(t){var e=String(t);if(x(gr,e))return gr[e];var n=cr(e);return gr[e]=n,mr[n]=e,n},keyFor:function(t){if(!Sr(t))throw TypeError(t+" is not a symbol");if(x(mr,t))return mr[t]},useSetter:function(){wr=!0},useSimple:function(){wr=!1}}),Ct({target:"Object",stat:!0,forced:!Gt,sham:!i},{create:function(t,e){return void 0===e?Ue(t):kr(Ue(t),e)},defineProperty:Ar,defineProperties:kr,getOwnPropertyDescriptor:Or}),Ct({target:"Object",stat:!0,forced:!Gt},{getOwnPropertyNames:Tr,getOwnPropertySymbols:jr}),Ct({target:"Object",stat:!0,forced:o((function(){wt.f(1)}))},{getOwnPropertySymbols:function(t){return wt.f(Lt(t))}}),lr){var Pr=!Gt||o((function(){var t=cr();return"[null]"!=lr([t])||"{}"!=lr({a:t})||"{}"!=lr(Object(t))}));Ct({target:"JSON",stat:!0,forced:Pr},{stringify:function(t,e,n){for(var r,o=[t],i=1;arguments.length>i;)o.push(arguments[i++]);if(r=e,(y(e)||void 0!==t)&&!Sr(t))return Pn(e)||(e=function(t,e){if("function"==typeof r&&(e=r.call(this,t,e)),!Sr(e))return e}),o[1]=e,lr.apply(null,o)}})}cr.prototype[ir]||T(cr.prototype,ir,cr.prototype.valueOf),He(cr,or),K[rr]=!0;var Cr=Dn.filter,Ir=Fn("filter");Ct({target:"Array",proto:!0,forced:!Ir},{filter:function(t){return Cr(this,t,arguments.length>1?arguments[1]:void 0)}});var Ur=A.f,Lr=o((function(){Ur(1)}));function Dr(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}Ct({target:"Object",stat:!0,forced:!i||Lr,sham:!i},{getOwnPropertyDescriptor:function(t,e){return Ur(v(t),e)}}),Ct({target:"Object",stat:!0,sham:!i},{getOwnPropertyDescriptors:function(t){for(var e,n,r=v(t),o=A.f,i=_t(r),a={},u=0;i.length>u;)void 0!==(n=o(r,e=i[u++]))&&te(a,e,n);return a}});var Hr=Vt("isConcatSpreadable"),Fr=9007199254740991,Mr="Maximum allowed index exceeded",Nr=qt>=51||!o((function(){var t=[];return t[Hr]=!1,t.concat()[0]!==t})),Br=Fn("concat"),zr=function(t){if(!y(t))return!1;var e=t[Hr];return void 0!==e?!!e:Pn(t)};Ct({target:"Array",proto:!0,forced:!Nr||!Br},{concat:function(t){var e,n,r,o,i,a=Lt(this),u=In(a,0),s=0;for(e=-1,r=arguments.length;e<r;e++)if(zr(i=-1===e?a:arguments[e])){if(s+(o=lt(i.length))>Fr)throw TypeError(Mr);for(n=0;n<o;n++,s++)n in i&&te(u,s,i[n])}else{if(s>=Fr)throw TypeError(Mr);te(u,s++,i)}return u.length=s,u}});var qr,Gr=Vt("match"),$r=function(t){var e;return y(t)&&(void 0!==(e=t[Gr])?!!e:"RegExp"==f(t))},Kr=function(t){if($r(t))throw TypeError("The method doesn't accept regular expressions");return t},Wr=Vt("match"),Xr=function(t){var e=/./;try{"/./"[t](e)}catch(n){try{return e[Wr]=!1,"/./"[t](e)}catch(t){}}return!1},Vr=A.f,Yr="".startsWith,Qr=Math.min,Jr=Xr("startsWith"),Zr=!(Jr||(qr=Vr(String.prototype,"startsWith"),!qr||qr.writable));Ct({target:"String",proto:!0,forced:!Zr&&!Jr},{startsWith:function(t){var e=String(h(this));Kr(t);var n=lt(Qr(arguments.length>1?arguments[1]:void 0,e.length)),r=String(t);return Yr?Yr.call(e,r,n):e.slice(n,n+r.length)===r}});var to=O.f,eo=Function.prototype,no=eo.toString,ro=/^\s*function ([^ (]*)/,oo="name";function io(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,r=new Array(e);n<e;n++)r[n]=t[n];return r}i&&!(oo in eo)&&to(eo,oo,{configurable:!0,get:function(){try{return no.call(this).match(ro)[1]}catch(t){return""}}});var ao=O.f,uo=r.Symbol;if(i&&"function"==typeof uo&&(!("description"in uo.prototype)||void 0!==uo().description)){var so={},co=function(){var t=arguments.length<1||void 0===arguments[0]?void 0:String(arguments[0]),e=this instanceof co?new uo(t):void 0===t?uo():uo(t);return""===t&&(so[e]=!0),e};Et(co,uo);var lo=co.prototype=uo.prototype;lo.constructor=co;var fo=lo.toString,po="Symbol(test)"==String(uo("test")),ho=/^Symbol\((.*)\)[^)]+$/;ao(lo,"description",{configurable:!0,get:function(){var t=y(this)?this.valueOf():this,e=fo.call(t);if(x(so,t))return"";var n=po?e.slice(7,-1):e.replace(ho,"$1");return""===n?void 0:n}}),Ct({global:!0,forced:!0},{Symbol:co})}er("iterator");var vo=Fn("slice"),yo=Vt("species"),go=[].slice,mo=Math.max;function xo(t,e){if(t){if("string"==typeof t)return io(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);return"Object"===n&&t.constructor&&(n=t.constructor.name),"Map"===n||"Set"===n?Array.from(t):"Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)?io(t,e):void 0}}function bo(t){return function(t){if(Array.isArray(t))return io(t)}(t)||function(t){if("undefined"!=typeof Symbol&&Symbol.iterator in Object(t))return Array.from(t)}(t)||xo(t)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}Ct({target:"Array",proto:!0,forced:!vo},{slice:function(t,e){var n,r,o,i=v(this),a=lt(i.length),u=dt(t,a),s=dt(void 0===e?a:e,a);if(Pn(i)&&("function"!=typeof(n=i.constructor)||n!==Array&&!Pn(n.prototype)?y(n)&&null===(n=n[yo])&&(n=void 0):n=void 0,n===Array||void 0===n))return go.call(i,u,s);for(r=new(void 0===n?Array:n)(mo(s-u,0)),o=0;u<s;u++,o++)u in i&&te(r,o,i[u]);return r.length=o,r}});var wo,_o,Eo,So=r.Promise,Ao=Vt("species"),ko=function(t,e){var n,r=k(t).constructor;return void 0===r||null==(n=k(r)[Ao])?e:It(n)},Ro=/(iphone|ipod|ipad).*applewebkit/i.test(Mt),Oo=r.location,To=r.setImmediate,jo=r.clearImmediate,Po=r.process,Co=r.MessageChannel,Io=r.Dispatch,Uo=0,Lo={},Do="onreadystatechange",Ho=function(t){if(Lo.hasOwnProperty(t)){var e=Lo[t];delete Lo[t],e()}},Fo=function(t){return function(){Ho(t)}},Mo=function(t){Ho(t.data)},No=function(t){r.postMessage(t+"",Oo.protocol+"//"+Oo.host)};To&&jo||(To=function(t){for(var e=[],n=1;arguments.length>n;)e.push(arguments[n++]);return Lo[++Uo]=function(){("function"==typeof t?t:Function(t)).apply(void 0,e)},wo(Uo),Uo},jo=function(t){delete Lo[t]},Ft?wo=function(t){Po.nextTick(Fo(t))}:Io&&Io.now?wo=function(t){Io.now(Fo(t))}:Co&&!Ro?(Eo=(_o=new Co).port2,_o.port1.onmessage=Mo,wo=Ut(Eo.postMessage,Eo,1)):r.addEventListener&&"function"==typeof postMessage&&!r.importScripts&&Oo&&"file:"!==Oo.protocol&&!o(No)?(wo=No,r.addEventListener("message",Mo,!1)):wo=Do in _("script")?function(t){Te.appendChild(_("script")).onreadystatechange=function(){Te.removeChild(this),Ho(t)}}:function(t){setTimeout(Fo(t),0)});var Bo,zo,qo,Go,$o,Ko,Wo,Xo,Vo={set:To,clear:jo},Yo=/web0s(?!.*chrome)/i.test(Mt),Qo=A.f,Jo=Vo.set,Zo=r.MutationObserver||r.WebKitMutationObserver,ti=r.document,ei=r.process,ni=r.Promise,ri=Qo(r,"queueMicrotask"),oi=ri&&ri.value;oi||(Bo=function(){var t,e;for(Ft&&(t=ei.domain)&&t.exit();zo;){e=zo.fn,zo=zo.next;try{e()}catch(t){throw zo?Go():qo=void 0,t}}qo=void 0,t&&t.enter()},Ro||Ft||Yo||!Zo||!ti?ni&&ni.resolve?(Wo=ni.resolve(void 0),Xo=Wo.then,Go=function(){Xo.call(Wo,Bo)}):Go=Ft?function(){ei.nextTick(Bo)}:function(){Jo.call(r,Bo)}:($o=!0,Ko=ti.createTextNode(""),new Zo(Bo).observe(Ko,{characterData:!0}),Go=function(){Ko.data=$o=!$o}));var ii,ai,ui,si,ci=oi||function(t){var e={fn:t,next:void 0};qo&&(qo.next=e),zo||(zo=e,Go()),qo=e},li=function(t){var e,n;this.promise=new t((function(t,r){if(void 0!==e||void 0!==n)throw TypeError("Bad Promise constructor");e=t,n=r})),this.resolve=It(e),this.reject=It(n)},fi={f:function(t){return new li(t)}},pi=function(t,e){if(k(t),y(e)&&e.constructor===t)return e;var n=fi.f(t);return(0,n.resolve)(e),n.promise},di=function(t){try{return{error:!1,value:t()}}catch(t){return{error:!0,value:t}}},hi=Vo.set,vi=Vt("species"),yi="Promise",gi=et.get,mi=et.set,xi=et.getterFor(yi),bi=So,wi=r.TypeError,_i=r.document,Ei=r.process,Si=it("fetch"),Ai=fi.f,ki=Ai,Ri=!!(_i&&_i.createEvent&&r.dispatchEvent),Oi="function"==typeof PromiseRejectionEvent,Ti="unhandledrejection",ji=jt(yi,(function(){if(!(H(bi)!==String(bi))){if(66===qt)return!0;if(!Ft&&!Oi)return!0}if(qt>=51&&/native code/.test(bi))return!1;var t=bi.resolve(1),e=function(t){t((function(){}),(function(){}))};return(t.constructor={})[vi]=e,!(t.then((function(){}))instanceof e)})),Pi=ji||!pe((function(t){bi.all(t).catch((function(){}))})),Ci=function(t){var e;return!(!y(t)||"function"!=typeof(e=t.then))&&e},Ii=function(t,e){if(!t.notified){t.notified=!0;var n=t.reactions;ci((function(){for(var r=t.value,o=1==t.state,i=0;n.length>i;){var a,u,s,c=n[i++],l=o?c.ok:c.fail,f=c.resolve,p=c.reject,d=c.domain;try{l?(o||(2===t.rejection&&Hi(t),t.rejection=1),!0===l?a=r:(d&&d.enter(),a=l(r),d&&(d.exit(),s=!0)),a===c.promise?p(wi("Promise-chain cycle")):(u=Ci(a))?u.call(a,f,p):f(a)):p(r)}catch(t){d&&!s&&d.exit(),p(t)}}t.reactions=[],t.notified=!1,e&&!t.rejection&&Li(t)}))}},Ui=function(t,e,n){var o,i;Ri?((o=_i.createEvent("Event")).promise=e,o.reason=n,o.initEvent(t,!1,!0),r.dispatchEvent(o)):o={promise:e,reason:n},!Oi&&(i=r["on"+t])?i(o):t===Ti&&function(t,e){var n=r.console;n&&n.error&&(1===arguments.length?n.error(t):n.error(t,e))}("Unhandled promise rejection",n)},Li=function(t){hi.call(r,(function(){var e,n=t.facade,r=t.value;if(Di(t)&&(e=di((function(){Ft?Ei.emit("unhandledRejection",r,n):Ui(Ti,n,r)})),t.rejection=Ft||Di(t)?2:1,e.error))throw e.value}))},Di=function(t){return 1!==t.rejection&&!t.parent},Hi=function(t){hi.call(r,(function(){var e=t.facade;Ft?Ei.emit("rejectionHandled",e):Ui("rejectionhandled",e,t.value)}))},Fi=function(t,e,n){return function(r){t(e,r,n)}},Mi=function(t,e,n){t.done||(t.done=!0,n&&(t=n),t.value=e,t.state=2,Ii(t,!0))},Ni=function(t,e,n){if(!t.done){t.done=!0,n&&(t=n);try{if(t.facade===e)throw wi("Promise can't be resolved itself");var r=Ci(e);r?ci((function(){var n={done:!1};try{r.call(e,Fi(Ni,n,t),Fi(Mi,n,t))}catch(e){Mi(n,e,t)}})):(t.value=e,t.state=1,Ii(t,!1))}catch(e){Mi({done:!1},e,t)}}};function Bi(t,e,n,r,o,i,a){try{var u=t[i](a),s=u.value}catch(t){return void n(t)}u.done?e(s):Promise.resolve(s).then(r,o)}function zi(t){return function(){var e=this,n=arguments;return new Promise((function(r,o){var i=t.apply(e,n);function a(t){Bi(i,r,o,a,u,"next",t)}function u(t){Bi(i,r,o,a,u,"throw",t)}a(void 0)}))}}function qi(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}function Gi(t,e){for(var n=0;n<e.length;n++){var r=e[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(t,r.key,r)}}function $i(t,e,n){return e&&Gi(t.prototype,e),n&&Gi(t,n),t}function Ki(t){var e;if("undefined"!=typeof Symbol){if(Symbol.asyncIterator&&null!=(e=t[Symbol.asyncIterator]))return e.call(t);if(Symbol.iterator&&null!=(e=t[Symbol.iterator]))return e.call(t)}throw new TypeError("Object is not async iterable")}ji&&(bi=function(t){bn(this,bi,yi),It(t),ii.call(this);var e=gi(this);try{t(Fi(Ni,e),Fi(Mi,e))}catch(t){Mi(e,t)}},(ii=function(t){mi(this,{type:yi,done:!1,notified:!1,parent:!1,reactions:[],rejection:!1,state:0,value:void 0})}).prototype=En(bi.prototype,{then:function(t,e){var n=xi(this),r=Ai(ko(this,bi));return r.ok="function"!=typeof t||t,r.fail="function"==typeof e&&e,r.domain=Ft?Ei.domain:void 0,n.parent=!0,n.reactions.push(r),0!=n.state&&Ii(n,!1),r.promise},catch:function(t){return this.then(void 0,t)}}),ai=function(){var t=new ii,e=gi(t);this.promise=t,this.resolve=Fi(Ni,e),this.reject=Fi(Mi,e)},fi.f=Ai=function(t){return t===bi||t===ui?new ai(t):ki(t)},"function"==typeof So&&(si=So.prototype.then,nt(So.prototype,"then",(function(t,e){var n=this;return new bi((function(t,e){si.call(n,t,e)})).then(t,e)}),{unsafe:!0}),"function"==typeof Si&&Ct({global:!0,enumerable:!0,forced:!0},{fetch:function(t){return pi(bi,Si.apply(r,arguments))}}))),Ct({global:!0,wrap:!0,forced:ji},{Promise:bi}),He(bi,yi,!1),An(yi),ui=it(yi),Ct({target:yi,stat:!0,forced:ji},{reject:function(t){var e=Ai(this);return e.reject.call(void 0,t),e.promise}}),Ct({target:yi,stat:!0,forced:ji},{resolve:function(t){return pi(this,t)}}),Ct({target:yi,stat:!0,forced:Pi},{all:function(t){var e=this,n=Ai(e),r=n.resolve,o=n.reject,i=di((function(){var n=It(e.resolve),i=[],a=0,u=1;xn(t,(function(t){var s=a++,c=!1;i.push(void 0),u++,n.call(e,t).then((function(t){c||(c=!0,i[s]=t,--u||r(i))}),o)})),--u||r(i)}));return i.error&&o(i.value),n.promise},race:function(t){var e=this,n=Ai(e),r=n.reject,o=di((function(){var o=It(e.resolve);xn(t,(function(t){o.call(e,t).then(n.resolve,r)}))}));return o.error&&r(o.value),n.promise}}),er("asyncIterator");var Wi="".repeat||function(t){var e=String(h(this)),n="",r=st(t);if(r<0||r==1/0)throw RangeError("Wrong number of repetitions");for(;r>0;(r>>>=1)&&(e+=e))1&r&&(n+=e);return n},Xi=1..toFixed,Vi=Math.floor,Yi=function(t,e,n){return 0===e?n:e%2==1?Yi(t,e-1,n*t):Yi(t*t,e/2,n)},Qi=function(t,e,n){for(var r=-1,o=n;++r<6;)o+=e*t[r],t[r]=o%1e7,o=Vi(o/1e7)},Ji=function(t,e){for(var n=6,r=0;--n>=0;)r+=t[n],t[n]=Vi(r/e),r=r%e*1e7},Zi=function(t){for(var e=6,n="";--e>=0;)if(""!==n||0===e||0!==t[e]){var r=String(t[e]);n=""===n?r:n+Wi.call("0",7-r.length)+r}return n},ta=Xi&&("0.000"!==8e-5.toFixed(3)||"1"!==.9.toFixed(0)||"1.25"!==1.255.toFixed(2)||"1000000000000000128"!==(0xde0b6b3a7640080).toFixed(0))||!o((function(){Xi.call({})}));Ct({target:"Number",proto:!0,forced:ta},{toFixed:function(t){var e,n,r,o,i=function(t){if("number"!=typeof t&&"Number"!=f(t))throw TypeError("Incorrect invocation");return+t}(this),a=st(t),u=[0,0,0,0,0,0],s="",c="0";if(a<0||a>20)throw RangeError("Incorrect fraction digits");if(i!=i)return"NaN";if(i<=-1e21||i>=1e21)return String(i);if(i<0&&(s="-",i=-i),i>1e-21)if(n=(e=function(t){for(var e=0,n=t;n>=4096;)e+=12,n/=4096;for(;n>=2;)e+=1,n/=2;return e}(i*Yi(2,69,1))-69)<0?i*Yi(2,-e,1):i/Yi(2,e,1),n*=4503599627370496,(e=52-e)>0){for(Qi(u,0,n),r=a;r>=7;)Qi(u,1e7,0),r-=7;for(Qi(u,Yi(10,r,1),0),r=e-1;r>=23;)Ji(u,1<<23),r-=23;Ji(u,1<<r),Qi(u,1,1),Ji(u,2),c=Zi(u)}else Qi(u,0,n),Qi(u,1<<-e,0),c=Zi(u)+Wi.call("0",a);return c=a>0?s+((o=c.length)<=a?"0."+Wi.call("0",a-o)+c:c.slice(0,o-a)+"."+c.slice(o-a)):s+c}});var ea=Dn.find,na="find",ra=!0;na in[]&&Array(1).find((function(){ra=!1})),Ct({target:"Array",proto:!0,forced:ra},{find:function(t){return ea(this,t,arguments.length>1?arguments[1]:void 0)}}),nn(na);var oa=Fn("splice"),ia=Math.max,aa=Math.min,ua=9007199254740991,sa="Maximum allowed length exceeded";function ca(t){return(ca="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}Ct({target:"Array",proto:!0,forced:!oa},{splice:function(t,e){var n,r,o,i,a,u,s=Lt(this),c=lt(s.length),l=dt(t,c),f=arguments.length;if(0===f?n=r=0:1===f?(n=0,r=c-l):(n=f-2,r=aa(ia(st(e),0),c-l)),c+n-r>ua)throw TypeError(sa);for(o=In(s,r),i=0;i<r;i++)(a=l+i)in s&&te(o,i,s[a]);if(o.length=r,n<r){for(i=l;i<c-r;i++)u=i+n,(a=i+r)in s?s[u]=s[a]:delete s[u];for(i=c;i>c-r+n;i--)delete s[i-1]}else if(n>r)for(i=c-r;i>l;i--)u=i+n-1,(a=i+r-1)in s?s[u]=s[a]:delete s[u];for(i=0;i<n;i++)s[i+l]=arguments[i+2];return s.length=c-r+n,o}}),er("toStringTag"),He(Math,"Math",!0),He(r.JSON,"JSON",!0);var la=o((function(){_e(1)}));Ct({target:"Object",stat:!0,forced:la,sham:!xe},{getPrototypeOf:function(t){return _e(Lt(t))}});var fa=function(){var t=k(this),e="";return t.global&&(e+="g"),t.ignoreCase&&(e+="i"),t.multiline&&(e+="m"),t.dotAll&&(e+="s"),t.unicode&&(e+="u"),t.sticky&&(e+="y"),e},pa="toString",da=RegExp.prototype,ha=da.toString,va=o((function(){return"/a/b"!=ha.call({source:"a",flags:"b"})})),ya=ha.name!=pa;(va||ya)&&nt(RegExp.prototype,pa,(function(){var t=k(this),e=String(t.source),n=t.flags;return"/"+e+"/"+String(void 0===n&&t instanceof RegExp&&!("flags"in da)?fa.call(t):n)}),{unsafe:!0});var ga=e((function(t){var e=function(t){var e,n=Object.prototype,r=n.hasOwnProperty,o="function"==typeof Symbol?Symbol:{},i=o.iterator||"@@iterator",a=o.asyncIterator||"@@asyncIterator",u=o.toStringTag||"@@toStringTag";function s(t,e,n){return Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}),t[e]}try{s({},"")}catch(t){s=function(t,e,n){return t[e]=n}}function c(t,e,n,r){var o=e&&e.prototype instanceof y?e:y,i=Object.create(o.prototype),a=new O(r||[]);return i._invoke=function(t,e,n){var r=f;return function(o,i){if(r===d)throw new Error("Generator is already running");if(r===h){if("throw"===o)throw i;return j()}for(n.method=o,n.arg=i;;){var a=n.delegate;if(a){var u=A(a,n);if(u){if(u===v)continue;return u}}if("next"===n.method)n.sent=n._sent=n.arg;else if("throw"===n.method){if(r===f)throw r=h,n.arg;n.dispatchException(n.arg)}else"return"===n.method&&n.abrupt("return",n.arg);r=d;var s=l(t,e,n);if("normal"===s.type){if(r=n.done?h:p,s.arg===v)continue;return{value:s.arg,done:n.done}}"throw"===s.type&&(r=h,n.method="throw",n.arg=s.arg)}}}(t,n,a),i}function l(t,e,n){try{return{type:"normal",arg:t.call(e,n)}}catch(t){return{type:"throw",arg:t}}}t.wrap=c;var f="suspendedStart",p="suspendedYield",d="executing",h="completed",v={};function y(){}function g(){}function m(){}var x={};x[i]=function(){return this};var b=Object.getPrototypeOf,w=b&&b(b(T([])));w&&w!==n&&r.call(w,i)&&(x=w);var _=m.prototype=y.prototype=Object.create(x);function E(t){["next","throw","return"].forEach((function(e){s(t,e,(function(t){return this._invoke(e,t)}))}))}function S(t,e){function n(o,i,a,u){var s=l(t[o],t,i);if("throw"!==s.type){var c=s.arg,f=c.value;return f&&"object"===ca(f)&&r.call(f,"__await")?e.resolve(f.__await).then((function(t){n("next",t,a,u)}),(function(t){n("throw",t,a,u)})):e.resolve(f).then((function(t){c.value=t,a(c)}),(function(t){return n("throw",t,a,u)}))}u(s.arg)}var o;this._invoke=function(t,r){function i(){return new e((function(e,o){n(t,r,e,o)}))}return o=o?o.then(i,i):i()}}function A(t,n){var r=t.iterator[n.method];if(r===e){if(n.delegate=null,"throw"===n.method){if(t.iterator.return&&(n.method="return",n.arg=e,A(t,n),"throw"===n.method))return v;n.method="throw",n.arg=new TypeError("The iterator does not provide a 'throw' method")}return v}var o=l(r,t.iterator,n.arg);if("throw"===o.type)return n.method="throw",n.arg=o.arg,n.delegate=null,v;var i=o.arg;return i?i.done?(n[t.resultName]=i.value,n.next=t.nextLoc,"return"!==n.method&&(n.method="next",n.arg=e),n.delegate=null,v):i:(n.method="throw",n.arg=new TypeError("iterator result is not an object"),n.delegate=null,v)}function k(t){var e={tryLoc:t[0]};1 in t&&(e.catchLoc=t[1]),2 in t&&(e.finallyLoc=t[2],e.afterLoc=t[3]),this.tryEntries.push(e)}function R(t){var e=t.completion||{};e.type="normal",delete e.arg,t.completion=e}function O(t){this.tryEntries=[{tryLoc:"root"}],t.forEach(k,this),this.reset(!0)}function T(t){if(t){var n=t[i];if(n)return n.call(t);if("function"==typeof t.next)return t;if(!isNaN(t.length)){var o=-1,a=function n(){for(;++o<t.length;)if(r.call(t,o))return n.value=t[o],n.done=!1,n;return n.value=e,n.done=!0,n};return a.next=a}}return{next:j}}function j(){return{value:e,done:!0}}return g.prototype=_.constructor=m,m.constructor=g,g.displayName=s(m,u,"GeneratorFunction"),t.isGeneratorFunction=function(t){var e="function"==typeof t&&t.constructor;return!!e&&(e===g||"GeneratorFunction"===(e.displayName||e.name))},t.mark=function(t){return Object.setPrototypeOf?Object.setPrototypeOf(t,m):(t.__proto__=m,s(t,u,"GeneratorFunction")),t.prototype=Object.create(_),t},t.awrap=function(t){return{__await:t}},E(S.prototype),S.prototype[a]=function(){return this},t.AsyncIterator=S,t.async=function(e,n,r,o,i){void 0===i&&(i=Promise);var a=new S(c(e,n,r,o),i);return t.isGeneratorFunction(n)?a:a.next().then((function(t){return t.done?t.value:a.next()}))},E(_),s(_,u,"Generator"),_[i]=function(){return this},_.toString=function(){return"[object Generator]"},t.keys=function(t){var e=[];for(var n in t)e.push(n);return e.reverse(),function n(){for(;e.length;){var r=e.pop();if(r in t)return n.value=r,n.done=!1,n}return n.done=!0,n}},t.values=T,O.prototype={constructor:O,reset:function(t){if(this.prev=0,this.next=0,this.sent=this._sent=e,this.done=!1,this.delegate=null,this.method="next",this.arg=e,this.tryEntries.forEach(R),!t)for(var n in this)"t"===n.charAt(0)&&r.call(this,n)&&!isNaN(+n.slice(1))&&(this[n]=e)},stop:function(){this.done=!0;var t=this.tryEntries[0].completion;if("throw"===t.type)throw t.arg;return this.rval},dispatchException:function(t){if(this.done)throw t;var n=this;function o(r,o){return u.type="throw",u.arg=t,n.next=r,o&&(n.method="next",n.arg=e),!!o}for(var i=this.tryEntries.length-1;i>=0;--i){var a=this.tryEntries[i],u=a.completion;if("root"===a.tryLoc)return o("end");if(a.tryLoc<=this.prev){var s=r.call(a,"catchLoc"),c=r.call(a,"finallyLoc");if(s&&c){if(this.prev<a.catchLoc)return o(a.catchLoc,!0);if(this.prev<a.finallyLoc)return o(a.finallyLoc)}else if(s){if(this.prev<a.catchLoc)return o(a.catchLoc,!0)}else{if(!c)throw new Error("try statement without catch or finally");if(this.prev<a.finallyLoc)return o(a.finallyLoc)}}}},abrupt:function(t,e){for(var n=this.tryEntries.length-1;n>=0;--n){var o=this.tryEntries[n];if(o.tryLoc<=this.prev&&r.call(o,"finallyLoc")&&this.prev<o.finallyLoc){var i=o;break}}i&&("break"===t||"continue"===t)&&i.tryLoc<=e&&e<=i.finallyLoc&&(i=null);var a=i?i.completion:{};return a.type=t,a.arg=e,i?(this.method="next",this.next=i.finallyLoc,v):this.complete(a)},complete:function(t,e){if("throw"===t.type)throw t.arg;return"break"===t.type||"continue"===t.type?this.next=t.arg:"return"===t.type?(this.rval=this.arg=t.arg,this.method="return",this.next="end"):"normal"===t.type&&e&&(this.next=e),v},finish:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var n=this.tryEntries[e];if(n.finallyLoc===t)return this.complete(n.completion,n.afterLoc),R(n),v}},catch:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var n=this.tryEntries[e];if(n.tryLoc===t){var r=n.completion;if("throw"===r.type){var o=r.arg;R(n)}return o}}throw new Error("illegal catch attempt")},delegateYield:function(t,n,r){return this.delegate={iterator:T(t),resultName:n,nextLoc:r},"next"===this.method&&(this.arg=e),v}},t}(t.exports);try{regeneratorRuntime=e}catch(t){Function("r","regeneratorRuntime = r")(e)}})),ma=function(t,e){return e?"".concat(e,"-").concat(t):t},xa=function(t,e){return e?t.slice(e.length+1):t},ba=function(t,e,n){var r=ma(e,n),o=t.querySelector('[name="'.concat(r,'"]'));return o||null},wa=function(t,e){return"".concat(xa(t,e),"-uploads")},_a=function(t,e,n){var r;return null===(r=ba(t,e,n))||void 0===r?void 0:r.value};function Ea(t,e){return RegExp(t,e)}var Sa={UNSUPPORTED_Y:o((function(){var t=Ea("a","y");return t.lastIndex=2,null!=t.exec("abcd")})),BROKEN_CARET:o((function(){var t=Ea("^r","gy");return t.lastIndex=2,null!=t.exec("str")}))},Aa=RegExp.prototype.exec,ka=String.prototype.replace,Ra=Aa,Oa=function(){var t=/a/,e=/b*/g;return Aa.call(t,"a"),Aa.call(e,"a"),0!==t.lastIndex||0!==e.lastIndex}(),Ta=Sa.UNSUPPORTED_Y||Sa.BROKEN_CARET,ja=void 0!==/()??/.exec("")[1];(Oa||ja||Ta)&&(Ra=function(t){var e,n,r,o,i=this,a=Ta&&i.sticky,u=fa.call(i),s=i.source,c=0,l=t;return a&&(-1===(u=u.replace("y","")).indexOf("g")&&(u+="g"),l=String(t).slice(i.lastIndex),i.lastIndex>0&&(!i.multiline||i.multiline&&"\n"!==t[i.lastIndex-1])&&(s="(?: "+s+")",l=" "+l,c++),n=new RegExp("^(?:"+s+")",u)),ja&&(n=new RegExp("^"+s+"$(?!\\s)",u)),Oa&&(e=i.lastIndex),r=Aa.call(a?n:i,l),a?r?(r.input=r.input.slice(c),r[0]=r[0].slice(c),r.index=i.lastIndex,i.lastIndex+=r[0].length):i.lastIndex=0:Oa&&r&&(i.lastIndex=i.global?r.index+r[0].length:e),ja&&r&&r.length>1&&ka.call(r[0],n,(function(){for(o=1;o<arguments.length-2;o++)void 0===arguments[o]&&(r[o]=void 0)})),r});var Pa=Ra;Ct({target:"RegExp",proto:!0,forced:/./.exec!==Pa},{exec:Pa});
 /*!
 	 * escape-html
 	 * Copyright(c) 2012-2013 TJ Holowaychuk
@@ -6,5 +6,5 @@
 	 * Copyright(c) 2015 Tiancheng "Timothy" Gu
 	 * MIT Licensed
 	 */
-var Ca=/["'&<>]/,Ia=function(t){var e,n=""+t,r=Ca.exec(n);if(!r)return n;var o="",i=0,a=0;for(i=r.index;i<n.length;i++){switch(n.charCodeAt(i)){case 34:e="&quot;";break;case 38:e="&amp;";break;case 39:e="&#39;";break;case 60:e="&lt;";break;case 62:e="&gt;";break;default:continue}a!==i&&(o+=n.substring(a,i)),a=i+1,o+=e}return a!==i?o+n.substring(a,i):o};var Ua=function(){function t(e){var n=e.parent,r=e.input,o=e.skipRequired,i=e.translations;qi(this,t),Dr(this,"container",void 0),Dr(this,"input",void 0),Dr(this,"translations",void 0),Dr(this,"createFilesContainer",(function(t){var e=document.createElement("div");return e.className="dff-files",t.appendChild(e),e})),this.container=this.createFilesContainer(n),this.input=r,this.translations=i,o&&(this.input.required=!1)}return $i(t,[{key:"addNewUpload",value:function(t,e){var n=this.addFile(t,e),r=document.createElement("span");r.className="dff-progress";var o=document.createElement("span");o.className="dff-progress-inner",r.appendChild(o),n.appendChild(r);var i=document.createElement("a");return i.className="dff-cancel",i.innerHTML=this.translations.Cancel,i.setAttribute("data-index","".concat(e)),i.href="#",n.appendChild(i),n}},{key:"addUploadedFile",value:function(t,e,n){var r=this.addFile(t,e);return this.setSuccess(e,n),r}},{key:"clearInput",value:function(){this.input.value=""}},{key:"deleteFile",value:function(t){var e=this.findFileDiv(t);e&&e.remove()}},{key:"disableCancel",value:function(t){var e=this.findCancelSpan(t);e&&e.classList.add("dff-disabled")}},{key:"disableDelete",value:function(t){var e=this.findDeleteLink(t);e&&e.classList.add("dff-disabled")}},{key:"findFileDiv",value:function(t){return this.container.querySelector(".dff-file-id-".concat(t))}},{key:"removeDropHint",value:function(){var t=this.container.querySelector(".dff-drop-hint");t&&t.remove()}},{key:"renderDropHint",value:function(){if(!this.container.querySelector(".dff-drop-hint")){var t=document.createElement("div");t.className="dff-drop-hint",t.innerHTML=this.translations["Drop your files here"],this.container.appendChild(t)}}},{key:"setDeleteFailed",value:function(t){this.setErrorMessage(t,this.translations["Delete failed"]),this.enableDelete(t)}},{key:"setError",value:function(t){this.setErrorMessage(t,this.translations["Upload failed"]);var e=this.findFileDiv(t);e&&e.classList.add("dff-upload-fail"),this.removeProgress(t),this.removeCancel(t)}},{key:"setSuccess",value:function(t,e){var n=this.translations,r=this.findFileDiv(t);if(r){if(r.classList.add("dff-upload-success"),null!=e){var o=document.createElement("span");o.innerHTML=function(t,e){if(0===t)return"0 Bytes";var n=e<=0?0:e||2,r=Math.floor(Math.log(t)/Math.log(1024)),o=parseFloat((t/Math.pow(1024,r)).toFixed(n)),i=["Bytes","KB","MB","GB"][r];return"".concat(o," ").concat(i)}(e,2),o.className="dff-filesize",r.appendChild(o)}var i=document.createElement("a");i.innerHTML=n.Delete,i.className="dff-delete",i.setAttribute("data-index","".concat(t)),i.href="#",r.appendChild(i)}this.removeProgress(t),this.removeCancel(t)}},{key:"updateProgress",value:function(t,e){var n=this.container.querySelector(".dff-file-id-".concat(t));if(n){var r=n.querySelector(".dff-progress-inner");r&&(r.style.width="".concat(e,"%"))}}},{key:"addFile",value:function(t,e){var n=document.createElement("div");n.className="dff-file dff-file-id-".concat(e);var r=document.createElement("span");return r.innerHTML=Ia(t),n.appendChild(r),this.container.appendChild(n),this.input.required=!1,n}},{key:"removeProgress",value:function(t){var e=this.findFileDiv(t);if(e){var n=e.querySelector(".dff-progress");n&&n.remove()}}},{key:"removeCancel",value:function(t){var e=this.findCancelSpan(t);e&&e.remove()}},{key:"findCancelSpan",value:function(t){var e=this.findFileDiv(t);return e?e.querySelector(".dff-cancel"):null}},{key:"enableDelete",value:function(t){var e=this.findDeleteLink(t);e&&e.classList.remove("dff-disabled")}},{key:"findDeleteLink",value:function(t){var e=this.findFileDiv(t);return e?e.querySelector(".dff-delete"):e}},{key:"setErrorMessage",value:function(t,e){var n=this.findFileDiv(t);if(n){var r=document.querySelector(".dff-error");r&&r.remove();var o=document.createElement("span");o.classList.add("dff-error"),o.innerHTML=e,n.appendChild(o)}}}]),t}();function La(t,e){return function(t){if(Array.isArray(t))return t}(t)||function(t,e){if("undefined"!=typeof Symbol&&Symbol.iterator in Object(t)){var n=[],r=!0,o=!1,i=void 0;try{for(var a,u=t[Symbol.iterator]();!(r=(a=u.next()).done)&&(n.push(a.value),!e||n.length!==e);r=!0);}catch(t){o=!0,i=t}finally{try{r||null==u.return||u.return()}finally{if(o)throw i}}return n}}(t,e)||xo(t,e)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}var Da=Vt("species"),Ha=!o((function(){var t=/./;return t.exec=function(){var t=[];return t.groups={a:"7"},t},"7"!=="".replace(t,"$<a>")})),Fa="$0"==="a".replace(/./,"$0"),Ma=Vt("replace"),Na=!!/./[Ma]&&""===/./[Ma]("a","$0"),Ba=!o((function(){var t=/(?:)/,e=t.exec;t.exec=function(){return e.apply(this,arguments)};var n="ab".split(t);return 2!==n.length||"a"!==n[0]||"b"!==n[1]})),za=function(t,e,n,r){var i=Vt(t),a=!o((function(){var e={};return e[i]=function(){return 7},7!=""[t](e)})),u=a&&!o((function(){var e=!1,n=/a/;return"split"===t&&((n={}).constructor={},n.constructor[Da]=function(){return n},n.flags="",n[i]=/./[i]),n.exec=function(){return e=!0,null},n[i](""),!e}));if(!a||!u||"replace"===t&&(!Ha||!Fa||Na)||"split"===t&&!Ba){var s=/./[i],c=n(i,""[t],(function(t,e,n,r,o){return e.exec===Pa?a&&!o?{done:!0,value:s.call(e,n,r)}:{done:!0,value:t.call(n,e,r)}:{done:!1}}),{REPLACE_KEEPS_$0:Fa,REGEXP_REPLACE_SUBSTITUTES_UNDEFINED_CAPTURE:Na}),l=c[0],f=c[1];nt(String.prototype,t,l),nt(RegExp.prototype,i,2==e?function(t,e){return f.call(t,this,e)}:function(t){return f.call(t,this)})}r&&T(RegExp.prototype[i],"sham",!0)},qa=me.charAt,Ga=function(t,e,n){return e+(n?qa(t,e).length:1)},$a=function(t,e){var n=t.exec;if("function"==typeof n){var r=n.call(t,e);if("object"!=typeof r)throw TypeError("RegExp exec method returned something other than an Object or null");return r}if("RegExp"!==f(t))throw TypeError("RegExp#exec called on incompatible receiver");return Pa.call(t,e)},Ka=[].push,Wa=Math.min,Xa=4294967295,Va=!o((function(){return!RegExp(Xa,"y")}));za("split",2,(function(t,e,n){var r;return r="c"=="abbc".split(/(b)*/)[1]||4!="test".split(/(?:)/,-1).length||2!="ab".split(/(?:ab)*/).length||4!=".".split(/(.?)(.?)/).length||".".split(/()()/).length>1||"".split(/.?/).length?function(t,n){var r=String(h(this)),o=void 0===n?Xa:n>>>0;if(0===o)return[];if(void 0===t)return[r];if(!$r(t))return e.call(r,t,o);for(var i,a,u,s=[],c=(t.ignoreCase?"i":"")+(t.multiline?"m":"")+(t.unicode?"u":"")+(t.sticky?"y":""),l=0,f=new RegExp(t.source,c+"g");(i=Pa.call(f,r))&&!((a=f.lastIndex)>l&&(s.push(r.slice(l,i.index)),i.length>1&&i.index<r.length&&Ka.apply(s,i.slice(1)),u=i[0].length,l=a,s.length>=o));)f.lastIndex===i.index&&f.lastIndex++;return l===r.length?!u&&f.test("")||s.push(""):s.push(r.slice(l)),s.length>o?s.slice(0,o):s}:"0".split(void 0,0).length?function(t,n){return void 0===t&&0===n?[]:e.call(this,t,n)}:e,[function(e,n){var o=h(this),i=null==e?void 0:e[t];return void 0!==i?i.call(e,o,n):r.call(String(o),e,n)},function(t,o){var i=n(r,t,this,o,r!==e);if(i.done)return i.value;var a=k(t),u=String(this),s=ko(a,RegExp),c=a.unicode,l=(a.ignoreCase?"i":"")+(a.multiline?"m":"")+(a.unicode?"u":"")+(Va?"y":"g"),f=new s(Va?a:"^(?:"+a.source+")",l),p=void 0===o?Xa:o>>>0;if(0===p)return[];if(0===u.length)return null===$a(f,u)?[u]:[];for(var d=0,h=0,v=[];h<u.length;){f.lastIndex=Va?h:0;var y,g=$a(f,Va?u:u.slice(h));if(null===g||(y=Wa(lt(f.lastIndex+(Va?0:h)),u.length))===d)h=Ga(u,h,c);else{if(v.push(u.slice(d,h)),v.length===p)return v;for(var m=1;m<=g.length-1;m++)if(v.push(g[m]),v.length===p)return v;h=d=y}}return v.push(u.slice(d)),v}]}),!Va);var Ya="\t\n\v\f\r                　\u2028\u2029\ufeff",Qa="["+Ya+"]",Ja=RegExp("^"+Qa+Qa+"*"),Za=RegExp(Qa+Qa+"*$"),tu=function(t){return function(e){var n=String(h(e));return 1&t&&(n=n.replace(Ja,"")),2&t&&(n=n.replace(Za,"")),n}},eu={start:tu(1),end:tu(2),trim:tu(3)},nu=function(t){return o((function(){return!!Ya[t]()||"​᠎"!="​᠎"[t]()||Ya[t].name!==t}))},ru=eu.trim;Ct({target:"String",proto:!0,forced:nu("trim")},{trim:function(){return ru(this)}});var ou=Math.floor,iu="".replace,au=/\$([$&'`]|\d{1,2}|<[^>]*>)/g,uu=/\$([$&'`]|\d{1,2})/g,su=function(t,e,n,r,o,i){var a=n+t.length,u=r.length,s=uu;return void 0!==o&&(o=Lt(o),s=au),iu.call(i,s,(function(i,s){var c;switch(s.charAt(0)){case"$":return"$";case"&":return t;case"`":return e.slice(0,n);case"'":return e.slice(a);case"<":c=o[s.slice(1,-1)];break;default:var l=+s;if(0===l)return i;if(l>u){var f=ou(l/10);return 0===f?i:f<=u?void 0===r[f-1]?s.charAt(1):r[f-1]+s.charAt(1):i}c=r[l-1]}return void 0===c?"":c}))},cu=Math.max,lu=Math.min;za("replace",2,(function(t,e,n,r){var o=r.REGEXP_REPLACE_SUBSTITUTES_UNDEFINED_CAPTURE,i=r.REPLACE_KEEPS_$0,a=o?"$":"$0";return[function(n,r){var o=h(this),i=null==n?void 0:n[t];return void 0!==i?i.call(n,o,r):e.call(String(o),n,r)},function(t,r){if(!o&&i||"string"==typeof r&&-1===r.indexOf(a)){var u=n(e,t,this,r);if(u.done)return u.value}var s=k(t),c=String(this),l="function"==typeof r;l||(r=String(r));var f=s.global;if(f){var p=s.unicode;s.lastIndex=0}for(var d=[];;){var h=$a(s,c);if(null===h)break;if(d.push(h),!f)break;""===String(h[0])&&(s.lastIndex=Ga(c,lt(s.lastIndex),p))}for(var v,y="",g=0,m=0;m<d.length;m++){h=d[m];for(var x=String(h[0]),b=cu(lu(st(h.index),c.length),0),w=[],_=1;_<h.length;_++)w.push(void 0===(v=h[_])?v:String(v));var S=h.groups;if(l){var E=[x].concat(w,b,c);void 0!==S&&E.push(S);var A=String(r.apply(void 0,E))}else A=su(x,c,b,w,S,r);b>=g&&(y+=c.slice(g,b)+A,g=b+x.length)}return y+c.slice(g)}]}));var fu=O.f,pu=bt.f,du=et.set,hu=Vt("match"),vu=r.RegExp,yu=vu.prototype,gu=/a/g,mu=/a/g,xu=new vu(gu)!==gu,bu=Ea.UNSUPPORTED_Y;if(i&&jt("RegExp",!xu||bu||o((function(){return mu[hu]=!1,vu(gu)!=gu||vu(mu)==mu||"/a/i"!=vu(gu,"i")})))){for(var wu=function(t,e){var n,r=this instanceof wu,o=$r(t),i=void 0===e;if(!r&&o&&t.constructor===wu&&i)return t;xu?o&&!i&&(t=t.source):t instanceof wu&&(i&&(e=fa.call(t)),t=t.source),bu&&(n=!!e&&e.indexOf("y")>-1)&&(e=e.replace(/y/g,""));var a=wn(xu?new vu(t,e):vu(t,e),r?this:yu,wu);return bu&&n&&du(a,{sticky:n}),a},_u=function(t){t in wu||fu(wu,t,{configurable:!0,get:function(){return vu[t]},set:function(e){vu[t]=e}})},Su=pu(vu),Eu=0;Su.length>Eu;)_u(Su[Eu++]);yu.constructor=wu,wu.prototype=yu,nt(r,"RegExp",wu)}function Au(){this._types=Object.create(null),this._extensions=Object.create(null);for(var t=0;t<arguments.length;t++)this.define(arguments[t]);this.define=this.define.bind(this),this.getType=this.getType.bind(this),this.getExtension=this.getExtension.bind(this)}An("RegExp"),Au.prototype.define=function(t,e){for(var n in t){var r=t[n].map((function(t){return t.toLowerCase()}));n=n.toLowerCase();for(var o=0;o<r.length;o++){var i=r[o];if("*"!==i[0]){if(!e&&i in this._types)throw new Error('Attempt to change mapping for "'+i+'" extension from "'+this._types[i]+'" to "'+n+'". Pass `force=true` to allow this, otherwise remove "'+i+'" from the list of extensions for "'+n+'".');this._types[i]=n}}if(e||!this._extensions[n]){var a=r[0];this._extensions[n]="*"!==a[0]?a:a.substr(1)}}},Au.prototype.getType=function(t){var e=(t=String(t)).replace(/^.*[/\\]/,"").toLowerCase(),n=e.replace(/^.*\./,"").toLowerCase(),r=e.length<t.length;return(n.length<e.length-1||!r)&&this._types[n]||null},Au.prototype.getExtension=function(t){return(t=/^\s*([^;\s]*)/.test(t)&&RegExp.$1)&&this._extensions[t.toLowerCase()]||null};var ku=new Au({"application/andrew-inset":["ez"],"application/applixware":["aw"],"application/atom+xml":["atom"],"application/atomcat+xml":["atomcat"],"application/atomdeleted+xml":["atomdeleted"],"application/atomsvc+xml":["atomsvc"],"application/atsc-dwd+xml":["dwd"],"application/atsc-held+xml":["held"],"application/atsc-rsat+xml":["rsat"],"application/bdoc":["bdoc"],"application/calendar+xml":["xcs"],"application/ccxml+xml":["ccxml"],"application/cdfx+xml":["cdfx"],"application/cdmi-capability":["cdmia"],"application/cdmi-container":["cdmic"],"application/cdmi-domain":["cdmid"],"application/cdmi-object":["cdmio"],"application/cdmi-queue":["cdmiq"],"application/cu-seeme":["cu"],"application/dash+xml":["mpd"],"application/davmount+xml":["davmount"],"application/docbook+xml":["dbk"],"application/dssc+der":["dssc"],"application/dssc+xml":["xdssc"],"application/ecmascript":["ecma","es"],"application/emma+xml":["emma"],"application/emotionml+xml":["emotionml"],"application/epub+zip":["epub"],"application/exi":["exi"],"application/fdt+xml":["fdt"],"application/font-tdpfr":["pfr"],"application/geo+json":["geojson"],"application/gml+xml":["gml"],"application/gpx+xml":["gpx"],"application/gxf":["gxf"],"application/gzip":["gz"],"application/hjson":["hjson"],"application/hyperstudio":["stk"],"application/inkml+xml":["ink","inkml"],"application/ipfix":["ipfix"],"application/its+xml":["its"],"application/java-archive":["jar","war","ear"],"application/java-serialized-object":["ser"],"application/java-vm":["class"],"application/javascript":["js","mjs"],"application/json":["json","map"],"application/json5":["json5"],"application/jsonml+json":["jsonml"],"application/ld+json":["jsonld"],"application/lgr+xml":["lgr"],"application/lost+xml":["lostxml"],"application/mac-binhex40":["hqx"],"application/mac-compactpro":["cpt"],"application/mads+xml":["mads"],"application/manifest+json":["webmanifest"],"application/marc":["mrc"],"application/marcxml+xml":["mrcx"],"application/mathematica":["ma","nb","mb"],"application/mathml+xml":["mathml"],"application/mbox":["mbox"],"application/mediaservercontrol+xml":["mscml"],"application/metalink+xml":["metalink"],"application/metalink4+xml":["meta4"],"application/mets+xml":["mets"],"application/mmt-aei+xml":["maei"],"application/mmt-usd+xml":["musd"],"application/mods+xml":["mods"],"application/mp21":["m21","mp21"],"application/mp4":["mp4s","m4p"],"application/mrb-consumer+xml":["*xdf"],"application/mrb-publish+xml":["*xdf"],"application/msword":["doc","dot"],"application/mxf":["mxf"],"application/n-quads":["nq"],"application/n-triples":["nt"],"application/node":["cjs"],"application/octet-stream":["bin","dms","lrf","mar","so","dist","distz","pkg","bpk","dump","elc","deploy","exe","dll","deb","dmg","iso","img","msi","msp","msm","buffer"],"application/oda":["oda"],"application/oebps-package+xml":["opf"],"application/ogg":["ogx"],"application/omdoc+xml":["omdoc"],"application/onenote":["onetoc","onetoc2","onetmp","onepkg"],"application/oxps":["oxps"],"application/p2p-overlay+xml":["relo"],"application/patch-ops-error+xml":["*xer"],"application/pdf":["pdf"],"application/pgp-encrypted":["pgp"],"application/pgp-signature":["asc","sig"],"application/pics-rules":["prf"],"application/pkcs10":["p10"],"application/pkcs7-mime":["p7m","p7c"],"application/pkcs7-signature":["p7s"],"application/pkcs8":["p8"],"application/pkix-attr-cert":["ac"],"application/pkix-cert":["cer"],"application/pkix-crl":["crl"],"application/pkix-pkipath":["pkipath"],"application/pkixcmp":["pki"],"application/pls+xml":["pls"],"application/postscript":["ai","eps","ps"],"application/provenance+xml":["provx"],"application/pskc+xml":["pskcxml"],"application/raml+yaml":["raml"],"application/rdf+xml":["rdf","owl"],"application/reginfo+xml":["rif"],"application/relax-ng-compact-syntax":["rnc"],"application/resource-lists+xml":["rl"],"application/resource-lists-diff+xml":["rld"],"application/rls-services+xml":["rs"],"application/route-apd+xml":["rapd"],"application/route-s-tsid+xml":["sls"],"application/route-usd+xml":["rusd"],"application/rpki-ghostbusters":["gbr"],"application/rpki-manifest":["mft"],"application/rpki-roa":["roa"],"application/rsd+xml":["rsd"],"application/rss+xml":["rss"],"application/rtf":["rtf"],"application/sbml+xml":["sbml"],"application/scvp-cv-request":["scq"],"application/scvp-cv-response":["scs"],"application/scvp-vp-request":["spq"],"application/scvp-vp-response":["spp"],"application/sdp":["sdp"],"application/senml+xml":["senmlx"],"application/sensml+xml":["sensmlx"],"application/set-payment-initiation":["setpay"],"application/set-registration-initiation":["setreg"],"application/shf+xml":["shf"],"application/sieve":["siv","sieve"],"application/smil+xml":["smi","smil"],"application/sparql-query":["rq"],"application/sparql-results+xml":["srx"],"application/srgs":["gram"],"application/srgs+xml":["grxml"],"application/sru+xml":["sru"],"application/ssdl+xml":["ssdl"],"application/ssml+xml":["ssml"],"application/swid+xml":["swidtag"],"application/tei+xml":["tei","teicorpus"],"application/thraud+xml":["tfi"],"application/timestamped-data":["tsd"],"application/toml":["toml"],"application/ttml+xml":["ttml"],"application/ubjson":["ubj"],"application/urc-ressheet+xml":["rsheet"],"application/urc-targetdesc+xml":["td"],"application/voicexml+xml":["vxml"],"application/wasm":["wasm"],"application/widget":["wgt"],"application/winhlp":["hlp"],"application/wsdl+xml":["wsdl"],"application/wspolicy+xml":["wspolicy"],"application/xaml+xml":["xaml"],"application/xcap-att+xml":["xav"],"application/xcap-caps+xml":["xca"],"application/xcap-diff+xml":["xdf"],"application/xcap-el+xml":["xel"],"application/xcap-error+xml":["xer"],"application/xcap-ns+xml":["xns"],"application/xenc+xml":["xenc"],"application/xhtml+xml":["xhtml","xht"],"application/xliff+xml":["xlf"],"application/xml":["xml","xsl","xsd","rng"],"application/xml-dtd":["dtd"],"application/xop+xml":["xop"],"application/xproc+xml":["xpl"],"application/xslt+xml":["*xsl","xslt"],"application/xspf+xml":["xspf"],"application/xv+xml":["mxml","xhvml","xvml","xvm"],"application/yang":["yang"],"application/yin+xml":["yin"],"application/zip":["zip"],"audio/3gpp":["*3gpp"],"audio/adpcm":["adp"],"audio/amr":["amr"],"audio/basic":["au","snd"],"audio/midi":["mid","midi","kar","rmi"],"audio/mobile-xmf":["mxmf"],"audio/mp3":["*mp3"],"audio/mp4":["m4a","mp4a"],"audio/mpeg":["mpga","mp2","mp2a","mp3","m2a","m3a"],"audio/ogg":["oga","ogg","spx","opus"],"audio/s3m":["s3m"],"audio/silk":["sil"],"audio/wav":["wav"],"audio/wave":["*wav"],"audio/webm":["weba"],"audio/xm":["xm"],"font/collection":["ttc"],"font/otf":["otf"],"font/ttf":["ttf"],"font/woff":["woff"],"font/woff2":["woff2"],"image/aces":["exr"],"image/apng":["apng"],"image/avif":["avif"],"image/bmp":["bmp"],"image/cgm":["cgm"],"image/dicom-rle":["drle"],"image/emf":["emf"],"image/fits":["fits"],"image/g3fax":["g3"],"image/gif":["gif"],"image/heic":["heic"],"image/heic-sequence":["heics"],"image/heif":["heif"],"image/heif-sequence":["heifs"],"image/hej2k":["hej2"],"image/hsj2":["hsj2"],"image/ief":["ief"],"image/jls":["jls"],"image/jp2":["jp2","jpg2"],"image/jpeg":["jpeg","jpg","jpe"],"image/jph":["jph"],"image/jphc":["jhc"],"image/jpm":["jpm"],"image/jpx":["jpx","jpf"],"image/jxr":["jxr"],"image/jxra":["jxra"],"image/jxrs":["jxrs"],"image/jxs":["jxs"],"image/jxsc":["jxsc"],"image/jxsi":["jxsi"],"image/jxss":["jxss"],"image/ktx":["ktx"],"image/ktx2":["ktx2"],"image/png":["png"],"image/sgi":["sgi"],"image/svg+xml":["svg","svgz"],"image/t38":["t38"],"image/tiff":["tif","tiff"],"image/tiff-fx":["tfx"],"image/webp":["webp"],"image/wmf":["wmf"],"message/disposition-notification":["disposition-notification"],"message/global":["u8msg"],"message/global-delivery-status":["u8dsn"],"message/global-disposition-notification":["u8mdn"],"message/global-headers":["u8hdr"],"message/rfc822":["eml","mime"],"model/3mf":["3mf"],"model/gltf+json":["gltf"],"model/gltf-binary":["glb"],"model/iges":["igs","iges"],"model/mesh":["msh","mesh","silo"],"model/mtl":["mtl"],"model/obj":["obj"],"model/stl":["stl"],"model/vrml":["wrl","vrml"],"model/x3d+binary":["*x3db","x3dbz"],"model/x3d+fastinfoset":["x3db"],"model/x3d+vrml":["*x3dv","x3dvz"],"model/x3d+xml":["x3d","x3dz"],"model/x3d-vrml":["x3dv"],"text/cache-manifest":["appcache","manifest"],"text/calendar":["ics","ifb"],"text/coffeescript":["coffee","litcoffee"],"text/css":["css"],"text/csv":["csv"],"text/html":["html","htm","shtml"],"text/jade":["jade"],"text/jsx":["jsx"],"text/less":["less"],"text/markdown":["markdown","md"],"text/mathml":["mml"],"text/mdx":["mdx"],"text/n3":["n3"],"text/plain":["txt","text","conf","def","list","log","in","ini"],"text/richtext":["rtx"],"text/rtf":["*rtf"],"text/sgml":["sgml","sgm"],"text/shex":["shex"],"text/slim":["slim","slm"],"text/spdx":["spdx"],"text/stylus":["stylus","styl"],"text/tab-separated-values":["tsv"],"text/troff":["t","tr","roff","man","me","ms"],"text/turtle":["ttl"],"text/uri-list":["uri","uris","urls"],"text/vcard":["vcard"],"text/vtt":["vtt"],"text/xml":["*xml"],"text/yaml":["yaml","yml"],"video/3gpp":["3gp","3gpp"],"video/3gpp2":["3g2"],"video/h261":["h261"],"video/h263":["h263"],"video/h264":["h264"],"video/iso.segment":["m4s"],"video/jpeg":["jpgv"],"video/jpm":["*jpm","jpgm"],"video/mj2":["mj2","mjp2"],"video/mp2t":["ts"],"video/mp4":["mp4","mp4v","mpg4"],"video/mpeg":["mpeg","mpg","mpe","m1v","m2v"],"video/ogg":["ogv"],"video/quicktime":["qt","mov"],"video/webm":["webm"]});za("match",1,(function(t,e,n){return[function(e){var n=h(this),r=null==e?void 0:e[t];return void 0!==r?r.call(e,n):new RegExp(e)[t](String(n))},function(t){var r=n(e,t,this);if(r.done)return r.value;var o=k(t),i=String(this);if(!o.global)return $a(o,i);var a=o.unicode;o.lastIndex=0;for(var u,s=[],c=0;null!==(u=$a(o,i));){var l=String(u[0]);s[c]=l,""===l&&(o.lastIndex=Ga(i,lt(o.lastIndex),a)),c++}return 0===c?null:s}]})),i&&("g"!=/./g.flags||Ea.UNSUPPORTED_Y)&&O.f(RegExp.prototype,"flags",{configurable:!0,get:fa});var Ru=function(t){var e,n=0,r=-1,o=!0;for(e=t.length-1;e>=0;--e)if(47===t.charCodeAt(e)){if(!o){n=e+1;break}}else-1===r&&(o=!1,r=e+1);return-1===r?"":t.slice(n,r)};function Ou(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function Tu(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?Ou(Object(n),!0).forEach((function(e){Dr(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):Ou(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}var ju="\\\\/",Pu="[^".concat(ju,"]"),Cu="\\.",Iu="\\/",Uu="[^/]",Lu="(?:".concat(Iu,"|$)"),Du="(?:^|".concat(Iu,")"),Hu="".concat(Cu,"{1,2}").concat(Lu),Fu={DOT_LITERAL:Cu,PLUS_LITERAL:"\\+",QMARK_LITERAL:"\\?",SLASH_LITERAL:Iu,ONE_CHAR:"(?=.)",QMARK:Uu,END_ANCHOR:Lu,DOTS_SLASH:Hu,NO_DOT:"(?!".concat(Cu,")"),NO_DOTS:"(?!".concat(Du).concat(Hu,")"),NO_DOT_SLASH:"(?!".concat(Cu,"{0,1}").concat(Lu,")"),NO_DOTS_SLASH:"(?!".concat(Hu,")"),QMARK_NO_DOT:"[^.".concat(Iu,"]"),STAR:"".concat(Uu,"*?"),START_ANCHOR:Du},Mu=Tu(Tu({},Fu),{},{SLASH_LITERAL:"[".concat(ju,"]"),QMARK:Pu,STAR:"".concat(Pu,"*?"),DOTS_SLASH:"".concat(Cu,"{1,2}(?:[").concat(ju,"]|$)"),NO_DOT:"(?!".concat(Cu,")"),NO_DOTS:"(?!(?:^|[".concat(ju,"])").concat(Cu,"{1,2}(?:[").concat(ju,"]|$))"),NO_DOT_SLASH:"(?!".concat(Cu,"{0,1}(?:[").concat(ju,"]|$))"),NO_DOTS_SLASH:"(?!".concat(Cu,"{1,2}(?:[").concat(ju,"]|$))"),QMARK_NO_DOT:"[^.".concat(ju,"]"),START_ANCHOR:"(?:^|[".concat(ju,"])"),END_ANCHOR:"(?:[".concat(ju,"]|$)")}),Nu={MAX_LENGTH:65536,POSIX_REGEX_SOURCE:{alnum:"a-zA-Z0-9",alpha:"a-zA-Z",ascii:"\\x00-\\x7F",blank:" \\t",cntrl:"\\x00-\\x1F\\x7F",digit:"0-9",graph:"\\x21-\\x7E",lower:"a-z",print:"\\x20-\\x7E ",punct:"\\-!\"#$%&'()\\*+,./:;<=>?@[\\]^_`{|}~",space:" \\t\\r\\n\\v\\f",upper:"A-Z",word:"A-Za-z0-9_",xdigit:"A-Fa-f0-9"},REGEX_BACKSLASH:/\\(?![*+?^${}(|)[\]])/g,REGEX_NON_SPECIAL_CHARS:/^[^@![\].,$*+?^{}()|\\/]+/,REGEX_SPECIAL_CHARS:/[-*+?.^${}(|)[\]]/,REGEX_SPECIAL_CHARS_BACKREF:/(\\?)((\W)(\3*))/g,REGEX_SPECIAL_CHARS_GLOBAL:/([-*+?.^${}(|)[\]])/g,REGEX_REMOVE_BACKSLASH:/(?:\[.*?[^\\]\]|\\(?=.))/g,REPLACEMENTS:{"***":"*","**/**":"**","**/**/**":"**"},CHAR_0:48,CHAR_9:57,CHAR_UPPERCASE_A:65,CHAR_LOWERCASE_A:97,CHAR_UPPERCASE_Z:90,CHAR_LOWERCASE_Z:122,CHAR_LEFT_PARENTHESES:40,CHAR_RIGHT_PARENTHESES:41,CHAR_ASTERISK:42,CHAR_AMPERSAND:38,CHAR_AT:64,CHAR_BACKWARD_SLASH:92,CHAR_CARRIAGE_RETURN:13,CHAR_CIRCUMFLEX_ACCENT:94,CHAR_COLON:58,CHAR_COMMA:44,CHAR_DOT:46,CHAR_DOUBLE_QUOTE:34,CHAR_EQUAL:61,CHAR_EXCLAMATION_MARK:33,CHAR_FORM_FEED:12,CHAR_FORWARD_SLASH:47,CHAR_GRAVE_ACCENT:96,CHAR_HASH:35,CHAR_HYPHEN_MINUS:45,CHAR_LEFT_ANGLE_BRACKET:60,CHAR_LEFT_CURLY_BRACE:123,CHAR_LEFT_SQUARE_BRACKET:91,CHAR_LINE_FEED:10,CHAR_NO_BREAK_SPACE:160,CHAR_PERCENT:37,CHAR_PLUS:43,CHAR_QUESTION_MARK:63,CHAR_RIGHT_ANGLE_BRACKET:62,CHAR_RIGHT_CURLY_BRACE:125,CHAR_RIGHT_SQUARE_BRACKET:93,CHAR_SEMICOLON:59,CHAR_SINGLE_QUOTE:39,CHAR_SPACE:32,CHAR_TAB:9,CHAR_UNDERSCORE:95,CHAR_VERTICAL_LINE:124,CHAR_ZERO_WIDTH_NOBREAK_SPACE:65279,SEP:"/",extglobChars:function(t){return{"!":{type:"negate",open:"(?:(?!(?:",close:"))".concat(t.STAR,")")},"?":{type:"qmark",open:"(?:",close:")?"},"+":{type:"plus",open:"(?:",close:")+"},"*":{type:"star",open:"(?:",close:")*"},"@":{type:"at",open:"(?:",close:")"}}},globChars:function(){var t=arguments.length>0&&void 0!==arguments[0]&&arguments[0];return!0===t?Mu:Fu}},Bu=e((function(t,e){var n=Nu.REGEX_BACKSLASH,r=Nu.REGEX_REMOVE_BACKSLASH,o=Nu.REGEX_SPECIAL_CHARS,i=Nu.REGEX_SPECIAL_CHARS_GLOBAL;e.isObject=function(t){return null!==t&&"object"===ca(t)&&!Array.isArray(t)},e.hasRegexChars=function(t){return o.test(t)},e.isRegexChar=function(t){return 1===t.length&&e.hasRegexChars(t)},e.escapeRegex=function(t){return t.replace(i,"\\$1")},e.toPosixSlashes=function(t){return t.replace(n,"/")},e.removeBackslashes=function(t){return t.replace(r,(function(t){return"\\"===t?"":t}))},e.isWindows=function(t){return!(!t||"boolean"!=typeof t.windows)&&t.windows},e.escapeLast=function(t,n,r){var o=t.lastIndexOf(n,r);return-1===o?t:"\\"===t[o-1]?e.escapeLast(t,n,o-1):"".concat(t.slice(0,o),"\\").concat(t.slice(o))},e.removePrefix=function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},n=t;return n.startsWith("./")&&(n=n.slice(2),e.prefix="./"),n},e.wrapOutput=function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},n=arguments.length>2&&void 0!==arguments[2]?arguments[2]:{},r=n.contains?"":"^",o=n.contains?"":"$",i="".concat(r,"(?:").concat(t,")").concat(o);return!0===e.negated&&(i="(?:^(?!".concat(i,").*$)")),i}})),zu=Nu.CHAR_ASTERISK,qu=Nu.CHAR_AT,Gu=Nu.CHAR_BACKWARD_SLASH,$u=Nu.CHAR_COMMA,Ku=Nu.CHAR_DOT,Wu=Nu.CHAR_EXCLAMATION_MARK,Xu=Nu.CHAR_FORWARD_SLASH,Vu=Nu.CHAR_LEFT_CURLY_BRACE,Yu=Nu.CHAR_LEFT_PARENTHESES,Qu=Nu.CHAR_LEFT_SQUARE_BRACKET,Ju=Nu.CHAR_PLUS,Zu=Nu.CHAR_QUESTION_MARK,ts=Nu.CHAR_RIGHT_CURLY_BRACE,es=Nu.CHAR_RIGHT_PARENTHESES,ns=Nu.CHAR_RIGHT_SQUARE_BRACKET,rs=function(t){return t===Xu||t===Gu},os=function(t){!0!==t.isPrefix&&(t.depth=t.isGlobstar?1/0:1)},is=function(t,e){for(var n,r,o=e||{},i=t.length-1,a=!0===o.parts||!0===o.scanToEnd,u=[],s=[],c=[],l=t,f=-1,p=0,d=0,h=!1,v=!1,y=!1,g=!1,m=!1,x=!1,b=!1,w=!1,_=!1,S=0,E={value:"",depth:0,isGlob:!1},A=function(){return f>=i},k=function(){return n=r,l.charCodeAt(++f)};f<i;){var R=void 0;if((r=k())!==Gu){if(!0===x||r===Vu){for(S++;!0!==A()&&(r=k());)if(r!==Gu)if(r!==Vu){if(!0!==x&&r===Ku&&(r=k())===Ku){if(h=E.isBrace=!0,y=E.isGlob=!0,_=!0,!0===a)continue;break}if(!0!==x&&r===$u){if(h=E.isBrace=!0,y=E.isGlob=!0,_=!0,!0===a)continue;break}if(r===ts&&0===--S){x=!1,h=E.isBrace=!0,_=!0;break}}else S++;else b=E.backslashes=!0,k();if(!0===a)continue;break}if(r!==Xu){if(!0!==o.noext)if(!0===(r===Ju||r===qu||r===zu||r===Zu||r===Wu)&&l.charCodeAt(f+1)===Yu){if(y=E.isGlob=!0,g=E.isExtglob=!0,_=!0,!0===a){for(;!0!==A()&&(r=k());)if(r!==Gu){if(r===es){y=E.isGlob=!0,_=!0;break}}else b=E.backslashes=!0,r=k();continue}break}if(r===zu){if(n===zu&&(m=E.isGlobstar=!0),y=E.isGlob=!0,_=!0,!0===a)continue;break}if(r===Zu){if(y=E.isGlob=!0,_=!0,!0===a)continue;break}if(r===Qu)for(;!0!==A()&&(R=k());)if(R!==Gu){if(R===ns){if(v=E.isBracket=!0,y=E.isGlob=!0,_=!0,!0===a)continue;break}}else b=E.backslashes=!0,k();if(!0===o.nonegate||r!==Wu||f!==p){if(!0!==o.noparen&&r===Yu){if(y=E.isGlob=!0,!0===a){for(;!0!==A()&&(r=k());)if(r!==Yu){if(r===es){_=!0;break}}else b=E.backslashes=!0,r=k();continue}break}if(!0===y){if(_=!0,!0===a)continue;break}}else w=E.negated=!0,p++}else{if(u.push(f),s.push(E),E={value:"",depth:0,isGlob:!1},!0===_)continue;if(n===Ku&&f===p+1){p+=2;continue}d=f+1}}else b=E.backslashes=!0,(r=k())===Vu&&(x=!0)}!0===o.noext&&(g=!1,y=!1);var O=l,T="",j="";p>0&&(T=l.slice(0,p),l=l.slice(p),d-=p),O&&!0===y&&d>0?(O=l.slice(0,d),j=l.slice(d)):!0===y?(O="",j=l):O=l,O&&""!==O&&"/"!==O&&O!==l&&rs(O.charCodeAt(O.length-1))&&(O=O.slice(0,-1)),!0===o.unescape&&(j&&(j=Bu.removeBackslashes(j)),O&&!0===b&&(O=Bu.removeBackslashes(O)));var P={prefix:T,input:t,start:p,base:O,glob:j,isBrace:h,isBracket:v,isGlob:y,isExtglob:g,isGlobstar:m,negated:w};if(!0===o.tokens&&(P.maxDepth=0,rs(r)||s.push(E),P.tokens=s),!0===o.parts||!0===o.tokens){for(var C,I=0;I<u.length;I++){var U=C?C+1:p,L=u[I],D=t.slice(U,L);o.tokens&&(0===I&&0!==p?(s[I].isPrefix=!0,s[I].value=T):s[I].value=D,os(s[I]),P.maxDepth+=s[I].depth),0===I&&""===D||c.push(D),C=L}if(C&&C+1<t.length){var H=t.slice(C+1);c.push(H),o.tokens&&(s[s.length-1].value=H,os(s[s.length-1]),P.maxDepth+=s[s.length-1].depth)}P.slashes=u,P.parts=c}return P},as=[].join,us=d!=Object,ss=Bn("join",",");Ct({target:"Array",proto:!0,forced:us||!ss},{join:function(t){return as.call(v(this),void 0===t?",":t)}});var cs=vt.includes;function ls(t,e){var n;if("undefined"==typeof Symbol||null==t[Symbol.iterator]){if(Array.isArray(t)||(n=function(t,e){if(!t)return;if("string"==typeof t)return fs(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);"Object"===n&&t.constructor&&(n=t.constructor.name);if("Map"===n||"Set"===n)return Array.from(t);if("Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n))return fs(t,e)}(t))||e&&t&&"number"==typeof t.length){n&&(t=n);var r=0,o=function(){};return{s:o,n:function(){return r>=t.length?{done:!0}:{done:!1,value:t[r++]}},e:function(t){throw t},f:o}}throw new TypeError("Invalid attempt to iterate non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}var i,a=!0,u=!1;return{s:function(){n=t[Symbol.iterator]()},n:function(){var t=n.next();return a=t.done,t},e:function(t){u=!0,i=t},f:function(){try{a||null==n.return||n.return()}finally{if(u)throw i}}}}function fs(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,r=new Array(e);n<e;n++)r[n]=t[n];return r}function ps(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function ds(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?ps(Object(n),!0).forEach((function(e){Dr(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):ps(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}Ct({target:"Array",proto:!0},{includes:function(t){return cs(this,t,arguments.length>1?arguments[1]:void 0)}}),nn("includes"),Ct({target:"String",proto:!0,forced:!Xr("includes")},{includes:function(t){return!!~String(h(this)).indexOf(Kr(t),arguments.length>1?arguments[1]:void 0)}}),Ct({target:"String",proto:!0},{repeat:Wi});var hs=Nu.MAX_LENGTH,vs=Nu.POSIX_REGEX_SOURCE,ys=Nu.REGEX_NON_SPECIAL_CHARS,gs=Nu.REGEX_SPECIAL_CHARS_BACKREF,ms=Nu.REPLACEMENTS,xs=function(t,e){if("function"==typeof e.expandRange)return e.expandRange.apply(e,bo(t).concat([e]));t.sort();var n="[".concat(t.join("-"),"]");try{new RegExp(n)}catch(e){return t.map((function(t){return Bu.escapeRegex(t)})).join("..")}return n},bs=function(t,e){return"Missing ".concat(t,': "').concat(e,'" - use "\\\\').concat(e,'" to match literal characters')},ws=function(t,e){if("string"!=typeof t)throw new TypeError("Expected a string");t=ms[t]||t;var n=ds({},e),r="number"==typeof n.maxLength?Math.min(hs,n.maxLength):hs,o=t.length;if(o>r)throw new SyntaxError("Input length: ".concat(o,", exceeds maximum allowed length: ").concat(r));var i={type:"bos",value:"",output:n.prepend||""},a=[i],u=n.capture?"":"?:",s=Nu.globChars(),c=Nu.extglobChars(s),l=s.DOT_LITERAL,f=s.PLUS_LITERAL,p=s.SLASH_LITERAL,d=s.ONE_CHAR,h=s.DOTS_SLASH,v=s.NO_DOT,y=s.NO_DOT_SLASH,g=s.NO_DOTS_SLASH,m=s.QMARK,x=s.QMARK_NO_DOT,b=s.STAR,w=s.START_ANCHOR,_=function(t){return"(".concat(u,"(?:(?!").concat(w).concat(t.dot?h:l,").)*?)")},S=n.dot?"":v,E=n.dot?m:x,A=!0===n.bash?_(n):b;n.capture&&(A="(".concat(A,")")),"boolean"==typeof n.noext&&(n.noextglob=n.noext);var k={input:t,index:-1,start:0,dot:!0===n.dot,consumed:"",output:"",prefix:"",backtrack:!1,negated:!1,brackets:0,braces:0,parens:0,quotes:0,globstar:!1,tokens:a};t=Bu.removePrefix(t,k),o=t.length;var R,O=[],T=[],j=[],P=i,C=function(){return k.index===o-1},I=k.peek=function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:1;return t[k.index+e]},U=k.advance=function(){return t[++k.index]},L=function(){return t.slice(k.index+1)},D=function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:"",e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:0;k.consumed+=t,k.index+=e},H=function(t){k.output+=null!=t.output?t.output:t.value,D(t.value)},F=function(){for(var t=1;"!"===I()&&("("!==I(2)||"?"===I(3));)U(),k.start++,t++;return t%2!=0&&(k.negated=!0,k.start++,!0)},M=function(t){k[t]++,j.push(t)},N=function(t){k[t]--,j.pop()},B=function(t){if("globstar"===P.type){var e=k.braces>0&&("comma"===t.type||"brace"===t.type),n=!0===t.extglob||O.length&&("pipe"===t.type||"paren"===t.type);"slash"===t.type||"paren"===t.type||e||n||(k.output=k.output.slice(0,-P.output.length),P.type="star",P.value="*",P.output=A,k.output+=P.output)}if(O.length&&"paren"!==t.type&&!c[t.value]&&(O[O.length-1].inner+=t.value),(t.value||t.output)&&H(t),P&&"text"===P.type&&"text"===t.type)return P.value+=t.value,void(P.output=(P.output||"")+t.value);t.prev=P,a.push(t),P=t},z=function(t,e){var r=ds(ds({},c[e]),{},{conditions:1,inner:""});r.prev=P,r.parens=k.parens,r.output=k.output;var o=(n.capture?"(":"")+r.open;M("parens"),B({type:t,value:e,output:k.output?"":d}),B({type:"paren",extglob:!0,value:U(),output:o}),O.push(r)},q=function(t){var e=t.close+(n.capture?")":"");if("negate"===t.type){var r=A;t.inner&&t.inner.length>1&&t.inner.includes("/")&&(r=_(n)),(r!==A||C()||/^\)+$/.test(L()))&&(e=t.close=")$))".concat(r)),"bos"===t.prev.type&&C()&&(k.negatedExtglob=!0)}B({type:"paren",extglob:!0,value:R,output:e}),N("parens")};if(!1!==n.fastpaths&&!/(^[*!]|[/()[\]{}"])/.test(t)){var G=!1,$=t.replace(gs,(function(t,e,n,r,o,i){return"\\"===r?(G=!0,t):"?"===r?e?e+r+(o?m.repeat(o.length):""):0===i?E+(o?m.repeat(o.length):""):m.repeat(n.length):"."===r?l.repeat(n.length):"*"===r?e?e+r+(o?A:""):A:e?t:"\\".concat(t)}));return!0===G&&($=!0===n.unescape?$.replace(/\\/g,""):$.replace(/\\+/g,(function(t){return t.length%2==0?"\\\\":t?"\\":""}))),$===t&&!0===n.contains?(k.output=t,k):(k.output=Bu.wrapOutput($,k,e),k)}for(;!C();)if("\0"!==(R=U())){if("\\"===R){var K=I();if("/"===K&&!0!==n.bash)continue;if("."===K||";"===K)continue;if(!K){B({type:"text",value:R+="\\"});continue}var W=/^\\+/.exec(L()),X=0;if(W&&W[0].length>2&&(X=W[0].length,k.index+=X,X%2!=0&&(R+="\\")),!0===n.unescape?R=U()||"":R+=U()||"",0===k.brackets){B({type:"text",value:R});continue}}if(k.brackets>0&&("]"!==R||"["===P.value||"[^"===P.value)){if(!1!==n.posix&&":"===R){var V=P.value.slice(1);if(V.includes("[")&&(P.posix=!0,V.includes(":"))){var Y=P.value.lastIndexOf("["),Q=P.value.slice(0,Y),J=P.value.slice(Y+2),Z=vs[J];if(Z){P.value=Q+Z,k.backtrack=!0,U(),i.output||1!==a.indexOf(P)||(i.output=d);continue}}}("["===R&&":"!==I()||"-"===R&&"]"===I())&&(R="\\".concat(R)),"]"!==R||"["!==P.value&&"[^"!==P.value||(R="\\".concat(R)),!0===n.posix&&"!"===R&&"["===P.value&&(R="^"),P.value+=R,H({value:R})}else if(1!==k.quotes||'"'===R)if('"'!==R)if("("!==R)if(")"!==R)if("["!==R)if("]"!==R)if("{"!==R||!0===n.nobrace)if("}"!==R)if("|"!==R)if(","!==R)if("/"!==R)if("."!==R)if("?"!==R){if("!"===R){if(!0!==n.noextglob&&"("===I()&&("?"!==I(2)||!/[!=<:]/.test(I(3)))){z("negate",R);continue}if(!0!==n.nonegate&&0===k.index){F();continue}}if("+"!==R)if("@"!==R)if("*"===R)if(!P||"globstar"!==P.type&&!0!==P.star){var tt=L();if(!0!==n.noextglob&&/^\([^?]/.test(tt))z("star",R);else if("star"!==P.type){var et={type:"star",value:R,output:A};!0!==n.bash?!P||"bracket"!==P.type&&"paren"!==P.type||!0!==n.regex?(k.index!==k.start&&"slash"!==P.type&&"dot"!==P.type||("dot"===P.type?(k.output+=y,P.output+=y):!0===n.dot?(k.output+=g,P.output+=g):(k.output+=S,P.output+=S),"*"!==I()&&(k.output+=d,P.output+=d)),B(et)):(et.output=R,B(et)):(et.output=".*?","bos"!==P.type&&"slash"!==P.type||(et.output=S+et.output),B(et))}else{if(!0===n.noglobstar){D(R);continue}var nt=P.prev,rt=nt.prev,ot="slash"===nt.type||"bos"===nt.type,it=rt&&("star"===rt.type||"globstar"===rt.type);if(!0===n.bash&&(!ot||tt[0]&&"/"!==tt[0])){B({type:"star",value:R,output:""});continue}var at=k.braces>0&&("comma"===nt.type||"brace"===nt.type),ut=O.length&&("pipe"===nt.type||"paren"===nt.type);if(!ot&&"paren"!==nt.type&&!at&&!ut){B({type:"star",value:R,output:""});continue}for(;"/**"===tt.slice(0,3);){var st=t[k.index+4];if(st&&"/"!==st)break;tt=tt.slice(3),D("/**",3)}if("bos"===nt.type&&C()){P.type="globstar",P.value+=R,P.output=_(n),k.output=P.output,k.globstar=!0,D(R);continue}if("slash"===nt.type&&"bos"!==nt.prev.type&&!it&&C()){k.output=k.output.slice(0,-(nt.output+P.output).length),nt.output="(?:".concat(nt.output),P.type="globstar",P.output=_(n)+(n.strictSlashes?")":"|$)"),P.value+=R,k.globstar=!0,k.output+=nt.output+P.output,D(R);continue}if("slash"===nt.type&&"bos"!==nt.prev.type&&"/"===tt[0]){var ct=void 0!==tt[1]?"|$":"";k.output=k.output.slice(0,-(nt.output+P.output).length),nt.output="(?:".concat(nt.output),P.type="globstar",P.output="".concat(_(n)).concat(p,"|").concat(p).concat(ct,")"),P.value+=R,k.output+=nt.output+P.output,k.globstar=!0,D(R+U()),B({type:"slash",value:"/",output:""});continue}if("bos"===nt.type&&"/"===tt[0]){P.type="globstar",P.value+=R,P.output="(?:^|".concat(p,"|").concat(_(n)).concat(p,")"),k.output=P.output,k.globstar=!0,D(R+U()),B({type:"slash",value:"/",output:""});continue}k.output=k.output.slice(0,-P.output.length),P.type="globstar",P.output=_(n),P.value+=R,k.output+=P.output,k.globstar=!0,D(R)}}else P.type="star",P.star=!0,P.value+=R,P.output=A,k.backtrack=!0,k.globstar=!0,D(R);else{"$"!==R&&"^"!==R||(R="\\".concat(R));var lt=ys.exec(L());lt&&(R+=lt[0],k.index+=lt[0].length),B({type:"text",value:R})}else{if(!0!==n.noextglob&&"("===I()&&"?"!==I(2)){B({type:"at",extglob:!0,value:R,output:""});continue}B({type:"text",value:R})}else{if(!0!==n.noextglob&&"("===I()&&"?"!==I(2)){z("plus",R);continue}if(P&&"("===P.value||!1===n.regex){B({type:"plus",value:R,output:f});continue}if(P&&("bracket"===P.type||"paren"===P.type||"brace"===P.type)||k.parens>0){B({type:"plus",value:R});continue}B({type:"plus",value:f})}}else{if(!(P&&"("===P.value)&&!0!==n.noextglob&&"("===I()&&"?"!==I(2)){z("qmark",R);continue}if(P&&"paren"===P.type){var ft=I(),pt=R;("("===P.value&&!/[!=<:]/.test(ft)||"<"===ft&&!/<([!=]|\w+>)/.test(L()))&&(pt="\\".concat(R)),B({type:"text",value:R,output:pt});continue}if(!0!==n.dot&&("slash"===P.type||"bos"===P.type)){B({type:"qmark",value:R,output:x});continue}B({type:"qmark",value:R,output:m})}else{if(k.braces>0&&"dot"===P.type){"."===P.value&&(P.output=l);var dt=T[T.length-1];P.type="dots",P.output+=R,P.value+=R,dt.dots=!0;continue}if(k.braces+k.parens===0&&"bos"!==P.type&&"slash"!==P.type){B({type:"text",value:R,output:l});continue}B({type:"dot",value:R,output:l})}else{if("dot"===P.type&&k.index===k.start+1){k.start=k.index+1,k.consumed="",k.output="",a.pop(),P=i;continue}B({type:"slash",value:R,output:p})}else{var ht=R,vt=T[T.length-1];vt&&"braces"===j[j.length-1]&&(vt.comma=!0,ht="|"),B({type:"comma",value:R,output:ht})}else O.length>0&&O[O.length-1].conditions++,B({type:"text",value:R});else{var yt=T[T.length-1];if(!0===n.nobrace||!yt){B({type:"text",value:R,output:R});continue}var gt=")";if(!0===yt.dots){for(var mt=a.slice(),xt=[],bt=mt.length-1;bt>=0&&(a.pop(),"brace"!==mt[bt].type);bt--)"dots"!==mt[bt].type&&xt.unshift(mt[bt].value);gt=xs(xt,n),k.backtrack=!0}if(!0!==yt.comma&&!0!==yt.dots){var wt=k.output.slice(0,yt.outputIndex),_t=k.tokens.slice(yt.tokensIndex);yt.value=yt.output="\\{",R=gt="\\}",k.output=wt;var St,Et=ls(_t);try{for(Et.s();!(St=Et.n()).done;){var At=St.value;k.output+=At.output||At.value}}catch(t){Et.e(t)}finally{Et.f()}}B({type:"brace",value:R,output:gt}),N("braces"),T.pop()}else{M("braces");var kt={type:"brace",value:R,output:"(",outputIndex:k.output.length,tokensIndex:k.tokens.length};T.push(kt),B(kt)}else{if(!0===n.nobracket||P&&"bracket"===P.type&&1===P.value.length){B({type:"text",value:R,output:"\\".concat(R)});continue}if(0===k.brackets){if(!0===n.strictBrackets)throw new SyntaxError(bs("opening","["));B({type:"text",value:R,output:"\\".concat(R)});continue}N("brackets");var Rt=P.value.slice(1);if(!0===P.posix||"^"!==Rt[0]||Rt.includes("/")||(R="/".concat(R)),P.value+=R,H({value:R}),!1===n.literalBrackets||Bu.hasRegexChars(Rt))continue;var Ot=Bu.escapeRegex(P.value);if(k.output=k.output.slice(0,-P.value.length),!0===n.literalBrackets){k.output+=Ot,P.value=Ot;continue}P.value="(".concat(u).concat(Ot,"|").concat(P.value,")"),k.output+=P.value}else{if(!0!==n.nobracket&&L().includes("]"))M("brackets");else{if(!0!==n.nobracket&&!0===n.strictBrackets)throw new SyntaxError(bs("closing","]"));R="\\".concat(R)}B({type:"bracket",value:R})}else{if(0===k.parens&&!0===n.strictBrackets)throw new SyntaxError(bs("opening","("));var Tt=O[O.length-1];if(Tt&&k.parens===Tt.parens+1){q(O.pop());continue}B({type:"paren",value:R,output:k.parens?")":"\\)"}),N("parens")}else M("parens"),B({type:"paren",value:R});else k.quotes=1===k.quotes?0:1,!0===n.keepQuotes&&B({type:"text",value:R});else R=Bu.escapeRegex(R),P.value+=R,H({value:R})}for(;k.brackets>0;){if(!0===n.strictBrackets)throw new SyntaxError(bs("closing","]"));k.output=Bu.escapeLast(k.output,"["),N("brackets")}for(;k.parens>0;){if(!0===n.strictBrackets)throw new SyntaxError(bs("closing",")"));k.output=Bu.escapeLast(k.output,"("),N("parens")}for(;k.braces>0;){if(!0===n.strictBrackets)throw new SyntaxError(bs("closing","}"));k.output=Bu.escapeLast(k.output,"{"),N("braces")}if(!0===n.strictSlashes||"star"!==P.type&&"bracket"!==P.type||B({type:"maybe_slash",value:"",output:"".concat(p,"?")}),!0===k.backtrack){k.output="";var jt,Pt=ls(k.tokens);try{for(Pt.s();!(jt=Pt.n()).done;){var Ct=jt.value;k.output+=null!=Ct.output?Ct.output:Ct.value,Ct.suffix&&(k.output+=Ct.suffix)}}catch(t){Pt.e(t)}finally{Pt.f()}}return k};ws.fastpaths=function(t,e){var n=ds({},e),r="number"==typeof n.maxLength?Math.min(hs,n.maxLength):hs,o=t.length;if(o>r)throw new SyntaxError("Input length: ".concat(o,", exceeds maximum allowed length: ").concat(r));t=ms[t]||t;var i=Nu.globChars(),a=i.DOT_LITERAL,u=i.SLASH_LITERAL,s=i.ONE_CHAR,c=i.DOTS_SLASH,l=i.NO_DOT,f=i.NO_DOTS,p=i.NO_DOTS_SLASH,d=i.STAR,h=i.START_ANCHOR,v=n.dot?f:l,y=n.dot?p:l,g=n.capture?"":"?:",m=!0===n.bash?".*?":d;n.capture&&(m="(".concat(m,")"));var x=function(t){return!0===t.noglobstar?m:"(".concat(g,"(?:(?!").concat(h).concat(t.dot?c:a,").)*?)")},b=function t(e){switch(e){case"*":return"".concat(v).concat(s).concat(m);case".*":return"".concat(a).concat(s).concat(m);case"*.*":return"".concat(v).concat(m).concat(a).concat(s).concat(m);case"*/*":return"".concat(v).concat(m).concat(u).concat(s).concat(y).concat(m);case"**":return v+x(n);case"**/*":return"(?:".concat(v).concat(x(n)).concat(u,")?").concat(y).concat(s).concat(m);case"**/*.*":return"(?:".concat(v).concat(x(n)).concat(u,")?").concat(y).concat(m).concat(a).concat(s).concat(m);case"**/.*":return"(?:".concat(v).concat(x(n)).concat(u,")?").concat(a).concat(s).concat(m);default:var r=/^(.*?)\.(\w+)$/.exec(e);if(!r)return;var o=t(r[1]);if(!o)return;return o+a+r[2]}}(Bu.removePrefix(t,{negated:!1,prefix:""}));return b&&!0!==n.strictSlashes&&(b+="".concat(u,"?")),b};var _s=ws;function Ss(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function Es(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?Ss(Object(n),!0).forEach((function(e){Dr(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):Ss(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function As(t,e){var n;if("undefined"==typeof Symbol||null==t[Symbol.iterator]){if(Array.isArray(t)||(n=function(t,e){if(!t)return;if("string"==typeof t)return ks(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);"Object"===n&&t.constructor&&(n=t.constructor.name);if("Map"===n||"Set"===n)return Array.from(t);if("Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n))return ks(t,e)}(t))||e&&t&&"number"==typeof t.length){n&&(t=n);var r=0,o=function(){};return{s:o,n:function(){return r>=t.length?{done:!0}:{done:!1,value:t[r++]}},e:function(t){throw t},f:o}}throw new TypeError("Invalid attempt to iterate non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}var i,a=!0,u=!1;return{s:function(){n=t[Symbol.iterator]()},n:function(){var t=n.next();return a=t.done,t},e:function(t){u=!0,i=t},f:function(){try{a||null==n.return||n.return()}finally{if(u)throw i}}}}function ks(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,r=new Array(e);n<e;n++)r[n]=t[n];return r}var Rs=function(t){return t&&"object"===ca(t)&&!Array.isArray(t)},Os=function t(e,n){var r=arguments.length>2&&void 0!==arguments[2]&&arguments[2];if(Array.isArray(e)){var o=e.map((function(e){return t(e,n,r)})),i=function(t){var e,n=As(o);try{for(n.s();!(e=n.n()).done;){var r=(0,e.value)(t);if(r)return r}}catch(t){n.e(t)}finally{n.f()}return!1};return i}var a=Rs(e)&&e.tokens&&e.input;if(""===e||"string"!=typeof e&&!a)throw new TypeError("Expected pattern to be a non-empty string");var u=n||{},s=Bu.isWindows(n),c=a?t.compileRe(e,n):t.makeRe(e,n,!1,!0),l=c.state;delete c.state;var f=function(){return!1};if(u.ignore){var p=Es(Es({},n),{},{ignore:null,onMatch:null,onResult:null});f=t(u.ignore,p,r)}var d=function(r){var o=arguments.length>1&&void 0!==arguments[1]&&arguments[1],i=t.test(r,c,n,{glob:e,posix:s}),a=i.isMatch,p=i.match,d=i.output,h={glob:e,state:l,regex:c,posix:s,input:r,output:d,match:p,isMatch:a};return"function"==typeof u.onResult&&u.onResult(h),!1===a?(h.isMatch=!1,!!o&&h):f(r)?("function"==typeof u.onIgnore&&u.onIgnore(h),h.isMatch=!1,!!o&&h):("function"==typeof u.onMatch&&u.onMatch(h),!o||h)};return r&&(d.state=l),d};Os.test=function(t,e,n){var r=arguments.length>3&&void 0!==arguments[3]?arguments[3]:{},o=r.glob,i=r.posix;if("string"!=typeof t)throw new TypeError("Expected input to be a string");if(""===t)return{isMatch:!1,output:""};var a=n||{},u=a.format||(i?Bu.toPosixSlashes:null),s=t===o,c=s&&u?u(t):t;return!1===s&&(s=(c=u?u(t):t)===o),!1!==s&&!0!==a.capture||(s=!0===a.matchBase||!0===a.basename?Os.matchBase(t,e,n,i):e.exec(c)),{isMatch:Boolean(s),match:s,output:c}},Os.matchBase=function(t,e,n){arguments.length>3&&void 0!==arguments[3]||Bu.isWindows(n);var r=e instanceof RegExp?e:Os.makeRe(e,n);return r.test(Ru(t))},Os.isMatch=function(t,e,n){return Os(e,n)(t)},Os.parse=function(t,e){return Array.isArray(t)?t.map((function(t){return Os.parse(t,e)})):_s(t,Es(Es({},e),{},{fastpaths:!1}))},Os.scan=function(t,e){return is(t,e)},Os.compileRe=function(t,e){var n=arguments.length>2&&void 0!==arguments[2]&&arguments[2],r=arguments.length>3&&void 0!==arguments[3]&&arguments[3];if(!0===n)return t.output;var o=e||{},i=o.contains?"":"^",a=o.contains?"":"$",u="".concat(i,"(?:").concat(t.output,")").concat(a);t&&!0===t.negated&&(u="^(?!".concat(u,").*$"));var s=Os.toRegex(u,e);return!0===r&&(s.state=t),s},Os.makeRe=function(t,e){var n=arguments.length>2&&void 0!==arguments[2]&&arguments[2],r=arguments.length>3&&void 0!==arguments[3]&&arguments[3];if(!t||"string"!=typeof t)throw new TypeError("Expected a non-empty string");var o,i=e||{},a={negated:!1,fastpaths:!0},u="";return t.startsWith("./")&&(t=t.slice(2),u=a.prefix="./"),!1===i.fastpaths||"."!==t[0]&&"*"!==t[0]||(o=_s.fastpaths(t,e)),void 0===o?(a=_s(t,e)).prefix=u+(a.prefix||""):a.output=o,Os.compileRe(a,e,n,r)},Os.toRegex=function(t,e){try{var n=e||{};return new RegExp(t,n.flags||(n.nocase?"i":""))}catch(t){if(e&&!0===e.debug)throw t;return/$^/}},Os.constants=Nu;var Ts=Os,js=function(){function t(e){qi(this,t),Dr(this,"extensions",void 0),Dr(this,"mimeTypes",void 0);var n=La(function(t){var e=[],n=[];return t.split(",").map((function(t){return t.trim()})).filter(Boolean).forEach((function(t){t.startsWith(".")?e.push("*".concat(t)):n.push(t)})),[e,n]}(e),2),r=n[0],o=n[1];this.extensions=r,this.mimeTypes=o}return $i(t,[{key:"isAccepted",value:function(t){return 0===this.extensions.length&&0===this.mimeTypes.length||(this.isMimeTypeAccepted(ku.getType(t))||this.isExtensionAccepted(t))}},{key:"isMimeTypeAccepted",value:function(t){return 0!==this.mimeTypes.length&&Ts.isMatch(t,this.mimeTypes)}},{key:"isExtensionAccepted",value:function(t){return 0!==this.extensions.length&&Ts.isMatch(t,this.extensions)}}]),t}(),Ps=function(){var t=zi(ga.mark((function t(e){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.abrupt("return",new Promise((function(t,n){return e.createReader().readEntries(t,n)})));case 1:case"end":return t.stop()}}),t)})));return function(e){return t.apply(this,arguments)}}(),Cs=function(){var t=zi(ga.mark((function t(e){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.abrupt("return",new Promise((function(t,n){return e.file(t,n)})));case 1:case"end":return t.stop()}}),t)})));return function(e){return t.apply(this,arguments)}}(),Is=function(){var t=zi(ga.mark((function t(e){var n,r,o,i,a,u,s,c,l,f;return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:n=[],r=!0,o=!1,t.prev=3,a=Ki(e);case 5:return t.next=7,a.next();case 7:return u=t.sent,r=u.done,t.next=11,u.value;case 11:if(s=t.sent,r){t.next=32;break}if(!(c=s).isFile){t.next=21;break}return t.next=17,Cs(c);case 17:l=t.sent,n.push(l),t.next=29;break;case 21:if(!c.isDirectory){t.next=29;break}return t.next=24,Ps(c);case 24:return f=t.sent,t.next=27,Is(f);case 27:t.sent.forEach((function(t){return n.push(t)}));case 29:r=!0,t.next=5;break;case 32:t.next=38;break;case 34:t.prev=34,t.t0=t.catch(3),o=!0,i=t.t0;case 38:if(t.prev=38,t.prev=39,r||null==a.return){t.next=43;break}return t.next=43,a.return();case 43:if(t.prev=43,!o){t.next=46;break}throw i;case 46:return t.finish(43);case 47:return t.finish(38);case 48:return t.abrupt("return",n);case 49:case"end":return t.stop()}}),t,null,[[3,34,38,48],[39,,43,47]])})));return function(e){return t.apply(this,arguments)}}(),Us=function(){var t=zi(ga.mark((function t(e){var n,r;return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if(!e.items){t.next=8;break}return n=bo(e.items).map((function(t){return t.webkitGetAsEntry()})),t.next=4,Is(n);case 4:return r=t.sent,t.abrupt("return",r);case 8:return t.abrupt("return",bo(e.files));case 9:case"end":return t.stop()}}),t)})));return function(e){return t.apply(this,arguments)}}(),Ls=function t(e){var n=this,r=e.container,o=e.inputAccept,i=e.onUploadFiles;qi(this,t),Dr(this,"acceptedFileTypes",void 0),Dr(this,"container",void 0),Dr(this,"onUploadFiles",void 0),Dr(this,"onDrop",(function(t){var e=t;n.container.classList.remove("dff-dropping"),e.preventDefault(),e.stopPropagation(),function(){var t=zi(ga.mark((function t(){var r,o;return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if(t.prev=0,!e.dataTransfer){t.next=7;break}return t.next=4,Us(e.dataTransfer);case 4:r=t.sent,o=r.filter((function(t){return n.acceptedFileTypes.isAccepted(t.name)})),n.onUploadFiles(o);case 7:t.next=12;break;case 9:t.prev=9,t.t0=t.catch(0),console.error(t.t0);case 12:case"end":return t.stop()}}),t,null,[[0,9]])})));return function(){return t.apply(this,arguments)}}()()})),this.container=r,this.onUploadFiles=i,this.acceptedFileTypes=new js(o),r.addEventListener("dragenter",(function(){r.classList.add("dff-dropping")})),r.addEventListener("dragleave",(function(){r.classList.remove("dff-dropping")})),r.addEventListener("dragover",(function(t){r.classList.add("dff-dropping"),t.preventDefault()})),r.addEventListener("drop",this.onDrop)},Ds=[].slice,Hs={},Fs=function(t,e,n){if(!(e in Hs)){for(var r=[],o=0;o<e;o++)r[o]="a["+o+"]";Hs[e]=Function("C,a","return new C("+r.join(",")+")")}return Hs[e](t,n)},Ms=Function.bind||function(t){var e=It(this),n=Ds.call(arguments,1),r=function(){var o=n.concat(Ds.call(arguments));return this instanceof r?Fs(e,o.length,o):e.apply(t,o)};return y(e.prototype)&&(r.prototype=e.prototype),r},Ns=it("Reflect","construct"),Bs=o((function(){function t(){}return!(Ns((function(){}),[],t)instanceof t)})),zs=!o((function(){Ns((function(){}))})),qs=Bs||zs;function Gs(t){if(void 0===t)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return t}function $s(t,e){return($s=Object.setPrototypeOf||function(t,e){return t.__proto__=e,t})(t,e)}function Ks(t,e){if("function"!=typeof e&&null!==e)throw new TypeError("Super expression must either be null or a function");t.prototype=Object.create(e&&e.prototype,{constructor:{value:t,writable:!0,configurable:!0}}),e&&$s(t,e)}function Ws(t,e){return!e||"object"!==ca(e)&&"function"!=typeof e?Gs(t):e}function Xs(t){return(Xs=Object.setPrototypeOf?Object.getPrototypeOf:function(t){return t.__proto__||Object.getPrototypeOf(t)})(t)}Ct({target:"Reflect",stat:!0,forced:qs,sham:qs},{construct:function(t,e){It(t),k(e);var n=arguments.length<3?t:It(arguments[2]);if(zs&&!Bs)return Ns(t,e,n);if(t==n){switch(e.length){case 0:return new t;case 1:return new t(e[0]);case 2:return new t(e[0],e[1]);case 3:return new t(e[0],e[1],e[2]);case 4:return new t(e[0],e[1],e[2],e[3])}var r=[null];return r.push.apply(r,e),new(Ms.apply(t,r))}var o=n.prototype,i=Ue(y(o)?o:Object.prototype),a=Function.apply.call(t,i,e);return y(a)?a:i}});var Vs=function(t){return function(e,n,r,o){It(n);var i=Lt(e),a=d(i),u=lt(i.length),s=t?u-1:0,c=t?-1:1;if(r<2)for(;;){if(s in a){o=a[s],s+=c;break}if(s+=c,t?s<0:u<=s)throw TypeError("Reduce of empty array with no initial value")}for(;t?s>=0:u>s;s+=c)s in a&&(o=n(o,a[s],s,i));return o}},Ys={left:Vs(!1),right:Vs(!0)},Qs=Ys.left,Js=Bn("reduce");Ct({target:"Array",proto:!0,forced:!Js||!Ft&&qt>79&&qt<83},{reduce:function(t){return Qs(this,t,arguments.length,arguments.length>1?arguments[1]:void 0)}});var Zs=function(){function t(e){var n=e.name,r=e.status,o=e.type,i=e.uploadIndex;qi(this,t),Dr(this,"name",void 0),Dr(this,"status",void 0),Dr(this,"type",void 0),Dr(this,"uploadIndex",void 0),this.name=n,this.status=r,this.type=o,this.uploadIndex=i}var e,n;return $i(t,[{key:"abort",value:(n=zi(ga.mark((function t(){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:case"end":return t.stop()}}),t)}))),function(){return n.apply(this,arguments)})},{key:"delete",value:(e=zi(ga.mark((function t(){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:case"end":return t.stop()}}),t)}))),function(){return e.apply(this,arguments)})}]),t}(),tc=e((function(e){var n,r;n=t,r=function(){function t(t){var e=[];if(0===t.length)return"";if("string"!=typeof t[0])throw new TypeError("Url must be a string. Received "+t[0]);if(t[0].match(/^[^/:]+:\/*$/)&&t.length>1){var n=t.shift();t[0]=n+t[0]}t[0].match(/^file:\/\/\//)?t[0]=t[0].replace(/^([^/:]+):\/*/,"$1:///"):t[0]=t[0].replace(/^([^/:]+):\/*/,"$1://");for(var r=0;r<t.length;r++){var o=t[r];if("string"!=typeof o)throw new TypeError("Url must be a string. Received "+o);""!==o&&(r>0&&(o=o.replace(/^[\/]+/,"")),o=r<t.length-1?o.replace(/[\/]+$/,""):o.replace(/[\/]+$/,"/"),e.push(o))}var i=e.join("/"),a=(i=i.replace(/\/(\?|&|#[^!])/g,"$1")).split("?");return i=a.shift()+(a.length>0?"?":"")+a.join("&")}return function(){return t("object"===ca(arguments[0])?arguments[0]:[].slice.call(arguments))}},e.exports?e.exports=r():n.urljoin=r()})),ec=function(t){var e=t.csrfToken,n=t.key,r=t.uploadId,o=t.endpoint,i=encodeURIComponent(n),a=encodeURIComponent(r),u=new Headers({"X-CSRFToken":e}),s=tc(o,a,"?key=".concat(i));return fetch(s,{method:"delete",headers:u}).then((function(t){return t.json()}))},nc=function(t,e){var n=t.indexOf(e);-1!==n&&t.splice(n,1)};function rc(t){var e=function(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Boolean.prototype.valueOf.call(Reflect.construct(Boolean,[],(function(){}))),!0}catch(t){return!1}}();return function(){var n,r=Xs(t);if(e){var o=Xs(this).constructor;n=Reflect.construct(r,arguments,o)}else n=r.apply(this,arguments);return Ws(this,n)}}var oc=function(t){Ks(o,t);var e,n,r=rc(o);function o(t){var e,n=t.csrfToken,i=t.endpoint,a=t.file,u=t.s3UploadDir,s=t.uploadIndex;return qi(this,o),Dr(Gs(e=r.call(this,{name:a.name,status:"uploading",type:"s3",uploadIndex:s})),"onError",void 0),Dr(Gs(e),"onProgress",void 0),Dr(Gs(e),"onSuccess",void 0),Dr(Gs(e),"chunkState",void 0),Dr(Gs(e),"chunks",void 0),Dr(Gs(e),"createdPromise",void 0),Dr(Gs(e),"csrfToken",void 0),Dr(Gs(e),"endpoint",void 0),Dr(Gs(e),"file",void 0),Dr(Gs(e),"key",void 0),Dr(Gs(e),"parts",void 0),Dr(Gs(e),"s3UploadDir",void 0),Dr(Gs(e),"uploadId",void 0),Dr(Gs(e),"uploading",void 0),e.csrfToken=n,e.endpoint=i,e.file=a,e.s3UploadDir=u,e.key=null,e.uploadId=null,e.parts=[],e.createdPromise=Promise.reject(),e.chunks=[],e.chunkState=[],e.uploading=[],e.onError=void 0,e.onProgress=void 0,e.onSuccess=void 0,e.initChunks(),e.createdPromise.catch((function(){return{}})),e}return $i(o,[{key:"abort",value:(n=zi(ga.mark((function t(){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return this.uploading.slice().forEach((function(t){t.abort()})),this.uploading=[],t.next=4,this.createdPromise;case 4:if(!this.key||!this.uploadId){t.next=7;break}return t.next=7,ec({csrfToken:this.csrfToken,endpoint:this.endpoint,key:this.key,uploadId:this.uploadId});case 7:case"end":return t.stop()}}),t,this)}))),function(){return n.apply(this,arguments)})},{key:"delete",value:(e=zi(ga.mark((function t(){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.abrupt("return",Promise.resolve());case 1:case"end":return t.stop()}}),t)}))),function(){return e.apply(this,arguments)})},{key:"getInitialFile",value:function(){return{id:this.uploadId||"",name:this.key||"",size:this.file.size,original_name:this.file.name,type:"s3"}}},{key:"getSize",value:function(){return this.file.size}},{key:"start",value:function(){this.createUpload()}},{key:"initChunks",value:function(){for(var t,e=[],n=(t=this.file,Math.ceil(t.size/1e4)),r=Math.max(5242880,Math.ceil(this.file.size/1e4)),o=Math.max(n,r),i=0;i<this.file.size;i+=o){var a=Math.min(this.file.size,i+o);e.push(this.file.slice(i,a))}this.chunks=e,this.chunkState=e.map((function(){return{uploaded:0,busy:!1,done:!1}}))}},{key:"createUpload",value:function(){var t,e,n,r,o,i,a=this;return this.createdPromise=(t={csrfToken:this.csrfToken,endpoint:this.endpoint,file:this.file,s3UploadDir:this.s3UploadDir},e=t.csrfToken,n=t.endpoint,r=t.file,o=t.s3UploadDir,i=new Headers({accept:"application/json","content-type":"application/json","X-CSRFToken":e}),fetch(n,{method:"post",headers:i,body:JSON.stringify({filename:r.name,contentType:r.type,s3UploadDir:o})}).then((function(t){return t.json()})).then((function(t){return t}))),this.createdPromise.then((function(t){if(!("object"===ca(t)&&t&&"string"==typeof t.uploadId&&"string"==typeof t.key))throw new TypeError("AwsS3/Multipart: Got incorrect result from `createMultipartUpload()`, expected an object `{ uploadId, key }`.");a.key=t.key,a.uploadId=t.uploadId,a.uploadParts()})).catch((function(t){a.handleError(t)}))}},{key:"uploadParts",value:function(){var t=this,e=1-this.uploading.length;if(0!==e)if(this.chunkState.every((function(t){return t.done})))this.completeUpload();else{for(var n=[],r=0;r<this.chunkState.length;r++){var o=this.chunkState[r];if(!o.done&&!o.busy&&(n.push(r),n.length>=e))break}n.forEach((function(e){t.uploadPart(e)}))}}},{key:"uploadPart",value:function(t){var e=this;return this.chunkState[t].busy=!0,this.key&&this.uploadId?function(t){var e=t.csrfToken,n=t.endpoint,r=t.key,o=t.number,i=t.uploadId,a=encodeURIComponent(r),u=new Headers({"X-CSRFToken":e}),s=tc(n,i,"".concat(o),"?key=".concat(a));return fetch(s,{method:"get",headers:u}).then((function(t){return t.json()})).then((function(t){return t}))}({csrfToken:this.csrfToken,endpoint:this.endpoint,key:this.key,number:t+1,uploadId:this.uploadId}).then((function(t){if(!("object"===ca(t)&&t&&"string"==typeof t.url))throw new TypeError("AwsS3/Multipart: Got incorrect result from `prepareUploadPart()`, expected an object `{ url }`.");return t})).then((function(n){var r=n.url;e.uploadPartBytes(t,r)}),(function(t){e.handleError(t)})):Promise.resolve()}},{key:"onPartProgress",value:function(t,e){if(this.chunkState[t].uploaded=e,this.onProgress){var n=this.chunkState.reduce((function(t,e){return t+e.uploaded}),0);this.onProgress(n,this.file.size)}}},{key:"onPartComplete",value:function(t,e){this.chunkState[t].etag=e,this.chunkState[t].done=!0;var n={PartNumber:t+1,ETag:e};this.parts.push(n),this.uploadParts()}},{key:"uploadPartBytes",value:function(t,e){var n=this,r=this.chunks[t],o=new XMLHttpRequest;o.open("PUT",e,!0),o.responseType="text",this.uploading.push(o),o.upload.addEventListener("progress",(function(e){e.lengthComputable&&n.onPartProgress(t,e.loaded)})),o.addEventListener("abort",(function(e){nc(n.uploading,e.target),n.chunkState[t].busy=!1})),o.addEventListener("load",(function(e){var o=e.target;if(nc(n.uploading,o),n.chunkState[t].busy=!1,o.status<200||o.status>=300)n.handleError(new Error("Non 2xx"));else{n.onPartProgress(t,r.size);var i=o.getResponseHeader("ETag");null!==i?n.onPartComplete(t,i):n.handleError(new Error("AwsS3/Multipart: Could not read the ETag header. This likely means CORS is not configured correctly on the S3 Bucket. See https://uppy.io/docs/aws-s3-multipart#S3-Bucket-Configuration for instructions."))}})),o.addEventListener("error",(function(e){nc(n.uploading,e.target),n.chunkState[t].busy=!1;var r=new Error("Unknown error");n.handleError(r)})),o.send(r)}},{key:"completeUpload",value:function(){var t=this;return this.parts.sort((function(t,e){return t.PartNumber-e.PartNumber})),this.uploadId&&this.key?function(t){var e=t.csrfToken,n=t.key,r=t.uploadId,o=t.parts,i=t.endpoint,a=encodeURIComponent(n),u=encodeURIComponent(r),s=new Headers({"X-CSRFToken":e}),c=tc(i,u,"complete","?key=".concat(a));return fetch(c,{method:"post",headers:s,body:JSON.stringify({parts:o})}).then((function(t){return t.json()})).then((function(t){return t}))}({csrfToken:this.csrfToken,endpoint:this.endpoint,key:this.key,uploadId:this.uploadId,parts:this.parts}).then((function(){t.onSuccess&&t.onSuccess()}),(function(e){t.handleError(e)})):Promise.resolve()}},{key:"handleError",value:function(t){if(!this.onError)throw t;this.onError(t)}}]),o}(Zs),ic=function(){var t=zi(ga.mark((function t(e,n){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.abrupt("return",new Promise((function(t,r){var o=new XMLHttpRequest;o.open("DELETE",e),o.onload=function(){204===o.status?t():r()},o.setRequestHeader("Tus-Resumable","1.0.0"),o.setRequestHeader("X-CSRFToken",n),o.send(null)})));case 1:case"end":return t.stop()}}),t)})));return function(e,n){return t.apply(this,arguments)}}();function ac(t){var e=function(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Boolean.prototype.valueOf.call(Reflect.construct(Boolean,[],(function(){}))),!0}catch(t){return!1}}();return function(){var n,r=Xs(t);if(e){var o=Xs(this).constructor;n=Reflect.construct(r,arguments,o)}else n=r.apply(this,arguments);return Ws(this,n)}}var uc,sc=function(t){Ks(o,t);var e,n,r=ac(o);function o(t){var e,n=t.name,i=t.size,a=t.type,u=t.uploadIndex;return qi(this,o),Dr(Gs(e=r.call(this,{name:n,status:"done",type:a,uploadIndex:u})),"size",void 0),e.size=i,e}return $i(o,[{key:"abort",value:(n=zi(ga.mark((function t(){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.abrupt("return",Promise.resolve());case 1:case"end":return t.stop()}}),t)}))),function(){return n.apply(this,arguments)})},{key:"delete",value:(e=zi(ga.mark((function t(){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.abrupt("return",Promise.resolve());case 1:case"end":return t.stop()}}),t)}))),function(){return e.apply(this,arguments)})},{key:"getSize",value:function(){return this.size}}]),o}(Zs),cc=function(t){Ks(n,t);var e=ac(n);function n(t,r){var o;return qi(this,n),Dr(Gs(o=e.call(this,{name:t.name,size:t.size,type:"placeholder",uploadIndex:r})),"id",void 0),o.id=t.id,o}return $i(n,[{key:"getInitialFile",value:function(){return{id:this.id,name:this.name,size:this.size,type:"placeholder"}}}]),n}(sc),lc=function(t){Ks(n,t);var e=ac(n);function n(t,r){var o;return qi(this,n),Dr(Gs(o=e.call(this,{name:t.original_name||t.name,size:t.size,type:"uploadedS3",uploadIndex:r})),"id",void 0),Dr(Gs(o),"key",void 0),o.id=t.id,o.key=t.name,o}return $i(n,[{key:"getInitialFile",value:function(){return{id:this.id,name:this.key,original_name:this.name,size:this.size,type:"s3"}}}]),n}(sc),fc=function(t){Ks(n,t);var e=ac(n);function n(t,r){return qi(this,n),e.call(this,{name:t.name,size:t.size,type:"existing",uploadIndex:r})}return $i(n,[{key:"getInitialFile",value:function(){return{name:this.name,size:this.size,type:"existing"}}}]),n}(sc),pc=function(t){Ks(r,t);var e,n=ac(r);function r(t){var e,o=t.csrfToken,i=t.initialFile,a=t.uploadIndex,u=t.uploadUrl;return qi(this,r),Dr(Gs(e=n.call(this,{name:i.name,size:i.size,type:"uploadedTus",uploadIndex:a})),"csrfToken",void 0),Dr(Gs(e),"id",void 0),Dr(Gs(e),"url",void 0),e.csrfToken=o,e.id=i.id,e.url="".concat(u).concat(i.id),e}return $i(r,[{key:"delete",value:(e=zi(ga.mark((function t(){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,ic(this.url,this.csrfToken);case 2:case"end":return t.stop()}}),t,this)}))),function(){return e.apply(this,arguments)})},{key:"getInitialFile",value:function(){return{id:this.id,name:this.name,size:this.size,type:"tus",url:""}}}]),r}(sc),dc="undefined"!=typeof ArrayBuffer&&"undefined"!=typeof DataView,hc=O.f,vc=r.Int8Array,yc=vc&&vc.prototype,gc=r.Uint8ClampedArray,mc=gc&&gc.prototype,xc=vc&&_e(vc),bc=yc&&_e(yc),wc=Object.prototype,_c=wc.isPrototypeOf,Sc=Vt("toStringTag"),Ec=q("TYPED_ARRAY_TAG"),Ac=dc&&!!Ne&&"Opera"!==ie(r.opera),kc=!1,Rc={Int8Array:1,Uint8Array:1,Uint8ClampedArray:1,Int16Array:2,Uint16Array:2,Int32Array:4,Uint32Array:4,Float32Array:4,Float64Array:8},Oc={BigInt64Array:8,BigUint64Array:8},Tc=function(t){if(!y(t))return!1;var e=ie(t);return x(Rc,e)||x(Oc,e)};for(uc in Rc)r[uc]||(Ac=!1);if((!Ac||"function"!=typeof xc||xc===Function.prototype)&&(xc=function(){throw TypeError("Incorrect invocation")},Ac))for(uc in Rc)r[uc]&&Ne(r[uc],xc);if((!Ac||!bc||bc===wc)&&(bc=xc.prototype,Ac))for(uc in Rc)r[uc]&&Ne(r[uc].prototype,bc);if(Ac&&_e(mc)!==bc&&Ne(mc,bc),i&&!x(bc,Sc))for(uc in kc=!0,hc(bc,Sc,{get:function(){return y(this)?this[Ec]:void 0}}),Rc)r[uc]&&T(r[uc],Ec,uc);var jc={NATIVE_ARRAY_BUFFER_VIEWS:Ac,TYPED_ARRAY_TAG:kc&&Ec,aTypedArray:function(t){if(Tc(t))return t;throw TypeError("Target is not a typed array")},aTypedArrayConstructor:function(t){if(Ne){if(_c.call(xc,t))return t}else for(var e in Rc)if(x(Rc,uc)){var n=r[e];if(n&&(t===n||_c.call(n,t)))return t}throw TypeError("Target is not a typed array constructor")},exportTypedArrayMethod:function(t,e,n){if(i){if(n)for(var o in Rc){var a=r[o];a&&x(a.prototype,t)&&delete a.prototype[t]}bc[t]&&!n||nt(bc,t,n?e:Ac&&yc[t]||e)}},exportTypedArrayStaticMethod:function(t,e,n){var o,a;if(i){if(Ne){if(n)for(o in Rc)(a=r[o])&&x(a,t)&&delete a[t];if(xc[t]&&!n)return;try{return nt(xc,t,n?e:Ac&&vc[t]||e)}catch(t){}}for(o in Rc)!(a=r[o])||a[t]&&!n||nt(a,t,e)}},isView:function(t){if(!y(t))return!1;var e=ie(t);return"DataView"===e||x(Rc,e)||x(Oc,e)},isTypedArray:Tc,TypedArray:xc,TypedArrayPrototype:bc},Pc=jc.NATIVE_ARRAY_BUFFER_VIEWS,Cc=r.ArrayBuffer,Ic=r.Int8Array,Uc=!Pc||!o((function(){Ic(1)}))||!o((function(){new Ic(-1)}))||!pe((function(t){new Ic,new Ic(null),new Ic(1.5),new Ic(t)}),!0)||o((function(){return 1!==new Ic(new Cc(2),1,void 0).length})),Lc=function(t){if(void 0===t)return 0;var e=st(t),n=lt(e);if(e!==n)throw RangeError("Wrong length or index");return n},Dc=Math.abs,Hc=Math.pow,Fc=Math.floor,Mc=Math.log,Nc=Math.LN2,Bc=function(t,e,n){var r,o,i,a=new Array(n),u=8*n-e-1,s=(1<<u)-1,c=s>>1,l=23===e?Hc(2,-24)-Hc(2,-77):0,f=t<0||0===t&&1/t<0?1:0,p=0;for((t=Dc(t))!=t||t===1/0?(o=t!=t?1:0,r=s):(r=Fc(Mc(t)/Nc),t*(i=Hc(2,-r))<1&&(r--,i*=2),(t+=r+c>=1?l/i:l*Hc(2,1-c))*i>=2&&(r++,i/=2),r+c>=s?(o=0,r=s):r+c>=1?(o=(t*i-1)*Hc(2,e),r+=c):(o=t*Hc(2,c-1)*Hc(2,e),r=0));e>=8;a[p++]=255&o,o/=256,e-=8);for(r=r<<e|o,u+=e;u>0;a[p++]=255&r,r/=256,u-=8);return a[--p]|=128*f,a},zc=function(t,e){var n,r=t.length,o=8*r-e-1,i=(1<<o)-1,a=i>>1,u=o-7,s=r-1,c=t[s--],l=127&c;for(c>>=7;u>0;l=256*l+t[s],s--,u-=8);for(n=l&(1<<-u)-1,l>>=-u,u+=e;u>0;n=256*n+t[s],s--,u-=8);if(0===l)l=1-a;else{if(l===i)return n?NaN:c?-1/0:1/0;n+=Hc(2,e),l-=a}return(c?-1:1)*n*Hc(2,l-e)},qc=function(t){for(var e=Lt(this),n=lt(e.length),r=arguments.length,o=dt(r>1?arguments[1]:void 0,n),i=r>2?arguments[2]:void 0,a=void 0===i?n:dt(i,n);a>o;)e[o++]=t;return e},Gc=bt.f,$c=O.f,Kc=et.get,Wc=et.set,Xc="ArrayBuffer",Vc="DataView",Yc="Wrong index",Qc=r.ArrayBuffer,Jc=Qc,Zc=r.DataView,tl=Zc&&Zc.prototype,el=Object.prototype,nl=r.RangeError,rl=Bc,ol=zc,il=function(t){return[255&t]},al=function(t){return[255&t,t>>8&255]},ul=function(t){return[255&t,t>>8&255,t>>16&255,t>>24&255]},sl=function(t){return t[3]<<24|t[2]<<16|t[1]<<8|t[0]},cl=function(t){return rl(t,23,4)},ll=function(t){return rl(t,52,8)},fl=function(t,e){$c(t.prototype,e,{get:function(){return Kc(this)[e]}})},pl=function(t,e,n,r){var o=Lc(n),i=Kc(t);if(o+e>i.byteLength)throw nl(Yc);var a=Kc(i.buffer).bytes,u=o+i.byteOffset,s=a.slice(u,u+e);return r?s:s.reverse()},dl=function(t,e,n,r,o,i){var a=Lc(n),u=Kc(t);if(a+e>u.byteLength)throw nl(Yc);for(var s=Kc(u.buffer).bytes,c=a+u.byteOffset,l=r(+o),f=0;f<e;f++)s[c+f]=l[i?f:e-f-1]};if(dc){if(!o((function(){Qc(1)}))||!o((function(){new Qc(-1)}))||o((function(){return new Qc,new Qc(1.5),new Qc(NaN),Qc.name!=Xc}))){for(var hl,vl=(Jc=function(t){return bn(this,Jc),new Qc(Lc(t))}).prototype=Qc.prototype,yl=Gc(Qc),gl=0;yl.length>gl;)(hl=yl[gl++])in Jc||T(Jc,hl,Qc[hl]);vl.constructor=Jc}Ne&&_e(tl)!==el&&Ne(tl,el);var ml=new Zc(new Jc(2)),xl=tl.setInt8;ml.setInt8(0,2147483648),ml.setInt8(1,2147483649),!ml.getInt8(0)&&ml.getInt8(1)||Sn(tl,{setInt8:function(t,e){xl.call(this,t,e<<24>>24)},setUint8:function(t,e){xl.call(this,t,e<<24>>24)}},{unsafe:!0})}else Jc=function(t){bn(this,Jc,Xc);var e=Lc(t);Wc(this,{bytes:qc.call(new Array(e),0),byteLength:e}),i||(this.byteLength=e)},Zc=function(t,e,n){bn(this,Zc,Vc),bn(t,Jc,Vc);var r=Kc(t).byteLength,o=st(e);if(o<0||o>r)throw nl("Wrong offset");if(o+(n=void 0===n?r-o:lt(n))>r)throw nl("Wrong length");Wc(this,{buffer:t,byteLength:n,byteOffset:o}),i||(this.buffer=t,this.byteLength=n,this.byteOffset=o)},i&&(fl(Jc,"byteLength"),fl(Zc,"buffer"),fl(Zc,"byteLength"),fl(Zc,"byteOffset")),Sn(Zc.prototype,{getInt8:function(t){return pl(this,1,t)[0]<<24>>24},getUint8:function(t){return pl(this,1,t)[0]},getInt16:function(t){var e=pl(this,2,t,arguments.length>1?arguments[1]:void 0);return(e[1]<<8|e[0])<<16>>16},getUint16:function(t){var e=pl(this,2,t,arguments.length>1?arguments[1]:void 0);return e[1]<<8|e[0]},getInt32:function(t){return sl(pl(this,4,t,arguments.length>1?arguments[1]:void 0))},getUint32:function(t){return sl(pl(this,4,t,arguments.length>1?arguments[1]:void 0))>>>0},getFloat32:function(t){return ol(pl(this,4,t,arguments.length>1?arguments[1]:void 0),23)},getFloat64:function(t){return ol(pl(this,8,t,arguments.length>1?arguments[1]:void 0),52)},setInt8:function(t,e){dl(this,1,t,il,e)},setUint8:function(t,e){dl(this,1,t,il,e)},setInt16:function(t,e){dl(this,2,t,al,e,arguments.length>2?arguments[2]:void 0)},setUint16:function(t,e){dl(this,2,t,al,e,arguments.length>2?arguments[2]:void 0)},setInt32:function(t,e){dl(this,4,t,ul,e,arguments.length>2?arguments[2]:void 0)},setUint32:function(t,e){dl(this,4,t,ul,e,arguments.length>2?arguments[2]:void 0)},setFloat32:function(t,e){dl(this,4,t,cl,e,arguments.length>2?arguments[2]:void 0)},setFloat64:function(t,e){dl(this,8,t,ll,e,arguments.length>2?arguments[2]:void 0)}});He(Jc,Xc),He(Zc,Vc);var bl={ArrayBuffer:Jc,DataView:Zc},wl=function(t,e){var n=function(t){var e=st(t);if(e<0)throw RangeError("The argument can't be less than 0");return e}(t);if(n%e)throw RangeError("Wrong offset");return n},_l=jc.aTypedArrayConstructor,Sl=function(t){var e,n,r,o,i,a,u=Lt(t),s=arguments.length,c=s>1?arguments[1]:void 0,l=void 0!==c,f=ue(u);if(null!=f&&!Zt(f))for(a=(i=f.call(u)).next,u=[];!(o=a.call(i)).done;)u.push(o.value);for(l&&s>2&&(c=Ut(c,arguments[2],2)),n=lt(u.length),r=new(_l(this))(n),e=0;n>e;e++)r[e]=l?c(u[e],e):u[e];return r};e((function(t){var e=bt.f,n=Dn.forEach,o=et.get,a=et.set,u=O.f,s=A.f,l=Math.round,f=r.RangeError,p=bl.ArrayBuffer,d=bl.DataView,h=jc.NATIVE_ARRAY_BUFFER_VIEWS,v=jc.TYPED_ARRAY_TAG,m=jc.TypedArray,b=jc.TypedArrayPrototype,w=jc.aTypedArrayConstructor,_=jc.isTypedArray,S="BYTES_PER_ELEMENT",E="Wrong length",k=function(t,e){for(var n=0,r=e.length,o=new(w(t))(r);r>n;)o[n]=e[n++];return o},R=function(t,e){u(t,e,{get:function(){return o(this)[e]}})},j=function(t){var e;return t instanceof p||"ArrayBuffer"==(e=ie(t))||"SharedArrayBuffer"==e},P=function(t,e){return _(t)&&"symbol"!=typeof e&&e in t&&String(+e)==String(e)},C=function(t,e){return P(t,e=g(e,!0))?c(2,t[e]):s(t,e)},I=function(t,e,n){return!(P(t,e=g(e,!0))&&y(n)&&x(n,"value"))||x(n,"get")||x(n,"set")||n.configurable||x(n,"writable")&&!n.writable||x(n,"enumerable")&&!n.enumerable?u(t,e,n):(t[e]=n.value,t)};i?(h||(A.f=C,O.f=I,R(b,"buffer"),R(b,"byteOffset"),R(b,"byteLength"),R(b,"length")),Ct({target:"Object",stat:!0,forced:!h},{getOwnPropertyDescriptor:C,defineProperty:I}),t.exports=function(t,i,s){var c=t.match(/\d+$/)[0]/8,g=t+(s?"Clamped":"")+"Array",x="get"+t,w="set"+t,A=r[g],R=A,O=R&&R.prototype,P={},C=function(t,e){u(t,e,{get:function(){return function(t,e){var n=o(t);return n.view[x](e*c+n.byteOffset,!0)}(this,e)},set:function(t){return function(t,e,n){var r=o(t);s&&(n=(n=l(n))<0?0:n>255?255:255&n),r.view[w](e*c+r.byteOffset,n,!0)}(this,e,t)},enumerable:!0})};h?Uc&&(R=i((function(t,e,n,r){return bn(t,R,g),wn(y(e)?j(e)?void 0!==r?new A(e,wl(n,c),r):void 0!==n?new A(e,wl(n,c)):new A(e):_(e)?k(R,e):Sl.call(R,e):new A(Lc(e)),t,R)})),Ne&&Ne(R,m),n(e(A),(function(t){t in R||T(R,t,A[t])})),R.prototype=O):(R=i((function(t,e,n,r){bn(t,R,g);var o,i,u,s=0,l=0;if(y(e)){if(!j(e))return _(e)?k(R,e):Sl.call(R,e);o=e,l=wl(n,c);var h=e.byteLength;if(void 0===r){if(h%c)throw f(E);if((i=h-l)<0)throw f(E)}else if((i=lt(r)*c)+l>h)throw f(E);u=i/c}else u=Lc(e),o=new p(i=u*c);for(a(t,{buffer:o,byteOffset:l,byteLength:i,length:u,view:new d(o)});s<u;)C(t,s++)})),Ne&&Ne(R,m),O=R.prototype=Ue(b)),O.constructor!==R&&T(O,"constructor",R),v&&T(O,v,g),P[g]=R,Ct({global:!0,forced:R!=A,sham:!h},P),S in R||T(R,S,c),S in O||T(O,S,c),An(g)}):t.exports=function(){}}))("Uint8",(function(t){return function(e,n,r){return t(this,e,n,r)}}));var El=Math.min,Al=[].copyWithin||function(t,e){var n=Lt(this),r=lt(n.length),o=dt(t,r),i=dt(e,r),a=arguments.length>2?arguments[2]:void 0,u=El((void 0===a?r:dt(a,r))-i,r-o),s=1;for(i<o&&o<i+u&&(s=-1,i+=u-1,o+=u-1);u-- >0;)i in n?n[o]=n[i]:delete n[o],o+=s,i+=s;return n},kl=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("copyWithin",(function(t,e){return Al.call(kl(this),t,e,arguments.length>2?arguments[2]:void 0)}));var Rl=Dn.every,Ol=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("every",(function(t){return Rl(Ol(this),t,arguments.length>1?arguments[1]:void 0)}));var Tl=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("fill",(function(t){return qc.apply(Tl(this),arguments)}));var jl=jc.aTypedArrayConstructor,Pl=function(t,e){for(var n=ko(t,t.constructor),r=0,o=e.length,i=new(jl(n))(o);o>r;)i[r]=e[r++];return i},Cl=Dn.filter,Il=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("filter",(function(t){var e=Cl(Il(this),t,arguments.length>1?arguments[1]:void 0);return Pl(this,e)}));var Ul=Dn.find,Ll=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("find",(function(t){return Ul(Ll(this),t,arguments.length>1?arguments[1]:void 0)}));var Dl=Dn.findIndex,Hl=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("findIndex",(function(t){return Dl(Hl(this),t,arguments.length>1?arguments[1]:void 0)}));var Fl=Dn.forEach,Ml=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("forEach",(function(t){Fl(Ml(this),t,arguments.length>1?arguments[1]:void 0)}));var Nl=vt.includes,Bl=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("includes",(function(t){return Nl(Bl(this),t,arguments.length>1?arguments[1]:void 0)}));var zl=vt.indexOf,ql=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("indexOf",(function(t){return zl(ql(this),t,arguments.length>1?arguments[1]:void 0)}));var Gl=Vt("iterator"),$l=r.Uint8Array,Kl=un.values,Wl=un.keys,Xl=un.entries,Vl=jc.aTypedArray,Yl=jc.exportTypedArrayMethod,Ql=$l&&$l.prototype[Gl],Jl=!!Ql&&("values"==Ql.name||null==Ql.name),Zl=function(){return Kl.call(Vl(this))};Yl("entries",(function(){return Xl.call(Vl(this))})),Yl("keys",(function(){return Wl.call(Vl(this))})),Yl("values",Zl,!Jl),Yl(Gl,Zl,!Jl);var tf=jc.aTypedArray,ef=[].join;(0,jc.exportTypedArrayMethod)("join",(function(t){return ef.apply(tf(this),arguments)}));var nf=Math.min,rf=[].lastIndexOf,of=!!rf&&1/[1].lastIndexOf(1,-0)<0,af=Bn("lastIndexOf"),uf=of||!af?function(t){if(of)return rf.apply(this,arguments)||0;var e=v(this),n=lt(e.length),r=n-1;for(arguments.length>1&&(r=nf(r,st(arguments[1]))),r<0&&(r=n+r);r>=0;r--)if(r in e&&e[r]===t)return r||0;return-1}:rf,sf=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("lastIndexOf",(function(t){return uf.apply(sf(this),arguments)}));var cf=Dn.map,lf=jc.aTypedArray,ff=jc.aTypedArrayConstructor;(0,jc.exportTypedArrayMethod)("map",(function(t){return cf(lf(this),t,arguments.length>1?arguments[1]:void 0,(function(t,e){return new(ff(ko(t,t.constructor)))(e)}))}));var pf=Ys.left,df=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("reduce",(function(t){return pf(df(this),t,arguments.length,arguments.length>1?arguments[1]:void 0)}));var hf=Ys.right,vf=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("reduceRight",(function(t){return hf(vf(this),t,arguments.length,arguments.length>1?arguments[1]:void 0)}));var yf=jc.aTypedArray,gf=jc.exportTypedArrayMethod,mf=Math.floor;gf("reverse",(function(){for(var t,e=this,n=yf(e).length,r=mf(n/2),o=0;o<r;)t=e[o],e[o++]=e[--n],e[n]=t;return e}));var xf=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("set",(function(t){xf(this);var e=wl(arguments.length>1?arguments[1]:void 0,1),n=this.length,r=Lt(t),o=lt(r.length),i=0;if(o+e>n)throw RangeError("Wrong length");for(;i<o;)this[e+i]=r[i++]}),o((function(){new Int8Array(1).set({})})));var bf=jc.aTypedArray,wf=jc.aTypedArrayConstructor,_f=[].slice;(0,jc.exportTypedArrayMethod)("slice",(function(t,e){for(var n=_f.call(bf(this),t,e),r=ko(this,this.constructor),o=0,i=n.length,a=new(wf(r))(i);i>o;)a[o]=n[o++];return a}),o((function(){new Int8Array(1).slice()})));var Sf=Dn.some,Ef=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("some",(function(t){return Sf(Ef(this),t,arguments.length>1?arguments[1]:void 0)}));var Af=jc.aTypedArray,kf=[].sort;(0,jc.exportTypedArrayMethod)("sort",(function(t){return kf.call(Af(this),t)}));var Rf=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("subarray",(function(t,e){var n=Rf(this),r=n.length,o=dt(t,r);return new(ko(n,n.constructor))(n.buffer,n.byteOffset+o*n.BYTES_PER_ELEMENT,lt((void 0===e?r:dt(e,r))-o))}));var Of=r.Int8Array,Tf=jc.aTypedArray,jf=jc.exportTypedArrayMethod,Pf=[].toLocaleString,Cf=[].slice,If=!!Of&&o((function(){Pf.call(new Of(1))}));jf("toLocaleString",(function(){return Pf.apply(If?Cf.call(Tf(this)):Tf(this),arguments)}),o((function(){return[1,2].toLocaleString()!=new Of([1,2]).toLocaleString()}))||!o((function(){Of.prototype.toLocaleString.call([1,2])})));var Uf=jc.exportTypedArrayMethod,Lf=r.Uint8Array,Df=Lf&&Lf.prototype||{},Hf=[].toString,Ff=[].join;o((function(){Hf.call({})}))&&(Hf=function(){return Ff.call(this)});var Mf=Df.toString!=Hf;Uf("toString",Hf,Mf),(0,jc.exportTypedArrayStaticMethod)("from",Sl,Uc);var Nf=e((function(e,n){!function(t,n){e.exports=function(t){var n,r=(t=t||{}).Base64,o="2.6.4",i="ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/",a=function(t){for(var e={},n=0,r=t.length;n<r;n++)e[t.charAt(n)]=n;return e}(i),u=String.fromCharCode,s=function(t){if(t.length<2)return(e=t.charCodeAt(0))<128?t:e<2048?u(192|e>>>6)+u(128|63&e):u(224|e>>>12&15)+u(128|e>>>6&63)+u(128|63&e);var e=65536+1024*(t.charCodeAt(0)-55296)+(t.charCodeAt(1)-56320);return u(240|e>>>18&7)+u(128|e>>>12&63)+u(128|e>>>6&63)+u(128|63&e)},c=/[\uD800-\uDBFF][\uDC00-\uDFFFF]|[^\x00-\x7F]/g,l=function(t){return t.replace(c,s)},f=function(t){var e=[0,2,1][t.length%3],n=t.charCodeAt(0)<<16|(t.length>1?t.charCodeAt(1):0)<<8|(t.length>2?t.charCodeAt(2):0);return[i.charAt(n>>>18),i.charAt(n>>>12&63),e>=2?"=":i.charAt(n>>>6&63),e>=1?"=":i.charAt(63&n)].join("")},p=t.btoa&&"function"==typeof t.btoa?function(e){return t.btoa(e)}:function(t){if(t.match(/[^\x00-\xFF]/))throw new RangeError("The string contains invalid characters.");return t.replace(/[\s\S]{1,3}/g,f)},d=function(t){return p(l(String(t)))},h=function(t){return t.replace(/[+\/]/g,(function(t){return"+"==t?"-":"_"})).replace(/=/g,"")},v=function(t,e){return e?h(d(t)):d(t)},y=function(t){return v(t,!0)};t.Uint8Array&&(n=function(t,e){for(var n="",r=0,o=t.length;r<o;r+=3){var a=t[r],u=t[r+1],s=t[r+2],c=a<<16|u<<8|s;n+=i.charAt(c>>>18)+i.charAt(c>>>12&63)+(void 0!==u?i.charAt(c>>>6&63):"=")+(void 0!==s?i.charAt(63&c):"=")}return e?h(n):n});var g,m=/[\xC0-\xDF][\x80-\xBF]|[\xE0-\xEF][\x80-\xBF]{2}|[\xF0-\xF7][\x80-\xBF]{3}/g,x=function(t){switch(t.length){case 4:var e=((7&t.charCodeAt(0))<<18|(63&t.charCodeAt(1))<<12|(63&t.charCodeAt(2))<<6|63&t.charCodeAt(3))-65536;return u(55296+(e>>>10))+u(56320+(1023&e));case 3:return u((15&t.charCodeAt(0))<<12|(63&t.charCodeAt(1))<<6|63&t.charCodeAt(2));default:return u((31&t.charCodeAt(0))<<6|63&t.charCodeAt(1))}},b=function(t){return t.replace(m,x)},w=function(t){var e=t.length,n=e%4,r=(e>0?a[t.charAt(0)]<<18:0)|(e>1?a[t.charAt(1)]<<12:0)|(e>2?a[t.charAt(2)]<<6:0)|(e>3?a[t.charAt(3)]:0),o=[u(r>>>16),u(r>>>8&255),u(255&r)];return o.length-=[0,0,2,1][n],o.join("")},_=t.atob&&"function"==typeof t.atob?function(e){return t.atob(e)}:function(t){return t.replace(/\S{1,4}/g,w)},S=function(t){return _(String(t).replace(/[^A-Za-z0-9\+\/]/g,""))},E=function(t){return b(_(t))},A=function(t){return String(t).replace(/[-_]/g,(function(t){return"-"==t?"+":"/"})).replace(/[^A-Za-z0-9\+\/]/g,"")},k=function(t){return E(A(t))};t.Uint8Array&&(g=function(t){return Uint8Array.from(S(A(t)),(function(t){return t.charCodeAt(0)}))});var R=function(){var e=t.Base64;return t.Base64=r,e};if(t.Base64={VERSION:o,atob:S,btoa:p,fromBase64:k,toBase64:v,utob:l,encode:v,encodeURI:y,btou:b,decode:k,noConflict:R,fromUint8Array:n,toUint8Array:g},"function"==typeof Object.defineProperty){var O=function(t){return{value:t,enumerable:!1,writable:!0,configurable:!0}};t.Base64.extendString=function(){Object.defineProperty(String.prototype,"fromBase64",O((function(){return k(this)}))),Object.defineProperty(String.prototype,"toBase64",O((function(t){return v(this,t)}))),Object.defineProperty(String.prototype,"toBase64URI",O((function(){return v(this,!0)})))}}t.Meteor&&(Base64=t.Base64);e.exports&&(e.exports.Base64=t.Base64);return{Base64:t.Base64}}(t)}("undefined"!=typeof self?self:"undefined"!=typeof window?window:t)})),Bf=eu.start,zf=nu("trimStart"),qf=zf?function(){return Bf(this)}:"".trimStart;Ct({target:"String",proto:!0,forced:zf},{trimStart:qf,trimLeft:qf});var Gf=function(t,e){if(e=e.split(":")[0],!(t=+t))return!1;switch(e){case"http":case"ws":return 80!==t;case"https":case"wss":return 443!==t;case"ftp":return 21!==t;case"gopher":return 70!==t;case"file":return!1}return 0!==t},$f=Object.prototype.hasOwnProperty;function Kf(t){try{return decodeURIComponent(t.replace(/\+/g," "))}catch(t){return null}}function Wf(t){try{return encodeURIComponent(t)}catch(t){return null}}var Xf={stringify:function(t,e){e=e||"";var n,r,o=[];for(r in"string"!=typeof e&&(e="?"),t)if($f.call(t,r)){if((n=t[r])||null!=n&&!isNaN(n)||(n=""),r=Wf(r),n=Wf(n),null===r||null===n)continue;o.push(r+"="+n)}return o.length?e+o.join("&"):""},parse:function(t){for(var e,n=/([^=?#&]+)=?([^&]*)/g,r={};e=n.exec(t);){var o=Kf(e[1]),i=Kf(e[2]);null===o||null===i||o in r||(r[o]=i)}return r}},Vf=/^[A-Za-z][A-Za-z0-9+-.]*:[\\/]+/,Yf=/^([a-z][a-z0-9.+-]*:)?([\\/]{1,})?([\S\s]*)/i,Qf=new RegExp("^[\\x09\\x0A\\x0B\\x0C\\x0D\\x20\\xA0\\u1680\\u180E\\u2000\\u2001\\u2002\\u2003\\u2004\\u2005\\u2006\\u2007\\u2008\\u2009\\u200A\\u202F\\u205F\\u3000\\u2028\\u2029\\uFEFF]+");function Jf(t){return(t||"").toString().replace(Qf,"")}var Zf=[["#","hash"],["?","query"],function(t){return t.replace("\\","/")},["/","pathname"],["@","auth",1],[NaN,"host",void 0,1,1],[/:(\d+)$/,"port",void 0,1],[NaN,"hostname",void 0,1,1]],tp={hash:1,query:1};function ep(e){var n,r=("undefined"!=typeof window?window:void 0!==t?t:"undefined"!=typeof self?self:{}).location||{},o={},i=ca(e=e||r);if("blob:"===e.protocol)o=new rp(unescape(e.pathname),{});else if("string"===i)for(n in o=new rp(e,{}),tp)delete o[n];else if("object"===i){for(n in e)n in tp||(o[n]=e[n]);void 0===o.slashes&&(o.slashes=Vf.test(e.href))}return o}function np(t){t=Jf(t);var e=Yf.exec(t);return{protocol:e[1]?e[1].toLowerCase():"",slashes:!!(e[2]&&e[2].length>=2),rest:e[2]&&1===e[2].length?"/"+e[3]:e[3]}}function rp(t,e,n){if(t=Jf(t),!(this instanceof rp))return new rp(t,e,n);var r,o,i,a,u,s,c=Zf.slice(),l=ca(e),f=this,p=0;for("object"!==l&&"string"!==l&&(n=e,e=null),n&&"function"!=typeof n&&(n=Xf.parse),e=ep(e),r=!(o=np(t||"")).protocol&&!o.slashes,f.slashes=o.slashes||r&&e.slashes,f.protocol=o.protocol||e.protocol||"",t=o.rest,o.slashes||(c[3]=[/(.*)/,"pathname"]);p<c.length;p++)"function"!=typeof(a=c[p])?(i=a[0],s=a[1],i!=i?f[s]=t:"string"==typeof i?~(u=t.indexOf(i))&&("number"==typeof a[2]?(f[s]=t.slice(0,u),t=t.slice(u+a[2])):(f[s]=t.slice(u),t=t.slice(0,u))):(u=i.exec(t))&&(f[s]=u[1],t=t.slice(0,u.index)),f[s]=f[s]||r&&a[3]&&e[s]||"",a[4]&&(f[s]=f[s].toLowerCase())):t=a(t);n&&(f.query=n(f.query)),r&&e.slashes&&"/"!==f.pathname.charAt(0)&&(""!==f.pathname||""!==e.pathname)&&(f.pathname=function(t,e){if(""===t)return e;for(var n=(e||"/").split("/").slice(0,-1).concat(t.split("/")),r=n.length,o=n[r-1],i=!1,a=0;r--;)"."===n[r]?n.splice(r,1):".."===n[r]?(n.splice(r,1),a++):a&&(0===r&&(i=!0),n.splice(r,1),a--);return i&&n.unshift(""),"."!==o&&".."!==o||n.push(""),n.join("/")}(f.pathname,e.pathname)),"/"!==f.pathname.charAt(0)&&f.hostname&&(f.pathname="/"+f.pathname),Gf(f.port,f.protocol)||(f.host=f.hostname,f.port=""),f.username=f.password="",f.auth&&(a=f.auth.split(":"),f.username=a[0]||"",f.password=a[1]||""),f.origin=f.protocol&&f.host&&"file:"!==f.protocol?f.protocol+"//"+f.host:"null",f.href=f.toString()}rp.prototype={set:function(t,e,n){var r=this;switch(t){case"query":"string"==typeof e&&e.length&&(e=(n||Xf.parse)(e)),r[t]=e;break;case"port":r[t]=e,Gf(e,r.protocol)?e&&(r.host=r.hostname+":"+e):(r.host=r.hostname,r[t]="");break;case"hostname":r[t]=e,r.port&&(e+=":"+r.port),r.host=e;break;case"host":r[t]=e,/:\d+$/.test(e)?(e=e.split(":"),r.port=e.pop(),r.hostname=e.join(":")):(r.hostname=e,r.port="");break;case"protocol":r.protocol=e.toLowerCase(),r.slashes=!n;break;case"pathname":case"hash":if(e){var o="pathname"===t?"/":"#";r[t]=e.charAt(0)!==o?o+e:e}else r[t]=e;break;default:r[t]=e}for(var i=0;i<Zf.length;i++){var a=Zf[i];a[4]&&(r[a[1]]=r[a[1]].toLowerCase())}return r.origin=r.protocol&&r.host&&"file:"!==r.protocol?r.protocol+"//"+r.host:"null",r.href=r.toString(),r},toString:function(t){t&&"function"==typeof t||(t=Xf.stringify);var e,n=this,r=n.protocol;r&&":"!==r.charAt(r.length-1)&&(r+=":");var o=r+(n.slashes?"//":"");return n.username&&(o+=n.username,n.password&&(o+=":"+n.password),o+="@"),o+=n.host+n.pathname,(e="object"===ca(n.query)?t(n.query):n.query)&&(o+="?"!==e.charAt(0)?"?"+e:e),n.hash&&(o+=n.hash),o}},rp.extractProtocol=np,rp.location=ep,rp.trimLeft=Jf,rp.qs=Xf;var op=rp;function ip(t){return(ip="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}function ap(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}function up(t,e){return!e||"object"!==ip(e)&&"function"!=typeof e?function(t){if(void 0===t)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return t}(t):e}function sp(t){var e="function"==typeof Map?new Map:void 0;return(sp=function(t){if(null===t||(n=t,-1===Function.toString.call(n).indexOf("[native code]")))return t;var n;if("function"!=typeof t)throw new TypeError("Super expression must either be null or a function");if(void 0!==e){if(e.has(t))return e.get(t);e.set(t,r)}function r(){return cp(t,arguments,pp(this).constructor)}return r.prototype=Object.create(t.prototype,{constructor:{value:r,enumerable:!1,writable:!0,configurable:!0}}),fp(r,t)})(t)}function cp(t,e,n){return(cp=lp()?Reflect.construct:function(t,e,n){var r=[null];r.push.apply(r,e);var o=new(Function.bind.apply(t,r));return n&&fp(o,n.prototype),o}).apply(null,arguments)}function lp(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],(function(){}))),!0}catch(t){return!1}}function fp(t,e){return(fp=Object.setPrototypeOf||function(t,e){return t.__proto__=e,t})(t,e)}function pp(t){return(pp=Object.setPrototypeOf?Object.getPrototypeOf:function(t){return t.__proto__||Object.getPrototypeOf(t)})(t)}_n("Map",(function(t){return function(){return t(this,arguments.length?arguments[0]:void 0)}}),jn);var dp=function(t){!function(t,e){if("function"!=typeof e&&null!==e)throw new TypeError("Super expression must either be null or a function");t.prototype=Object.create(e&&e.prototype,{constructor:{value:t,writable:!0,configurable:!0}}),e&&fp(t,e)}(r,t);var e,n=(e=r,function(){var t,n=pp(e);if(lp()){var r=pp(this).constructor;t=Reflect.construct(n,arguments,r)}else t=n.apply(this,arguments);return up(this,t)});function r(t){var e,o=arguments.length>1&&void 0!==arguments[1]?arguments[1]:null,i=arguments.length>2&&void 0!==arguments[2]?arguments[2]:null,a=arguments.length>3&&void 0!==arguments[3]?arguments[3]:null;if(ap(this,r),(e=n.call(this,t)).originalRequest=i,e.originalResponse=a,e.causingError=o,null!=o&&(t+=", caused by ".concat(o.toString())),null!=i){var u=i.getHeader("X-Request-ID")||"n/a",s=i.getMethod(),c=i.getURL(),l=a?a.getStatus():"n/a",f=a?a.getBody()||"":"n/a";t+=", originated from request (method: ".concat(s,", url: ").concat(c,", response code: ").concat(l,", response text: ").concat(f,", request id: ").concat(u,")")}return e.message=t,e}return r}(sp(Error));function hp(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function vp(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?hp(Object(n),!0).forEach((function(e){yp(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):hp(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function yp(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}function gp(t,e){for(var n=0;n<e.length;n++){var r=e[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(t,r.key,r)}}var mp={endpoint:null,uploadUrl:null,metadata:{},fingerprint:null,uploadSize:null,onProgress:null,onChunkComplete:null,onSuccess:null,onError:null,_onUploadUrlAvailable:null,overridePatchMethod:!1,headers:{},addRequestId:!1,onBeforeRequest:null,onAfterResponse:null,onShouldRetry:null,chunkSize:1/0,retryDelays:[0,1e3,3e3,5e3],parallelUploads:1,storeFingerprintForResuming:!0,removeFingerprintOnSuccess:!1,uploadLengthDeferred:!1,uploadDataDuringCreation:!1,urlStorage:null,fileReader:null,httpStack:null},xp=function(){function t(e,n){!function(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}(this,t),"resume"in n&&console.log("tus: The `resume` option has been removed in tus-js-client v2. Please use the URL storage API instead."),this.options=n,this._urlStorage=this.options.urlStorage,this.file=e,this.url=null,this._req=null,this._fingerprint=null,this._urlStorageKey=null,this._offset=null,this._aborted=!1,this._size=null,this._source=null,this._retryAttempt=0,this._retryTimeout=null,this._offsetBeforeRetry=0,this._parallelUploads=null,this._parallelUploadUrls=null}var e,n,r;return e=t,r=[{key:"terminate",value:function(e){var n=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},r=arguments.length>2?arguments[2]:void 0;if("function"==typeof n||"function"==typeof r)throw new Error("tus: the terminate function does not accept a callback since v2 anymore; please use the returned Promise instead");var o=_p("DELETE",e,n);return Sp(o,null,n).then((function(t){if(204!==t.getStatus())throw new dp("tus: unexpected response while terminating upload",null,o,t)})).catch((function(r){if(r instanceof dp||(r=new dp("tus: failed to terminate upload",r,o,null)),!Ep(r,0,n))throw r;var i=n.retryDelays[0],a=n.retryDelays.slice(1),u=vp({},n,{retryDelays:a});return new Promise((function(t){return setTimeout(t,i)})).then((function(){return t.terminate(e,u)}))}))}}],(n=[{key:"findPreviousUploads",value:function(){var t=this;return this.options.fingerprint(this.file,this.options).then((function(e){return t._urlStorage.findUploadsByFingerprint(e)}))}},{key:"resumeFromPreviousUpload",value:function(t){this.url=t.uploadUrl||null,this._parallelUploadUrls=t.parallelUploadUrls||null,this._urlStorageKey=t.urlStorageKey}},{key:"start",value:function(){var t=this,e=this.file;if(e)if(this.options.endpoint||this.options.uploadUrl){var n=this.options.retryDelays;null==n||"[object Array]"===Object.prototype.toString.call(n)?(this.options.parallelUploads>1&&["uploadUrl","uploadSize","uploadLengthDeferred"].forEach((function(e){t.options[e]&&t._emitError(new Error("tus: cannot use the ".concat(e," option when parallelUploads is enabled")))})),this.options.fingerprint(e,this.options).then((function(n){return t._fingerprint=n,t._source?t._source:t.options.fileReader.openFile(e,t.options.chunkSize)})).then((function(e){t._source=e,t.options.parallelUploads>1||null!=t._parallelUploadUrls?t._startParallelUpload():t._startSingleUpload()})).catch((function(e){t._emitError(e)}))):this._emitError(new Error("tus: the `retryDelays` option must either be an array or null"))}else this._emitError(new Error("tus: neither an endpoint or an upload URL is provided"));else this._emitError(new Error("tus: no file or stream to upload provided"))}},{key:"_startParallelUpload",value:function(){var e=this,n=this._size=this._source.size,r=0;this._parallelUploads=[];var o=null!=this._parallelUploadUrls?this._parallelUploadUrls.length:this.options.parallelUploads,i=function(t,e,n){for(var r=Math.floor(t/e),o=[],i=0;i<e;i++)o.push({start:r*i,end:r*(i+1)});return o[e-1].end=t,n&&o.forEach((function(t,e){t.uploadUrl=n[e]||null})),o}(this._source.size,o,this._parallelUploadUrls);this._parallelUploadUrls=new Array(i.length);var a,u=i.map((function(o,a){var u=0;return e._source.slice(o.start,o.end).then((function(s){var c=s.value;return new Promise((function(s,l){var f=vp({},e.options,{uploadUrl:o.uploadUrl||null,storeFingerprintForResuming:!1,removeFingerprintOnSuccess:!1,parallelUploads:1,metadata:{},headers:vp({},e.options.headers,{"Upload-Concat":"partial"}),onSuccess:s,onError:l,onProgress:function(t){r=r-u+t,u=t,e._emitProgress(r,n)},_onUploadUrlAvailable:function(){e._parallelUploadUrls[a]=p.url,e._parallelUploadUrls.filter((function(t){return!!t})).length===i.length&&e._saveUploadInUrlStorage()}}),p=new t(c,f);p.start(),e._parallelUploads.push(p)}))}))}));Promise.all(u).then((function(){(a=e._openRequest("POST",e.options.endpoint)).setHeader("Upload-Concat","final;".concat(e._parallelUploadUrls.join(" ")));var t=bp(e.options.metadata);return""!==t&&a.setHeader("Upload-Metadata",t),e._sendRequest(a,null)})).then((function(t){if(wp(t.getStatus(),200)){var n=t.getHeader("Location");null!=n?(e.url=Ap(e.options.endpoint,n),"Created upload at ".concat(e.url),e._emitSuccess()):e._emitHttpError(a,t,"tus: invalid or missing Location header")}else e._emitHttpError(a,t,"tus: unexpected response while creating upload")})).catch((function(t){e._emitError(t)}))}},{key:"_startSingleUpload",value:function(){if(this.options.uploadLengthDeferred)this._size=null;else if(null!=this.options.uploadSize){if(this._size=+this.options.uploadSize,isNaN(this._size))return void this._emitError(new Error("tus: cannot convert `uploadSize` option into a number"))}else if(this._size=this._source.size,null==this._size)return void this._emitError(new Error("tus: cannot automatically derive upload's size from input and must be specified manually using the `uploadSize` option"));return this._aborted=!1,null!=this.url?("Resuming upload from previous URL: ".concat(this.url),void this._resumeUpload()):null!=this.options.uploadUrl?("Resuming upload from provided URL: ".concat(this.options.url),this.url=this.options.uploadUrl,void this._resumeUpload()):void this._createUpload()}},{key:"abort",value:function(e,n){var r=this;if("function"==typeof n)throw new Error("tus: the abort function does not accept a callback since v2 anymore; please use the returned Promise instead");return null!=this._parallelUploads&&this._parallelUploads.forEach((function(t){t.abort(e)})),null!==this._req&&(this._req.abort(),this._source.close()),this._aborted=!0,null!=this._retryTimeout&&(clearTimeout(this._retryTimeout),this._retryTimeout=null),e&&null!=this.url?t.terminate(this.url,this.options).then((function(){return r._removeFromUrlStorage()})):Promise.resolve()}},{key:"_emitHttpError",value:function(t,e,n,r){this._emitError(new dp(n,r,t,e))}},{key:"_emitError",value:function(t){var e=this;if(!this._aborted){if(null!=this.options.retryDelays&&(null!=this._offset&&this._offset>this._offsetBeforeRetry&&(this._retryAttempt=0),Ep(t,this._retryAttempt,this.options))){var n=this.options.retryDelays[this._retryAttempt++];return this._offsetBeforeRetry=this._offset,void(this._retryTimeout=setTimeout((function(){e.start()}),n))}if("function"!=typeof this.options.onError)throw t;this.options.onError(t)}}},{key:"_emitSuccess",value:function(){this.options.removeFingerprintOnSuccess&&this._removeFromUrlStorage(),"function"==typeof this.options.onSuccess&&this.options.onSuccess()}},{key:"_emitProgress",value:function(t,e){"function"==typeof this.options.onProgress&&this.options.onProgress(t,e)}},{key:"_emitChunkComplete",value:function(t,e,n){"function"==typeof this.options.onChunkComplete&&this.options.onChunkComplete(t,e,n)}},{key:"_createUpload",value:function(){var t=this;if(this.options.endpoint){var e=this._openRequest("POST",this.options.endpoint);this.options.uploadLengthDeferred?e.setHeader("Upload-Defer-Length",1):e.setHeader("Upload-Length",this._size);var n,r=bp(this.options.metadata);""!==r&&e.setHeader("Upload-Metadata",r),this.options.uploadDataDuringCreation&&!this.options.uploadLengthDeferred?(this._offset=0,n=this._addChunkToRequest(e)):n=this._sendRequest(e,null),n.then((function(n){if(wp(n.getStatus(),200)){var r=n.getHeader("Location");if(null!=r){if(t.url=Ap(t.options.endpoint,r),"Created upload at ".concat(t.url),"function"==typeof t.options._onUploadUrlAvailable&&t.options._onUploadUrlAvailable(),0===t._size)return t._emitSuccess(),void t._source.close();t._saveUploadInUrlStorage(),t.options.uploadDataDuringCreation?t._handleUploadResponse(e,n):(t._offset=0,t._performUpload())}else t._emitHttpError(e,n,"tus: invalid or missing Location header")}else t._emitHttpError(e,n,"tus: unexpected response while creating upload")})).catch((function(n){t._emitHttpError(e,null,"tus: failed to create upload",n)}))}else this._emitError(new Error("tus: unable to create upload because no endpoint is provided"))}},{key:"_resumeUpload",value:function(){var t=this,e=this._openRequest("HEAD",this.url);this._sendRequest(e,null).then((function(n){var r=n.getStatus();if(!wp(r,200))return wp(r,400)&&t._removeFromUrlStorage(),423===r?void t._emitHttpError(e,n,"tus: upload is currently locked; retry later"):t.options.endpoint?(t.url=null,void t._createUpload()):void t._emitHttpError(e,n,"tus: unable to resume upload (new upload cannot be created without an endpoint)");var o=parseInt(n.getHeader("Upload-Offset"),10);if(isNaN(o))t._emitHttpError(e,n,"tus: invalid or missing offset value");else{var i=parseInt(n.getHeader("Upload-Length"),10);if(!isNaN(i)||t.options.uploadLengthDeferred){if("function"==typeof t.options._onUploadUrlAvailable&&t.options._onUploadUrlAvailable(),o===i)return t._emitProgress(i,i),void t._emitSuccess();t._offset=o,t._performUpload()}else t._emitHttpError(e,n,"tus: invalid or missing length value")}})).catch((function(n){t._emitHttpError(e,null,"tus: failed to resume upload",n)}))}},{key:"_performUpload",value:function(){var t,e=this;this._aborted||(this.options.overridePatchMethod?(t=this._openRequest("POST",this.url)).setHeader("X-HTTP-Method-Override","PATCH"):t=this._openRequest("PATCH",this.url),t.setHeader("Upload-Offset",this._offset),this._addChunkToRequest(t).then((function(n){wp(n.getStatus(),200)?e._handleUploadResponse(t,n):e._emitHttpError(t,n,"tus: unexpected response while uploading chunk")})).catch((function(n){e._aborted||e._emitHttpError(t,null,"tus: failed to upload chunk at offset "+e._offset,n)})))}},{key:"_addChunkToRequest",value:function(t){var e=this,n=this._offset,r=this._offset+this.options.chunkSize;return t.setProgressHandler((function(t){e._emitProgress(n+t,e._size)})),t.setHeader("Content-Type","application/offset+octet-stream"),(r===1/0||r>this._size)&&!this.options.uploadLengthDeferred&&(r=this._size),this._source.slice(n,r).then((function(n){var r=n.value,o=n.done;return e.options.uploadLengthDeferred&&o&&(e._size=e._offset+(r&&r.size?r.size:0),t.setHeader("Upload-Length",e._size)),null===r?e._sendRequest(t):(e._emitProgress(e._offset,e._size),e._sendRequest(t,r))}))}},{key:"_handleUploadResponse",value:function(t,e){var n=parseInt(e.getHeader("Upload-Offset"),10);if(isNaN(n))this._emitHttpError(t,e,"tus: invalid or missing offset value");else{if(this._emitProgress(n,this._size),this._emitChunkComplete(n-this._offset,n,this._size),this._offset=n,n==this._size)return this._emitSuccess(),void this._source.close();this._performUpload()}}},{key:"_openRequest",value:function(t,e){var n=_p(t,e,this.options);return this._req=n,n}},{key:"_removeFromUrlStorage",value:function(){var t=this;this._urlStorageKey&&(this._urlStorage.removeUpload(this._urlStorageKey).catch((function(e){t._emitError(e)})),this._urlStorageKey=null)}},{key:"_saveUploadInUrlStorage",value:function(){var t=this;if(this.options.storeFingerprintForResuming&&this._fingerprint){var e={size:this._size,metadata:this.options.metadata,creationTime:(new Date).toString()};this._parallelUploads?e.parallelUploadUrls=this._parallelUploadUrls:e.uploadUrl=this.url,this._urlStorage.addUpload(this._fingerprint,e).then((function(e){return t._urlStorageKey=e})).catch((function(e){t._emitError(e)}))}}},{key:"_sendRequest",value:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:null;return Sp(t,e,this.options)}}])&&gp(e.prototype,n),r&&gp(e,r),t}();function bp(t){var e=[];for(var n in t)e.push(n+" "+Nf.Base64.encode(t[n]));return e.join(",")}function wp(t,e){return t>=e&&t<e+100}function _p(t,e,n){var r=n.httpStack.createRequest(t,e);r.setHeader("Tus-Resumable","1.0.0");var o=n.headers||{};for(var i in o)r.setHeader(i,o[i]);if(n.addRequestId){var a="xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx".replace(/[xy]/g,(function(t){var e=16*Math.random()|0;return("x"==t?e:3&e|8).toString(16)}));r.setHeader("X-Request-ID",a)}return r}function Sp(t,e,n){return("function"==typeof n.onBeforeRequest?Promise.resolve(n.onBeforeRequest(t)):Promise.resolve()).then((function(){return t.send(e).then((function(e){return("function"==typeof n.onAfterResponse?Promise.resolve(n.onAfterResponse(t,e)):Promise.resolve()).then((function(){return e}))}))}))}function Ep(t,e,n){if(null==n.retryDelays||e>=n.retryDelays.length||null==t.originalRequest)return!1;if(n&&"function"==typeof n.onShouldRetry)return n.onShouldRetry(t,e,n);var r,o=t.originalResponse?t.originalResponse.getStatus():0;return(!wp(o,400)||409===o||423===o)&&(r=!0,"undefined"!=typeof window&&"navigator"in window&&!1===window.navigator.onLine&&(r=!1),r)}function Ap(t,e){return new op(e,t).toString()}function kp(t,e){for(var n=0;n<e.length;n++){var r=e[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(t,r.key,r)}}xp.defaultOptions=mp;var Rp=function(){function t(){!function(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}(this,t)}var e,n,r;return e=t,(n=[{key:"listAllUploads",value:function(){return Promise.resolve([])}},{key:"findUploadsByFingerprint",value:function(t){return Promise.resolve([])}},{key:"removeUpload",value:function(t){return Promise.resolve()}},{key:"addUpload",value:function(t,e){return Promise.resolve(null)}}])&&kp(e.prototype,n),r&&kp(e,r),t}();function Op(t,e){for(var n=0;n<e.length;n++){var r=e[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(t,r.key,r)}}var Tp=!1;try{Tp="localStorage"in window;var jp="tusSupport";localStorage.setItem(jp,localStorage.getItem(jp))}catch(t){if(t.code!==t.SECURITY_ERR&&t.code!==t.QUOTA_EXCEEDED_ERR)throw t;Tp=!1}var Pp=Tp,Cp=function(){function t(){!function(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}(this,t)}var e,n,r;return e=t,(n=[{key:"findAllUploads",value:function(){var t=this._findEntries("tus::");return Promise.resolve(t)}},{key:"findUploadsByFingerprint",value:function(t){var e=this._findEntries("tus::".concat(t,"::"));return Promise.resolve(e)}},{key:"removeUpload",value:function(t){return localStorage.removeItem(t),Promise.resolve()}},{key:"addUpload",value:function(t,e){var n=Math.round(1e12*Math.random()),r="tus::".concat(t,"::").concat(n);return localStorage.setItem(r,JSON.stringify(e)),Promise.resolve(r)}},{key:"_findEntries",value:function(t){for(var e=[],n=0;n<localStorage.length;n++){var r=localStorage.key(n);if(0===r.indexOf(t))try{var o=JSON.parse(localStorage.getItem(r));o.urlStorageKey=r,e.push(o)}catch(t){}}return e}}])&&Op(e.prototype,n),r&&Op(e,r),t}();function Ip(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}function Up(t,e){for(var n=0;n<e.length;n++){var r=e[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(t,r.key,r)}}function Lp(t,e,n){return e&&Up(t.prototype,e),n&&Up(t,n),t}var Dp=function(){function t(){Ip(this,t)}return Lp(t,[{key:"createRequest",value:function(t,e){return new Hp(t,e)}},{key:"getName",value:function(){return"XHRHttpStack"}}]),t}(),Hp=function(){function t(e,n){Ip(this,t),this._xhr=new XMLHttpRequest,this._xhr.open(e,n,!0),this._method=e,this._url=n,this._headers={}}return Lp(t,[{key:"getMethod",value:function(){return this._method}},{key:"getURL",value:function(){return this._url}},{key:"setHeader",value:function(t,e){this._xhr.setRequestHeader(t,e),this._headers[t]=e}},{key:"getHeader",value:function(t){return this._headers[t]}},{key:"setProgressHandler",value:function(t){"upload"in this._xhr&&(this._xhr.upload.onprogress=function(e){e.lengthComputable&&t(e.loaded)})}},{key:"send",value:function(){var t=this,e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:null;return new Promise((function(n,r){t._xhr.onload=function(){n(new Fp(t._xhr))},t._xhr.onerror=function(t){r(t)},t._xhr.send(e)}))}},{key:"abort",value:function(){return this._xhr.abort(),Promise.resolve()}},{key:"getUnderlyingObject",value:function(){return this._xhr}}]),t}(),Fp=function(){function t(e){Ip(this,t),this._xhr=e}return Lp(t,[{key:"getStatus",value:function(){return this._xhr.status}},{key:"getHeader",value:function(t){return this._xhr.getResponseHeader(t)}},{key:"getBody",value:function(){return this._xhr.responseText}},{key:"getUnderlyingObject",value:function(){return this._xhr}}]),t}(),Mp=function(){return"undefined"!=typeof navigator&&"string"==typeof navigator.product&&"reactnative"===navigator.product.toLowerCase()};function Np(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}function Bp(t,e){for(var n=0;n<e.length;n++){var r=e[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(t,r.key,r)}}function zp(t,e,n){return e&&Bp(t.prototype,e),n&&Bp(t,n),t}var qp=function(){function t(e){Np(this,t),this._file=e,this.size=e.size}return zp(t,[{key:"slice",value:function(t,e){if("undefined"!=typeof window&&(void 0!==window.PhoneGap||void 0!==window.Cordova||void 0!==window.cordova))return n=this._file.slice(t,e),new Promise((function(t,e){var r=new FileReader;r.onload=function(){var e=new Uint8Array(r.result);t({value:e})},r.onerror=function(t){e(t)},r.readAsArrayBuffer(n)}));var n,r=this._file.slice(t,e);return Promise.resolve({value:r})}},{key:"close",value:function(){}}]),t}(),Gp=function(){function t(e,n){Np(this,t),this._chunkSize=n,this._buffer=void 0,this._bufferOffset=0,this._reader=e,this._done=!1}return zp(t,[{key:"slice",value:function(t,e){return t<this._bufferOffset?Promise.reject(new Error("Requested data is before the reader's current offset")):this._readUntilEnoughDataOrDone(t,e)}},{key:"_readUntilEnoughDataOrDone",value:function(t,e){var n=this,r=e<=this._bufferOffset+$p(this._buffer);if(this._done||r){var o=this._getDataFromBuffer(t,e),i=null==o&&this._done;return Promise.resolve({value:o,done:i})}return this._reader.read().then((function(r){var o=r.value;return r.done?n._done=!0:void 0===n._buffer?n._buffer=o:n._buffer=function(t,e){if(t.concat)return t.concat(e);if(t instanceof Blob)return new Blob([t,e],{type:t.type});if(t.set){var n=new t.constructor(t.length+e.length);return n.set(t),n.set(e,t.length),n}throw new Error("Unknown data type")}(n._buffer,o),n._readUntilEnoughDataOrDone(t,e)}))}},{key:"_getDataFromBuffer",value:function(t,e){t>this._bufferOffset&&(this._buffer=this._buffer.slice(t-this._bufferOffset),this._bufferOffset=t);var n=0===$p(this._buffer);return this._done&&n?null:this._buffer.slice(0,e-t)}},{key:"close",value:function(){this._reader.cancel&&this._reader.cancel()}}]),t}();function $p(t){return void 0===t?0:void 0!==t.size?t.size:t.length}var Kp=function(){function t(){Np(this,t)}return zp(t,[{key:"openFile",value:function(t,e){return Mp()&&t&&void 0!==t.uri?(n=t.uri,new Promise((function(t,e){var r=new XMLHttpRequest;r.responseType="blob",r.onload=function(){var e=r.response;t(e)},r.onerror=function(t){e(t)},r.open("GET",n),r.send()}))).then((function(t){return new qp(t)})).catch((function(t){throw new Error("tus: cannot fetch `file.uri` as Blob, make sure the uri is correct and accessible. "+t)})):"function"==typeof t.slice&&void 0!==t.size?Promise.resolve(new qp(t)):"function"==typeof t.read?(e=+e,isFinite(e)?Promise.resolve(new Gp(t,e)):Promise.reject(new Error("cannot create source for stream without a finite value for the `chunkSize` option"))):Promise.reject(new Error("source object may only be an instance of File, Blob, or Reader in this environment"));var n}}]),t}();function Wp(t){return(Wp="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}function Xp(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}function Vp(t,e){for(var n=0;n<e.length;n++){var r=e[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(t,r.key,r)}}function Yp(t,e){return(Yp=Object.setPrototypeOf||function(t,e){return t.__proto__=e,t})(t,e)}function Qp(t,e){return!e||"object"!==Wp(e)&&"function"!=typeof e?function(t){if(void 0===t)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return t}(t):e}function Jp(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],(function(){}))),!0}catch(t){return!1}}function Zp(t){return(Zp=Object.setPrototypeOf?Object.getPrototypeOf:function(t){return t.__proto__||Object.getPrototypeOf(t)})(t)}function td(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function ed(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?td(Object(n),!0).forEach((function(e){nd(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):td(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function nd(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var rd=ed({},xp.defaultOptions,{httpStack:new Dp,fileReader:new Kp,urlStorage:Pp?new Cp:new Rp,fingerprint:function(t,e){return Mp()?Promise.resolve(function(t,e){var n=t.exif?function(t){var e=0;if(0===t.length)return e;for(var n=0;n<t.length;n++){e=(e<<5)-e+t.charCodeAt(n),e&=e}return e}(JSON.stringify(t.exif)):"noexif";return["tus-rn",t.name||"noname",t.size||"nosize",n,e.endpoint].join("/")}(t,e)):Promise.resolve(["tus-br",t.name,t.type,t.size,t.lastModified,e.endpoint].join("-"))}}),od=function(t){!function(t,e){if("function"!=typeof e&&null!==e)throw new TypeError("Super expression must either be null or a function");t.prototype=Object.create(e&&e.prototype,{constructor:{value:t,writable:!0,configurable:!0}}),e&&Yp(t,e)}(a,t);var e,n,r,o,i=(e=a,function(){var t,n=Zp(e);if(Jp()){var r=Zp(this).constructor;t=Reflect.construct(n,arguments,r)}else t=n.apply(this,arguments);return Qp(this,t)});function a(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:null,e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{};return Xp(this,a),e=ed({},rd,{},e),i.call(this,t,e)}return n=a,o=[{key:"terminate",value:function(t,e,n){return e=ed({},rd,{},e),xp.terminate(t,e,n)}}],(r=null)&&Vp(n.prototype,r),o&&Vp(n,o),a}(xp);function id(t){var e=function(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Boolean.prototype.valueOf.call(Reflect.construct(Boolean,[],(function(){}))),!0}catch(t){return!1}}();return function(){var n,r=Xs(t);if(e){var o=Xs(this).constructor;n=Reflect.construct(r,arguments,o)}else n=r.apply(this,arguments);return Ws(this,n)}}var ad=function(t){Ks(o,t);var e,n,r=id(o);function o(t){var e,n=t.chunkSize,i=t.csrfToken,a=t.fieldName,u=t.file,s=t.formId,c=t.retryDelays,l=t.uploadIndex,f=t.uploadUrl;return qi(this,o),Dr(Gs(e=r.call(this,{name:u.name,status:"uploading",type:"tus",uploadIndex:l})),"onError",void 0),Dr(Gs(e),"onProgress",void 0),Dr(Gs(e),"onSuccess",void 0),Dr(Gs(e),"upload",void 0),Dr(Gs(e),"csrfToken",void 0),Dr(Gs(e),"handleError",(function(t){if(!e.onError)throw t;e.onError(t)})),Dr(Gs(e),"handleProgress",(function(t,n){e.onProgress&&e.onProgress(t,n)})),Dr(Gs(e),"handleSucces",(function(){e.onSuccess&&e.onSuccess()})),Dr(Gs(e),"addCsrTokenToRequest",(function(t){t.setHeader("X-CSRFToken",e.csrfToken)})),e.csrfToken=i,e.upload=new od(u,{chunkSize:n,endpoint:f,metadata:{fieldName:a,filename:u.name,formId:s},onBeforeRequest:e.addCsrTokenToRequest,onError:e.handleError,onProgress:e.handleProgress,onSuccess:e.handleSucces,retryDelays:c||[0,1e3,3e3,5e3]}),e.onError=void 0,e.onProgress=void 0,e.onSuccess=void 0,e}return $i(o,[{key:"abort",value:(n=zi(ga.mark((function t(){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,this.upload.abort(!0);case 2:case"end":return t.stop()}}),t,this)}))),function(){return n.apply(this,arguments)})},{key:"delete",value:(e=zi(ga.mark((function t(){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if(this.upload.url){t.next=2;break}return t.abrupt("return",Promise.resolve());case 2:return t.next=4,ic(this.upload.url,this.csrfToken);case 4:case"end":return t.stop()}}),t,this)}))),function(){return e.apply(this,arguments)})},{key:"getSize",value:function(){return this.upload.file.size}},{key:"start",value:function(){this.upload.start()}},{key:"getInitialFile",value:function(){return{id:this.name,name:this.name,size:this.getSize(),type:"tus",url:""}}}]),o}(Zs),ud=function(){function t(e){var n=this,r=e.callbacks,o=e.chunkSize,i=e.csrfToken,a=e.eventEmitter,u=e.fieldName,s=e.form,c=e.formId,l=e.initial,f=e.input,p=e.multiple,d=e.parent,h=e.prefix,v=e.retryDelays,y=e.s3UploadDir,g=e.skipRequired,m=e.supportDropArea,x=e.translations,b=e.uploadUrl;qi(this,t),Dr(this,"callbacks",void 0),Dr(this,"chunkSize",void 0),Dr(this,"csrfToken",void 0),Dr(this,"eventEmitter",void 0),Dr(this,"fieldName",void 0),Dr(this,"form",void 0),Dr(this,"formId",void 0),Dr(this,"multiple",void 0),Dr(this,"nextUploadIndex",void 0),Dr(this,"prefix",void 0),Dr(this,"renderer",void 0),Dr(this,"retryDelays",void 0),Dr(this,"s3UploadDir",void 0),Dr(this,"supportDropArea",void 0),Dr(this,"uploads",void 0),Dr(this,"uploadUrl",void 0),Dr(this,"uploadFiles",function(){var t=zi(ga.mark((function t(e){var r,o,i,a,u,s,c;return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if(0!==e.length){t.next=2;break}return t.abrupt("return");case 2:n.multiple||0===n.uploads.length||(n.renderer.deleteFile(0),n.uploads=[]),r=!0,o=!1,t.prev=5,a=Ki(e);case 7:return t.next=9,a.next();case 9:return u=t.sent,r=u.done,t.next=13,u.value;case 13:if(s=t.sent,r){t.next=21;break}return c=s,t.next=18,n.uploadFile(c);case 18:r=!0,t.next=7;break;case 21:t.next=27;break;case 23:t.prev=23,t.t0=t.catch(5),o=!0,i=t.t0;case 27:if(t.prev=27,t.prev=28,r||null==a.return){t.next=32;break}return t.next=32,a.return();case 32:if(t.prev=32,!o){t.next=35;break}throw i;case 35:return t.finish(32);case 36:return t.finish(27);case 37:n.checkDropHint();case 38:case"end":return t.stop()}}),t,null,[[5,23,27,37],[28,,32,36]])})));return function(e){return t.apply(this,arguments)}}()),Dr(this,"onChange",(function(t){var e=t.target.files;e&&n.uploadFiles(bo(e))})),Dr(this,"onClick",(function(t){var e=t.target,r=function(){var t=e.getAttribute("data-index");if(t){var r=parseInt(t,10);return n.getUploadByIndex(r)}};if(e.classList.contains("dff-delete")&&!e.classList.contains("dff-disabled")){var o=r();o&&n.removeExistingUpload(o),t.preventDefault()}else if(e.classList.contains("dff-cancel")){var i=r();i&&n.handleCancel(i),t.preventDefault()}})),Dr(this,"handleProgress",(function(t,e,r){var o=(e/r*100).toFixed(2);n.renderer.updateProgress(t.uploadIndex,o);var i=n.callbacks.onProgress;i&&t instanceof ad&&i(e,r,t)})),Dr(this,"handleError",(function(t,e){n.renderer.setError(t.uploadIndex),t.status="error";var r=n.callbacks.onError;r&&t instanceof ad&&r(e,t)})),Dr(this,"handleSuccess",(function(t){var e=n.renderer;n.updatePlaceholderInput(),e.clearInput(),e.setSuccess(t.uploadIndex,t.getSize()),t.status="done";var r=n.callbacks.onSuccess,o=document.getElementsByClassName("dff-file-id-".concat(t.uploadIndex))[0];n.emitEvent("uploadComplete",o,t),r&&"tus"===t.type&&r(t)})),this.callbacks=r,this.chunkSize=o,this.csrfToken=i,this.eventEmitter=a,this.fieldName=u,this.form=s,this.formId=c,this.multiple=p,this.prefix=h,this.retryDelays=v,this.s3UploadDir=y,this.supportDropArea=m,this.uploadUrl=b,this.uploads=[],this.nextUploadIndex=0,this.renderer=new Ua({parent:d,input:f,skipRequired:g,translations:x});var w=this.renderer.container;m&&this.initDropArea(w,f.accept),l&&this.addInitialFiles(l),this.checkDropHint(),f.addEventListener("change",this.onChange),w.addEventListener("click",this.onClick)}var e,n,r;return $i(t,[{key:"addInitialFiles",value:function(t){var e=this;if(0!==t.length){var n=this.multiple,r=this.renderer,o=function(t){var n=t.size,o="s3"===t.type&&t.original_name?t.original_name:t.name,i=r.addUploadedFile(o,e.nextUploadIndex,n),a=function(t){var e=t.csrfToken,n=t.initialFile,r=t.uploadIndex,o=t.uploadUrl;switch(n.type){case"existing":return new fc(n,r);case"placeholder":return new cc(n,r);case"s3":return new lc(n,r);case"tus":return new pc({csrfToken:e,initialFile:n,uploadUrl:o,uploadIndex:r})}}({csrfToken:e.csrfToken,initialFile:t,uploadIndex:e.nextUploadIndex,uploadUrl:e.uploadUrl});e.uploads.push(a),e.emitEvent("addUpload",i,a)};n?t.forEach((function(t){o(t),e.nextUploadIndex+=1})):o(t[0])}}},{key:"uploadFile",value:(r=zi(ga.mark((function t(e){var n,r,o,i,a,u,s,c,l,f,p=this;return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if(n=function(){var t=p.csrfToken,n=p.s3UploadDir;return null!=n?new oc({csrfToken:t,endpoint:a,file:e,s3UploadDir:n,uploadIndex:c}):new ad({chunkSize:p.chunkSize,csrfToken:p.csrfToken,fieldName:r,file:e,formId:o,retryDelays:p.retryDelays,uploadIndex:c,uploadUrl:a})},r=this.fieldName,o=this.formId,i=this.renderer,a=this.uploadUrl,u=e.name,s=this.findUploadByName(u),c=s?s.uploadIndex:this.nextUploadIndex,s||(this.nextUploadIndex+=1),!s){t.next=9;break}return t.next=9,this.removeExistingUpload(s);case 9:(l=n()).onError=function(t){return p.handleError(l,t)},l.onProgress=function(t,e){return p.handleProgress(l,t,e)},l.onSuccess=function(){return p.handleSuccess(l)},l.start(),this.uploads.push(l),f=i.addNewUpload(u,c),this.emitEvent("addUpload",f,l);case 17:case"end":return t.stop()}}),t,this)}))),function(t){return r.apply(this,arguments)})},{key:"getUploadByIndex",value:function(t){return this.uploads.find((function(e){return e.uploadIndex===t}))}},{key:"findUploadByName",value:function(t){return this.uploads.find((function(e){return e.name===t}))}},{key:"removeExistingUpload",value:(n=zi(ga.mark((function t(e){var n;return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if((n=this.renderer.findFileDiv(e.uploadIndex))&&this.emitEvent("removeUpload",n,e),"uploading"!==e.status){t.next=8;break}return this.renderer.disableCancel(e.uploadIndex),t.next=6,e.abort();case 6:t.next=19;break;case 8:if("done"!==e.status){t.next=19;break}return this.renderer.disableDelete(e.uploadIndex),t.prev=10,t.next=13,e.delete();case 13:t.next=19;break;case 15:return t.prev=15,t.t0=t.catch(10),this.renderer.setDeleteFailed(e.uploadIndex),t.abrupt("return");case 19:this.removeUploadFromList(e),this.updatePlaceholderInput();case 21:case"end":return t.stop()}}),t,this,[[10,15]])}))),function(t){return n.apply(this,arguments)})},{key:"removeUploadFromList",value:function(t){this.renderer.deleteFile(t.uploadIndex);var e=this.uploads.indexOf(t);e>=0&&this.uploads.splice(e,1),this.checkDropHint();var n=this.callbacks.onDelete;n&&n(t)}},{key:"handleCancel",value:(e=zi(ga.mark((function t(e){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return this.renderer.disableCancel(e.uploadIndex),t.next=3,e.abort();case 3:this.removeUploadFromList(e);case 4:case"end":return t.stop()}}),t,this)}))),function(t){return e.apply(this,arguments)})},{key:"initDropArea",value:function(t,e){new Ls({container:t,inputAccept:e,onUploadFiles:this.uploadFiles})}},{key:"checkDropHint",value:function(){this.supportDropArea&&(0===this.uploads.filter((function(t){return t})).length?this.renderer.renderDropHint():this.renderer.removeDropHint())}},{key:"updatePlaceholderInput",value:function(){var t=ba(this.form,wa(this.fieldName,this.prefix),this.prefix);if(t){var e=this.uploads.filter((function(t){return"placeholder"===t.type||"s3"===t.type||"uploadedS3"===t.type})).map((function(t){return t.getInitialFile()}));t.value=JSON.stringify(e)}}},{key:"getMetaDataField",value:function(){return ba(this.form,(t=this.fieldName,e=this.prefix,"".concat(xa(t,e),"-metadata")),this.prefix);var t,e}},{key:"emitEvent",value:function(t,e,n){this.eventEmitter&&this.eventEmitter.emit(t,{element:e,fieldName:this.fieldName,fileName:n.name,metaDataField:this.getMetaDataField(),upload:n})}}]),t}(),sd=function(t){var e,n=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},r=function(t){return!n||!n.prefix||t.startsWith("".concat(n.prefix,"-"))},o=function(){return n&&n.prefix?n.prefix:null},i=function(e){return _a(t,e,o())},a=function(t){var e=t.dataset.files;return e?JSON.parse(e):[]},u=function(t){var e=i(wa(t,o()));return e?JSON.parse(e):[]},s=i("upload_url"),c=i("form_id"),l=i("s3_upload_dir"),f=n.skipRequired||!1,p=o(),d=null===(e=ba(t,"csrfmiddlewaretoken",null))||void 0===e?void 0:e.value;if(!d)throw Error("Csrf token not found");c&&s&&t.querySelectorAll(".dff-uploader").forEach((function(e){var o=e.querySelector(".dff-container");if(o){var i=o.querySelector("input[type=file]");if(i&&r(i.name)){var h=i.name,v=i.multiple,y=a(o).concat(u(h)),g=o.getAttribute("data-translations"),m=g?JSON.parse(g):{},x=!(!1===n.supportDropArea);new ud({callbacks:n.callbacks||{},chunkSize:n.chunkSize||2621440,csrfToken:d,eventEmitter:n.eventEmitter,fieldName:h,form:t,formId:c,s3UploadDir:l||null,initial:y,input:i,multiple:v,parent:o,prefix:p,retryDelays:n.retryDelays||null,skipRequired:f,supportDropArea:x,translations:m,uploadUrl:s})}}}))};function cd(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function ld(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?cd(Object(n),!0).forEach((function(e){Dr(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):cd(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}window.autoInitFileForms=function(){var t,e=window.initUploadFields;(t=Array.from(document.querySelectorAll(".dff-uploader")).map(Wn),Array.from(new Set(t).values())).forEach(e)},window.initFormSet=function(t,e){var n,r=(n="string"==typeof e?{prefix:e}:e).prefix||"form",o=_a(t,"TOTAL_FORMS",r);if(o)for(var i=parseInt(o,10),a=0;a<i;a+=1){var u=ma("".concat(a),null);sd(t,ld(ld({},n),{},{prefix:"".concat(r,"-").concat(u)}))}},window.initUploadFields=sd}();
+var Ca=/["'&<>]/,Ia=function(t){var e,n=""+t,r=Ca.exec(n);if(!r)return n;var o="",i=0,a=0;for(i=r.index;i<n.length;i++){switch(n.charCodeAt(i)){case 34:e="&quot;";break;case 38:e="&amp;";break;case 39:e="&#39;";break;case 60:e="&lt;";break;case 62:e="&gt;";break;default:continue}a!==i&&(o+=n.substring(a,i)),a=i+1,o+=e}return a!==i?o+n.substring(a,i):o};var Ua=function(){function t(e){var n=e.parent,r=e.input,o=e.skipRequired,i=e.translations;qi(this,t),Dr(this,"container",void 0),Dr(this,"input",void 0),Dr(this,"translations",void 0),Dr(this,"createFilesContainer",(function(t){var e=document.createElement("div");return e.className="dff-files",t.appendChild(e),e})),this.container=this.createFilesContainer(n),this.input=r,this.translations=i,o&&(this.input.required=!1)}return $i(t,[{key:"addNewUpload",value:function(t,e){var n=this.addFile(t,e),r=document.createElement("span");r.className="dff-progress";var o=document.createElement("span");o.className="dff-progress-inner",r.appendChild(o),n.appendChild(r);var i=document.createElement("a");return i.className="dff-cancel",i.innerHTML=this.translations.Cancel,i.setAttribute("data-index","".concat(e)),i.href="#",n.appendChild(i),n}},{key:"addUploadedFile",value:function(t,e,n){var r=this.addFile(t,e);return this.setSuccess(e,n),r}},{key:"clearInput",value:function(){this.input.value=""}},{key:"deleteFile",value:function(t){var e=this.findFileDiv(t);e&&e.remove()}},{key:"disableCancel",value:function(t){var e=this.findCancelSpan(t);e&&e.classList.add("dff-disabled")}},{key:"disableDelete",value:function(t){var e=this.findDeleteLink(t);e&&e.classList.add("dff-disabled")}},{key:"findFileDiv",value:function(t){return this.container.querySelector(".dff-file-id-".concat(t))}},{key:"removeDropHint",value:function(){var t=this.container.querySelector(".dff-drop-hint");t&&t.remove()}},{key:"renderDropHint",value:function(){if(!this.container.querySelector(".dff-drop-hint")){var t=document.createElement("div");t.className="dff-drop-hint",t.innerHTML=this.translations["Drop your files here"],this.container.appendChild(t)}}},{key:"setDeleteFailed",value:function(t){this.setErrorMessage(t,this.translations["Delete failed"]),this.enableDelete(t)}},{key:"setError",value:function(t){this.setErrorMessage(t,this.translations["Upload failed"]);var e=this.findFileDiv(t);e&&e.classList.add("dff-upload-fail"),this.removeProgress(t),this.removeCancel(t)}},{key:"setSuccess",value:function(t,e){var n=this.translations,r=this.findFileDiv(t);if(r){if(r.classList.add("dff-upload-success"),null!=e){var o=document.createElement("span");o.innerHTML=function(t,e){if(0===t)return"0 Bytes";var n=e<=0?0:e||2,r=Math.floor(Math.log(t)/Math.log(1024)),o=parseFloat((t/Math.pow(1024,r)).toFixed(n)),i=["Bytes","KB","MB","GB"][r];return"".concat(o," ").concat(i)}(e,2),o.className="dff-filesize",r.appendChild(o)}var i=document.createElement("a");i.innerHTML=n.Delete,i.className="dff-delete",i.setAttribute("data-index","".concat(t)),i.href="#",r.appendChild(i)}this.removeProgress(t),this.removeCancel(t)}},{key:"updateProgress",value:function(t,e){var n=this.container.querySelector(".dff-file-id-".concat(t));if(n){var r=n.querySelector(".dff-progress-inner");r&&(r.style.width="".concat(e,"%"))}}},{key:"addFile",value:function(t,e){var n=document.createElement("div");n.className="dff-file dff-file-id-".concat(e);var r=document.createElement("span");return r.innerHTML=Ia(t),n.appendChild(r),this.container.appendChild(n),this.input.required=!1,n}},{key:"removeProgress",value:function(t){var e=this.findFileDiv(t);if(e){var n=e.querySelector(".dff-progress");n&&n.remove()}}},{key:"removeCancel",value:function(t){var e=this.findCancelSpan(t);e&&e.remove()}},{key:"findCancelSpan",value:function(t){var e=this.findFileDiv(t);return e?e.querySelector(".dff-cancel"):null}},{key:"enableDelete",value:function(t){var e=this.findDeleteLink(t);e&&e.classList.remove("dff-disabled")}},{key:"findDeleteLink",value:function(t){var e=this.findFileDiv(t);return e?e.querySelector(".dff-delete"):e}},{key:"setErrorMessage",value:function(t,e){var n=this.findFileDiv(t);if(n){var r=document.querySelector(".dff-error");r&&r.remove();var o=document.createElement("span");o.classList.add("dff-error"),o.innerHTML=e,n.appendChild(o)}}}]),t}();function La(t,e){return function(t){if(Array.isArray(t))return t}(t)||function(t,e){if("undefined"!=typeof Symbol&&Symbol.iterator in Object(t)){var n=[],r=!0,o=!1,i=void 0;try{for(var a,u=t[Symbol.iterator]();!(r=(a=u.next()).done)&&(n.push(a.value),!e||n.length!==e);r=!0);}catch(t){o=!0,i=t}finally{try{r||null==u.return||u.return()}finally{if(o)throw i}}return n}}(t,e)||xo(t,e)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}var Da=Vt("species"),Ha=!o((function(){var t=/./;return t.exec=function(){var t=[];return t.groups={a:"7"},t},"7"!=="".replace(t,"$<a>")})),Fa="$0"==="a".replace(/./,"$0"),Ma=Vt("replace"),Na=!!/./[Ma]&&""===/./[Ma]("a","$0"),Ba=!o((function(){var t=/(?:)/,e=t.exec;t.exec=function(){return e.apply(this,arguments)};var n="ab".split(t);return 2!==n.length||"a"!==n[0]||"b"!==n[1]})),za=function(t,e,n,r){var i=Vt(t),a=!o((function(){var e={};return e[i]=function(){return 7},7!=""[t](e)})),u=a&&!o((function(){var e=!1,n=/a/;return"split"===t&&((n={}).constructor={},n.constructor[Da]=function(){return n},n.flags="",n[i]=/./[i]),n.exec=function(){return e=!0,null},n[i](""),!e}));if(!a||!u||"replace"===t&&(!Ha||!Fa||Na)||"split"===t&&!Ba){var s=/./[i],c=n(i,""[t],(function(t,e,n,r,o){return e.exec===Pa?a&&!o?{done:!0,value:s.call(e,n,r)}:{done:!0,value:t.call(n,e,r)}:{done:!1}}),{REPLACE_KEEPS_$0:Fa,REGEXP_REPLACE_SUBSTITUTES_UNDEFINED_CAPTURE:Na}),l=c[0],f=c[1];nt(String.prototype,t,l),nt(RegExp.prototype,i,2==e?function(t,e){return f.call(t,this,e)}:function(t){return f.call(t,this)})}r&&T(RegExp.prototype[i],"sham",!0)},qa=me.charAt,Ga=function(t,e,n){return e+(n?qa(t,e).length:1)},$a=function(t,e){var n=t.exec;if("function"==typeof n){var r=n.call(t,e);if("object"!=typeof r)throw TypeError("RegExp exec method returned something other than an Object or null");return r}if("RegExp"!==f(t))throw TypeError("RegExp#exec called on incompatible receiver");return Pa.call(t,e)},Ka=[].push,Wa=Math.min,Xa=4294967295,Va=!o((function(){return!RegExp(Xa,"y")}));za("split",2,(function(t,e,n){var r;return r="c"=="abbc".split(/(b)*/)[1]||4!="test".split(/(?:)/,-1).length||2!="ab".split(/(?:ab)*/).length||4!=".".split(/(.?)(.?)/).length||".".split(/()()/).length>1||"".split(/.?/).length?function(t,n){var r=String(h(this)),o=void 0===n?Xa:n>>>0;if(0===o)return[];if(void 0===t)return[r];if(!$r(t))return e.call(r,t,o);for(var i,a,u,s=[],c=(t.ignoreCase?"i":"")+(t.multiline?"m":"")+(t.unicode?"u":"")+(t.sticky?"y":""),l=0,f=new RegExp(t.source,c+"g");(i=Pa.call(f,r))&&!((a=f.lastIndex)>l&&(s.push(r.slice(l,i.index)),i.length>1&&i.index<r.length&&Ka.apply(s,i.slice(1)),u=i[0].length,l=a,s.length>=o));)f.lastIndex===i.index&&f.lastIndex++;return l===r.length?!u&&f.test("")||s.push(""):s.push(r.slice(l)),s.length>o?s.slice(0,o):s}:"0".split(void 0,0).length?function(t,n){return void 0===t&&0===n?[]:e.call(this,t,n)}:e,[function(e,n){var o=h(this),i=null==e?void 0:e[t];return void 0!==i?i.call(e,o,n):r.call(String(o),e,n)},function(t,o){var i=n(r,t,this,o,r!==e);if(i.done)return i.value;var a=k(t),u=String(this),s=ko(a,RegExp),c=a.unicode,l=(a.ignoreCase?"i":"")+(a.multiline?"m":"")+(a.unicode?"u":"")+(Va?"y":"g"),f=new s(Va?a:"^(?:"+a.source+")",l),p=void 0===o?Xa:o>>>0;if(0===p)return[];if(0===u.length)return null===$a(f,u)?[u]:[];for(var d=0,h=0,v=[];h<u.length;){f.lastIndex=Va?h:0;var y,g=$a(f,Va?u:u.slice(h));if(null===g||(y=Wa(lt(f.lastIndex+(Va?0:h)),u.length))===d)h=Ga(u,h,c);else{if(v.push(u.slice(d,h)),v.length===p)return v;for(var m=1;m<=g.length-1;m++)if(v.push(g[m]),v.length===p)return v;h=d=y}}return v.push(u.slice(d)),v}]}),!Va);var Ya="\t\n\v\f\r                　\u2028\u2029\ufeff",Qa="["+Ya+"]",Ja=RegExp("^"+Qa+Qa+"*"),Za=RegExp(Qa+Qa+"*$"),tu=function(t){return function(e){var n=String(h(e));return 1&t&&(n=n.replace(Ja,"")),2&t&&(n=n.replace(Za,"")),n}},eu={start:tu(1),end:tu(2),trim:tu(3)},nu=function(t){return o((function(){return!!Ya[t]()||"​᠎"!="​᠎"[t]()||Ya[t].name!==t}))},ru=eu.trim;Ct({target:"String",proto:!0,forced:nu("trim")},{trim:function(){return ru(this)}});var ou=Math.floor,iu="".replace,au=/\$([$&'`]|\d{1,2}|<[^>]*>)/g,uu=/\$([$&'`]|\d{1,2})/g,su=function(t,e,n,r,o,i){var a=n+t.length,u=r.length,s=uu;return void 0!==o&&(o=Lt(o),s=au),iu.call(i,s,(function(i,s){var c;switch(s.charAt(0)){case"$":return"$";case"&":return t;case"`":return e.slice(0,n);case"'":return e.slice(a);case"<":c=o[s.slice(1,-1)];break;default:var l=+s;if(0===l)return i;if(l>u){var f=ou(l/10);return 0===f?i:f<=u?void 0===r[f-1]?s.charAt(1):r[f-1]+s.charAt(1):i}c=r[l-1]}return void 0===c?"":c}))},cu=Math.max,lu=Math.min;za("replace",2,(function(t,e,n,r){var o=r.REGEXP_REPLACE_SUBSTITUTES_UNDEFINED_CAPTURE,i=r.REPLACE_KEEPS_$0,a=o?"$":"$0";return[function(n,r){var o=h(this),i=null==n?void 0:n[t];return void 0!==i?i.call(n,o,r):e.call(String(o),n,r)},function(t,r){if(!o&&i||"string"==typeof r&&-1===r.indexOf(a)){var u=n(e,t,this,r);if(u.done)return u.value}var s=k(t),c=String(this),l="function"==typeof r;l||(r=String(r));var f=s.global;if(f){var p=s.unicode;s.lastIndex=0}for(var d=[];;){var h=$a(s,c);if(null===h)break;if(d.push(h),!f)break;""===String(h[0])&&(s.lastIndex=Ga(c,lt(s.lastIndex),p))}for(var v,y="",g=0,m=0;m<d.length;m++){h=d[m];for(var x=String(h[0]),b=cu(lu(st(h.index),c.length),0),w=[],_=1;_<h.length;_++)w.push(void 0===(v=h[_])?v:String(v));var E=h.groups;if(l){var S=[x].concat(w,b,c);void 0!==E&&S.push(E);var A=String(r.apply(void 0,S))}else A=su(x,c,b,w,E,r);b>=g&&(y+=c.slice(g,b)+A,g=b+x.length)}return y+c.slice(g)}]}));var fu=O.f,pu=bt.f,du=et.set,hu=Vt("match"),vu=r.RegExp,yu=vu.prototype,gu=/a/g,mu=/a/g,xu=new vu(gu)!==gu,bu=Sa.UNSUPPORTED_Y;if(i&&jt("RegExp",!xu||bu||o((function(){return mu[hu]=!1,vu(gu)!=gu||vu(mu)==mu||"/a/i"!=vu(gu,"i")})))){for(var wu=function(t,e){var n,r=this instanceof wu,o=$r(t),i=void 0===e;if(!r&&o&&t.constructor===wu&&i)return t;xu?o&&!i&&(t=t.source):t instanceof wu&&(i&&(e=fa.call(t)),t=t.source),bu&&(n=!!e&&e.indexOf("y")>-1)&&(e=e.replace(/y/g,""));var a=wn(xu?new vu(t,e):vu(t,e),r?this:yu,wu);return bu&&n&&du(a,{sticky:n}),a},_u=function(t){t in wu||fu(wu,t,{configurable:!0,get:function(){return vu[t]},set:function(e){vu[t]=e}})},Eu=pu(vu),Su=0;Eu.length>Su;)_u(Eu[Su++]);yu.constructor=wu,wu.prototype=yu,nt(r,"RegExp",wu)}function Au(){this._types=Object.create(null),this._extensions=Object.create(null);for(var t=0;t<arguments.length;t++)this.define(arguments[t]);this.define=this.define.bind(this),this.getType=this.getType.bind(this),this.getExtension=this.getExtension.bind(this)}An("RegExp"),Au.prototype.define=function(t,e){for(var n in t){var r=t[n].map((function(t){return t.toLowerCase()}));n=n.toLowerCase();for(var o=0;o<r.length;o++){var i=r[o];if("*"!==i[0]){if(!e&&i in this._types)throw new Error('Attempt to change mapping for "'+i+'" extension from "'+this._types[i]+'" to "'+n+'". Pass `force=true` to allow this, otherwise remove "'+i+'" from the list of extensions for "'+n+'".');this._types[i]=n}}if(e||!this._extensions[n]){var a=r[0];this._extensions[n]="*"!==a[0]?a:a.substr(1)}}},Au.prototype.getType=function(t){var e=(t=String(t)).replace(/^.*[/\\]/,"").toLowerCase(),n=e.replace(/^.*\./,"").toLowerCase(),r=e.length<t.length;return(n.length<e.length-1||!r)&&this._types[n]||null},Au.prototype.getExtension=function(t){return(t=/^\s*([^;\s]*)/.test(t)&&RegExp.$1)&&this._extensions[t.toLowerCase()]||null};var ku=new Au({"application/andrew-inset":["ez"],"application/applixware":["aw"],"application/atom+xml":["atom"],"application/atomcat+xml":["atomcat"],"application/atomdeleted+xml":["atomdeleted"],"application/atomsvc+xml":["atomsvc"],"application/atsc-dwd+xml":["dwd"],"application/atsc-held+xml":["held"],"application/atsc-rsat+xml":["rsat"],"application/bdoc":["bdoc"],"application/calendar+xml":["xcs"],"application/ccxml+xml":["ccxml"],"application/cdfx+xml":["cdfx"],"application/cdmi-capability":["cdmia"],"application/cdmi-container":["cdmic"],"application/cdmi-domain":["cdmid"],"application/cdmi-object":["cdmio"],"application/cdmi-queue":["cdmiq"],"application/cu-seeme":["cu"],"application/dash+xml":["mpd"],"application/davmount+xml":["davmount"],"application/docbook+xml":["dbk"],"application/dssc+der":["dssc"],"application/dssc+xml":["xdssc"],"application/ecmascript":["ecma","es"],"application/emma+xml":["emma"],"application/emotionml+xml":["emotionml"],"application/epub+zip":["epub"],"application/exi":["exi"],"application/fdt+xml":["fdt"],"application/font-tdpfr":["pfr"],"application/geo+json":["geojson"],"application/gml+xml":["gml"],"application/gpx+xml":["gpx"],"application/gxf":["gxf"],"application/gzip":["gz"],"application/hjson":["hjson"],"application/hyperstudio":["stk"],"application/inkml+xml":["ink","inkml"],"application/ipfix":["ipfix"],"application/its+xml":["its"],"application/java-archive":["jar","war","ear"],"application/java-serialized-object":["ser"],"application/java-vm":["class"],"application/javascript":["js","mjs"],"application/json":["json","map"],"application/json5":["json5"],"application/jsonml+json":["jsonml"],"application/ld+json":["jsonld"],"application/lgr+xml":["lgr"],"application/lost+xml":["lostxml"],"application/mac-binhex40":["hqx"],"application/mac-compactpro":["cpt"],"application/mads+xml":["mads"],"application/manifest+json":["webmanifest"],"application/marc":["mrc"],"application/marcxml+xml":["mrcx"],"application/mathematica":["ma","nb","mb"],"application/mathml+xml":["mathml"],"application/mbox":["mbox"],"application/mediaservercontrol+xml":["mscml"],"application/metalink+xml":["metalink"],"application/metalink4+xml":["meta4"],"application/mets+xml":["mets"],"application/mmt-aei+xml":["maei"],"application/mmt-usd+xml":["musd"],"application/mods+xml":["mods"],"application/mp21":["m21","mp21"],"application/mp4":["mp4s","m4p"],"application/mrb-consumer+xml":["*xdf"],"application/mrb-publish+xml":["*xdf"],"application/msword":["doc","dot"],"application/mxf":["mxf"],"application/n-quads":["nq"],"application/n-triples":["nt"],"application/node":["cjs"],"application/octet-stream":["bin","dms","lrf","mar","so","dist","distz","pkg","bpk","dump","elc","deploy","exe","dll","deb","dmg","iso","img","msi","msp","msm","buffer"],"application/oda":["oda"],"application/oebps-package+xml":["opf"],"application/ogg":["ogx"],"application/omdoc+xml":["omdoc"],"application/onenote":["onetoc","onetoc2","onetmp","onepkg"],"application/oxps":["oxps"],"application/p2p-overlay+xml":["relo"],"application/patch-ops-error+xml":["*xer"],"application/pdf":["pdf"],"application/pgp-encrypted":["pgp"],"application/pgp-signature":["asc","sig"],"application/pics-rules":["prf"],"application/pkcs10":["p10"],"application/pkcs7-mime":["p7m","p7c"],"application/pkcs7-signature":["p7s"],"application/pkcs8":["p8"],"application/pkix-attr-cert":["ac"],"application/pkix-cert":["cer"],"application/pkix-crl":["crl"],"application/pkix-pkipath":["pkipath"],"application/pkixcmp":["pki"],"application/pls+xml":["pls"],"application/postscript":["ai","eps","ps"],"application/provenance+xml":["provx"],"application/pskc+xml":["pskcxml"],"application/raml+yaml":["raml"],"application/rdf+xml":["rdf","owl"],"application/reginfo+xml":["rif"],"application/relax-ng-compact-syntax":["rnc"],"application/resource-lists+xml":["rl"],"application/resource-lists-diff+xml":["rld"],"application/rls-services+xml":["rs"],"application/route-apd+xml":["rapd"],"application/route-s-tsid+xml":["sls"],"application/route-usd+xml":["rusd"],"application/rpki-ghostbusters":["gbr"],"application/rpki-manifest":["mft"],"application/rpki-roa":["roa"],"application/rsd+xml":["rsd"],"application/rss+xml":["rss"],"application/rtf":["rtf"],"application/sbml+xml":["sbml"],"application/scvp-cv-request":["scq"],"application/scvp-cv-response":["scs"],"application/scvp-vp-request":["spq"],"application/scvp-vp-response":["spp"],"application/sdp":["sdp"],"application/senml+xml":["senmlx"],"application/sensml+xml":["sensmlx"],"application/set-payment-initiation":["setpay"],"application/set-registration-initiation":["setreg"],"application/shf+xml":["shf"],"application/sieve":["siv","sieve"],"application/smil+xml":["smi","smil"],"application/sparql-query":["rq"],"application/sparql-results+xml":["srx"],"application/srgs":["gram"],"application/srgs+xml":["grxml"],"application/sru+xml":["sru"],"application/ssdl+xml":["ssdl"],"application/ssml+xml":["ssml"],"application/swid+xml":["swidtag"],"application/tei+xml":["tei","teicorpus"],"application/thraud+xml":["tfi"],"application/timestamped-data":["tsd"],"application/toml":["toml"],"application/ttml+xml":["ttml"],"application/ubjson":["ubj"],"application/urc-ressheet+xml":["rsheet"],"application/urc-targetdesc+xml":["td"],"application/voicexml+xml":["vxml"],"application/wasm":["wasm"],"application/widget":["wgt"],"application/winhlp":["hlp"],"application/wsdl+xml":["wsdl"],"application/wspolicy+xml":["wspolicy"],"application/xaml+xml":["xaml"],"application/xcap-att+xml":["xav"],"application/xcap-caps+xml":["xca"],"application/xcap-diff+xml":["xdf"],"application/xcap-el+xml":["xel"],"application/xcap-error+xml":["xer"],"application/xcap-ns+xml":["xns"],"application/xenc+xml":["xenc"],"application/xhtml+xml":["xhtml","xht"],"application/xliff+xml":["xlf"],"application/xml":["xml","xsl","xsd","rng"],"application/xml-dtd":["dtd"],"application/xop+xml":["xop"],"application/xproc+xml":["xpl"],"application/xslt+xml":["*xsl","xslt"],"application/xspf+xml":["xspf"],"application/xv+xml":["mxml","xhvml","xvml","xvm"],"application/yang":["yang"],"application/yin+xml":["yin"],"application/zip":["zip"],"audio/3gpp":["*3gpp"],"audio/adpcm":["adp"],"audio/amr":["amr"],"audio/basic":["au","snd"],"audio/midi":["mid","midi","kar","rmi"],"audio/mobile-xmf":["mxmf"],"audio/mp3":["*mp3"],"audio/mp4":["m4a","mp4a"],"audio/mpeg":["mpga","mp2","mp2a","mp3","m2a","m3a"],"audio/ogg":["oga","ogg","spx","opus"],"audio/s3m":["s3m"],"audio/silk":["sil"],"audio/wav":["wav"],"audio/wave":["*wav"],"audio/webm":["weba"],"audio/xm":["xm"],"font/collection":["ttc"],"font/otf":["otf"],"font/ttf":["ttf"],"font/woff":["woff"],"font/woff2":["woff2"],"image/aces":["exr"],"image/apng":["apng"],"image/avif":["avif"],"image/bmp":["bmp"],"image/cgm":["cgm"],"image/dicom-rle":["drle"],"image/emf":["emf"],"image/fits":["fits"],"image/g3fax":["g3"],"image/gif":["gif"],"image/heic":["heic"],"image/heic-sequence":["heics"],"image/heif":["heif"],"image/heif-sequence":["heifs"],"image/hej2k":["hej2"],"image/hsj2":["hsj2"],"image/ief":["ief"],"image/jls":["jls"],"image/jp2":["jp2","jpg2"],"image/jpeg":["jpeg","jpg","jpe"],"image/jph":["jph"],"image/jphc":["jhc"],"image/jpm":["jpm"],"image/jpx":["jpx","jpf"],"image/jxr":["jxr"],"image/jxra":["jxra"],"image/jxrs":["jxrs"],"image/jxs":["jxs"],"image/jxsc":["jxsc"],"image/jxsi":["jxsi"],"image/jxss":["jxss"],"image/ktx":["ktx"],"image/ktx2":["ktx2"],"image/png":["png"],"image/sgi":["sgi"],"image/svg+xml":["svg","svgz"],"image/t38":["t38"],"image/tiff":["tif","tiff"],"image/tiff-fx":["tfx"],"image/webp":["webp"],"image/wmf":["wmf"],"message/disposition-notification":["disposition-notification"],"message/global":["u8msg"],"message/global-delivery-status":["u8dsn"],"message/global-disposition-notification":["u8mdn"],"message/global-headers":["u8hdr"],"message/rfc822":["eml","mime"],"model/3mf":["3mf"],"model/gltf+json":["gltf"],"model/gltf-binary":["glb"],"model/iges":["igs","iges"],"model/mesh":["msh","mesh","silo"],"model/mtl":["mtl"],"model/obj":["obj"],"model/stl":["stl"],"model/vrml":["wrl","vrml"],"model/x3d+binary":["*x3db","x3dbz"],"model/x3d+fastinfoset":["x3db"],"model/x3d+vrml":["*x3dv","x3dvz"],"model/x3d+xml":["x3d","x3dz"],"model/x3d-vrml":["x3dv"],"text/cache-manifest":["appcache","manifest"],"text/calendar":["ics","ifb"],"text/coffeescript":["coffee","litcoffee"],"text/css":["css"],"text/csv":["csv"],"text/html":["html","htm","shtml"],"text/jade":["jade"],"text/jsx":["jsx"],"text/less":["less"],"text/markdown":["markdown","md"],"text/mathml":["mml"],"text/mdx":["mdx"],"text/n3":["n3"],"text/plain":["txt","text","conf","def","list","log","in","ini"],"text/richtext":["rtx"],"text/rtf":["*rtf"],"text/sgml":["sgml","sgm"],"text/shex":["shex"],"text/slim":["slim","slm"],"text/spdx":["spdx"],"text/stylus":["stylus","styl"],"text/tab-separated-values":["tsv"],"text/troff":["t","tr","roff","man","me","ms"],"text/turtle":["ttl"],"text/uri-list":["uri","uris","urls"],"text/vcard":["vcard"],"text/vtt":["vtt"],"text/xml":["*xml"],"text/yaml":["yaml","yml"],"video/3gpp":["3gp","3gpp"],"video/3gpp2":["3g2"],"video/h261":["h261"],"video/h263":["h263"],"video/h264":["h264"],"video/iso.segment":["m4s"],"video/jpeg":["jpgv"],"video/jpm":["*jpm","jpgm"],"video/mj2":["mj2","mjp2"],"video/mp2t":["ts"],"video/mp4":["mp4","mp4v","mpg4"],"video/mpeg":["mpeg","mpg","mpe","m1v","m2v"],"video/ogg":["ogv"],"video/quicktime":["qt","mov"],"video/webm":["webm"]});za("match",1,(function(t,e,n){return[function(e){var n=h(this),r=null==e?void 0:e[t];return void 0!==r?r.call(e,n):new RegExp(e)[t](String(n))},function(t){var r=n(e,t,this);if(r.done)return r.value;var o=k(t),i=String(this);if(!o.global)return $a(o,i);var a=o.unicode;o.lastIndex=0;for(var u,s=[],c=0;null!==(u=$a(o,i));){var l=String(u[0]);s[c]=l,""===l&&(o.lastIndex=Ga(i,lt(o.lastIndex),a)),c++}return 0===c?null:s}]})),i&&("g"!=/./g.flags||Sa.UNSUPPORTED_Y)&&O.f(RegExp.prototype,"flags",{configurable:!0,get:fa});var Ru=function(t){var e,n=0,r=-1,o=!0;for(e=t.length-1;e>=0;--e)if(47===t.charCodeAt(e)){if(!o){n=e+1;break}}else-1===r&&(o=!1,r=e+1);return-1===r?"":t.slice(n,r)};function Ou(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function Tu(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?Ou(Object(n),!0).forEach((function(e){Dr(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):Ou(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}var ju="\\\\/",Pu="[^".concat(ju,"]"),Cu="\\.",Iu="\\/",Uu="[^/]",Lu="(?:".concat(Iu,"|$)"),Du="(?:^|".concat(Iu,")"),Hu="".concat(Cu,"{1,2}").concat(Lu),Fu={DOT_LITERAL:Cu,PLUS_LITERAL:"\\+",QMARK_LITERAL:"\\?",SLASH_LITERAL:Iu,ONE_CHAR:"(?=.)",QMARK:Uu,END_ANCHOR:Lu,DOTS_SLASH:Hu,NO_DOT:"(?!".concat(Cu,")"),NO_DOTS:"(?!".concat(Du).concat(Hu,")"),NO_DOT_SLASH:"(?!".concat(Cu,"{0,1}").concat(Lu,")"),NO_DOTS_SLASH:"(?!".concat(Hu,")"),QMARK_NO_DOT:"[^.".concat(Iu,"]"),STAR:"".concat(Uu,"*?"),START_ANCHOR:Du},Mu=Tu(Tu({},Fu),{},{SLASH_LITERAL:"[".concat(ju,"]"),QMARK:Pu,STAR:"".concat(Pu,"*?"),DOTS_SLASH:"".concat(Cu,"{1,2}(?:[").concat(ju,"]|$)"),NO_DOT:"(?!".concat(Cu,")"),NO_DOTS:"(?!(?:^|[".concat(ju,"])").concat(Cu,"{1,2}(?:[").concat(ju,"]|$))"),NO_DOT_SLASH:"(?!".concat(Cu,"{0,1}(?:[").concat(ju,"]|$))"),NO_DOTS_SLASH:"(?!".concat(Cu,"{1,2}(?:[").concat(ju,"]|$))"),QMARK_NO_DOT:"[^.".concat(ju,"]"),START_ANCHOR:"(?:^|[".concat(ju,"])"),END_ANCHOR:"(?:[".concat(ju,"]|$)")}),Nu={MAX_LENGTH:65536,POSIX_REGEX_SOURCE:{alnum:"a-zA-Z0-9",alpha:"a-zA-Z",ascii:"\\x00-\\x7F",blank:" \\t",cntrl:"\\x00-\\x1F\\x7F",digit:"0-9",graph:"\\x21-\\x7E",lower:"a-z",print:"\\x20-\\x7E ",punct:"\\-!\"#$%&'()\\*+,./:;<=>?@[\\]^_`{|}~",space:" \\t\\r\\n\\v\\f",upper:"A-Z",word:"A-Za-z0-9_",xdigit:"A-Fa-f0-9"},REGEX_BACKSLASH:/\\(?![*+?^${}(|)[\]])/g,REGEX_NON_SPECIAL_CHARS:/^[^@![\].,$*+?^{}()|\\/]+/,REGEX_SPECIAL_CHARS:/[-*+?.^${}(|)[\]]/,REGEX_SPECIAL_CHARS_BACKREF:/(\\?)((\W)(\3*))/g,REGEX_SPECIAL_CHARS_GLOBAL:/([-*+?.^${}(|)[\]])/g,REGEX_REMOVE_BACKSLASH:/(?:\[.*?[^\\]\]|\\(?=.))/g,REPLACEMENTS:{"***":"*","**/**":"**","**/**/**":"**"},CHAR_0:48,CHAR_9:57,CHAR_UPPERCASE_A:65,CHAR_LOWERCASE_A:97,CHAR_UPPERCASE_Z:90,CHAR_LOWERCASE_Z:122,CHAR_LEFT_PARENTHESES:40,CHAR_RIGHT_PARENTHESES:41,CHAR_ASTERISK:42,CHAR_AMPERSAND:38,CHAR_AT:64,CHAR_BACKWARD_SLASH:92,CHAR_CARRIAGE_RETURN:13,CHAR_CIRCUMFLEX_ACCENT:94,CHAR_COLON:58,CHAR_COMMA:44,CHAR_DOT:46,CHAR_DOUBLE_QUOTE:34,CHAR_EQUAL:61,CHAR_EXCLAMATION_MARK:33,CHAR_FORM_FEED:12,CHAR_FORWARD_SLASH:47,CHAR_GRAVE_ACCENT:96,CHAR_HASH:35,CHAR_HYPHEN_MINUS:45,CHAR_LEFT_ANGLE_BRACKET:60,CHAR_LEFT_CURLY_BRACE:123,CHAR_LEFT_SQUARE_BRACKET:91,CHAR_LINE_FEED:10,CHAR_NO_BREAK_SPACE:160,CHAR_PERCENT:37,CHAR_PLUS:43,CHAR_QUESTION_MARK:63,CHAR_RIGHT_ANGLE_BRACKET:62,CHAR_RIGHT_CURLY_BRACE:125,CHAR_RIGHT_SQUARE_BRACKET:93,CHAR_SEMICOLON:59,CHAR_SINGLE_QUOTE:39,CHAR_SPACE:32,CHAR_TAB:9,CHAR_UNDERSCORE:95,CHAR_VERTICAL_LINE:124,CHAR_ZERO_WIDTH_NOBREAK_SPACE:65279,SEP:"/",extglobChars:function(t){return{"!":{type:"negate",open:"(?:(?!(?:",close:"))".concat(t.STAR,")")},"?":{type:"qmark",open:"(?:",close:")?"},"+":{type:"plus",open:"(?:",close:")+"},"*":{type:"star",open:"(?:",close:")*"},"@":{type:"at",open:"(?:",close:")"}}},globChars:function(){var t=arguments.length>0&&void 0!==arguments[0]&&arguments[0];return!0===t?Mu:Fu}},Bu=e((function(t,e){var n=Nu.REGEX_BACKSLASH,r=Nu.REGEX_REMOVE_BACKSLASH,o=Nu.REGEX_SPECIAL_CHARS,i=Nu.REGEX_SPECIAL_CHARS_GLOBAL;e.isObject=function(t){return null!==t&&"object"===ca(t)&&!Array.isArray(t)},e.hasRegexChars=function(t){return o.test(t)},e.isRegexChar=function(t){return 1===t.length&&e.hasRegexChars(t)},e.escapeRegex=function(t){return t.replace(i,"\\$1")},e.toPosixSlashes=function(t){return t.replace(n,"/")},e.removeBackslashes=function(t){return t.replace(r,(function(t){return"\\"===t?"":t}))},e.isWindows=function(t){return!(!t||"boolean"!=typeof t.windows)&&t.windows},e.escapeLast=function(t,n,r){var o=t.lastIndexOf(n,r);return-1===o?t:"\\"===t[o-1]?e.escapeLast(t,n,o-1):"".concat(t.slice(0,o),"\\").concat(t.slice(o))},e.removePrefix=function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},n=t;return n.startsWith("./")&&(n=n.slice(2),e.prefix="./"),n},e.wrapOutput=function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},n=arguments.length>2&&void 0!==arguments[2]?arguments[2]:{},r=n.contains?"":"^",o=n.contains?"":"$",i="".concat(r,"(?:").concat(t,")").concat(o);return!0===e.negated&&(i="(?:^(?!".concat(i,").*$)")),i}})),zu=Nu.CHAR_ASTERISK,qu=Nu.CHAR_AT,Gu=Nu.CHAR_BACKWARD_SLASH,$u=Nu.CHAR_COMMA,Ku=Nu.CHAR_DOT,Wu=Nu.CHAR_EXCLAMATION_MARK,Xu=Nu.CHAR_FORWARD_SLASH,Vu=Nu.CHAR_LEFT_CURLY_BRACE,Yu=Nu.CHAR_LEFT_PARENTHESES,Qu=Nu.CHAR_LEFT_SQUARE_BRACKET,Ju=Nu.CHAR_PLUS,Zu=Nu.CHAR_QUESTION_MARK,ts=Nu.CHAR_RIGHT_CURLY_BRACE,es=Nu.CHAR_RIGHT_PARENTHESES,ns=Nu.CHAR_RIGHT_SQUARE_BRACKET,rs=function(t){return t===Xu||t===Gu},os=function(t){!0!==t.isPrefix&&(t.depth=t.isGlobstar?1/0:1)},is=function(t,e){for(var n,r,o=e||{},i=t.length-1,a=!0===o.parts||!0===o.scanToEnd,u=[],s=[],c=[],l=t,f=-1,p=0,d=0,h=!1,v=!1,y=!1,g=!1,m=!1,x=!1,b=!1,w=!1,_=!1,E=0,S={value:"",depth:0,isGlob:!1},A=function(){return f>=i},k=function(){return n=r,l.charCodeAt(++f)};f<i;){var R=void 0;if((r=k())!==Gu){if(!0===x||r===Vu){for(E++;!0!==A()&&(r=k());)if(r!==Gu)if(r!==Vu){if(!0!==x&&r===Ku&&(r=k())===Ku){if(h=S.isBrace=!0,y=S.isGlob=!0,_=!0,!0===a)continue;break}if(!0!==x&&r===$u){if(h=S.isBrace=!0,y=S.isGlob=!0,_=!0,!0===a)continue;break}if(r===ts&&0===--E){x=!1,h=S.isBrace=!0,_=!0;break}}else E++;else b=S.backslashes=!0,k();if(!0===a)continue;break}if(r!==Xu){if(!0!==o.noext)if(!0===(r===Ju||r===qu||r===zu||r===Zu||r===Wu)&&l.charCodeAt(f+1)===Yu){if(y=S.isGlob=!0,g=S.isExtglob=!0,_=!0,!0===a){for(;!0!==A()&&(r=k());)if(r!==Gu){if(r===es){y=S.isGlob=!0,_=!0;break}}else b=S.backslashes=!0,r=k();continue}break}if(r===zu){if(n===zu&&(m=S.isGlobstar=!0),y=S.isGlob=!0,_=!0,!0===a)continue;break}if(r===Zu){if(y=S.isGlob=!0,_=!0,!0===a)continue;break}if(r===Qu)for(;!0!==A()&&(R=k());)if(R!==Gu){if(R===ns){if(v=S.isBracket=!0,y=S.isGlob=!0,_=!0,!0===a)continue;break}}else b=S.backslashes=!0,k();if(!0===o.nonegate||r!==Wu||f!==p){if(!0!==o.noparen&&r===Yu){if(y=S.isGlob=!0,!0===a){for(;!0!==A()&&(r=k());)if(r!==Yu){if(r===es){_=!0;break}}else b=S.backslashes=!0,r=k();continue}break}if(!0===y){if(_=!0,!0===a)continue;break}}else w=S.negated=!0,p++}else{if(u.push(f),s.push(S),S={value:"",depth:0,isGlob:!1},!0===_)continue;if(n===Ku&&f===p+1){p+=2;continue}d=f+1}}else b=S.backslashes=!0,(r=k())===Vu&&(x=!0)}!0===o.noext&&(g=!1,y=!1);var O=l,T="",j="";p>0&&(T=l.slice(0,p),l=l.slice(p),d-=p),O&&!0===y&&d>0?(O=l.slice(0,d),j=l.slice(d)):!0===y?(O="",j=l):O=l,O&&""!==O&&"/"!==O&&O!==l&&rs(O.charCodeAt(O.length-1))&&(O=O.slice(0,-1)),!0===o.unescape&&(j&&(j=Bu.removeBackslashes(j)),O&&!0===b&&(O=Bu.removeBackslashes(O)));var P={prefix:T,input:t,start:p,base:O,glob:j,isBrace:h,isBracket:v,isGlob:y,isExtglob:g,isGlobstar:m,negated:w};if(!0===o.tokens&&(P.maxDepth=0,rs(r)||s.push(S),P.tokens=s),!0===o.parts||!0===o.tokens){for(var C,I=0;I<u.length;I++){var U=C?C+1:p,L=u[I],D=t.slice(U,L);o.tokens&&(0===I&&0!==p?(s[I].isPrefix=!0,s[I].value=T):s[I].value=D,os(s[I]),P.maxDepth+=s[I].depth),0===I&&""===D||c.push(D),C=L}if(C&&C+1<t.length){var H=t.slice(C+1);c.push(H),o.tokens&&(s[s.length-1].value=H,os(s[s.length-1]),P.maxDepth+=s[s.length-1].depth)}P.slashes=u,P.parts=c}return P},as=[].join,us=d!=Object,ss=Bn("join",",");Ct({target:"Array",proto:!0,forced:us||!ss},{join:function(t){return as.call(v(this),void 0===t?",":t)}});var cs=vt.includes;function ls(t,e){var n;if("undefined"==typeof Symbol||null==t[Symbol.iterator]){if(Array.isArray(t)||(n=function(t,e){if(!t)return;if("string"==typeof t)return fs(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);"Object"===n&&t.constructor&&(n=t.constructor.name);if("Map"===n||"Set"===n)return Array.from(t);if("Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n))return fs(t,e)}(t))||e&&t&&"number"==typeof t.length){n&&(t=n);var r=0,o=function(){};return{s:o,n:function(){return r>=t.length?{done:!0}:{done:!1,value:t[r++]}},e:function(t){throw t},f:o}}throw new TypeError("Invalid attempt to iterate non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}var i,a=!0,u=!1;return{s:function(){n=t[Symbol.iterator]()},n:function(){var t=n.next();return a=t.done,t},e:function(t){u=!0,i=t},f:function(){try{a||null==n.return||n.return()}finally{if(u)throw i}}}}function fs(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,r=new Array(e);n<e;n++)r[n]=t[n];return r}function ps(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function ds(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?ps(Object(n),!0).forEach((function(e){Dr(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):ps(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}Ct({target:"Array",proto:!0},{includes:function(t){return cs(this,t,arguments.length>1?arguments[1]:void 0)}}),nn("includes"),Ct({target:"String",proto:!0,forced:!Xr("includes")},{includes:function(t){return!!~String(h(this)).indexOf(Kr(t),arguments.length>1?arguments[1]:void 0)}}),Ct({target:"String",proto:!0},{repeat:Wi});var hs=Nu.MAX_LENGTH,vs=Nu.POSIX_REGEX_SOURCE,ys=Nu.REGEX_NON_SPECIAL_CHARS,gs=Nu.REGEX_SPECIAL_CHARS_BACKREF,ms=Nu.REPLACEMENTS,xs=function(t,e){if("function"==typeof e.expandRange)return e.expandRange.apply(e,bo(t).concat([e]));t.sort();var n="[".concat(t.join("-"),"]");try{new RegExp(n)}catch(e){return t.map((function(t){return Bu.escapeRegex(t)})).join("..")}return n},bs=function(t,e){return"Missing ".concat(t,': "').concat(e,'" - use "\\\\').concat(e,'" to match literal characters')},ws=function(t,e){if("string"!=typeof t)throw new TypeError("Expected a string");t=ms[t]||t;var n=ds({},e),r="number"==typeof n.maxLength?Math.min(hs,n.maxLength):hs,o=t.length;if(o>r)throw new SyntaxError("Input length: ".concat(o,", exceeds maximum allowed length: ").concat(r));var i={type:"bos",value:"",output:n.prepend||""},a=[i],u=n.capture?"":"?:",s=Nu.globChars(),c=Nu.extglobChars(s),l=s.DOT_LITERAL,f=s.PLUS_LITERAL,p=s.SLASH_LITERAL,d=s.ONE_CHAR,h=s.DOTS_SLASH,v=s.NO_DOT,y=s.NO_DOT_SLASH,g=s.NO_DOTS_SLASH,m=s.QMARK,x=s.QMARK_NO_DOT,b=s.STAR,w=s.START_ANCHOR,_=function(t){return"(".concat(u,"(?:(?!").concat(w).concat(t.dot?h:l,").)*?)")},E=n.dot?"":v,S=n.dot?m:x,A=!0===n.bash?_(n):b;n.capture&&(A="(".concat(A,")")),"boolean"==typeof n.noext&&(n.noextglob=n.noext);var k={input:t,index:-1,start:0,dot:!0===n.dot,consumed:"",output:"",prefix:"",backtrack:!1,negated:!1,brackets:0,braces:0,parens:0,quotes:0,globstar:!1,tokens:a};t=Bu.removePrefix(t,k),o=t.length;var R,O=[],T=[],j=[],P=i,C=function(){return k.index===o-1},I=k.peek=function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:1;return t[k.index+e]},U=k.advance=function(){return t[++k.index]},L=function(){return t.slice(k.index+1)},D=function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:"",e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:0;k.consumed+=t,k.index+=e},H=function(t){k.output+=null!=t.output?t.output:t.value,D(t.value)},F=function(){for(var t=1;"!"===I()&&("("!==I(2)||"?"===I(3));)U(),k.start++,t++;return t%2!=0&&(k.negated=!0,k.start++,!0)},M=function(t){k[t]++,j.push(t)},N=function(t){k[t]--,j.pop()},B=function(t){if("globstar"===P.type){var e=k.braces>0&&("comma"===t.type||"brace"===t.type),n=!0===t.extglob||O.length&&("pipe"===t.type||"paren"===t.type);"slash"===t.type||"paren"===t.type||e||n||(k.output=k.output.slice(0,-P.output.length),P.type="star",P.value="*",P.output=A,k.output+=P.output)}if(O.length&&"paren"!==t.type&&!c[t.value]&&(O[O.length-1].inner+=t.value),(t.value||t.output)&&H(t),P&&"text"===P.type&&"text"===t.type)return P.value+=t.value,void(P.output=(P.output||"")+t.value);t.prev=P,a.push(t),P=t},z=function(t,e){var r=ds(ds({},c[e]),{},{conditions:1,inner:""});r.prev=P,r.parens=k.parens,r.output=k.output;var o=(n.capture?"(":"")+r.open;M("parens"),B({type:t,value:e,output:k.output?"":d}),B({type:"paren",extglob:!0,value:U(),output:o}),O.push(r)},q=function(t){var e=t.close+(n.capture?")":"");if("negate"===t.type){var r=A;t.inner&&t.inner.length>1&&t.inner.includes("/")&&(r=_(n)),(r!==A||C()||/^\)+$/.test(L()))&&(e=t.close=")$))".concat(r)),"bos"===t.prev.type&&C()&&(k.negatedExtglob=!0)}B({type:"paren",extglob:!0,value:R,output:e}),N("parens")};if(!1!==n.fastpaths&&!/(^[*!]|[/()[\]{}"])/.test(t)){var G=!1,$=t.replace(gs,(function(t,e,n,r,o,i){return"\\"===r?(G=!0,t):"?"===r?e?e+r+(o?m.repeat(o.length):""):0===i?S+(o?m.repeat(o.length):""):m.repeat(n.length):"."===r?l.repeat(n.length):"*"===r?e?e+r+(o?A:""):A:e?t:"\\".concat(t)}));return!0===G&&($=!0===n.unescape?$.replace(/\\/g,""):$.replace(/\\+/g,(function(t){return t.length%2==0?"\\\\":t?"\\":""}))),$===t&&!0===n.contains?(k.output=t,k):(k.output=Bu.wrapOutput($,k,e),k)}for(;!C();)if("\0"!==(R=U())){if("\\"===R){var K=I();if("/"===K&&!0!==n.bash)continue;if("."===K||";"===K)continue;if(!K){B({type:"text",value:R+="\\"});continue}var W=/^\\+/.exec(L()),X=0;if(W&&W[0].length>2&&(X=W[0].length,k.index+=X,X%2!=0&&(R+="\\")),!0===n.unescape?R=U()||"":R+=U()||"",0===k.brackets){B({type:"text",value:R});continue}}if(k.brackets>0&&("]"!==R||"["===P.value||"[^"===P.value)){if(!1!==n.posix&&":"===R){var V=P.value.slice(1);if(V.includes("[")&&(P.posix=!0,V.includes(":"))){var Y=P.value.lastIndexOf("["),Q=P.value.slice(0,Y),J=P.value.slice(Y+2),Z=vs[J];if(Z){P.value=Q+Z,k.backtrack=!0,U(),i.output||1!==a.indexOf(P)||(i.output=d);continue}}}("["===R&&":"!==I()||"-"===R&&"]"===I())&&(R="\\".concat(R)),"]"!==R||"["!==P.value&&"[^"!==P.value||(R="\\".concat(R)),!0===n.posix&&"!"===R&&"["===P.value&&(R="^"),P.value+=R,H({value:R})}else if(1!==k.quotes||'"'===R)if('"'!==R)if("("!==R)if(")"!==R)if("["!==R)if("]"!==R)if("{"!==R||!0===n.nobrace)if("}"!==R)if("|"!==R)if(","!==R)if("/"!==R)if("."!==R)if("?"!==R){if("!"===R){if(!0!==n.noextglob&&"("===I()&&("?"!==I(2)||!/[!=<:]/.test(I(3)))){z("negate",R);continue}if(!0!==n.nonegate&&0===k.index){F();continue}}if("+"!==R)if("@"!==R)if("*"===R)if(!P||"globstar"!==P.type&&!0!==P.star){var tt=L();if(!0!==n.noextglob&&/^\([^?]/.test(tt))z("star",R);else if("star"!==P.type){var et={type:"star",value:R,output:A};!0!==n.bash?!P||"bracket"!==P.type&&"paren"!==P.type||!0!==n.regex?(k.index!==k.start&&"slash"!==P.type&&"dot"!==P.type||("dot"===P.type?(k.output+=y,P.output+=y):!0===n.dot?(k.output+=g,P.output+=g):(k.output+=E,P.output+=E),"*"!==I()&&(k.output+=d,P.output+=d)),B(et)):(et.output=R,B(et)):(et.output=".*?","bos"!==P.type&&"slash"!==P.type||(et.output=E+et.output),B(et))}else{if(!0===n.noglobstar){D(R);continue}var nt=P.prev,rt=nt.prev,ot="slash"===nt.type||"bos"===nt.type,it=rt&&("star"===rt.type||"globstar"===rt.type);if(!0===n.bash&&(!ot||tt[0]&&"/"!==tt[0])){B({type:"star",value:R,output:""});continue}var at=k.braces>0&&("comma"===nt.type||"brace"===nt.type),ut=O.length&&("pipe"===nt.type||"paren"===nt.type);if(!ot&&"paren"!==nt.type&&!at&&!ut){B({type:"star",value:R,output:""});continue}for(;"/**"===tt.slice(0,3);){var st=t[k.index+4];if(st&&"/"!==st)break;tt=tt.slice(3),D("/**",3)}if("bos"===nt.type&&C()){P.type="globstar",P.value+=R,P.output=_(n),k.output=P.output,k.globstar=!0,D(R);continue}if("slash"===nt.type&&"bos"!==nt.prev.type&&!it&&C()){k.output=k.output.slice(0,-(nt.output+P.output).length),nt.output="(?:".concat(nt.output),P.type="globstar",P.output=_(n)+(n.strictSlashes?")":"|$)"),P.value+=R,k.globstar=!0,k.output+=nt.output+P.output,D(R);continue}if("slash"===nt.type&&"bos"!==nt.prev.type&&"/"===tt[0]){var ct=void 0!==tt[1]?"|$":"";k.output=k.output.slice(0,-(nt.output+P.output).length),nt.output="(?:".concat(nt.output),P.type="globstar",P.output="".concat(_(n)).concat(p,"|").concat(p).concat(ct,")"),P.value+=R,k.output+=nt.output+P.output,k.globstar=!0,D(R+U()),B({type:"slash",value:"/",output:""});continue}if("bos"===nt.type&&"/"===tt[0]){P.type="globstar",P.value+=R,P.output="(?:^|".concat(p,"|").concat(_(n)).concat(p,")"),k.output=P.output,k.globstar=!0,D(R+U()),B({type:"slash",value:"/",output:""});continue}k.output=k.output.slice(0,-P.output.length),P.type="globstar",P.output=_(n),P.value+=R,k.output+=P.output,k.globstar=!0,D(R)}}else P.type="star",P.star=!0,P.value+=R,P.output=A,k.backtrack=!0,k.globstar=!0,D(R);else{"$"!==R&&"^"!==R||(R="\\".concat(R));var lt=ys.exec(L());lt&&(R+=lt[0],k.index+=lt[0].length),B({type:"text",value:R})}else{if(!0!==n.noextglob&&"("===I()&&"?"!==I(2)){B({type:"at",extglob:!0,value:R,output:""});continue}B({type:"text",value:R})}else{if(!0!==n.noextglob&&"("===I()&&"?"!==I(2)){z("plus",R);continue}if(P&&"("===P.value||!1===n.regex){B({type:"plus",value:R,output:f});continue}if(P&&("bracket"===P.type||"paren"===P.type||"brace"===P.type)||k.parens>0){B({type:"plus",value:R});continue}B({type:"plus",value:f})}}else{if(!(P&&"("===P.value)&&!0!==n.noextglob&&"("===I()&&"?"!==I(2)){z("qmark",R);continue}if(P&&"paren"===P.type){var ft=I(),pt=R;("("===P.value&&!/[!=<:]/.test(ft)||"<"===ft&&!/<([!=]|\w+>)/.test(L()))&&(pt="\\".concat(R)),B({type:"text",value:R,output:pt});continue}if(!0!==n.dot&&("slash"===P.type||"bos"===P.type)){B({type:"qmark",value:R,output:x});continue}B({type:"qmark",value:R,output:m})}else{if(k.braces>0&&"dot"===P.type){"."===P.value&&(P.output=l);var dt=T[T.length-1];P.type="dots",P.output+=R,P.value+=R,dt.dots=!0;continue}if(k.braces+k.parens===0&&"bos"!==P.type&&"slash"!==P.type){B({type:"text",value:R,output:l});continue}B({type:"dot",value:R,output:l})}else{if("dot"===P.type&&k.index===k.start+1){k.start=k.index+1,k.consumed="",k.output="",a.pop(),P=i;continue}B({type:"slash",value:R,output:p})}else{var ht=R,vt=T[T.length-1];vt&&"braces"===j[j.length-1]&&(vt.comma=!0,ht="|"),B({type:"comma",value:R,output:ht})}else O.length>0&&O[O.length-1].conditions++,B({type:"text",value:R});else{var yt=T[T.length-1];if(!0===n.nobrace||!yt){B({type:"text",value:R,output:R});continue}var gt=")";if(!0===yt.dots){for(var mt=a.slice(),xt=[],bt=mt.length-1;bt>=0&&(a.pop(),"brace"!==mt[bt].type);bt--)"dots"!==mt[bt].type&&xt.unshift(mt[bt].value);gt=xs(xt,n),k.backtrack=!0}if(!0!==yt.comma&&!0!==yt.dots){var wt=k.output.slice(0,yt.outputIndex),_t=k.tokens.slice(yt.tokensIndex);yt.value=yt.output="\\{",R=gt="\\}",k.output=wt;var Et,St=ls(_t);try{for(St.s();!(Et=St.n()).done;){var At=Et.value;k.output+=At.output||At.value}}catch(t){St.e(t)}finally{St.f()}}B({type:"brace",value:R,output:gt}),N("braces"),T.pop()}else{M("braces");var kt={type:"brace",value:R,output:"(",outputIndex:k.output.length,tokensIndex:k.tokens.length};T.push(kt),B(kt)}else{if(!0===n.nobracket||P&&"bracket"===P.type&&1===P.value.length){B({type:"text",value:R,output:"\\".concat(R)});continue}if(0===k.brackets){if(!0===n.strictBrackets)throw new SyntaxError(bs("opening","["));B({type:"text",value:R,output:"\\".concat(R)});continue}N("brackets");var Rt=P.value.slice(1);if(!0===P.posix||"^"!==Rt[0]||Rt.includes("/")||(R="/".concat(R)),P.value+=R,H({value:R}),!1===n.literalBrackets||Bu.hasRegexChars(Rt))continue;var Ot=Bu.escapeRegex(P.value);if(k.output=k.output.slice(0,-P.value.length),!0===n.literalBrackets){k.output+=Ot,P.value=Ot;continue}P.value="(".concat(u).concat(Ot,"|").concat(P.value,")"),k.output+=P.value}else{if(!0!==n.nobracket&&L().includes("]"))M("brackets");else{if(!0!==n.nobracket&&!0===n.strictBrackets)throw new SyntaxError(bs("closing","]"));R="\\".concat(R)}B({type:"bracket",value:R})}else{if(0===k.parens&&!0===n.strictBrackets)throw new SyntaxError(bs("opening","("));var Tt=O[O.length-1];if(Tt&&k.parens===Tt.parens+1){q(O.pop());continue}B({type:"paren",value:R,output:k.parens?")":"\\)"}),N("parens")}else M("parens"),B({type:"paren",value:R});else k.quotes=1===k.quotes?0:1,!0===n.keepQuotes&&B({type:"text",value:R});else R=Bu.escapeRegex(R),P.value+=R,H({value:R})}for(;k.brackets>0;){if(!0===n.strictBrackets)throw new SyntaxError(bs("closing","]"));k.output=Bu.escapeLast(k.output,"["),N("brackets")}for(;k.parens>0;){if(!0===n.strictBrackets)throw new SyntaxError(bs("closing",")"));k.output=Bu.escapeLast(k.output,"("),N("parens")}for(;k.braces>0;){if(!0===n.strictBrackets)throw new SyntaxError(bs("closing","}"));k.output=Bu.escapeLast(k.output,"{"),N("braces")}if(!0===n.strictSlashes||"star"!==P.type&&"bracket"!==P.type||B({type:"maybe_slash",value:"",output:"".concat(p,"?")}),!0===k.backtrack){k.output="";var jt,Pt=ls(k.tokens);try{for(Pt.s();!(jt=Pt.n()).done;){var Ct=jt.value;k.output+=null!=Ct.output?Ct.output:Ct.value,Ct.suffix&&(k.output+=Ct.suffix)}}catch(t){Pt.e(t)}finally{Pt.f()}}return k};ws.fastpaths=function(t,e){var n=ds({},e),r="number"==typeof n.maxLength?Math.min(hs,n.maxLength):hs,o=t.length;if(o>r)throw new SyntaxError("Input length: ".concat(o,", exceeds maximum allowed length: ").concat(r));t=ms[t]||t;var i=Nu.globChars(),a=i.DOT_LITERAL,u=i.SLASH_LITERAL,s=i.ONE_CHAR,c=i.DOTS_SLASH,l=i.NO_DOT,f=i.NO_DOTS,p=i.NO_DOTS_SLASH,d=i.STAR,h=i.START_ANCHOR,v=n.dot?f:l,y=n.dot?p:l,g=n.capture?"":"?:",m=!0===n.bash?".*?":d;n.capture&&(m="(".concat(m,")"));var x=function(t){return!0===t.noglobstar?m:"(".concat(g,"(?:(?!").concat(h).concat(t.dot?c:a,").)*?)")},b=function t(e){switch(e){case"*":return"".concat(v).concat(s).concat(m);case".*":return"".concat(a).concat(s).concat(m);case"*.*":return"".concat(v).concat(m).concat(a).concat(s).concat(m);case"*/*":return"".concat(v).concat(m).concat(u).concat(s).concat(y).concat(m);case"**":return v+x(n);case"**/*":return"(?:".concat(v).concat(x(n)).concat(u,")?").concat(y).concat(s).concat(m);case"**/*.*":return"(?:".concat(v).concat(x(n)).concat(u,")?").concat(y).concat(m).concat(a).concat(s).concat(m);case"**/.*":return"(?:".concat(v).concat(x(n)).concat(u,")?").concat(a).concat(s).concat(m);default:var r=/^(.*?)\.(\w+)$/.exec(e);if(!r)return;var o=t(r[1]);if(!o)return;return o+a+r[2]}}(Bu.removePrefix(t,{negated:!1,prefix:""}));return b&&!0!==n.strictSlashes&&(b+="".concat(u,"?")),b};var _s=ws;function Es(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function Ss(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?Es(Object(n),!0).forEach((function(e){Dr(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):Es(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function As(t,e){var n;if("undefined"==typeof Symbol||null==t[Symbol.iterator]){if(Array.isArray(t)||(n=function(t,e){if(!t)return;if("string"==typeof t)return ks(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);"Object"===n&&t.constructor&&(n=t.constructor.name);if("Map"===n||"Set"===n)return Array.from(t);if("Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n))return ks(t,e)}(t))||e&&t&&"number"==typeof t.length){n&&(t=n);var r=0,o=function(){};return{s:o,n:function(){return r>=t.length?{done:!0}:{done:!1,value:t[r++]}},e:function(t){throw t},f:o}}throw new TypeError("Invalid attempt to iterate non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}var i,a=!0,u=!1;return{s:function(){n=t[Symbol.iterator]()},n:function(){var t=n.next();return a=t.done,t},e:function(t){u=!0,i=t},f:function(){try{a||null==n.return||n.return()}finally{if(u)throw i}}}}function ks(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,r=new Array(e);n<e;n++)r[n]=t[n];return r}var Rs=function(t){return t&&"object"===ca(t)&&!Array.isArray(t)},Os=function t(e,n){var r=arguments.length>2&&void 0!==arguments[2]&&arguments[2];if(Array.isArray(e)){var o=e.map((function(e){return t(e,n,r)})),i=function(t){var e,n=As(o);try{for(n.s();!(e=n.n()).done;){var r=(0,e.value)(t);if(r)return r}}catch(t){n.e(t)}finally{n.f()}return!1};return i}var a=Rs(e)&&e.tokens&&e.input;if(""===e||"string"!=typeof e&&!a)throw new TypeError("Expected pattern to be a non-empty string");var u=n||{},s=Bu.isWindows(n),c=a?t.compileRe(e,n):t.makeRe(e,n,!1,!0),l=c.state;delete c.state;var f=function(){return!1};if(u.ignore){var p=Ss(Ss({},n),{},{ignore:null,onMatch:null,onResult:null});f=t(u.ignore,p,r)}var d=function(r){var o=arguments.length>1&&void 0!==arguments[1]&&arguments[1],i=t.test(r,c,n,{glob:e,posix:s}),a=i.isMatch,p=i.match,d=i.output,h={glob:e,state:l,regex:c,posix:s,input:r,output:d,match:p,isMatch:a};return"function"==typeof u.onResult&&u.onResult(h),!1===a?(h.isMatch=!1,!!o&&h):f(r)?("function"==typeof u.onIgnore&&u.onIgnore(h),h.isMatch=!1,!!o&&h):("function"==typeof u.onMatch&&u.onMatch(h),!o||h)};return r&&(d.state=l),d};Os.test=function(t,e,n){var r=arguments.length>3&&void 0!==arguments[3]?arguments[3]:{},o=r.glob,i=r.posix;if("string"!=typeof t)throw new TypeError("Expected input to be a string");if(""===t)return{isMatch:!1,output:""};var a=n||{},u=a.format||(i?Bu.toPosixSlashes:null),s=t===o,c=s&&u?u(t):t;return!1===s&&(s=(c=u?u(t):t)===o),!1!==s&&!0!==a.capture||(s=!0===a.matchBase||!0===a.basename?Os.matchBase(t,e,n,i):e.exec(c)),{isMatch:Boolean(s),match:s,output:c}},Os.matchBase=function(t,e,n){arguments.length>3&&void 0!==arguments[3]||Bu.isWindows(n);var r=e instanceof RegExp?e:Os.makeRe(e,n);return r.test(Ru(t))},Os.isMatch=function(t,e,n){return Os(e,n)(t)},Os.parse=function(t,e){return Array.isArray(t)?t.map((function(t){return Os.parse(t,e)})):_s(t,Ss(Ss({},e),{},{fastpaths:!1}))},Os.scan=function(t,e){return is(t,e)},Os.compileRe=function(t,e){var n=arguments.length>2&&void 0!==arguments[2]&&arguments[2],r=arguments.length>3&&void 0!==arguments[3]&&arguments[3];if(!0===n)return t.output;var o=e||{},i=o.contains?"":"^",a=o.contains?"":"$",u="".concat(i,"(?:").concat(t.output,")").concat(a);t&&!0===t.negated&&(u="^(?!".concat(u,").*$"));var s=Os.toRegex(u,e);return!0===r&&(s.state=t),s},Os.makeRe=function(t,e){var n=arguments.length>2&&void 0!==arguments[2]&&arguments[2],r=arguments.length>3&&void 0!==arguments[3]&&arguments[3];if(!t||"string"!=typeof t)throw new TypeError("Expected a non-empty string");var o,i=e||{},a={negated:!1,fastpaths:!0},u="";return t.startsWith("./")&&(t=t.slice(2),u=a.prefix="./"),!1===i.fastpaths||"."!==t[0]&&"*"!==t[0]||(o=_s.fastpaths(t,e)),void 0===o?(a=_s(t,e)).prefix=u+(a.prefix||""):a.output=o,Os.compileRe(a,e,n,r)},Os.toRegex=function(t,e){try{var n=e||{};return new RegExp(t,n.flags||(n.nocase?"i":""))}catch(t){if(e&&!0===e.debug)throw t;return/$^/}},Os.constants=Nu;var Ts=Os,js=function(){function t(e){qi(this,t),Dr(this,"extensions",void 0),Dr(this,"mimeTypes",void 0);var n=La(function(t){var e=[],n=[];return t.split(",").map((function(t){return t.trim()})).filter(Boolean).forEach((function(t){t.startsWith(".")?e.push("*".concat(t)):n.push(t)})),[e,n]}(e),2),r=n[0],o=n[1];this.extensions=r,this.mimeTypes=o}return $i(t,[{key:"isAccepted",value:function(t){return 0===this.extensions.length&&0===this.mimeTypes.length||(this.isMimeTypeAccepted(ku.getType(t))||this.isExtensionAccepted(t))}},{key:"isMimeTypeAccepted",value:function(t){return 0!==this.mimeTypes.length&&Ts.isMatch(t,this.mimeTypes)}},{key:"isExtensionAccepted",value:function(t){return 0!==this.extensions.length&&Ts.isMatch(t,this.extensions)}}]),t}(),Ps=function(){var t=zi(ga.mark((function t(e){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.abrupt("return",new Promise((function(t,n){return e.createReader().readEntries(t,n)})));case 1:case"end":return t.stop()}}),t)})));return function(e){return t.apply(this,arguments)}}(),Cs=function(){var t=zi(ga.mark((function t(e){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.abrupt("return",new Promise((function(t,n){return e.file(t,n)})));case 1:case"end":return t.stop()}}),t)})));return function(e){return t.apply(this,arguments)}}(),Is=function(){var t=zi(ga.mark((function t(e){var n,r,o,i,a,u,s,c,l,f;return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:n=[],r=!0,o=!1,t.prev=3,a=Ki(e);case 5:return t.next=7,a.next();case 7:return u=t.sent,r=u.done,t.next=11,u.value;case 11:if(s=t.sent,r){t.next=32;break}if(!(c=s).isFile){t.next=21;break}return t.next=17,Cs(c);case 17:l=t.sent,n.push(l),t.next=29;break;case 21:if(!c.isDirectory){t.next=29;break}return t.next=24,Ps(c);case 24:return f=t.sent,t.next=27,Is(f);case 27:t.sent.forEach((function(t){return n.push(t)}));case 29:r=!0,t.next=5;break;case 32:t.next=38;break;case 34:t.prev=34,t.t0=t.catch(3),o=!0,i=t.t0;case 38:if(t.prev=38,t.prev=39,r||null==a.return){t.next=43;break}return t.next=43,a.return();case 43:if(t.prev=43,!o){t.next=46;break}throw i;case 46:return t.finish(43);case 47:return t.finish(38);case 48:return t.abrupt("return",n);case 49:case"end":return t.stop()}}),t,null,[[3,34,38,48],[39,,43,47]])})));return function(e){return t.apply(this,arguments)}}(),Us=function(){var t=zi(ga.mark((function t(e){var n,r;return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if(!e.items){t.next=8;break}return n=bo(e.items).map((function(t){return t.webkitGetAsEntry()})),t.next=4,Is(n);case 4:return r=t.sent,t.abrupt("return",r);case 8:return t.abrupt("return",bo(e.files));case 9:case"end":return t.stop()}}),t)})));return function(e){return t.apply(this,arguments)}}(),Ls=function t(e){var n=this,r=e.container,o=e.inputAccept,i=e.onUploadFiles;qi(this,t),Dr(this,"acceptedFileTypes",void 0),Dr(this,"container",void 0),Dr(this,"onUploadFiles",void 0),Dr(this,"onDrop",(function(t){var e=t;n.container.classList.remove("dff-dropping"),e.preventDefault(),e.stopPropagation(),function(){var t=zi(ga.mark((function t(){var r,o;return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if(t.prev=0,!e.dataTransfer){t.next=7;break}return t.next=4,Us(e.dataTransfer);case 4:r=t.sent,o=r.filter((function(t){return n.acceptedFileTypes.isAccepted(t.name)})),n.onUploadFiles(o);case 7:t.next=12;break;case 9:t.prev=9,t.t0=t.catch(0),console.error(t.t0);case 12:case"end":return t.stop()}}),t,null,[[0,9]])})));return function(){return t.apply(this,arguments)}}()()})),this.container=r,this.onUploadFiles=i,this.acceptedFileTypes=new js(o),r.addEventListener("dragenter",(function(){r.classList.add("dff-dropping")})),r.addEventListener("dragleave",(function(){r.classList.remove("dff-dropping")})),r.addEventListener("dragover",(function(t){r.classList.add("dff-dropping"),t.preventDefault()})),r.addEventListener("drop",this.onDrop)},Ds=[].slice,Hs={},Fs=function(t,e,n){if(!(e in Hs)){for(var r=[],o=0;o<e;o++)r[o]="a["+o+"]";Hs[e]=Function("C,a","return new C("+r.join(",")+")")}return Hs[e](t,n)},Ms=Function.bind||function(t){var e=It(this),n=Ds.call(arguments,1),r=function(){var o=n.concat(Ds.call(arguments));return this instanceof r?Fs(e,o.length,o):e.apply(t,o)};return y(e.prototype)&&(r.prototype=e.prototype),r},Ns=it("Reflect","construct"),Bs=o((function(){function t(){}return!(Ns((function(){}),[],t)instanceof t)})),zs=!o((function(){Ns((function(){}))})),qs=Bs||zs;function Gs(t){if(void 0===t)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return t}function $s(t,e){return($s=Object.setPrototypeOf||function(t,e){return t.__proto__=e,t})(t,e)}function Ks(t,e){if("function"!=typeof e&&null!==e)throw new TypeError("Super expression must either be null or a function");t.prototype=Object.create(e&&e.prototype,{constructor:{value:t,writable:!0,configurable:!0}}),e&&$s(t,e)}function Ws(t,e){return!e||"object"!==ca(e)&&"function"!=typeof e?Gs(t):e}function Xs(t){return(Xs=Object.setPrototypeOf?Object.getPrototypeOf:function(t){return t.__proto__||Object.getPrototypeOf(t)})(t)}Ct({target:"Reflect",stat:!0,forced:qs,sham:qs},{construct:function(t,e){It(t),k(e);var n=arguments.length<3?t:It(arguments[2]);if(zs&&!Bs)return Ns(t,e,n);if(t==n){switch(e.length){case 0:return new t;case 1:return new t(e[0]);case 2:return new t(e[0],e[1]);case 3:return new t(e[0],e[1],e[2]);case 4:return new t(e[0],e[1],e[2],e[3])}var r=[null];return r.push.apply(r,e),new(Ms.apply(t,r))}var o=n.prototype,i=Ue(y(o)?o:Object.prototype),a=Function.apply.call(t,i,e);return y(a)?a:i}});var Vs=function(t){return function(e,n,r,o){It(n);var i=Lt(e),a=d(i),u=lt(i.length),s=t?u-1:0,c=t?-1:1;if(r<2)for(;;){if(s in a){o=a[s],s+=c;break}if(s+=c,t?s<0:u<=s)throw TypeError("Reduce of empty array with no initial value")}for(;t?s>=0:u>s;s+=c)s in a&&(o=n(o,a[s],s,i));return o}},Ys={left:Vs(!1),right:Vs(!0)},Qs=Ys.left,Js=Bn("reduce");Ct({target:"Array",proto:!0,forced:!Js||!Ft&&qt>79&&qt<83},{reduce:function(t){return Qs(this,t,arguments.length,arguments.length>1?arguments[1]:void 0)}});var Zs=function(){function t(e){var n=e.name,r=e.status,o=e.type,i=e.uploadIndex;qi(this,t),Dr(this,"name",void 0),Dr(this,"status",void 0),Dr(this,"type",void 0),Dr(this,"uploadIndex",void 0),this.name=n,this.status=r,this.type=o,this.uploadIndex=i}var e,n;return $i(t,[{key:"abort",value:(n=zi(ga.mark((function t(){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:case"end":return t.stop()}}),t)}))),function(){return n.apply(this,arguments)})},{key:"delete",value:(e=zi(ga.mark((function t(){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:case"end":return t.stop()}}),t)}))),function(){return e.apply(this,arguments)})}]),t}(),tc=e((function(e){var n,r;n=t,r=function(){function t(t){var e=[];if(0===t.length)return"";if("string"!=typeof t[0])throw new TypeError("Url must be a string. Received "+t[0]);if(t[0].match(/^[^/:]+:\/*$/)&&t.length>1){var n=t.shift();t[0]=n+t[0]}t[0].match(/^file:\/\/\//)?t[0]=t[0].replace(/^([^/:]+):\/*/,"$1:///"):t[0]=t[0].replace(/^([^/:]+):\/*/,"$1://");for(var r=0;r<t.length;r++){var o=t[r];if("string"!=typeof o)throw new TypeError("Url must be a string. Received "+o);""!==o&&(r>0&&(o=o.replace(/^[\/]+/,"")),o=r<t.length-1?o.replace(/[\/]+$/,""):o.replace(/[\/]+$/,"/"),e.push(o))}var i=e.join("/"),a=(i=i.replace(/\/(\?|&|#[^!])/g,"$1")).split("?");return i=a.shift()+(a.length>0?"?":"")+a.join("&")}return function(){return t("object"===ca(arguments[0])?arguments[0]:[].slice.call(arguments))}},e.exports?e.exports=r():n.urljoin=r()})),ec=function(t){var e=t.csrfToken,n=t.key,r=t.uploadId,o=t.endpoint,i=encodeURIComponent(n),a=encodeURIComponent(r),u=new Headers({"X-CSRFToken":e}),s=tc(o,a,"?key=".concat(i));return fetch(s,{method:"delete",headers:u}).then((function(t){return t.json()}))},nc=function(t,e){var n=t.indexOf(e);-1!==n&&t.splice(n,1)};function rc(t){var e=function(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Boolean.prototype.valueOf.call(Reflect.construct(Boolean,[],(function(){}))),!0}catch(t){return!1}}();return function(){var n,r=Xs(t);if(e){var o=Xs(this).constructor;n=Reflect.construct(r,arguments,o)}else n=r.apply(this,arguments);return Ws(this,n)}}var oc=function(t){Ks(o,t);var e,n,r=rc(o);function o(t){var e,n=t.csrfToken,i=t.endpoint,a=t.file,u=t.s3UploadDir,s=t.uploadIndex;return qi(this,o),Dr(Gs(e=r.call(this,{name:a.name,status:"uploading",type:"s3",uploadIndex:s})),"onError",void 0),Dr(Gs(e),"onProgress",void 0),Dr(Gs(e),"onSuccess",void 0),Dr(Gs(e),"chunkState",void 0),Dr(Gs(e),"chunks",void 0),Dr(Gs(e),"createdPromise",void 0),Dr(Gs(e),"csrfToken",void 0),Dr(Gs(e),"endpoint",void 0),Dr(Gs(e),"file",void 0),Dr(Gs(e),"key",void 0),Dr(Gs(e),"parts",void 0),Dr(Gs(e),"s3UploadDir",void 0),Dr(Gs(e),"uploadId",void 0),Dr(Gs(e),"uploading",void 0),e.csrfToken=n,e.endpoint=i,e.file=a,e.s3UploadDir=u,e.key=null,e.uploadId=null,e.parts=[],e.createdPromise=Promise.reject(),e.chunks=[],e.chunkState=[],e.uploading=[],e.onError=void 0,e.onProgress=void 0,e.onSuccess=void 0,e.initChunks(),e.createdPromise.catch((function(){return{}})),e}return $i(o,[{key:"abort",value:(n=zi(ga.mark((function t(){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return this.uploading.slice().forEach((function(t){t.abort()})),this.uploading=[],t.next=4,this.createdPromise;case 4:if(!this.key||!this.uploadId){t.next=7;break}return t.next=7,ec({csrfToken:this.csrfToken,endpoint:this.endpoint,key:this.key,uploadId:this.uploadId});case 7:case"end":return t.stop()}}),t,this)}))),function(){return n.apply(this,arguments)})},{key:"delete",value:(e=zi(ga.mark((function t(){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.abrupt("return",Promise.resolve());case 1:case"end":return t.stop()}}),t)}))),function(){return e.apply(this,arguments)})},{key:"getInitialFile",value:function(){return{id:this.uploadId||"",name:this.key||"",size:this.file.size,original_name:this.file.name,type:"s3"}}},{key:"getSize",value:function(){return this.file.size}},{key:"start",value:function(){this.createUpload()}},{key:"initChunks",value:function(){for(var t,e=[],n=(t=this.file,Math.ceil(t.size/1e4)),r=Math.max(5242880,Math.ceil(this.file.size/1e4)),o=Math.max(n,r),i=0;i<this.file.size;i+=o){var a=Math.min(this.file.size,i+o);e.push(this.file.slice(i,a))}this.chunks=e,this.chunkState=e.map((function(){return{uploaded:0,busy:!1,done:!1}}))}},{key:"createUpload",value:function(){var t,e,n,r,o,i,a=this;return this.createdPromise=(t={csrfToken:this.csrfToken,endpoint:this.endpoint,file:this.file,s3UploadDir:this.s3UploadDir},e=t.csrfToken,n=t.endpoint,r=t.file,o=t.s3UploadDir,i=new Headers({accept:"application/json","content-type":"application/json","X-CSRFToken":e}),fetch(n,{method:"post",headers:i,body:JSON.stringify({filename:r.name,contentType:r.type,s3UploadDir:o})}).then((function(t){return t.json()})).then((function(t){return t}))),this.createdPromise.then((function(t){if(!("object"===ca(t)&&t&&"string"==typeof t.uploadId&&"string"==typeof t.key))throw new TypeError("AwsS3/Multipart: Got incorrect result from `createMultipartUpload()`, expected an object `{ uploadId, key }`.");a.key=t.key,a.uploadId=t.uploadId,a.uploadParts()})).catch((function(t){a.handleError(t)}))}},{key:"uploadParts",value:function(){var t=this,e=1-this.uploading.length;if(0!==e)if(this.chunkState.every((function(t){return t.done})))this.completeUpload();else{for(var n=[],r=0;r<this.chunkState.length;r++){var o=this.chunkState[r];if(!o.done&&!o.busy&&(n.push(r),n.length>=e))break}n.forEach((function(e){t.uploadPart(e)}))}}},{key:"uploadPart",value:function(t){var e=this;return this.chunkState[t].busy=!0,this.key&&this.uploadId?function(t){var e=t.csrfToken,n=t.endpoint,r=t.key,o=t.number,i=t.uploadId,a=encodeURIComponent(r),u=new Headers({"X-CSRFToken":e}),s=tc(n,i,"".concat(o),"?key=".concat(a));return fetch(s,{method:"get",headers:u}).then((function(t){return t.json()})).then((function(t){return t}))}({csrfToken:this.csrfToken,endpoint:this.endpoint,key:this.key,number:t+1,uploadId:this.uploadId}).then((function(t){if(!("object"===ca(t)&&t&&"string"==typeof t.url))throw new TypeError("AwsS3/Multipart: Got incorrect result from `prepareUploadPart()`, expected an object `{ url }`.");return t})).then((function(n){var r=n.url;e.uploadPartBytes(t,r)}),(function(t){e.handleError(t)})):Promise.resolve()}},{key:"onPartProgress",value:function(t,e){if(this.chunkState[t].uploaded=e,this.onProgress){var n=this.chunkState.reduce((function(t,e){return t+e.uploaded}),0);this.onProgress(n,this.file.size)}}},{key:"onPartComplete",value:function(t,e){this.chunkState[t].etag=e,this.chunkState[t].done=!0;var n={PartNumber:t+1,ETag:e};this.parts.push(n),this.uploadParts()}},{key:"uploadPartBytes",value:function(t,e){var n=this,r=this.chunks[t],o=new XMLHttpRequest;o.open("PUT",e,!0),o.responseType="text",this.uploading.push(o),o.upload.addEventListener("progress",(function(e){e.lengthComputable&&n.onPartProgress(t,e.loaded)})),o.addEventListener("abort",(function(e){nc(n.uploading,e.target),n.chunkState[t].busy=!1})),o.addEventListener("load",(function(e){var o=e.target;if(nc(n.uploading,o),n.chunkState[t].busy=!1,o.status<200||o.status>=300)n.handleError(new Error("Non 2xx"));else{n.onPartProgress(t,r.size);var i=o.getResponseHeader("ETag");null!==i?n.onPartComplete(t,i):n.handleError(new Error("AwsS3/Multipart: Could not read the ETag header. This likely means CORS is not configured correctly on the S3 Bucket. See https://uppy.io/docs/aws-s3-multipart#S3-Bucket-Configuration for instructions."))}})),o.addEventListener("error",(function(e){nc(n.uploading,e.target),n.chunkState[t].busy=!1;var r=new Error("Unknown error");n.handleError(r)})),o.send(r)}},{key:"completeUpload",value:function(){var t=this;return this.parts.sort((function(t,e){return t.PartNumber-e.PartNumber})),this.uploadId&&this.key?function(t){var e=t.csrfToken,n=t.key,r=t.uploadId,o=t.parts,i=t.endpoint,a=encodeURIComponent(n),u=encodeURIComponent(r),s=new Headers({"X-CSRFToken":e}),c=tc(i,u,"complete","?key=".concat(a));return fetch(c,{method:"post",headers:s,body:JSON.stringify({parts:o})}).then((function(t){return t.json()})).then((function(t){return t}))}({csrfToken:this.csrfToken,endpoint:this.endpoint,key:this.key,uploadId:this.uploadId,parts:this.parts}).then((function(){t.onSuccess&&t.onSuccess()}),(function(e){t.handleError(e)})):Promise.resolve()}},{key:"handleError",value:function(t){if(!this.onError)throw t;this.onError(t)}}]),o}(Zs),ic=function(){var t=zi(ga.mark((function t(e,n){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.abrupt("return",new Promise((function(t,r){var o=new XMLHttpRequest;o.open("DELETE",e),o.onload=function(){204===o.status?t():r()},o.setRequestHeader("Tus-Resumable","1.0.0"),o.setRequestHeader("X-CSRFToken",n),o.send(null)})));case 1:case"end":return t.stop()}}),t)})));return function(e,n){return t.apply(this,arguments)}}();function ac(t){var e=function(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Boolean.prototype.valueOf.call(Reflect.construct(Boolean,[],(function(){}))),!0}catch(t){return!1}}();return function(){var n,r=Xs(t);if(e){var o=Xs(this).constructor;n=Reflect.construct(r,arguments,o)}else n=r.apply(this,arguments);return Ws(this,n)}}var uc,sc=function(t){Ks(o,t);var e,n,r=ac(o);function o(t){var e,n=t.name,i=t.size,a=t.type,u=t.uploadIndex;return qi(this,o),Dr(Gs(e=r.call(this,{name:n,status:"done",type:a,uploadIndex:u})),"size",void 0),e.size=i,e}return $i(o,[{key:"abort",value:(n=zi(ga.mark((function t(){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.abrupt("return",Promise.resolve());case 1:case"end":return t.stop()}}),t)}))),function(){return n.apply(this,arguments)})},{key:"delete",value:(e=zi(ga.mark((function t(){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.abrupt("return",Promise.resolve());case 1:case"end":return t.stop()}}),t)}))),function(){return e.apply(this,arguments)})},{key:"getSize",value:function(){return this.size}}]),o}(Zs),cc=function(t){Ks(n,t);var e=ac(n);function n(t,r){var o;return qi(this,n),Dr(Gs(o=e.call(this,{name:t.name,size:t.size,type:"placeholder",uploadIndex:r})),"id",void 0),o.id=t.id,o}return $i(n,[{key:"getInitialFile",value:function(){return{id:this.id,name:this.name,size:this.size,type:"placeholder"}}}]),n}(sc),lc=function(t){Ks(n,t);var e=ac(n);function n(t,r){var o;return qi(this,n),Dr(Gs(o=e.call(this,{name:t.original_name||t.name,size:t.size,type:"uploadedS3",uploadIndex:r})),"id",void 0),Dr(Gs(o),"key",void 0),o.id=t.id,o.key=t.name,o}return $i(n,[{key:"getInitialFile",value:function(){return{id:this.id,name:this.key,original_name:this.name,size:this.size,type:"s3"}}}]),n}(sc),fc=function(t){Ks(n,t);var e=ac(n);function n(t,r){return qi(this,n),e.call(this,{name:t.name,size:t.size,type:"existing",uploadIndex:r})}return $i(n,[{key:"getInitialFile",value:function(){return{name:this.name,size:this.size,type:"existing"}}}]),n}(sc),pc=function(t){Ks(r,t);var e,n=ac(r);function r(t){var e,o=t.csrfToken,i=t.initialFile,a=t.uploadIndex,u=t.uploadUrl;return qi(this,r),Dr(Gs(e=n.call(this,{name:i.name,size:i.size,type:"uploadedTus",uploadIndex:a})),"csrfToken",void 0),Dr(Gs(e),"id",void 0),Dr(Gs(e),"url",void 0),e.csrfToken=o,e.id=i.id,e.url="".concat(u).concat(i.id),e}return $i(r,[{key:"delete",value:(e=zi(ga.mark((function t(){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,ic(this.url,this.csrfToken);case 2:case"end":return t.stop()}}),t,this)}))),function(){return e.apply(this,arguments)})},{key:"getInitialFile",value:function(){return{id:this.id,name:this.name,size:this.size,type:"tus",url:""}}}]),r}(sc),dc="undefined"!=typeof ArrayBuffer&&"undefined"!=typeof DataView,hc=O.f,vc=r.Int8Array,yc=vc&&vc.prototype,gc=r.Uint8ClampedArray,mc=gc&&gc.prototype,xc=vc&&_e(vc),bc=yc&&_e(yc),wc=Object.prototype,_c=wc.isPrototypeOf,Ec=Vt("toStringTag"),Sc=q("TYPED_ARRAY_TAG"),Ac=dc&&!!Ne&&"Opera"!==ie(r.opera),kc=!1,Rc={Int8Array:1,Uint8Array:1,Uint8ClampedArray:1,Int16Array:2,Uint16Array:2,Int32Array:4,Uint32Array:4,Float32Array:4,Float64Array:8},Oc={BigInt64Array:8,BigUint64Array:8},Tc=function(t){if(!y(t))return!1;var e=ie(t);return x(Rc,e)||x(Oc,e)};for(uc in Rc)r[uc]||(Ac=!1);if((!Ac||"function"!=typeof xc||xc===Function.prototype)&&(xc=function(){throw TypeError("Incorrect invocation")},Ac))for(uc in Rc)r[uc]&&Ne(r[uc],xc);if((!Ac||!bc||bc===wc)&&(bc=xc.prototype,Ac))for(uc in Rc)r[uc]&&Ne(r[uc].prototype,bc);if(Ac&&_e(mc)!==bc&&Ne(mc,bc),i&&!x(bc,Ec))for(uc in kc=!0,hc(bc,Ec,{get:function(){return y(this)?this[Sc]:void 0}}),Rc)r[uc]&&T(r[uc],Sc,uc);var jc={NATIVE_ARRAY_BUFFER_VIEWS:Ac,TYPED_ARRAY_TAG:kc&&Sc,aTypedArray:function(t){if(Tc(t))return t;throw TypeError("Target is not a typed array")},aTypedArrayConstructor:function(t){if(Ne){if(_c.call(xc,t))return t}else for(var e in Rc)if(x(Rc,uc)){var n=r[e];if(n&&(t===n||_c.call(n,t)))return t}throw TypeError("Target is not a typed array constructor")},exportTypedArrayMethod:function(t,e,n){if(i){if(n)for(var o in Rc){var a=r[o];a&&x(a.prototype,t)&&delete a.prototype[t]}bc[t]&&!n||nt(bc,t,n?e:Ac&&yc[t]||e)}},exportTypedArrayStaticMethod:function(t,e,n){var o,a;if(i){if(Ne){if(n)for(o in Rc)(a=r[o])&&x(a,t)&&delete a[t];if(xc[t]&&!n)return;try{return nt(xc,t,n?e:Ac&&vc[t]||e)}catch(t){}}for(o in Rc)!(a=r[o])||a[t]&&!n||nt(a,t,e)}},isView:function(t){if(!y(t))return!1;var e=ie(t);return"DataView"===e||x(Rc,e)||x(Oc,e)},isTypedArray:Tc,TypedArray:xc,TypedArrayPrototype:bc},Pc=jc.NATIVE_ARRAY_BUFFER_VIEWS,Cc=r.ArrayBuffer,Ic=r.Int8Array,Uc=!Pc||!o((function(){Ic(1)}))||!o((function(){new Ic(-1)}))||!pe((function(t){new Ic,new Ic(null),new Ic(1.5),new Ic(t)}),!0)||o((function(){return 1!==new Ic(new Cc(2),1,void 0).length})),Lc=function(t){if(void 0===t)return 0;var e=st(t),n=lt(e);if(e!==n)throw RangeError("Wrong length or index");return n},Dc=Math.abs,Hc=Math.pow,Fc=Math.floor,Mc=Math.log,Nc=Math.LN2,Bc=function(t,e,n){var r,o,i,a=new Array(n),u=8*n-e-1,s=(1<<u)-1,c=s>>1,l=23===e?Hc(2,-24)-Hc(2,-77):0,f=t<0||0===t&&1/t<0?1:0,p=0;for((t=Dc(t))!=t||t===1/0?(o=t!=t?1:0,r=s):(r=Fc(Mc(t)/Nc),t*(i=Hc(2,-r))<1&&(r--,i*=2),(t+=r+c>=1?l/i:l*Hc(2,1-c))*i>=2&&(r++,i/=2),r+c>=s?(o=0,r=s):r+c>=1?(o=(t*i-1)*Hc(2,e),r+=c):(o=t*Hc(2,c-1)*Hc(2,e),r=0));e>=8;a[p++]=255&o,o/=256,e-=8);for(r=r<<e|o,u+=e;u>0;a[p++]=255&r,r/=256,u-=8);return a[--p]|=128*f,a},zc=function(t,e){var n,r=t.length,o=8*r-e-1,i=(1<<o)-1,a=i>>1,u=o-7,s=r-1,c=t[s--],l=127&c;for(c>>=7;u>0;l=256*l+t[s],s--,u-=8);for(n=l&(1<<-u)-1,l>>=-u,u+=e;u>0;n=256*n+t[s],s--,u-=8);if(0===l)l=1-a;else{if(l===i)return n?NaN:c?-1/0:1/0;n+=Hc(2,e),l-=a}return(c?-1:1)*n*Hc(2,l-e)},qc=function(t){for(var e=Lt(this),n=lt(e.length),r=arguments.length,o=dt(r>1?arguments[1]:void 0,n),i=r>2?arguments[2]:void 0,a=void 0===i?n:dt(i,n);a>o;)e[o++]=t;return e},Gc=bt.f,$c=O.f,Kc=et.get,Wc=et.set,Xc="ArrayBuffer",Vc="DataView",Yc="Wrong index",Qc=r.ArrayBuffer,Jc=Qc,Zc=r.DataView,tl=Zc&&Zc.prototype,el=Object.prototype,nl=r.RangeError,rl=Bc,ol=zc,il=function(t){return[255&t]},al=function(t){return[255&t,t>>8&255]},ul=function(t){return[255&t,t>>8&255,t>>16&255,t>>24&255]},sl=function(t){return t[3]<<24|t[2]<<16|t[1]<<8|t[0]},cl=function(t){return rl(t,23,4)},ll=function(t){return rl(t,52,8)},fl=function(t,e){$c(t.prototype,e,{get:function(){return Kc(this)[e]}})},pl=function(t,e,n,r){var o=Lc(n),i=Kc(t);if(o+e>i.byteLength)throw nl(Yc);var a=Kc(i.buffer).bytes,u=o+i.byteOffset,s=a.slice(u,u+e);return r?s:s.reverse()},dl=function(t,e,n,r,o,i){var a=Lc(n),u=Kc(t);if(a+e>u.byteLength)throw nl(Yc);for(var s=Kc(u.buffer).bytes,c=a+u.byteOffset,l=r(+o),f=0;f<e;f++)s[c+f]=l[i?f:e-f-1]};if(dc){if(!o((function(){Qc(1)}))||!o((function(){new Qc(-1)}))||o((function(){return new Qc,new Qc(1.5),new Qc(NaN),Qc.name!=Xc}))){for(var hl,vl=(Jc=function(t){return bn(this,Jc),new Qc(Lc(t))}).prototype=Qc.prototype,yl=Gc(Qc),gl=0;yl.length>gl;)(hl=yl[gl++])in Jc||T(Jc,hl,Qc[hl]);vl.constructor=Jc}Ne&&_e(tl)!==el&&Ne(tl,el);var ml=new Zc(new Jc(2)),xl=tl.setInt8;ml.setInt8(0,2147483648),ml.setInt8(1,2147483649),!ml.getInt8(0)&&ml.getInt8(1)||En(tl,{setInt8:function(t,e){xl.call(this,t,e<<24>>24)},setUint8:function(t,e){xl.call(this,t,e<<24>>24)}},{unsafe:!0})}else Jc=function(t){bn(this,Jc,Xc);var e=Lc(t);Wc(this,{bytes:qc.call(new Array(e),0),byteLength:e}),i||(this.byteLength=e)},Zc=function(t,e,n){bn(this,Zc,Vc),bn(t,Jc,Vc);var r=Kc(t).byteLength,o=st(e);if(o<0||o>r)throw nl("Wrong offset");if(o+(n=void 0===n?r-o:lt(n))>r)throw nl("Wrong length");Wc(this,{buffer:t,byteLength:n,byteOffset:o}),i||(this.buffer=t,this.byteLength=n,this.byteOffset=o)},i&&(fl(Jc,"byteLength"),fl(Zc,"buffer"),fl(Zc,"byteLength"),fl(Zc,"byteOffset")),En(Zc.prototype,{getInt8:function(t){return pl(this,1,t)[0]<<24>>24},getUint8:function(t){return pl(this,1,t)[0]},getInt16:function(t){var e=pl(this,2,t,arguments.length>1?arguments[1]:void 0);return(e[1]<<8|e[0])<<16>>16},getUint16:function(t){var e=pl(this,2,t,arguments.length>1?arguments[1]:void 0);return e[1]<<8|e[0]},getInt32:function(t){return sl(pl(this,4,t,arguments.length>1?arguments[1]:void 0))},getUint32:function(t){return sl(pl(this,4,t,arguments.length>1?arguments[1]:void 0))>>>0},getFloat32:function(t){return ol(pl(this,4,t,arguments.length>1?arguments[1]:void 0),23)},getFloat64:function(t){return ol(pl(this,8,t,arguments.length>1?arguments[1]:void 0),52)},setInt8:function(t,e){dl(this,1,t,il,e)},setUint8:function(t,e){dl(this,1,t,il,e)},setInt16:function(t,e){dl(this,2,t,al,e,arguments.length>2?arguments[2]:void 0)},setUint16:function(t,e){dl(this,2,t,al,e,arguments.length>2?arguments[2]:void 0)},setInt32:function(t,e){dl(this,4,t,ul,e,arguments.length>2?arguments[2]:void 0)},setUint32:function(t,e){dl(this,4,t,ul,e,arguments.length>2?arguments[2]:void 0)},setFloat32:function(t,e){dl(this,4,t,cl,e,arguments.length>2?arguments[2]:void 0)},setFloat64:function(t,e){dl(this,8,t,ll,e,arguments.length>2?arguments[2]:void 0)}});He(Jc,Xc),He(Zc,Vc);var bl={ArrayBuffer:Jc,DataView:Zc},wl=function(t,e){var n=function(t){var e=st(t);if(e<0)throw RangeError("The argument can't be less than 0");return e}(t);if(n%e)throw RangeError("Wrong offset");return n},_l=jc.aTypedArrayConstructor,El=function(t){var e,n,r,o,i,a,u=Lt(t),s=arguments.length,c=s>1?arguments[1]:void 0,l=void 0!==c,f=ue(u);if(null!=f&&!Zt(f))for(a=(i=f.call(u)).next,u=[];!(o=a.call(i)).done;)u.push(o.value);for(l&&s>2&&(c=Ut(c,arguments[2],2)),n=lt(u.length),r=new(_l(this))(n),e=0;n>e;e++)r[e]=l?c(u[e],e):u[e];return r};e((function(t){var e=bt.f,n=Dn.forEach,o=et.get,a=et.set,u=O.f,s=A.f,l=Math.round,f=r.RangeError,p=bl.ArrayBuffer,d=bl.DataView,h=jc.NATIVE_ARRAY_BUFFER_VIEWS,v=jc.TYPED_ARRAY_TAG,m=jc.TypedArray,b=jc.TypedArrayPrototype,w=jc.aTypedArrayConstructor,_=jc.isTypedArray,E="BYTES_PER_ELEMENT",S="Wrong length",k=function(t,e){for(var n=0,r=e.length,o=new(w(t))(r);r>n;)o[n]=e[n++];return o},R=function(t,e){u(t,e,{get:function(){return o(this)[e]}})},j=function(t){var e;return t instanceof p||"ArrayBuffer"==(e=ie(t))||"SharedArrayBuffer"==e},P=function(t,e){return _(t)&&"symbol"!=typeof e&&e in t&&String(+e)==String(e)},C=function(t,e){return P(t,e=g(e,!0))?c(2,t[e]):s(t,e)},I=function(t,e,n){return!(P(t,e=g(e,!0))&&y(n)&&x(n,"value"))||x(n,"get")||x(n,"set")||n.configurable||x(n,"writable")&&!n.writable||x(n,"enumerable")&&!n.enumerable?u(t,e,n):(t[e]=n.value,t)};i?(h||(A.f=C,O.f=I,R(b,"buffer"),R(b,"byteOffset"),R(b,"byteLength"),R(b,"length")),Ct({target:"Object",stat:!0,forced:!h},{getOwnPropertyDescriptor:C,defineProperty:I}),t.exports=function(t,i,s){var c=t.match(/\d+$/)[0]/8,g=t+(s?"Clamped":"")+"Array",x="get"+t,w="set"+t,A=r[g],R=A,O=R&&R.prototype,P={},C=function(t,e){u(t,e,{get:function(){return function(t,e){var n=o(t);return n.view[x](e*c+n.byteOffset,!0)}(this,e)},set:function(t){return function(t,e,n){var r=o(t);s&&(n=(n=l(n))<0?0:n>255?255:255&n),r.view[w](e*c+r.byteOffset,n,!0)}(this,e,t)},enumerable:!0})};h?Uc&&(R=i((function(t,e,n,r){return bn(t,R,g),wn(y(e)?j(e)?void 0!==r?new A(e,wl(n,c),r):void 0!==n?new A(e,wl(n,c)):new A(e):_(e)?k(R,e):El.call(R,e):new A(Lc(e)),t,R)})),Ne&&Ne(R,m),n(e(A),(function(t){t in R||T(R,t,A[t])})),R.prototype=O):(R=i((function(t,e,n,r){bn(t,R,g);var o,i,u,s=0,l=0;if(y(e)){if(!j(e))return _(e)?k(R,e):El.call(R,e);o=e,l=wl(n,c);var h=e.byteLength;if(void 0===r){if(h%c)throw f(S);if((i=h-l)<0)throw f(S)}else if((i=lt(r)*c)+l>h)throw f(S);u=i/c}else u=Lc(e),o=new p(i=u*c);for(a(t,{buffer:o,byteOffset:l,byteLength:i,length:u,view:new d(o)});s<u;)C(t,s++)})),Ne&&Ne(R,m),O=R.prototype=Ue(b)),O.constructor!==R&&T(O,"constructor",R),v&&T(O,v,g),P[g]=R,Ct({global:!0,forced:R!=A,sham:!h},P),E in R||T(R,E,c),E in O||T(O,E,c),An(g)}):t.exports=function(){}}))("Uint8",(function(t){return function(e,n,r){return t(this,e,n,r)}}));var Sl=Math.min,Al=[].copyWithin||function(t,e){var n=Lt(this),r=lt(n.length),o=dt(t,r),i=dt(e,r),a=arguments.length>2?arguments[2]:void 0,u=Sl((void 0===a?r:dt(a,r))-i,r-o),s=1;for(i<o&&o<i+u&&(s=-1,i+=u-1,o+=u-1);u-- >0;)i in n?n[o]=n[i]:delete n[o],o+=s,i+=s;return n},kl=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("copyWithin",(function(t,e){return Al.call(kl(this),t,e,arguments.length>2?arguments[2]:void 0)}));var Rl=Dn.every,Ol=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("every",(function(t){return Rl(Ol(this),t,arguments.length>1?arguments[1]:void 0)}));var Tl=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("fill",(function(t){return qc.apply(Tl(this),arguments)}));var jl=jc.aTypedArrayConstructor,Pl=function(t,e){for(var n=ko(t,t.constructor),r=0,o=e.length,i=new(jl(n))(o);o>r;)i[r]=e[r++];return i},Cl=Dn.filter,Il=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("filter",(function(t){var e=Cl(Il(this),t,arguments.length>1?arguments[1]:void 0);return Pl(this,e)}));var Ul=Dn.find,Ll=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("find",(function(t){return Ul(Ll(this),t,arguments.length>1?arguments[1]:void 0)}));var Dl=Dn.findIndex,Hl=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("findIndex",(function(t){return Dl(Hl(this),t,arguments.length>1?arguments[1]:void 0)}));var Fl=Dn.forEach,Ml=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("forEach",(function(t){Fl(Ml(this),t,arguments.length>1?arguments[1]:void 0)}));var Nl=vt.includes,Bl=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("includes",(function(t){return Nl(Bl(this),t,arguments.length>1?arguments[1]:void 0)}));var zl=vt.indexOf,ql=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("indexOf",(function(t){return zl(ql(this),t,arguments.length>1?arguments[1]:void 0)}));var Gl=Vt("iterator"),$l=r.Uint8Array,Kl=un.values,Wl=un.keys,Xl=un.entries,Vl=jc.aTypedArray,Yl=jc.exportTypedArrayMethod,Ql=$l&&$l.prototype[Gl],Jl=!!Ql&&("values"==Ql.name||null==Ql.name),Zl=function(){return Kl.call(Vl(this))};Yl("entries",(function(){return Xl.call(Vl(this))})),Yl("keys",(function(){return Wl.call(Vl(this))})),Yl("values",Zl,!Jl),Yl(Gl,Zl,!Jl);var tf=jc.aTypedArray,ef=[].join;(0,jc.exportTypedArrayMethod)("join",(function(t){return ef.apply(tf(this),arguments)}));var nf=Math.min,rf=[].lastIndexOf,of=!!rf&&1/[1].lastIndexOf(1,-0)<0,af=Bn("lastIndexOf"),uf=of||!af?function(t){if(of)return rf.apply(this,arguments)||0;var e=v(this),n=lt(e.length),r=n-1;for(arguments.length>1&&(r=nf(r,st(arguments[1]))),r<0&&(r=n+r);r>=0;r--)if(r in e&&e[r]===t)return r||0;return-1}:rf,sf=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("lastIndexOf",(function(t){return uf.apply(sf(this),arguments)}));var cf=Dn.map,lf=jc.aTypedArray,ff=jc.aTypedArrayConstructor;(0,jc.exportTypedArrayMethod)("map",(function(t){return cf(lf(this),t,arguments.length>1?arguments[1]:void 0,(function(t,e){return new(ff(ko(t,t.constructor)))(e)}))}));var pf=Ys.left,df=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("reduce",(function(t){return pf(df(this),t,arguments.length,arguments.length>1?arguments[1]:void 0)}));var hf=Ys.right,vf=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("reduceRight",(function(t){return hf(vf(this),t,arguments.length,arguments.length>1?arguments[1]:void 0)}));var yf=jc.aTypedArray,gf=jc.exportTypedArrayMethod,mf=Math.floor;gf("reverse",(function(){for(var t,e=this,n=yf(e).length,r=mf(n/2),o=0;o<r;)t=e[o],e[o++]=e[--n],e[n]=t;return e}));var xf=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("set",(function(t){xf(this);var e=wl(arguments.length>1?arguments[1]:void 0,1),n=this.length,r=Lt(t),o=lt(r.length),i=0;if(o+e>n)throw RangeError("Wrong length");for(;i<o;)this[e+i]=r[i++]}),o((function(){new Int8Array(1).set({})})));var bf=jc.aTypedArray,wf=jc.aTypedArrayConstructor,_f=[].slice;(0,jc.exportTypedArrayMethod)("slice",(function(t,e){for(var n=_f.call(bf(this),t,e),r=ko(this,this.constructor),o=0,i=n.length,a=new(wf(r))(i);i>o;)a[o]=n[o++];return a}),o((function(){new Int8Array(1).slice()})));var Ef=Dn.some,Sf=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("some",(function(t){return Ef(Sf(this),t,arguments.length>1?arguments[1]:void 0)}));var Af=jc.aTypedArray,kf=[].sort;(0,jc.exportTypedArrayMethod)("sort",(function(t){return kf.call(Af(this),t)}));var Rf=jc.aTypedArray;(0,jc.exportTypedArrayMethod)("subarray",(function(t,e){var n=Rf(this),r=n.length,o=dt(t,r);return new(ko(n,n.constructor))(n.buffer,n.byteOffset+o*n.BYTES_PER_ELEMENT,lt((void 0===e?r:dt(e,r))-o))}));var Of=r.Int8Array,Tf=jc.aTypedArray,jf=jc.exportTypedArrayMethod,Pf=[].toLocaleString,Cf=[].slice,If=!!Of&&o((function(){Pf.call(new Of(1))}));jf("toLocaleString",(function(){return Pf.apply(If?Cf.call(Tf(this)):Tf(this),arguments)}),o((function(){return[1,2].toLocaleString()!=new Of([1,2]).toLocaleString()}))||!o((function(){Of.prototype.toLocaleString.call([1,2])})));var Uf=jc.exportTypedArrayMethod,Lf=r.Uint8Array,Df=Lf&&Lf.prototype||{},Hf=[].toString,Ff=[].join;o((function(){Hf.call({})}))&&(Hf=function(){return Ff.call(this)});var Mf=Df.toString!=Hf;Uf("toString",Hf,Mf),(0,jc.exportTypedArrayStaticMethod)("from",El,Uc);var Nf=e((function(e,n){!function(t,n){e.exports=function(t){var n,r=(t=t||{}).Base64,o="2.6.4",i="ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/",a=function(t){for(var e={},n=0,r=t.length;n<r;n++)e[t.charAt(n)]=n;return e}(i),u=String.fromCharCode,s=function(t){if(t.length<2)return(e=t.charCodeAt(0))<128?t:e<2048?u(192|e>>>6)+u(128|63&e):u(224|e>>>12&15)+u(128|e>>>6&63)+u(128|63&e);var e=65536+1024*(t.charCodeAt(0)-55296)+(t.charCodeAt(1)-56320);return u(240|e>>>18&7)+u(128|e>>>12&63)+u(128|e>>>6&63)+u(128|63&e)},c=/[\uD800-\uDBFF][\uDC00-\uDFFFF]|[^\x00-\x7F]/g,l=function(t){return t.replace(c,s)},f=function(t){var e=[0,2,1][t.length%3],n=t.charCodeAt(0)<<16|(t.length>1?t.charCodeAt(1):0)<<8|(t.length>2?t.charCodeAt(2):0);return[i.charAt(n>>>18),i.charAt(n>>>12&63),e>=2?"=":i.charAt(n>>>6&63),e>=1?"=":i.charAt(63&n)].join("")},p=t.btoa&&"function"==typeof t.btoa?function(e){return t.btoa(e)}:function(t){if(t.match(/[^\x00-\xFF]/))throw new RangeError("The string contains invalid characters.");return t.replace(/[\s\S]{1,3}/g,f)},d=function(t){return p(l(String(t)))},h=function(t){return t.replace(/[+\/]/g,(function(t){return"+"==t?"-":"_"})).replace(/=/g,"")},v=function(t,e){return e?h(d(t)):d(t)},y=function(t){return v(t,!0)};t.Uint8Array&&(n=function(t,e){for(var n="",r=0,o=t.length;r<o;r+=3){var a=t[r],u=t[r+1],s=t[r+2],c=a<<16|u<<8|s;n+=i.charAt(c>>>18)+i.charAt(c>>>12&63)+(void 0!==u?i.charAt(c>>>6&63):"=")+(void 0!==s?i.charAt(63&c):"=")}return e?h(n):n});var g,m=/[\xC0-\xDF][\x80-\xBF]|[\xE0-\xEF][\x80-\xBF]{2}|[\xF0-\xF7][\x80-\xBF]{3}/g,x=function(t){switch(t.length){case 4:var e=((7&t.charCodeAt(0))<<18|(63&t.charCodeAt(1))<<12|(63&t.charCodeAt(2))<<6|63&t.charCodeAt(3))-65536;return u(55296+(e>>>10))+u(56320+(1023&e));case 3:return u((15&t.charCodeAt(0))<<12|(63&t.charCodeAt(1))<<6|63&t.charCodeAt(2));default:return u((31&t.charCodeAt(0))<<6|63&t.charCodeAt(1))}},b=function(t){return t.replace(m,x)},w=function(t){var e=t.length,n=e%4,r=(e>0?a[t.charAt(0)]<<18:0)|(e>1?a[t.charAt(1)]<<12:0)|(e>2?a[t.charAt(2)]<<6:0)|(e>3?a[t.charAt(3)]:0),o=[u(r>>>16),u(r>>>8&255),u(255&r)];return o.length-=[0,0,2,1][n],o.join("")},_=t.atob&&"function"==typeof t.atob?function(e){return t.atob(e)}:function(t){return t.replace(/\S{1,4}/g,w)},E=function(t){return _(String(t).replace(/[^A-Za-z0-9\+\/]/g,""))},S=function(t){return b(_(t))},A=function(t){return String(t).replace(/[-_]/g,(function(t){return"-"==t?"+":"/"})).replace(/[^A-Za-z0-9\+\/]/g,"")},k=function(t){return S(A(t))};t.Uint8Array&&(g=function(t){return Uint8Array.from(E(A(t)),(function(t){return t.charCodeAt(0)}))});var R=function(){var e=t.Base64;return t.Base64=r,e};if(t.Base64={VERSION:o,atob:E,btoa:p,fromBase64:k,toBase64:v,utob:l,encode:v,encodeURI:y,btou:b,decode:k,noConflict:R,fromUint8Array:n,toUint8Array:g},"function"==typeof Object.defineProperty){var O=function(t){return{value:t,enumerable:!1,writable:!0,configurable:!0}};t.Base64.extendString=function(){Object.defineProperty(String.prototype,"fromBase64",O((function(){return k(this)}))),Object.defineProperty(String.prototype,"toBase64",O((function(t){return v(this,t)}))),Object.defineProperty(String.prototype,"toBase64URI",O((function(){return v(this,!0)})))}}t.Meteor&&(Base64=t.Base64);e.exports&&(e.exports.Base64=t.Base64);return{Base64:t.Base64}}(t)}("undefined"!=typeof self?self:"undefined"!=typeof window?window:t)})),Bf=eu.start,zf=nu("trimStart"),qf=zf?function(){return Bf(this)}:"".trimStart;Ct({target:"String",proto:!0,forced:zf},{trimStart:qf,trimLeft:qf});var Gf=function(t,e){if(e=e.split(":")[0],!(t=+t))return!1;switch(e){case"http":case"ws":return 80!==t;case"https":case"wss":return 443!==t;case"ftp":return 21!==t;case"gopher":return 70!==t;case"file":return!1}return 0!==t},$f=Object.prototype.hasOwnProperty;function Kf(t){try{return decodeURIComponent(t.replace(/\+/g," "))}catch(t){return null}}function Wf(t){try{return encodeURIComponent(t)}catch(t){return null}}var Xf={stringify:function(t,e){e=e||"";var n,r,o=[];for(r in"string"!=typeof e&&(e="?"),t)if($f.call(t,r)){if((n=t[r])||null!=n&&!isNaN(n)||(n=""),r=Wf(r),n=Wf(n),null===r||null===n)continue;o.push(r+"="+n)}return o.length?e+o.join("&"):""},parse:function(t){for(var e,n=/([^=?#&]+)=?([^&]*)/g,r={};e=n.exec(t);){var o=Kf(e[1]),i=Kf(e[2]);null===o||null===i||o in r||(r[o]=i)}return r}},Vf=/^[A-Za-z][A-Za-z0-9+-.]*:[\\/]+/,Yf=/^([a-z][a-z0-9.+-]*:)?([\\/]{1,})?([\S\s]*)/i,Qf=new RegExp("^[\\x09\\x0A\\x0B\\x0C\\x0D\\x20\\xA0\\u1680\\u180E\\u2000\\u2001\\u2002\\u2003\\u2004\\u2005\\u2006\\u2007\\u2008\\u2009\\u200A\\u202F\\u205F\\u3000\\u2028\\u2029\\uFEFF]+");function Jf(t){return(t||"").toString().replace(Qf,"")}var Zf=[["#","hash"],["?","query"],function(t){return t.replace("\\","/")},["/","pathname"],["@","auth",1],[NaN,"host",void 0,1,1],[/:(\d+)$/,"port",void 0,1],[NaN,"hostname",void 0,1,1]],tp={hash:1,query:1};function ep(e){var n,r=("undefined"!=typeof window?window:void 0!==t?t:"undefined"!=typeof self?self:{}).location||{},o={},i=ca(e=e||r);if("blob:"===e.protocol)o=new rp(unescape(e.pathname),{});else if("string"===i)for(n in o=new rp(e,{}),tp)delete o[n];else if("object"===i){for(n in e)n in tp||(o[n]=e[n]);void 0===o.slashes&&(o.slashes=Vf.test(e.href))}return o}function np(t){t=Jf(t);var e=Yf.exec(t);return{protocol:e[1]?e[1].toLowerCase():"",slashes:!!(e[2]&&e[2].length>=2),rest:e[2]&&1===e[2].length?"/"+e[3]:e[3]}}function rp(t,e,n){if(t=Jf(t),!(this instanceof rp))return new rp(t,e,n);var r,o,i,a,u,s,c=Zf.slice(),l=ca(e),f=this,p=0;for("object"!==l&&"string"!==l&&(n=e,e=null),n&&"function"!=typeof n&&(n=Xf.parse),e=ep(e),r=!(o=np(t||"")).protocol&&!o.slashes,f.slashes=o.slashes||r&&e.slashes,f.protocol=o.protocol||e.protocol||"",t=o.rest,o.slashes||(c[3]=[/(.*)/,"pathname"]);p<c.length;p++)"function"!=typeof(a=c[p])?(i=a[0],s=a[1],i!=i?f[s]=t:"string"==typeof i?~(u=t.indexOf(i))&&("number"==typeof a[2]?(f[s]=t.slice(0,u),t=t.slice(u+a[2])):(f[s]=t.slice(u),t=t.slice(0,u))):(u=i.exec(t))&&(f[s]=u[1],t=t.slice(0,u.index)),f[s]=f[s]||r&&a[3]&&e[s]||"",a[4]&&(f[s]=f[s].toLowerCase())):t=a(t);n&&(f.query=n(f.query)),r&&e.slashes&&"/"!==f.pathname.charAt(0)&&(""!==f.pathname||""!==e.pathname)&&(f.pathname=function(t,e){if(""===t)return e;for(var n=(e||"/").split("/").slice(0,-1).concat(t.split("/")),r=n.length,o=n[r-1],i=!1,a=0;r--;)"."===n[r]?n.splice(r,1):".."===n[r]?(n.splice(r,1),a++):a&&(0===r&&(i=!0),n.splice(r,1),a--);return i&&n.unshift(""),"."!==o&&".."!==o||n.push(""),n.join("/")}(f.pathname,e.pathname)),"/"!==f.pathname.charAt(0)&&f.hostname&&(f.pathname="/"+f.pathname),Gf(f.port,f.protocol)||(f.host=f.hostname,f.port=""),f.username=f.password="",f.auth&&(a=f.auth.split(":"),f.username=a[0]||"",f.password=a[1]||""),f.origin=f.protocol&&f.host&&"file:"!==f.protocol?f.protocol+"//"+f.host:"null",f.href=f.toString()}rp.prototype={set:function(t,e,n){var r=this;switch(t){case"query":"string"==typeof e&&e.length&&(e=(n||Xf.parse)(e)),r[t]=e;break;case"port":r[t]=e,Gf(e,r.protocol)?e&&(r.host=r.hostname+":"+e):(r.host=r.hostname,r[t]="");break;case"hostname":r[t]=e,r.port&&(e+=":"+r.port),r.host=e;break;case"host":r[t]=e,/:\d+$/.test(e)?(e=e.split(":"),r.port=e.pop(),r.hostname=e.join(":")):(r.hostname=e,r.port="");break;case"protocol":r.protocol=e.toLowerCase(),r.slashes=!n;break;case"pathname":case"hash":if(e){var o="pathname"===t?"/":"#";r[t]=e.charAt(0)!==o?o+e:e}else r[t]=e;break;default:r[t]=e}for(var i=0;i<Zf.length;i++){var a=Zf[i];a[4]&&(r[a[1]]=r[a[1]].toLowerCase())}return r.origin=r.protocol&&r.host&&"file:"!==r.protocol?r.protocol+"//"+r.host:"null",r.href=r.toString(),r},toString:function(t){t&&"function"==typeof t||(t=Xf.stringify);var e,n=this,r=n.protocol;r&&":"!==r.charAt(r.length-1)&&(r+=":");var o=r+(n.slashes?"//":"");return n.username&&(o+=n.username,n.password&&(o+=":"+n.password),o+="@"),o+=n.host+n.pathname,(e="object"===ca(n.query)?t(n.query):n.query)&&(o+="?"!==e.charAt(0)?"?"+e:e),n.hash&&(o+=n.hash),o}},rp.extractProtocol=np,rp.location=ep,rp.trimLeft=Jf,rp.qs=Xf;var op=rp;function ip(t){return(ip="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}function ap(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}function up(t,e){return!e||"object"!==ip(e)&&"function"!=typeof e?function(t){if(void 0===t)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return t}(t):e}function sp(t){var e="function"==typeof Map?new Map:void 0;return(sp=function(t){if(null===t||(n=t,-1===Function.toString.call(n).indexOf("[native code]")))return t;var n;if("function"!=typeof t)throw new TypeError("Super expression must either be null or a function");if(void 0!==e){if(e.has(t))return e.get(t);e.set(t,r)}function r(){return cp(t,arguments,pp(this).constructor)}return r.prototype=Object.create(t.prototype,{constructor:{value:r,enumerable:!1,writable:!0,configurable:!0}}),fp(r,t)})(t)}function cp(t,e,n){return(cp=lp()?Reflect.construct:function(t,e,n){var r=[null];r.push.apply(r,e);var o=new(Function.bind.apply(t,r));return n&&fp(o,n.prototype),o}).apply(null,arguments)}function lp(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],(function(){}))),!0}catch(t){return!1}}function fp(t,e){return(fp=Object.setPrototypeOf||function(t,e){return t.__proto__=e,t})(t,e)}function pp(t){return(pp=Object.setPrototypeOf?Object.getPrototypeOf:function(t){return t.__proto__||Object.getPrototypeOf(t)})(t)}_n("Map",(function(t){return function(){return t(this,arguments.length?arguments[0]:void 0)}}),jn);var dp=function(t){!function(t,e){if("function"!=typeof e&&null!==e)throw new TypeError("Super expression must either be null or a function");t.prototype=Object.create(e&&e.prototype,{constructor:{value:t,writable:!0,configurable:!0}}),e&&fp(t,e)}(r,t);var e,n=(e=r,function(){var t,n=pp(e);if(lp()){var r=pp(this).constructor;t=Reflect.construct(n,arguments,r)}else t=n.apply(this,arguments);return up(this,t)});function r(t){var e,o=arguments.length>1&&void 0!==arguments[1]?arguments[1]:null,i=arguments.length>2&&void 0!==arguments[2]?arguments[2]:null,a=arguments.length>3&&void 0!==arguments[3]?arguments[3]:null;if(ap(this,r),(e=n.call(this,t)).originalRequest=i,e.originalResponse=a,e.causingError=o,null!=o&&(t+=", caused by ".concat(o.toString())),null!=i){var u=i.getHeader("X-Request-ID")||"n/a",s=i.getMethod(),c=i.getURL(),l=a?a.getStatus():"n/a",f=a?a.getBody()||"":"n/a";t+=", originated from request (method: ".concat(s,", url: ").concat(c,", response code: ").concat(l,", response text: ").concat(f,", request id: ").concat(u,")")}return e.message=t,e}return r}(sp(Error));function hp(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function vp(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?hp(Object(n),!0).forEach((function(e){yp(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):hp(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function yp(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}function gp(t,e){for(var n=0;n<e.length;n++){var r=e[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(t,r.key,r)}}var mp={endpoint:null,uploadUrl:null,metadata:{},fingerprint:null,uploadSize:null,onProgress:null,onChunkComplete:null,onSuccess:null,onError:null,_onUploadUrlAvailable:null,overridePatchMethod:!1,headers:{},addRequestId:!1,onBeforeRequest:null,onAfterResponse:null,onShouldRetry:null,chunkSize:1/0,retryDelays:[0,1e3,3e3,5e3],parallelUploads:1,storeFingerprintForResuming:!0,removeFingerprintOnSuccess:!1,uploadLengthDeferred:!1,uploadDataDuringCreation:!1,urlStorage:null,fileReader:null,httpStack:null},xp=function(){function t(e,n){!function(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}(this,t),"resume"in n&&console.log("tus: The `resume` option has been removed in tus-js-client v2. Please use the URL storage API instead."),this.options=n,this._urlStorage=this.options.urlStorage,this.file=e,this.url=null,this._req=null,this._fingerprint=null,this._urlStorageKey=null,this._offset=null,this._aborted=!1,this._size=null,this._source=null,this._retryAttempt=0,this._retryTimeout=null,this._offsetBeforeRetry=0,this._parallelUploads=null,this._parallelUploadUrls=null}var e,n,r;return e=t,r=[{key:"terminate",value:function(e){var n=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},r=arguments.length>2?arguments[2]:void 0;if("function"==typeof n||"function"==typeof r)throw new Error("tus: the terminate function does not accept a callback since v2 anymore; please use the returned Promise instead");var o=_p("DELETE",e,n);return Ep(o,null,n).then((function(t){if(204!==t.getStatus())throw new dp("tus: unexpected response while terminating upload",null,o,t)})).catch((function(r){if(r instanceof dp||(r=new dp("tus: failed to terminate upload",r,o,null)),!Sp(r,0,n))throw r;var i=n.retryDelays[0],a=n.retryDelays.slice(1),u=vp({},n,{retryDelays:a});return new Promise((function(t){return setTimeout(t,i)})).then((function(){return t.terminate(e,u)}))}))}}],(n=[{key:"findPreviousUploads",value:function(){var t=this;return this.options.fingerprint(this.file,this.options).then((function(e){return t._urlStorage.findUploadsByFingerprint(e)}))}},{key:"resumeFromPreviousUpload",value:function(t){this.url=t.uploadUrl||null,this._parallelUploadUrls=t.parallelUploadUrls||null,this._urlStorageKey=t.urlStorageKey}},{key:"start",value:function(){var t=this,e=this.file;if(e)if(this.options.endpoint||this.options.uploadUrl){var n=this.options.retryDelays;null==n||"[object Array]"===Object.prototype.toString.call(n)?(this.options.parallelUploads>1&&["uploadUrl","uploadSize","uploadLengthDeferred"].forEach((function(e){t.options[e]&&t._emitError(new Error("tus: cannot use the ".concat(e," option when parallelUploads is enabled")))})),this.options.fingerprint(e,this.options).then((function(n){return t._fingerprint=n,t._source?t._source:t.options.fileReader.openFile(e,t.options.chunkSize)})).then((function(e){t._source=e,t.options.parallelUploads>1||null!=t._parallelUploadUrls?t._startParallelUpload():t._startSingleUpload()})).catch((function(e){t._emitError(e)}))):this._emitError(new Error("tus: the `retryDelays` option must either be an array or null"))}else this._emitError(new Error("tus: neither an endpoint or an upload URL is provided"));else this._emitError(new Error("tus: no file or stream to upload provided"))}},{key:"_startParallelUpload",value:function(){var e=this,n=this._size=this._source.size,r=0;this._parallelUploads=[];var o=null!=this._parallelUploadUrls?this._parallelUploadUrls.length:this.options.parallelUploads,i=function(t,e,n){for(var r=Math.floor(t/e),o=[],i=0;i<e;i++)o.push({start:r*i,end:r*(i+1)});return o[e-1].end=t,n&&o.forEach((function(t,e){t.uploadUrl=n[e]||null})),o}(this._source.size,o,this._parallelUploadUrls);this._parallelUploadUrls=new Array(i.length);var a,u=i.map((function(o,a){var u=0;return e._source.slice(o.start,o.end).then((function(s){var c=s.value;return new Promise((function(s,l){var f=vp({},e.options,{uploadUrl:o.uploadUrl||null,storeFingerprintForResuming:!1,removeFingerprintOnSuccess:!1,parallelUploads:1,metadata:{},headers:vp({},e.options.headers,{"Upload-Concat":"partial"}),onSuccess:s,onError:l,onProgress:function(t){r=r-u+t,u=t,e._emitProgress(r,n)},_onUploadUrlAvailable:function(){e._parallelUploadUrls[a]=p.url,e._parallelUploadUrls.filter((function(t){return!!t})).length===i.length&&e._saveUploadInUrlStorage()}}),p=new t(c,f);p.start(),e._parallelUploads.push(p)}))}))}));Promise.all(u).then((function(){(a=e._openRequest("POST",e.options.endpoint)).setHeader("Upload-Concat","final;".concat(e._parallelUploadUrls.join(" ")));var t=bp(e.options.metadata);return""!==t&&a.setHeader("Upload-Metadata",t),e._sendRequest(a,null)})).then((function(t){if(wp(t.getStatus(),200)){var n=t.getHeader("Location");null!=n?(e.url=Ap(e.options.endpoint,n),"Created upload at ".concat(e.url),e._emitSuccess()):e._emitHttpError(a,t,"tus: invalid or missing Location header")}else e._emitHttpError(a,t,"tus: unexpected response while creating upload")})).catch((function(t){e._emitError(t)}))}},{key:"_startSingleUpload",value:function(){if(this.options.uploadLengthDeferred)this._size=null;else if(null!=this.options.uploadSize){if(this._size=+this.options.uploadSize,isNaN(this._size))return void this._emitError(new Error("tus: cannot convert `uploadSize` option into a number"))}else if(this._size=this._source.size,null==this._size)return void this._emitError(new Error("tus: cannot automatically derive upload's size from input and must be specified manually using the `uploadSize` option"));return this._aborted=!1,null!=this.url?("Resuming upload from previous URL: ".concat(this.url),void this._resumeUpload()):null!=this.options.uploadUrl?("Resuming upload from provided URL: ".concat(this.options.url),this.url=this.options.uploadUrl,void this._resumeUpload()):void this._createUpload()}},{key:"abort",value:function(e,n){var r=this;if("function"==typeof n)throw new Error("tus: the abort function does not accept a callback since v2 anymore; please use the returned Promise instead");return null!=this._parallelUploads&&this._parallelUploads.forEach((function(t){t.abort(e)})),null!==this._req&&(this._req.abort(),this._source.close()),this._aborted=!0,null!=this._retryTimeout&&(clearTimeout(this._retryTimeout),this._retryTimeout=null),e&&null!=this.url?t.terminate(this.url,this.options).then((function(){return r._removeFromUrlStorage()})):Promise.resolve()}},{key:"_emitHttpError",value:function(t,e,n,r){this._emitError(new dp(n,r,t,e))}},{key:"_emitError",value:function(t){var e=this;if(!this._aborted){if(null!=this.options.retryDelays&&(null!=this._offset&&this._offset>this._offsetBeforeRetry&&(this._retryAttempt=0),Sp(t,this._retryAttempt,this.options))){var n=this.options.retryDelays[this._retryAttempt++];return this._offsetBeforeRetry=this._offset,void(this._retryTimeout=setTimeout((function(){e.start()}),n))}if("function"!=typeof this.options.onError)throw t;this.options.onError(t)}}},{key:"_emitSuccess",value:function(){this.options.removeFingerprintOnSuccess&&this._removeFromUrlStorage(),"function"==typeof this.options.onSuccess&&this.options.onSuccess()}},{key:"_emitProgress",value:function(t,e){"function"==typeof this.options.onProgress&&this.options.onProgress(t,e)}},{key:"_emitChunkComplete",value:function(t,e,n){"function"==typeof this.options.onChunkComplete&&this.options.onChunkComplete(t,e,n)}},{key:"_createUpload",value:function(){var t=this;if(this.options.endpoint){var e=this._openRequest("POST",this.options.endpoint);this.options.uploadLengthDeferred?e.setHeader("Upload-Defer-Length",1):e.setHeader("Upload-Length",this._size);var n,r=bp(this.options.metadata);""!==r&&e.setHeader("Upload-Metadata",r),this.options.uploadDataDuringCreation&&!this.options.uploadLengthDeferred?(this._offset=0,n=this._addChunkToRequest(e)):n=this._sendRequest(e,null),n.then((function(n){if(wp(n.getStatus(),200)){var r=n.getHeader("Location");if(null!=r){if(t.url=Ap(t.options.endpoint,r),"Created upload at ".concat(t.url),"function"==typeof t.options._onUploadUrlAvailable&&t.options._onUploadUrlAvailable(),0===t._size)return t._emitSuccess(),void t._source.close();t._saveUploadInUrlStorage(),t.options.uploadDataDuringCreation?t._handleUploadResponse(e,n):(t._offset=0,t._performUpload())}else t._emitHttpError(e,n,"tus: invalid or missing Location header")}else t._emitHttpError(e,n,"tus: unexpected response while creating upload")})).catch((function(n){t._emitHttpError(e,null,"tus: failed to create upload",n)}))}else this._emitError(new Error("tus: unable to create upload because no endpoint is provided"))}},{key:"_resumeUpload",value:function(){var t=this,e=this._openRequest("HEAD",this.url);this._sendRequest(e,null).then((function(n){var r=n.getStatus();if(!wp(r,200))return wp(r,400)&&t._removeFromUrlStorage(),423===r?void t._emitHttpError(e,n,"tus: upload is currently locked; retry later"):t.options.endpoint?(t.url=null,void t._createUpload()):void t._emitHttpError(e,n,"tus: unable to resume upload (new upload cannot be created without an endpoint)");var o=parseInt(n.getHeader("Upload-Offset"),10);if(isNaN(o))t._emitHttpError(e,n,"tus: invalid or missing offset value");else{var i=parseInt(n.getHeader("Upload-Length"),10);if(!isNaN(i)||t.options.uploadLengthDeferred){if("function"==typeof t.options._onUploadUrlAvailable&&t.options._onUploadUrlAvailable(),o===i)return t._emitProgress(i,i),void t._emitSuccess();t._offset=o,t._performUpload()}else t._emitHttpError(e,n,"tus: invalid or missing length value")}})).catch((function(n){t._emitHttpError(e,null,"tus: failed to resume upload",n)}))}},{key:"_performUpload",value:function(){var t,e=this;this._aborted||(this.options.overridePatchMethod?(t=this._openRequest("POST",this.url)).setHeader("X-HTTP-Method-Override","PATCH"):t=this._openRequest("PATCH",this.url),t.setHeader("Upload-Offset",this._offset),this._addChunkToRequest(t).then((function(n){wp(n.getStatus(),200)?e._handleUploadResponse(t,n):e._emitHttpError(t,n,"tus: unexpected response while uploading chunk")})).catch((function(n){e._aborted||e._emitHttpError(t,null,"tus: failed to upload chunk at offset "+e._offset,n)})))}},{key:"_addChunkToRequest",value:function(t){var e=this,n=this._offset,r=this._offset+this.options.chunkSize;return t.setProgressHandler((function(t){e._emitProgress(n+t,e._size)})),t.setHeader("Content-Type","application/offset+octet-stream"),(r===1/0||r>this._size)&&!this.options.uploadLengthDeferred&&(r=this._size),this._source.slice(n,r).then((function(n){var r=n.value,o=n.done;return e.options.uploadLengthDeferred&&o&&(e._size=e._offset+(r&&r.size?r.size:0),t.setHeader("Upload-Length",e._size)),null===r?e._sendRequest(t):(e._emitProgress(e._offset,e._size),e._sendRequest(t,r))}))}},{key:"_handleUploadResponse",value:function(t,e){var n=parseInt(e.getHeader("Upload-Offset"),10);if(isNaN(n))this._emitHttpError(t,e,"tus: invalid or missing offset value");else{if(this._emitProgress(n,this._size),this._emitChunkComplete(n-this._offset,n,this._size),this._offset=n,n==this._size)return this._emitSuccess(),void this._source.close();this._performUpload()}}},{key:"_openRequest",value:function(t,e){var n=_p(t,e,this.options);return this._req=n,n}},{key:"_removeFromUrlStorage",value:function(){var t=this;this._urlStorageKey&&(this._urlStorage.removeUpload(this._urlStorageKey).catch((function(e){t._emitError(e)})),this._urlStorageKey=null)}},{key:"_saveUploadInUrlStorage",value:function(){var t=this;if(this.options.storeFingerprintForResuming&&this._fingerprint){var e={size:this._size,metadata:this.options.metadata,creationTime:(new Date).toString()};this._parallelUploads?e.parallelUploadUrls=this._parallelUploadUrls:e.uploadUrl=this.url,this._urlStorage.addUpload(this._fingerprint,e).then((function(e){return t._urlStorageKey=e})).catch((function(e){t._emitError(e)}))}}},{key:"_sendRequest",value:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:null;return Ep(t,e,this.options)}}])&&gp(e.prototype,n),r&&gp(e,r),t}();function bp(t){var e=[];for(var n in t)e.push(n+" "+Nf.Base64.encode(t[n]));return e.join(",")}function wp(t,e){return t>=e&&t<e+100}function _p(t,e,n){var r=n.httpStack.createRequest(t,e);r.setHeader("Tus-Resumable","1.0.0");var o=n.headers||{};for(var i in o)r.setHeader(i,o[i]);if(n.addRequestId){var a="xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx".replace(/[xy]/g,(function(t){var e=16*Math.random()|0;return("x"==t?e:3&e|8).toString(16)}));r.setHeader("X-Request-ID",a)}return r}function Ep(t,e,n){return("function"==typeof n.onBeforeRequest?Promise.resolve(n.onBeforeRequest(t)):Promise.resolve()).then((function(){return t.send(e).then((function(e){return("function"==typeof n.onAfterResponse?Promise.resolve(n.onAfterResponse(t,e)):Promise.resolve()).then((function(){return e}))}))}))}function Sp(t,e,n){if(null==n.retryDelays||e>=n.retryDelays.length||null==t.originalRequest)return!1;if(n&&"function"==typeof n.onShouldRetry)return n.onShouldRetry(t,e,n);var r,o=t.originalResponse?t.originalResponse.getStatus():0;return(!wp(o,400)||409===o||423===o)&&(r=!0,"undefined"!=typeof window&&"navigator"in window&&!1===window.navigator.onLine&&(r=!1),r)}function Ap(t,e){return new op(e,t).toString()}function kp(t,e){for(var n=0;n<e.length;n++){var r=e[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(t,r.key,r)}}xp.defaultOptions=mp;var Rp=function(){function t(){!function(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}(this,t)}var e,n,r;return e=t,(n=[{key:"listAllUploads",value:function(){return Promise.resolve([])}},{key:"findUploadsByFingerprint",value:function(t){return Promise.resolve([])}},{key:"removeUpload",value:function(t){return Promise.resolve()}},{key:"addUpload",value:function(t,e){return Promise.resolve(null)}}])&&kp(e.prototype,n),r&&kp(e,r),t}();function Op(t,e){for(var n=0;n<e.length;n++){var r=e[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(t,r.key,r)}}var Tp=!1;try{Tp="localStorage"in window;var jp="tusSupport";localStorage.setItem(jp,localStorage.getItem(jp))}catch(t){if(t.code!==t.SECURITY_ERR&&t.code!==t.QUOTA_EXCEEDED_ERR)throw t;Tp=!1}var Pp=Tp,Cp=function(){function t(){!function(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}(this,t)}var e,n,r;return e=t,(n=[{key:"findAllUploads",value:function(){var t=this._findEntries("tus::");return Promise.resolve(t)}},{key:"findUploadsByFingerprint",value:function(t){var e=this._findEntries("tus::".concat(t,"::"));return Promise.resolve(e)}},{key:"removeUpload",value:function(t){return localStorage.removeItem(t),Promise.resolve()}},{key:"addUpload",value:function(t,e){var n=Math.round(1e12*Math.random()),r="tus::".concat(t,"::").concat(n);return localStorage.setItem(r,JSON.stringify(e)),Promise.resolve(r)}},{key:"_findEntries",value:function(t){for(var e=[],n=0;n<localStorage.length;n++){var r=localStorage.key(n);if(0===r.indexOf(t))try{var o=JSON.parse(localStorage.getItem(r));o.urlStorageKey=r,e.push(o)}catch(t){}}return e}}])&&Op(e.prototype,n),r&&Op(e,r),t}();function Ip(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}function Up(t,e){for(var n=0;n<e.length;n++){var r=e[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(t,r.key,r)}}function Lp(t,e,n){return e&&Up(t.prototype,e),n&&Up(t,n),t}var Dp=function(){function t(){Ip(this,t)}return Lp(t,[{key:"createRequest",value:function(t,e){return new Hp(t,e)}},{key:"getName",value:function(){return"XHRHttpStack"}}]),t}(),Hp=function(){function t(e,n){Ip(this,t),this._xhr=new XMLHttpRequest,this._xhr.open(e,n,!0),this._method=e,this._url=n,this._headers={}}return Lp(t,[{key:"getMethod",value:function(){return this._method}},{key:"getURL",value:function(){return this._url}},{key:"setHeader",value:function(t,e){this._xhr.setRequestHeader(t,e),this._headers[t]=e}},{key:"getHeader",value:function(t){return this._headers[t]}},{key:"setProgressHandler",value:function(t){"upload"in this._xhr&&(this._xhr.upload.onprogress=function(e){e.lengthComputable&&t(e.loaded)})}},{key:"send",value:function(){var t=this,e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:null;return new Promise((function(n,r){t._xhr.onload=function(){n(new Fp(t._xhr))},t._xhr.onerror=function(t){r(t)},t._xhr.send(e)}))}},{key:"abort",value:function(){return this._xhr.abort(),Promise.resolve()}},{key:"getUnderlyingObject",value:function(){return this._xhr}}]),t}(),Fp=function(){function t(e){Ip(this,t),this._xhr=e}return Lp(t,[{key:"getStatus",value:function(){return this._xhr.status}},{key:"getHeader",value:function(t){return this._xhr.getResponseHeader(t)}},{key:"getBody",value:function(){return this._xhr.responseText}},{key:"getUnderlyingObject",value:function(){return this._xhr}}]),t}(),Mp=function(){return"undefined"!=typeof navigator&&"string"==typeof navigator.product&&"reactnative"===navigator.product.toLowerCase()};function Np(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}function Bp(t,e){for(var n=0;n<e.length;n++){var r=e[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(t,r.key,r)}}function zp(t,e,n){return e&&Bp(t.prototype,e),n&&Bp(t,n),t}var qp=function(){function t(e){Np(this,t),this._file=e,this.size=e.size}return zp(t,[{key:"slice",value:function(t,e){if("undefined"!=typeof window&&(void 0!==window.PhoneGap||void 0!==window.Cordova||void 0!==window.cordova))return n=this._file.slice(t,e),new Promise((function(t,e){var r=new FileReader;r.onload=function(){var e=new Uint8Array(r.result);t({value:e})},r.onerror=function(t){e(t)},r.readAsArrayBuffer(n)}));var n,r=this._file.slice(t,e);return Promise.resolve({value:r})}},{key:"close",value:function(){}}]),t}(),Gp=function(){function t(e,n){Np(this,t),this._chunkSize=n,this._buffer=void 0,this._bufferOffset=0,this._reader=e,this._done=!1}return zp(t,[{key:"slice",value:function(t,e){return t<this._bufferOffset?Promise.reject(new Error("Requested data is before the reader's current offset")):this._readUntilEnoughDataOrDone(t,e)}},{key:"_readUntilEnoughDataOrDone",value:function(t,e){var n=this,r=e<=this._bufferOffset+$p(this._buffer);if(this._done||r){var o=this._getDataFromBuffer(t,e),i=null==o&&this._done;return Promise.resolve({value:o,done:i})}return this._reader.read().then((function(r){var o=r.value;return r.done?n._done=!0:void 0===n._buffer?n._buffer=o:n._buffer=function(t,e){if(t.concat)return t.concat(e);if(t instanceof Blob)return new Blob([t,e],{type:t.type});if(t.set){var n=new t.constructor(t.length+e.length);return n.set(t),n.set(e,t.length),n}throw new Error("Unknown data type")}(n._buffer,o),n._readUntilEnoughDataOrDone(t,e)}))}},{key:"_getDataFromBuffer",value:function(t,e){t>this._bufferOffset&&(this._buffer=this._buffer.slice(t-this._bufferOffset),this._bufferOffset=t);var n=0===$p(this._buffer);return this._done&&n?null:this._buffer.slice(0,e-t)}},{key:"close",value:function(){this._reader.cancel&&this._reader.cancel()}}]),t}();function $p(t){return void 0===t?0:void 0!==t.size?t.size:t.length}var Kp=function(){function t(){Np(this,t)}return zp(t,[{key:"openFile",value:function(t,e){return Mp()&&t&&void 0!==t.uri?(n=t.uri,new Promise((function(t,e){var r=new XMLHttpRequest;r.responseType="blob",r.onload=function(){var e=r.response;t(e)},r.onerror=function(t){e(t)},r.open("GET",n),r.send()}))).then((function(t){return new qp(t)})).catch((function(t){throw new Error("tus: cannot fetch `file.uri` as Blob, make sure the uri is correct and accessible. "+t)})):"function"==typeof t.slice&&void 0!==t.size?Promise.resolve(new qp(t)):"function"==typeof t.read?(e=+e,isFinite(e)?Promise.resolve(new Gp(t,e)):Promise.reject(new Error("cannot create source for stream without a finite value for the `chunkSize` option"))):Promise.reject(new Error("source object may only be an instance of File, Blob, or Reader in this environment"));var n}}]),t}();function Wp(t){return(Wp="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}function Xp(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}function Vp(t,e){for(var n=0;n<e.length;n++){var r=e[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(t,r.key,r)}}function Yp(t,e){return(Yp=Object.setPrototypeOf||function(t,e){return t.__proto__=e,t})(t,e)}function Qp(t,e){return!e||"object"!==Wp(e)&&"function"!=typeof e?function(t){if(void 0===t)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return t}(t):e}function Jp(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],(function(){}))),!0}catch(t){return!1}}function Zp(t){return(Zp=Object.setPrototypeOf?Object.getPrototypeOf:function(t){return t.__proto__||Object.getPrototypeOf(t)})(t)}function td(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function ed(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?td(Object(n),!0).forEach((function(e){nd(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):td(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function nd(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var rd=ed({},xp.defaultOptions,{httpStack:new Dp,fileReader:new Kp,urlStorage:Pp?new Cp:new Rp,fingerprint:function(t,e){return Mp()?Promise.resolve(function(t,e){var n=t.exif?function(t){var e=0;if(0===t.length)return e;for(var n=0;n<t.length;n++){e=(e<<5)-e+t.charCodeAt(n),e&=e}return e}(JSON.stringify(t.exif)):"noexif";return["tus-rn",t.name||"noname",t.size||"nosize",n,e.endpoint].join("/")}(t,e)):Promise.resolve(["tus-br",t.name,t.type,t.size,t.lastModified,e.endpoint].join("-"))}}),od=function(t){!function(t,e){if("function"!=typeof e&&null!==e)throw new TypeError("Super expression must either be null or a function");t.prototype=Object.create(e&&e.prototype,{constructor:{value:t,writable:!0,configurable:!0}}),e&&Yp(t,e)}(a,t);var e,n,r,o,i=(e=a,function(){var t,n=Zp(e);if(Jp()){var r=Zp(this).constructor;t=Reflect.construct(n,arguments,r)}else t=n.apply(this,arguments);return Qp(this,t)});function a(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:null,e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{};return Xp(this,a),e=ed({},rd,{},e),i.call(this,t,e)}return n=a,o=[{key:"terminate",value:function(t,e,n){return e=ed({},rd,{},e),xp.terminate(t,e,n)}}],(r=null)&&Vp(n.prototype,r),o&&Vp(n,o),a}(xp);function id(t){var e=function(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Boolean.prototype.valueOf.call(Reflect.construct(Boolean,[],(function(){}))),!0}catch(t){return!1}}();return function(){var n,r=Xs(t);if(e){var o=Xs(this).constructor;n=Reflect.construct(r,arguments,o)}else n=r.apply(this,arguments);return Ws(this,n)}}var ad=function(t){Ks(o,t);var e,n,r=id(o);function o(t){var e,n=t.chunkSize,i=t.csrfToken,a=t.fieldName,u=t.file,s=t.formId,c=t.retryDelays,l=t.uploadIndex,f=t.uploadUrl;return qi(this,o),Dr(Gs(e=r.call(this,{name:u.name,status:"uploading",type:"tus",uploadIndex:l})),"onError",void 0),Dr(Gs(e),"onProgress",void 0),Dr(Gs(e),"onSuccess",void 0),Dr(Gs(e),"upload",void 0),Dr(Gs(e),"csrfToken",void 0),Dr(Gs(e),"handleError",(function(t){if(!e.onError)throw t;e.onError(t)})),Dr(Gs(e),"handleProgress",(function(t,n){e.onProgress&&e.onProgress(t,n)})),Dr(Gs(e),"handleSucces",(function(){e.onSuccess&&e.onSuccess()})),Dr(Gs(e),"addCsrTokenToRequest",(function(t){t.setHeader("X-CSRFToken",e.csrfToken)})),e.csrfToken=i,e.upload=new od(u,{chunkSize:n,endpoint:f,metadata:{fieldName:a,filename:u.name,formId:s},onBeforeRequest:e.addCsrTokenToRequest,onError:e.handleError,onProgress:e.handleProgress,onSuccess:e.handleSucces,retryDelays:c||[0,1e3,3e3,5e3]}),e.onError=void 0,e.onProgress=void 0,e.onSuccess=void 0,e}return $i(o,[{key:"abort",value:(n=zi(ga.mark((function t(){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,this.upload.abort(!0);case 2:case"end":return t.stop()}}),t,this)}))),function(){return n.apply(this,arguments)})},{key:"delete",value:(e=zi(ga.mark((function t(){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if(this.upload.url){t.next=2;break}return t.abrupt("return",Promise.resolve());case 2:return t.next=4,ic(this.upload.url,this.csrfToken);case 4:case"end":return t.stop()}}),t,this)}))),function(){return e.apply(this,arguments)})},{key:"getSize",value:function(){return this.upload.file.size}},{key:"start",value:function(){this.upload.start()}},{key:"getInitialFile",value:function(){return{id:this.name,name:this.name,size:this.getSize(),type:"tus",url:""}}}]),o}(Zs),ud=function(){function t(e){var n=this,r=e.callbacks,o=e.chunkSize,i=e.csrfToken,a=e.eventEmitter,u=e.fieldName,s=e.form,c=e.formId,l=e.initial,f=e.input,p=e.multiple,d=e.parent,h=e.prefix,v=e.retryDelays,y=e.s3UploadDir,g=e.skipRequired,m=e.supportDropArea,x=e.translations,b=e.uploadUrl;qi(this,t),Dr(this,"callbacks",void 0),Dr(this,"chunkSize",void 0),Dr(this,"csrfToken",void 0),Dr(this,"eventEmitter",void 0),Dr(this,"fieldName",void 0),Dr(this,"form",void 0),Dr(this,"formId",void 0),Dr(this,"multiple",void 0),Dr(this,"nextUploadIndex",void 0),Dr(this,"prefix",void 0),Dr(this,"renderer",void 0),Dr(this,"retryDelays",void 0),Dr(this,"s3UploadDir",void 0),Dr(this,"supportDropArea",void 0),Dr(this,"uploads",void 0),Dr(this,"uploadUrl",void 0),Dr(this,"uploadFiles",function(){var t=zi(ga.mark((function t(e){var r,o,i,a,u,s,c;return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if(0!==e.length){t.next=2;break}return t.abrupt("return");case 2:n.multiple||0===n.uploads.length||(n.renderer.deleteFile(0),n.uploads=[]),r=!0,o=!1,t.prev=5,a=Ki(e);case 7:return t.next=9,a.next();case 9:return u=t.sent,r=u.done,t.next=13,u.value;case 13:if(s=t.sent,r){t.next=21;break}return c=s,t.next=18,n.uploadFile(c);case 18:r=!0,t.next=7;break;case 21:t.next=27;break;case 23:t.prev=23,t.t0=t.catch(5),o=!0,i=t.t0;case 27:if(t.prev=27,t.prev=28,r||null==a.return){t.next=32;break}return t.next=32,a.return();case 32:if(t.prev=32,!o){t.next=35;break}throw i;case 35:return t.finish(32);case 36:return t.finish(27);case 37:n.checkDropHint();case 38:case"end":return t.stop()}}),t,null,[[5,23,27,37],[28,,32,36]])})));return function(e){return t.apply(this,arguments)}}()),Dr(this,"onChange",(function(t){var e=t.target.files;e&&n.uploadFiles(bo(e))})),Dr(this,"onClick",(function(t){var e=t.target,r=function(){var t=e.getAttribute("data-index");if(t){var r=parseInt(t,10);return n.getUploadByIndex(r)}};if(e.classList.contains("dff-delete")&&!e.classList.contains("dff-disabled")){var o=r();o&&n.removeExistingUpload(o),t.preventDefault()}else if(e.classList.contains("dff-cancel")){var i=r();i&&n.handleCancel(i),t.preventDefault()}})),Dr(this,"handleProgress",(function(t,e,r){var o=(e/r*100).toFixed(2);n.renderer.updateProgress(t.uploadIndex,o);var i=n.callbacks.onProgress;i&&t instanceof ad&&i(e,r,t)})),Dr(this,"handleError",(function(t,e){n.renderer.setError(t.uploadIndex),t.status="error";var r=n.callbacks.onError;r&&t instanceof ad&&r(e,t)})),Dr(this,"handleSuccess",(function(t){var e=n.renderer;n.updatePlaceholderInput(),e.clearInput(),e.setSuccess(t.uploadIndex,t.getSize()),t.status="done";var r=n.callbacks.onSuccess,o=document.getElementsByClassName("dff-file-id-".concat(t.uploadIndex))[0];n.emitEvent("uploadComplete",o,t),r&&"tus"===t.type&&r(t)})),this.callbacks=r,this.chunkSize=o,this.csrfToken=i,this.eventEmitter=a,this.fieldName=u,this.form=s,this.formId=c,this.multiple=p,this.prefix=h,this.retryDelays=v,this.s3UploadDir=y,this.supportDropArea=m,this.uploadUrl=b,this.uploads=[],this.nextUploadIndex=0,this.renderer=new Ua({parent:d,input:f,skipRequired:g,translations:x});var w=this.renderer.container;m&&this.initDropArea(w,f.accept),l&&this.addInitialFiles(l),this.checkDropHint(),f.addEventListener("change",this.onChange),w.addEventListener("click",this.onClick)}var e,n,r;return $i(t,[{key:"addInitialFiles",value:function(t){var e=this;if(0!==t.length){var n=this.multiple,r=this.renderer,o=function(t){var n=t.size,o="s3"===t.type&&t.original_name?t.original_name:t.name,i=r.addUploadedFile(o,e.nextUploadIndex,n),a=function(t){var e=t.csrfToken,n=t.initialFile,r=t.uploadIndex,o=t.uploadUrl;switch(n.type){case"existing":return new fc(n,r);case"placeholder":return new cc(n,r);case"s3":return new lc(n,r);case"tus":return new pc({csrfToken:e,initialFile:n,uploadUrl:o,uploadIndex:r})}}({csrfToken:e.csrfToken,initialFile:t,uploadIndex:e.nextUploadIndex,uploadUrl:e.uploadUrl});e.uploads.push(a),e.emitEvent("addUpload",i,a)};n?t.forEach((function(t){o(t),e.nextUploadIndex+=1})):o(t[0])}}},{key:"uploadFile",value:(r=zi(ga.mark((function t(e){var n,r,o,i,a,u,s,c,l,f,p=this;return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if(n=function(){var t=p.csrfToken,n=p.s3UploadDir;return null!=n?new oc({csrfToken:t,endpoint:a,file:e,s3UploadDir:n,uploadIndex:c}):new ad({chunkSize:p.chunkSize,csrfToken:p.csrfToken,fieldName:r,file:e,formId:o,retryDelays:p.retryDelays,uploadIndex:c,uploadUrl:a})},r=this.fieldName,o=this.formId,i=this.renderer,a=this.uploadUrl,u=e.name,s=this.findUploadByName(u),c=s?s.uploadIndex:this.nextUploadIndex,s||(this.nextUploadIndex+=1),!s){t.next=9;break}return t.next=9,this.removeExistingUpload(s);case 9:(l=n()).onError=function(t){return p.handleError(l,t)},l.onProgress=function(t,e){return p.handleProgress(l,t,e)},l.onSuccess=function(){return p.handleSuccess(l)},l.start(),this.uploads.push(l),f=i.addNewUpload(u,c),this.emitEvent("addUpload",f,l);case 17:case"end":return t.stop()}}),t,this)}))),function(t){return r.apply(this,arguments)})},{key:"getUploadByIndex",value:function(t){return this.uploads.find((function(e){return e.uploadIndex===t}))}},{key:"findUploadByName",value:function(t){return this.uploads.find((function(e){return e.name===t}))}},{key:"removeExistingUpload",value:(n=zi(ga.mark((function t(e){var n;return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if((n=this.renderer.findFileDiv(e.uploadIndex))&&this.emitEvent("removeUpload",n,e),"uploading"!==e.status){t.next=8;break}return this.renderer.disableCancel(e.uploadIndex),t.next=6,e.abort();case 6:t.next=19;break;case 8:if("done"!==e.status){t.next=19;break}return this.renderer.disableDelete(e.uploadIndex),t.prev=10,t.next=13,e.delete();case 13:t.next=19;break;case 15:return t.prev=15,t.t0=t.catch(10),this.renderer.setDeleteFailed(e.uploadIndex),t.abrupt("return");case 19:this.removeUploadFromList(e),this.updatePlaceholderInput();case 21:case"end":return t.stop()}}),t,this,[[10,15]])}))),function(t){return n.apply(this,arguments)})},{key:"removeUploadFromList",value:function(t){this.renderer.deleteFile(t.uploadIndex);var e=this.uploads.indexOf(t);e>=0&&this.uploads.splice(e,1),this.checkDropHint();var n=this.callbacks.onDelete;n&&n(t)}},{key:"handleCancel",value:(e=zi(ga.mark((function t(e){return ga.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return this.renderer.disableCancel(e.uploadIndex),t.next=3,e.abort();case 3:this.removeUploadFromList(e);case 4:case"end":return t.stop()}}),t,this)}))),function(t){return e.apply(this,arguments)})},{key:"initDropArea",value:function(t,e){new Ls({container:t,inputAccept:e,onUploadFiles:this.uploadFiles})}},{key:"checkDropHint",value:function(){this.supportDropArea&&(0===this.uploads.filter((function(t){return t})).length?this.renderer.renderDropHint():this.renderer.removeDropHint())}},{key:"updatePlaceholderInput",value:function(){var t=ba(this.form,wa(this.fieldName,this.prefix),this.prefix);if(t){var e=this.uploads.map((function(t){return t.getInitialFile()}));t.value=JSON.stringify(e)}}},{key:"getMetaDataField",value:function(){return ba(this.form,(t=this.fieldName,e=this.prefix,"".concat(xa(t,e),"-metadata")),this.prefix);var t,e}},{key:"emitEvent",value:function(t,e,n){this.eventEmitter&&this.eventEmitter.emit(t,{element:e,fieldName:this.fieldName,fileName:n.name,metaDataField:this.getMetaDataField(),upload:n})}}]),t}(),sd=function(t){var e,n=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},r=function(t){return!n||!n.prefix||t.startsWith("".concat(n.prefix,"-"))},o=function(){return n&&n.prefix?n.prefix:null},i=function(e){return _a(t,e,o())},a=function(t){var e=i(wa(t,o()));return e?JSON.parse(e):[]},u=i("upload_url"),s=i("form_id"),c=i("s3_upload_dir"),l=n.skipRequired||!1,f=o(),p=null===(e=ba(t,"csrfmiddlewaretoken",null))||void 0===e?void 0:e.value;if(!p)throw Error("Csrf token not found");s&&u&&t.querySelectorAll(".dff-uploader").forEach((function(e){var o=e.querySelector(".dff-container");if(o){var i=o.querySelector("input[type=file]");if(i&&r(i.name)){var d=i.name,h=i.multiple,v=a(d),y=o.getAttribute("data-translations"),g=y?JSON.parse(y):{},m=!(!1===n.supportDropArea);new ud({callbacks:n.callbacks||{},chunkSize:n.chunkSize||2621440,csrfToken:p,eventEmitter:n.eventEmitter,fieldName:d,form:t,formId:s,s3UploadDir:c||null,initial:v,input:i,multiple:h,parent:o,prefix:f,retryDelays:n.retryDelays||null,skipRequired:l,supportDropArea:m,translations:g,uploadUrl:u})}}}))};function cd(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function ld(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?cd(Object(n),!0).forEach((function(e){Dr(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):cd(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}window.autoInitFileForms=function(){var t,e=window.initUploadFields;(t=Array.from(document.querySelectorAll(".dff-uploader")).map(Wn),Array.from(new Set(t).values())).forEach(e)},window.initFormSet=function(t,e){var n,r=(n="string"==typeof e?{prefix:e}:e).prefix||"form",o=_a(t,"TOTAL_FORMS",r);if(o)for(var i=parseInt(o,10),a=0;a<i;a+=1){var u=ma("".concat(a),null);sd(t,ld(ld({},n),{},{prefix:"".concat(r,"-").concat(u)}))}},window.initUploadFields=sd}();
 //# sourceMappingURL=file_form.js.map
