+++ django_file_form/widgets.py
@@ -1,18 +1,12 @@
 import json
-from typing import Dict, Union, List
+from typing import Dict, Union
 
-from django.db.models.fields.files import FieldFile
 from django.forms import ClearableFileInput
 from django.http import QueryDict
 from django.utils.datastructures import MultiValueDict
 from django.utils.translation import gettext as _
-from django_file_form.util import compact, get_list
-
-from django_file_form.models import (
-    PlaceholderUploadedFile,
-    S3UploadedFileWithId,
-    UploadedFileWithId,
-)
+from django_file_form.util import compact
+from .uploaded_file import PlaceholderUploadedFile, S3UploadedFileWithId
 
 
 TRANSLATIONS = {
@@ -22,39 +16,6 @@ TRANSLATIONS = {
     "Upload failed": _("Upload failed"),
     "Drop your files here": _("Drop your files here"),
 }
-
-UploadedFileTypes = Union[FieldFile, PlaceholderUploadedFile, UploadedFileWithId]
-
-UploadedFileTypesOrList = Union[
-    UploadedFileTypes,
-    List[Union[UploadedFileTypes]],
-]
-
-
-def get_uploaded_files(value: UploadedFileTypesOrList):
-    def must_include(file_info):
-        return not getattr(file_info, "is_placeholder", False) and not getattr(
-            file_info, "is_s3direct", False
-        )
-
-    def get_values(file_info: UploadedFileTypes):
-        if hasattr(file_info, "file_id"):
-            return file_info.get_values()
-        else:
-            return dict(
-                name=file_info.name,
-                size=file_info.size,
-                type="existing",
-            )
-
-    if not value:
-        return []
-
-    return [
-        get_values(file_info)
-        for file_info in get_list(value)
-        if must_include(file_info)
-    ]
 
 
 def get_upload(upload_data: Dict):
@@ -108,7 +69,6 @@ class BaseUploadWidget(ClearableFileInput):
         context = super().get_context(name, value, attrs)
 
         context["translations"] = json.dumps(TRANSLATIONS)
-        context["uploaded_files"] = json.dumps(get_uploaded_files(value))
 
         return context
